Module googapis::google::actions::sdk::v2::conversation

Modules

collection
Nested message and enum types in Collection.

content
Nested message and enum types in Content.

image
Nested message and enum types in Image.

list
Nested message and enum types in List.

media
Nested message and enum types in Media.

media_image
Nested message and enum types in MediaImage.

slot
Nested message and enum types in Slot.

table_column
Nested message and enum types in TableColumn.

Structs

Canvas
Represents an Interactive Canvas response to be sent to the user. This can be used in conjunction with the “first_simple” field in the containing prompt to speak to the user in addition to displaying a interactive canvas response. The maximum size of the response is 50k bytes.

Card
A basic card for displaying some information, e.g. an image and/or text.

Collection
A card for presenting a collection of options to select from.

Content
Content to be shown.

Image
An image displayed in the card.

Intent
Represents an intent.

IntentParameterValue
Represents a value for intent parameter.

Link
Link content.

List
A card for presenting a list of options to select from.

Media
Represents one media object. Contains information about the media, such as name, description, url, etc.

MediaImage
Image to show with the media card.

MediaObject
Represents a single media object

OpenUrl
Action taken when a user opens a link.

Prompt
Represent a response to a user.

Simple
Represents a simple prompt to be send to a user.

Slot
Represents a slot.

Suggestion
Input suggestion to be presented to the user.

Table
A table card for displaying a table of text.

TableCell
Describes a cell in a row.

TableColumn
Describes a column in a table.

TableRow
Describes a row in the table.

Enums

SlotFillingStatus
Represents the current status of slot filling.

UrlHint
Different types of url hints.