Enum googapis::google::actions::sdk::v2::execution_event::EventData

pub enum EventData {

    UserInput(UserConversationInput),
    IntentMatch(IntentMatch),
    ConditionsEvaluated(ConditionsEvaluated),
    OnSceneEnter(OnSceneEnter),
    WebhookRequest(WebhookRequest),
    WebhookResponse(WebhookResponse),
    WebhookInitiatedTransition(WebhookInitiatedTransition),
    SlotMatch(SlotMatch),
    SlotRequested(SlotRequested),
    SlotValidated(SlotValidated),
    FormFilled(FormFilled),
    WaitingUserInput(WaitingForUserInput),
    EndConversation(EndConversation),

}

Detailed information specific to different of events that may be involved in processing a conversation round. The field set here defines the type of this event.

Variants

UserInput(UserConversationInput)

User input handling event.

Tuple Fields of UserInput

0: UserConversationInput

IntentMatch(IntentMatch)

Intent matching event.

Tuple Fields of IntentMatch

0: IntentMatch

ConditionsEvaluated(ConditionsEvaluated)

Condition evaluation event.

Tuple Fields of ConditionsEvaluated

0: ConditionsEvaluated

OnSceneEnter(OnSceneEnter)

OnSceneEnter execution event.

Tuple Fields of OnSceneEnter

0: OnSceneEnter

WebhookRequest(WebhookRequest)

Webhook request dispatch event.

Tuple Fields of WebhookRequest

0: WebhookRequest

WebhookResponse(WebhookResponse)

Webhook response receipt event.

Tuple Fields of WebhookResponse

0: WebhookResponse

WebhookInitiatedTransition(WebhookInitiatedTransition)

Webhook-initiated transition event.

Tuple Fields of WebhookInitiatedTransition

0: WebhookInitiatedTransition

SlotMatch(SlotMatch)

Slot matching event.

Tuple Fields of SlotMatch

0: SlotMatch

SlotRequested(SlotRequested)

Slot requesting event.

Tuple Fields of SlotRequested

0: SlotRequested

SlotValidated(SlotValidated)

Slot validation event.

Tuple Fields of SlotValidated

0: SlotValidated

FormFilled(FormFilled)

Form filling event.

Tuple Fields of FormFilled

0: FormFilled

WaitingUserInput(WaitingForUserInput)

Waiting-for-user-input event.

Tuple Fields of WaitingUserInput

0: WaitingForUserInput

EndConversation(EndConversation)

End-of-conversation event.

Tuple Fields of EndConversation

0: EndConversation

Implementations

impl EventData

pub fn encode<B>(&self, buf: &mut B) where
B: BufMut,

pub fn merge<B>(
field: &mut Option<EventData>,
tag: u32,
wire_type: WireType,
buf: &mut B,
ctx: DecodeContext
) -> Result<(), DecodeError> where
B: Buf,

pub fn encoded_len(&self) -> usize