Module googapis::google::actions::sdk::v2::interactionmodel::prompt

Modules

media_image

Nested message and enum types in MediaImage.

static_collection_browse_prompt

Nested message and enum types in StaticCollectionBrowsePrompt.

static_collection_prompt

Nested message and enum types in StaticCollectionPrompt.

static_content_prompt

Nested message and enum types in StaticContentPrompt.

static_image_prompt

Nested message and enum types in StaticImagePrompt.

static_list_prompt

Nested message and enum types in StaticListPrompt.

static_media_prompt

Nested message and enum types in StaticMediaPrompt.

static_prompt

Nested message and enum types in StaticPrompt.

static_simple_prompt

Nested message and enum types in StaticSimplePrompt.

surface_capabilities

Nested message and enum types in SurfaceCapabilities.

table_column

Nested message and enum types in TableColumn.

Structs

MediaImage

Image to be shown inside a MediaPrompt.

MediaObject

Represents a single media object.

OpenUrl

Defines behavior when the user opens the link.

StaticCanvasPrompt

Represents a Interactive Canvas response to be sent to the user. This can be used in conjunction with the first_simple field in the containing prompt to speak to the user in addition to displaying a interactive canvas response.

StaticCardPrompt

A basic card for displaying some information, e.g. an image and/or text.

StaticCollectionBrowsePrompt

Presents a set of web documents as a collection of large-tile items. Items may be selected to launch their associated web document in a web viewer.

StaticCollectionPrompt

A card for presenting a collection of options to select from.

StaticContentPrompt

A placeholder for the Content part of a StaticPrompt.

StaticImagePrompt

An image displayed in the card.

StaticLinkPrompt

Defines a link which will be displayed as a suggestion chip and can be opened by the user.

StaticListPrompt

A card for presenting a list of options to select from.

StaticMediaPrompt

Contains information about the media, such as name, description, url, etc. Next id: 11

StaticPrompt

Represents a list of prompt candidates, one of which will be selected as the prompt to be shown in the response to the user. This message is localizable.

StaticSimplePrompt

Represents a simple prompt to be send to a user.

StaticTablePrompt

A table card for displaying a table of text.

Suggestion

Represents a suggestion chip, a UI element shown to the user for convenience.

SurfaceCapabilities

Represents the surface the user is using to make a request to the Action.

TableCell

Describes a cell in a row.

TableColumn

Describes a column in the table.

TableRow

Describes a row in the table.

Enums

UrlHint

Different types of url hints.