Module googapis::google::actions::sdk::v2::interactionmodel::type

Modules

regular_expression_type
Nested message and enum types in RegularExpressionType.

synonym_type
Nested message and enum types in SynonymType.

type
Nested message and enum types in Type.

Structs

ClassReference
A reference to a class which is used to declare the type of a field or return value. Enums are also a type of class that can be referenced using ClassReference.

EntityDisplay
Elements that will be displayed on the canvas once a particular type’s entity is extracted from a query. Only relevant for canvas enabled apps. This message is localizable.

FreeTextType
Type that matches any text if surrounding words context is close to provided training examples.

RegularExpressionType
Type that matches text by regular expressions. This message is localizable.

SynonymType
Type that matches text by set of synonyms.

Type
Declaration of a custom type, as opposed to built-in types. Types can be assigned to slots in a scene or parameters of an intent’s training phrases. Practically, Types can be thought of as enums. Note, type name is specified in the name of the file.