Struct googapis::google::ads::googleads::v7::common::LeadFormAsset

pub struct LeadFormAsset {

    pub business_name: String,
    pub call_to_action_type: i32,
    pub call_to_action_description: String,
    pub headline: String,
    pub description: String,
    pub privacy_policy_url: String,
    pub post_submit_headline: Option<String>,
    pub post_submit_description: Option<String>,
    pub fields: Vec<LeadFormField>,
    pub delivery_methods: Vec<LeadFormDeliveryMethod>,
    pub post_submit_call_to_action_type: i32,
    pub background_image_asset: Option<String>,
    pub desired_intent: i32,
    pub custom_disclosure: Option<String>,

}

A Lead Form asset.

Fields
business_name: String

Required. The name of the business being advertised.

call_to_action_type: i32

Required. Pre-defined display text that encourages user to expand the form.

call_to_action_description: String

Required. Text giving a clear value proposition of what users expect once they expand the form.

headline: String

Required. Headline of the expanded form to describe what the form is asking for or facilitating.

description: String

Required. Detailed description of the expanded form to describe what the form is asking for or facilitating.

privacy_policy_url: String

Required. Link to a page describing the policy on how the collected data is handled by the advertiser/business.

post_submit_headline: Option<String>

Headline of text shown after form submission that describes how the advertiser will follow up with the user.

post_submit_description: Option<String>

Detailed description shown after form submission that describes how the advertiser will follow up with the user.

fields: Vec<LeadFormField>

Ordered list of input fields.

delivery_methods: Vec<LeadFormDeliveryMethod>

Configured methods for collected lead data to be delivered to advertiser. Only one method typed as WebhookDelivery can be configured.

post_submit_call_to_action_type: i32

Pre-defined display text that encourages user action after the form is submitted.

background_image_asset: Option<String>

Asset resource name of the background image. The minimum size is 600x314 and the aspect ratio must be 1.91:1 (+-1%).

desired_intent: i32

Desired intent for the lead form, e.g. more volume or higher intent.

custom_disclosure: Option<String>

Custom disclosure shown along with Google disclaimer on the lead form. Accessible to allowed customers only.

Implementations

impl LeadFormAsset

pub fn post_submit_headline(&self) -> &str

Returns the value of post_submit_headline, or the default value if post_submit_headline is unset.

pub fn post_submit_description(&self) -> &str

Returns the value of post_submit_description, or the default value if post_submit_description is unset.

pub fn call_to_action_type(&self) -> LeadFormCallToActionType

Returns the enum value of call_to_action_type, or the default if the field is set to an invalid enum value.

pub fn set_call_to_action_type(&mut self, value: LeadFormCallToActionType)

Sets call_to_action_type to the provided enum value.

pub fn post_submit_call_to_action_type(
&self
) -> LeadFormPostSubmitCallToActionType

Returns the enum value of post_submit_call_to_action_type, or the default if the field is set to an invalid enum value.

pub fn set_post_submit_call_to_action_type(
&mut self,
value: LeadFormPostSubmitCallToActionType
)

Sets post_submit_call_to_action_type to the provided enum value.

pub fn background_image_asset(&self) -> &str

Returns the value of background_image_asset, or the default value if background_image_asset is unset.

pub fn desired_intent(&self) -> LeadFormDesiredIntent

Returns the enum value of desired_intent, or the default if the field is set to an invalid enum value.

pub fn set_desired_intent(&mut self, value: LeadFormDesiredIntent)

Sets desired_intent to the provided enum value.

pub fn custom_disclosure(&self) -> &str

Returns the value of custom_disclosure, or the default value if custom_disclosure is unset.