Enum googapis::google::ads::googleads::v7::enums::external_conversion_source_enum::ExternalConversionSource

#[repr(i32)]
pub enum ExternalConversionSource {

    Unspecified,
    Unknown,
    Webpage,
    Analytics,
    Upload,
    AdCallMetrics,
    WebsiteCallMetrics,
    StoreVisits,
    AndroidInApp,
    IosInApp,
    IosFirstOpen,
    AppUnspecified,
    AndroidFirstOpen,
    UploadCalls,
    Firebase,
    ClickToCall,
    Salesforce,
    StoreSalesCrm,
    StoreSalesPaymentNetwork,
    GooglePlay,
    ThirdPartyAppAnalytics,
    GoogleAttribution,
    StoreSalesDirectUpload,
    StoreSales,
    SearchAds360,
    GoogleHosted,
    Floodlight,
    AnalyticsSearchAds360,
    FirebaseSearchAds360,

}

The external conversion source that is associated with a ConversionAction.

Variants

Unspecified

Not specified.

Unknown

Represents value unknown in this version.

Webpage

Conversion that occurs when a user navigates to a particular webpage after viewing an ad; Displayed in Google Ads UI as ‘Website’.

Analytics

Conversion that comes from linked Google Analytics goal or transaction; Displayed in Google Ads UI as ‘Analytics’.

Upload

Website conversion that is uploaded through ConversionUploadService; Displayed in Google Ads UI as ‘Import from clicks’.

AdCallMetrics

Conversion that occurs when a user clicks on a call extension directly on an ad; Displayed in Google Ads UI as ‘Calls from ads’.

WebsiteCallMetrics

Conversion that occurs when a user calls a dynamically-generated phone number (by installed javascript) from an advertiser’s website after clicking on an ad; Displayed in Google Ads UI as ‘Calls from website’.

StoreVisits

Conversion that occurs when a user visits an advertiser’s retail store after clicking on a Google ad; Displayed in Google Ads UI as ‘Store visits’.

AndroidInApp

Conversion that occurs when a user takes an in-app action such as a purchase in an Android app; Displayed in Google Ads UI as ‘Android in-app action’.

IosInApp

Conversion that occurs when a user takes an in-app action such as a purchase in an iOS app; Displayed in Google Ads UI as ‘iOS in-app action’.

IosFirstOpen

Conversion that occurs when a user opens an iOS app for the first time; Displayed in Google Ads UI as ‘iOS app install (first open)’.

AppUnspecified

Legacy app conversions that do not have an AppPlatform provided; Displayed in Google Ads UI as ‘Mobile app’.

AndroidFirstOpen

Conversion that occurs when a user opens an Android app for the first time; Displayed in Google Ads UI as ‘Android app install (first open)’.

UploadCalls

Call conversion that is uploaded through ConversionUploadService; Displayed in Google Ads UI as ‘Import from calls’.

Firebase

Conversion that comes from a linked Firebase event; Displayed in Google Ads UI as ‘Firebase’.

ClickToCall

Conversion that occurs when a user clicks on a mobile phone number; Displayed in Google Ads UI as ‘Phone number clicks’.

Salesforce

Conversion that comes from Salesforce; Displayed in Google Ads UI as ‘Salesforce.com’.

StoreSalesCrm

Conversion that comes from in-store purchases recorded by CRM; Displayed in Google Ads UI as ‘Store sales (data partner)’.

StoreSalesPaymentNetwork

Conversion that comes from in-store purchases from payment network; Displayed in Google Ads UI as ‘Store sales (payment network)’.

GooglePlay

Codeless Google Play conversion; Displayed in Google Ads UI as ‘Google Play’.

ThirdPartyAppAnalytics

Conversion that comes from a linked third-party app analytics event; Displayed in Google Ads UI as ‘Third-party app analytics’.

GoogleAttribution

Conversion that is controlled by Google Attribution.

StoreSalesDirectUpload

Store Sales conversion based on first-party or third-party merchant data uploads. Displayed in Google Ads UI as ‘Store sales (direct upload)’.

StoreSales

Store Sales conversion based on first-party or third-party merchant data uploads and/or from in-store purchases using cards from payment networks. Displayed in Google Ads UI as ‘Store sales’.

SearchAds360

Conversions imported from Search Ads 360 Floodlight data.

GoogleHosted

Conversions that track local actions from Google’s products and services after interacting with an ad.

Floodlight

Conversions reported by Floodlight tags.

AnalyticsSearchAds360

Conversions that come from Google Analytics specifically for Search Ads 360. Displayed in Google Ads UI as Analytics (SA360).

FirebaseSearchAds360

Conversion that comes from a linked Firebase event for Search Ads 360.

Implementations

impl ExternalConversionSource

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ExternalConversionSource.

pub fn from_i32(value: i32) -> Option<ExternalConversionSource>

Converts an i32 to a ExternalConversionSource, or None if value is not a valid variant.