Enum googapis::google::ads::googleads::v7::errors::ad_error_enum::AdError

#[repr(i32)]
pub enum AdError {

    Unspecified,
    Unknown,
    AdCustomizersNotSupportedForAdType,
    ApproximatelyTooLong,
    ApproximatelyTooShort,
    BadSnippet,
    CannotModifyAd,
    CannotSetBusinessNameIfUrlSet,
    CannotSetField,
    CannotSetFieldWithOriginAdIdSet,
    CannotSetFieldWithAdIdSetForSharing,
    CannotSetAllowFlexibleColorFalse,
    CannotSetColorControlWhenNativeFormatSetting,
    CannotSetUrl,
    CannotSetWithoutFinalUrls,
    CannotSetWithFinalUrls,
    CannotSetWithUrlData,
    CannotUseAdSubclassForOperator,
    CustomerNotApprovedMobileads,
    CustomerNotApprovedThirdpartyAds,
    CustomerNotApprovedThirdpartyRedirectAds,
    CustomerNotEligible,
    CustomerNotEligibleForUpdatingBeaconUrl,
    DimensionAlreadyInUnion,
    DimensionMustBeSet,
    DimensionNotInUnion,
    DisplayUrlCannotBeSpecified,
    DomesticPhoneNumberFormat,
    EmergencyPhoneNumber,
    EmptyField,
    FeedAttributeMustHaveMappingForTypeId,
    FeedAttributeMappingTypeMismatch,
    IllegalAdCustomizerTagUse,
    IllegalTagUse,
    InconsistentDimensions,
    InconsistentStatusInTemplateUnion,
    IncorrectLength,
    IneligibleForUpgrade,
    InvalidAdAddressCampaignTarget,
    InvalidAdType,
    InvalidAttributesForMobileImage,
    InvalidAttributesForMobileText,
    InvalidCallToActionText,
    InvalidCharacterForUrl,
    InvalidCountryCode,
    InvalidExpandedDynamicSearchAdTag,
    InvalidInput,
    InvalidMarkupLanguage,
    InvalidMobileCarrier,
    InvalidMobileCarrierTarget,
    InvalidNumberOfElements,
    InvalidPhoneNumberFormat,
    InvalidRichMediaCertifiedVendorFormatId,
    InvalidTemplateData,
    InvalidTemplateElementFieldType,
    InvalidTemplateId,
    LineTooWide,
    MissingAdCustomizerMapping,
    MissingAddressComponent,
    MissingAdvertisementName,
    MissingBusinessName,
    MissingDescription1,
    MissingDescription2,
    MissingDestinationUrlTag,
    MissingLandingPageUrlTag,
    MissingDimension,
    MissingDisplayUrl,
    MissingHeadline,
    MissingHeight,
    MissingImage,
    MissingMarketingImageOrProductVideos,
    MissingMarkupLanguages,
    MissingMobileCarrier,
    MissingPhone,
    MissingRequiredTemplateFields,
    MissingTemplateFieldValue,
    MissingText,
    MissingVisibleUrl,
    MissingWidth,
    MultipleDistinctFeedsUnsupported,
    MustUseTempAdUnionIdOnAdd,
    TooLong,
    TooShort,
    UnionDimensionsCannotChange,
    UnknownAddressComponent,
    UnknownFieldName,
    UnknownUniqueName,
    UnsupportedDimensions,
    UrlInvalidScheme,
    UrlInvalidTopLevelDomain,
    UrlMalformed,
    UrlNoHost,
    UrlNotEquivalent,
    UrlHostNameTooLong,
    UrlNoScheme,
    UrlNoTopLevelDomain,
    UrlPathNotAllowed,
    UrlPortNotAllowed,
    UrlQueryNotAllowed,
    UrlSchemeBeforeExpandedDynamicSearchAdTag,
    UserDoesNotHaveAccessToTemplate,
    InconsistentExpandableSettings,
    InvalidFormat,
    InvalidFieldText,
    ElementNotPresent,
    ImageError,
    ValueNotInRange,
    FieldNotPresent,
    AddressNotComplete,
    AddressInvalid,
    VideoRetrievalError,
    AudioError,
    InvalidYoutubeDisplayUrl,
    TooManyProductImages,
    TooManyProductVideos,
    IncompatibleAdTypeAndDevicePreference,
    CalltrackingNotSupportedForCountry,
    CarrierSpecificShortNumberNotAllowed,
    DisallowedNumberType,
    PhoneNumberNotSupportedForCountry,
    PhoneNumberNotSupportedWithCalltrackingForCountry,
    PremiumRateNumberNotAllowed,
    VanityPhoneNumberNotAllowed,
    InvalidCallConversionTypeId,
    CannotDisableCallConversionAndSetConversionTypeId,
    CannotSetPath2WithoutPath1,
    MissingDynamicSearchAdsSettingDomainName,
    IncompatibleWithRestrictionType,
    CustomerConsentForCallRecordingRequired,
    MissingImageOrMediaBundle,
    ProductTypeNotSupportedInThisCampaign,
    PlaceholderCannotHaveEmptyDefaultValue,
    PlaceholderCountdownFunctionCannotHaveDefaultValue,
    PlaceholderDefaultValueMissing,
    UnexpectedPlaceholderDefaultValue,
    AdCustomizersMayNotBeAdjacent,
    UpdatingAdWithNoEnabledAssociation,
    TooManyAdCustomizers,
    InvalidAdCustomizerFormat,
    NestedAdCustomizerSyntax,
    UnsupportedAdCustomizerSyntax,
    UnpairedBraceInAdCustomizerTag,
    MoreThanOneCountdownTagTypeExists,
    DateTimeInCountdownTagIsInvalid,
    DateTimeInCountdownTagIsPast,
    UnrecognizedAdCustomizerTagFound,
    MissingRequiredImageAspectRatio,

}

Enum describing possible ad errors.

Variants

Unspecified

Enum unspecified.

Unknown

The received error code is not known in this version.

AdCustomizersNotSupportedForAdType

Ad customizers are not supported for ad type.

ApproximatelyTooLong

Estimating character sizes the string is too long.

ApproximatelyTooShort

Estimating character sizes the string is too short.

BadSnippet

There is a problem with the snippet.

CannotModifyAd

Cannot modify an ad.

CannotSetBusinessNameIfUrlSet

business name and url cannot be set at the same time

CannotSetField

The specified field is incompatible with this ad’s type or settings.

CannotSetFieldWithOriginAdIdSet

Cannot set field when originAdId is set.

CannotSetFieldWithAdIdSetForSharing

Cannot set field when an existing ad id is set for sharing.

CannotSetAllowFlexibleColorFalse

Cannot set allowFlexibleColor false if no color is provided by user.

CannotSetColorControlWhenNativeFormatSetting

When user select native, no color control is allowed because we will always respect publisher color for native format serving.

CannotSetUrl

Cannot specify a url for the ad type

CannotSetWithoutFinalUrls

Cannot specify a tracking or mobile url without also setting final urls

CannotSetWithFinalUrls

Cannot specify a legacy url and a final url simultaneously

CannotSetWithUrlData

Cannot specify a urls in UrlData and in template fields simultaneously.

CannotUseAdSubclassForOperator

This operator cannot be used with a subclass of Ad.

CustomerNotApprovedMobileads

Customer is not approved for mobile ads.

CustomerNotApprovedThirdpartyAds

Customer is not approved for 3PAS richmedia ads.

CustomerNotApprovedThirdpartyRedirectAds

Customer is not approved for 3PAS redirect richmedia (Ad Exchange) ads.

CustomerNotEligible

Not an eligible customer

CustomerNotEligibleForUpdatingBeaconUrl

Customer is not eligible for updating beacon url

DimensionAlreadyInUnion

There already exists an ad with the same dimensions in the union.

DimensionMustBeSet

Ad’s dimension must be set before setting union dimension.

DimensionNotInUnion

Ad’s dimension must be included in the union dimensions.

DisplayUrlCannotBeSpecified

Display Url cannot be specified (applies to Ad Exchange Ads)

DomesticPhoneNumberFormat

Telephone number contains invalid characters or invalid format. Please re-enter your number using digits (0-9), dashes (-), and parentheses only.

EmergencyPhoneNumber

Emergency telephone numbers are not allowed. Please enter a valid domestic phone number to connect customers to your business.

EmptyField

A required field was not specified or is an empty string.

FeedAttributeMustHaveMappingForTypeId

A feed attribute referenced in an ad customizer tag is not in the ad customizer mapping for the feed.

FeedAttributeMappingTypeMismatch

The ad customizer field mapping for the feed attribute does not match the expected field type.

IllegalAdCustomizerTagUse

The use of ad customizer tags in the ad text is disallowed. Details in trigger.

IllegalTagUse

Tags of the form {PH_x}, where x is a number, are disallowed in ad text.

InconsistentDimensions

The dimensions of the ad are specified or derived in multiple ways and are not consistent.

InconsistentStatusInTemplateUnion

The status cannot differ among template ads of the same union.

IncorrectLength

The length of the string is not valid.

IneligibleForUpgrade

The ad is ineligible for upgrade.

InvalidAdAddressCampaignTarget

User cannot create mobile ad for countries targeted in specified campaign.

InvalidAdType

Invalid Ad type. A specific type of Ad is required.

InvalidAttributesForMobileImage

Headline, description or phone cannot be present when creating mobile image ad.

InvalidAttributesForMobileText

Image cannot be present when creating mobile text ad.

InvalidCallToActionText

Invalid call to action text.

InvalidCharacterForUrl

Invalid character in URL.

InvalidCountryCode

Creative’s country code is not valid.

InvalidExpandedDynamicSearchAdTag

Invalid use of Expanded Dynamic Search Ads tags ({lpurl} etc.)

InvalidInput

An input error whose real reason was not properly mapped (should not happen).

InvalidMarkupLanguage

An invalid markup language was entered.

InvalidMobileCarrier

An invalid mobile carrier was entered.

InvalidMobileCarrierTarget

Specified mobile carriers target a country not targeted by the campaign.

InvalidNumberOfElements

Wrong number of elements for given element type

InvalidPhoneNumberFormat

The format of the telephone number is incorrect. Please re-enter the number using the correct format.

InvalidRichMediaCertifiedVendorFormatId

The certified vendor format id is incorrect.

InvalidTemplateData

The template ad data contains validation errors.

InvalidTemplateElementFieldType

The template field doesn’t have have the correct type.

InvalidTemplateId

Invalid template id.

LineTooWide

After substituting replacement strings, the line is too wide.

MissingAdCustomizerMapping

The feed referenced must have ad customizer mapping to be used in a customizer tag.

MissingAddressComponent

Missing address component in template element address field.

MissingAdvertisementName

An ad name must be entered.

MissingBusinessName

Business name must be entered.

MissingDescription1

Description (line 2) must be entered.

MissingDescription2

Description (line 3) must be entered.

MissingDestinationUrlTag

The destination url must contain at least one tag (e.g. {lpurl})

MissingLandingPageUrlTag

The tracking url template of ExpandedDynamicSearchAd must contain at least one tag. (e.g. {lpurl})

MissingDimension

A valid dimension must be specified for this ad.

MissingDisplayUrl

A display URL must be entered.

MissingHeadline

Headline must be entered.

MissingHeight

A height must be entered.

MissingImage

An image must be entered.

MissingMarketingImageOrProductVideos

Marketing image or product videos are required.

MissingMarkupLanguages

The markup language in which your site is written must be entered.

MissingMobileCarrier

A mobile carrier must be entered.

MissingPhone

Phone number must be entered.

MissingRequiredTemplateFields

Missing required template fields

MissingTemplateFieldValue

Missing a required field value

MissingText

The ad must have text.

MissingVisibleUrl

A visible URL must be entered.

MissingWidth

A width must be entered.

MultipleDistinctFeedsUnsupported

Only 1 feed can be used as the source of ad customizer substitutions in a single ad.

MustUseTempAdUnionIdOnAdd

TempAdUnionId must be use when adding template ads.

TooLong

The string has too many characters.

TooShort

The string has too few characters.

UnionDimensionsCannotChange

Ad union dimensions cannot change for saved ads.

UnknownAddressComponent

Address component is not {country, lat, lng}.

UnknownFieldName

Unknown unique field name

UnknownUniqueName

Unknown unique name (template element type specifier)

UnsupportedDimensions

Unsupported ad dimension

UrlInvalidScheme

URL starts with an invalid scheme.

UrlInvalidTopLevelDomain

URL ends with an invalid top-level domain name.

UrlMalformed

URL contains illegal characters.

UrlNoHost

URL must contain a host name.

UrlNotEquivalent

URL not equivalent during upgrade.

UrlHostNameTooLong

URL host name too long to be stored as visible URL (applies to Ad Exchange ads)

UrlNoScheme

URL must start with a scheme.

UrlNoTopLevelDomain

URL should end in a valid domain extension, such as .com or .net.

UrlPathNotAllowed

URL must not end with a path.

UrlPortNotAllowed

URL must not specify a port.

UrlQueryNotAllowed

URL must not contain a query.

UrlSchemeBeforeExpandedDynamicSearchAdTag

A url scheme is not allowed in front of tag in tracking url template (e.g. http://{lpurl})

UserDoesNotHaveAccessToTemplate

The user does not have permissions to create a template ad for the given template.

InconsistentExpandableSettings

Expandable setting is inconsistent/wrong. For example, an AdX ad is invalid if it has a expandable vendor format but no expanding directions specified, or expanding directions is specified, but the vendor format is not expandable.

InvalidFormat

Format is invalid

InvalidFieldText

The text of this field did not match a pattern of allowed values.

ElementNotPresent

Template element is mising

ImageError

Error occurred during image processing

ValueNotInRange

The value is not within the valid range

FieldNotPresent

Template element field is not present

AddressNotComplete

Address is incomplete

AddressInvalid

Invalid address

VideoRetrievalError

Error retrieving specified video

AudioError

Error processing audio

InvalidYoutubeDisplayUrl

Display URL is incorrect for YouTube PYV ads

TooManyProductImages

Too many product Images in GmailAd

TooManyProductVideos

Too many product Videos in GmailAd

IncompatibleAdTypeAndDevicePreference

The device preference is not compatible with the ad type

CalltrackingNotSupportedForCountry

Call tracking is not supported for specified country.

CarrierSpecificShortNumberNotAllowed

Carrier specific short number is not allowed.

DisallowedNumberType

Specified phone number type is disallowed.

PhoneNumberNotSupportedForCountry

Phone number not supported for country.

PhoneNumberNotSupportedWithCalltrackingForCountry

Phone number not supported with call tracking enabled for country.

PremiumRateNumberNotAllowed

Premium rate phone number is not allowed.

VanityPhoneNumberNotAllowed

Vanity phone number is not allowed.

InvalidCallConversionTypeId

Invalid call conversion type id.

CannotDisableCallConversionAndSetConversionTypeId

Cannot disable call conversion and set conversion type id.

CannotSetPath2WithoutPath1

Cannot set path2 without path1.

MissingDynamicSearchAdsSettingDomainName

Missing domain name in campaign setting when adding expanded dynamic search ad.

IncompatibleWithRestrictionType

The associated ad is not compatible with restriction type.

CustomerConsentForCallRecordingRequired

Consent for call recording is required for creating/updating call only ads. Please see https://support.google.com/google-ads/answer/7412639.

MissingImageOrMediaBundle

Either an image or a media bundle is required in a display upload ad.

ProductTypeNotSupportedInThisCampaign

The display upload product type is not supported in this campaign.

PlaceholderCannotHaveEmptyDefaultValue

The default value of an ad placeholder can not be the empty string.

PlaceholderCountdownFunctionCannotHaveDefaultValue

Ad placeholders with countdown functions must not have a default value.

PlaceholderDefaultValueMissing

A previous ad placeholder that had a default value was found which means that all (non-countdown) placeholders must have a default value. This ad placeholder does not have a default value.

UnexpectedPlaceholderDefaultValue

A previous ad placeholder that did not have a default value was found which means that no placeholders may have a default value. This ad placeholder does have a default value.

AdCustomizersMayNotBeAdjacent

Two ad customizers may not be directly adjacent in an ad text. They must be separated by at least one character.

UpdatingAdWithNoEnabledAssociation

The ad is not associated with any enabled AdGroupAd, and cannot be updated.

TooManyAdCustomizers

Too many ad customizers in one asset.

InvalidAdCustomizerFormat

The ad customizer tag is recognized, but the format is invalid.

NestedAdCustomizerSyntax

Customizer tags cannot be nested.

UnsupportedAdCustomizerSyntax

The ad customizer syntax used in the ad is not supported.

UnpairedBraceInAdCustomizerTag

There exists unpaired brace in the ad customizer tag.

MoreThanOneCountdownTagTypeExists

More than one type of countdown tag exists among all text lines.

DateTimeInCountdownTagIsInvalid

Date time in the countdown tag is invalid.

DateTimeInCountdownTagIsPast

Date time in the countdown tag is in the past.

UnrecognizedAdCustomizerTagFound

Cannot recognize the ad customizer tag.

MissingRequiredImageAspectRatio

Missing required image aspect ratio.

Implementations

impl AdError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of AdError.

pub fn from_i32(value: i32) -> Option<AdError>

Converts an i32 to a AdError, or None if value is not a valid variant.