Enum googapis::google::ads::googleads::v7::errors::criterion_error_enum::CriterionError

#[repr(i32)]
pub enum CriterionError {

    Unspecified,
    Unknown,
    ConcreteTypeRequired,
    InvalidExcludedCategory,
    InvalidKeywordText,
    KeywordTextTooLong,
    KeywordHasTooManyWords,
    KeywordHasInvalidChars,
    InvalidPlacementUrl,
    InvalidUserList,
    InvalidUserInterest,
    InvalidFormatForPlacementUrl,
    PlacementUrlIsTooLong,
    PlacementUrlHasIllegalChar,
    PlacementUrlHasMultipleSitesInLine,
    PlacementIsNotAvailableForTargetingOrExclusion,
    InvalidTopicPath,
    InvalidYoutubeChannelId,
    InvalidYoutubeVideoId,
    YoutubeVerticalChannelDeprecated,
    YoutubeDemographicChannelDeprecated,
    YoutubeUrlUnsupported,
    CannotExcludeCriteriaType,
    CannotAddCriteriaType,
    CannotExcludeSimilarUserList,
    CannotAddClosedUserList,
    CannotAddDisplayOnlyListsToSearchOnlyCampaigns,
    CannotAddDisplayOnlyListsToSearchCampaigns,
    CannotAddDisplayOnlyListsToShoppingCampaigns,
    CannotAddUserInterestsToSearchCampaigns,
    CannotSetBidsOnCriterionTypeInSearchCampaigns,
    CannotAddUrlsToCriterionTypeForCampaignType,
    InvalidCombinedAudience,
    InvalidCustomAffinity,
    InvalidCustomIntent,
    InvalidCustomAudience,
    InvalidIpAddress,
    InvalidIpFormat,
    InvalidMobileApp,
    InvalidMobileAppCategory,
    InvalidCriterionId,
    CannotTargetCriterion,
    CannotTargetObsoleteCriterion,
    CriterionIdAndTypeMismatch,
    InvalidProximityRadius,
    InvalidProximityRadiusUnits,
    InvalidStreetaddressLength,
    InvalidCitynameLength,
    InvalidRegioncodeLength,
    InvalidRegionnameLength,
    InvalidPostalcodeLength,
    InvalidCountryCode,
    InvalidLatitude,
    InvalidLongitude,
    ProximityGeopointAndAddressBothCannotBeNull,
    InvalidProximityAddress,
    InvalidUserDomainName,
    CriterionParameterTooLong,
    AdScheduleTimeIntervalsOverlap,
    AdScheduleIntervalCannotSpanMultipleDays,
    AdScheduleInvalidTimeInterval,
    AdScheduleExceededIntervalsPerDayLimit,
    AdScheduleCriterionIdMismatchingFields,
    CannotBidModifyCriterionType,
    CannotBidModifyCriterionCampaignOptedOut,
    CannotBidModifyNegativeCriterion,
    BidModifierAlreadyExists,
    FeedIdNotAllowed,
    AccountIneligibleForCriteriaType,
    CriteriaTypeInvalidForBiddingStrategy,
    CannotExcludeCriterion,
    CannotRemoveCriterion,
    InvalidProductBiddingCategory,
    MissingShoppingSetting,
    InvalidMatchingFunction,
    LocationFilterNotAllowed,
    InvalidFeedForLocationFilter,
    LocationFilterInvalid,
    CannotSetGeoTargetConstantsWithFeedItemSets,
    InvalidLocationGroupRadius,
    InvalidLocationGroupRadiusUnit,
    CannotAttachCriteriaAtCampaignAndAdgroup,
    HotelLengthOfStayOverlapsWithExistingCriterion,
    HotelAdvanceBookingWindowOverlapsWithExistingCriterion,
    FieldIncompatibleWithNegativeTargeting,
    InvalidWebpageCondition,
    InvalidWebpageConditionUrl,
    WebpageConditionUrlCannotBeEmpty,
    WebpageConditionUrlUnsupportedProtocol,
    WebpageConditionUrlCannotBeIpAddress,
    WebpageConditionUrlDomainNotConsistentWithCampaignSetting,
    WebpageConditionUrlCannotBePublicSuffix,
    WebpageConditionUrlInvalidPublicSuffix,
    WebpageConditionUrlValueTrackValueNotSupported,
    WebpageCriterionUrlEqualsCanHaveOnlyOneCondition,
    WebpageCriterionNotSupportedOnNonDsaAdGroup,
    CannotTargetUserListForSmartDisplayCampaigns,
    CannotTargetPlacementsForSearchCampaigns,
    ListingScopeTooManyDimensionTypes,
    ListingScopeTooManyInOperators,
    ListingScopeInOperatorNotSupported,
    DuplicateListingDimensionType,
    DuplicateListingDimensionValue,
    CannotSetBidsOnListingGroupSubdivision,
    InvalidListingGroupHierarchy,
    ListingGroupUnitCannotHaveChildren,
    ListingGroupSubdivisionRequiresOthersCase,
    ListingGroupRequiresSameDimensionTypeAsSiblings,
    ListingGroupAlreadyExists,
    ListingGroupDoesNotExist,
    ListingGroupCannotBeRemoved,
    InvalidListingGroupType,
    ListingGroupAddMayOnlyUseTempId,
    ListingScopeTooLong,
    ListingScopeTooManyDimensions,
    ListingGroupTooLong,
    ListingGroupTreeTooDeep,
    InvalidListingDimension,
    InvalidListingDimensionType,
    AdvertiserNotOnAllowlistForCombinedAudienceOnDisplay,
    CannotTargetRemovedCombinedAudience,
    InvalidCombinedAudienceId,
    CannotTargetRemovedCustomAudience,
    HotelCheckInDateRangeOverlapsWithExistingCriterion,
    HotelCheckInDateRangeStartDateTooEarly,
    HotelCheckInDateRangeEndDateTooLate,
    HotelCheckInDateRangeReversed,

}

Enum describing possible criterion errors.

Variants

Unspecified

Enum unspecified.

Unknown

The received error code is not known in this version.

ConcreteTypeRequired

Concrete type of criterion is required for CREATE and UPDATE operations.

InvalidExcludedCategory

The category requested for exclusion is invalid.

InvalidKeywordText

Invalid keyword criteria text.

KeywordTextTooLong

Keyword text should be less than 80 chars.

KeywordHasTooManyWords

Keyword text has too many words.

KeywordHasInvalidChars

Keyword text has invalid characters or symbols.

InvalidPlacementUrl

Invalid placement URL.

InvalidUserList

Invalid user list criterion.

InvalidUserInterest

Invalid user interest criterion.

InvalidFormatForPlacementUrl

Placement URL has wrong format.

PlacementUrlIsTooLong

Placement URL is too long.

PlacementUrlHasIllegalChar

Indicates the URL contains an illegal character.

PlacementUrlHasMultipleSitesInLine

Indicates the URL contains multiple comma separated URLs.

PlacementIsNotAvailableForTargetingOrExclusion

Indicates the domain is blocked.

InvalidTopicPath

Invalid topic path.

InvalidYoutubeChannelId

The YouTube Channel Id is invalid.

InvalidYoutubeVideoId

The YouTube Video Id is invalid.

YoutubeVerticalChannelDeprecated

Indicates the placement is a YouTube vertical channel, which is no longer supported.

YoutubeDemographicChannelDeprecated

Indicates the placement is a YouTube demographic channel, which is no longer supported.

YoutubeUrlUnsupported

YouTube urls are not supported in Placement criterion. Use YouTubeChannel and YouTubeVideo criterion instead.

CannotExcludeCriteriaType

Criteria type can not be excluded by the customer, like AOL account type cannot target site type criteria.

CannotAddCriteriaType

Criteria type can not be targeted.

CannotExcludeSimilarUserList

Not allowed to exclude similar user list.

CannotAddClosedUserList

Not allowed to target a closed user list.

CannotAddDisplayOnlyListsToSearchOnlyCampaigns

Not allowed to add display only UserLists to search only campaigns.

CannotAddDisplayOnlyListsToSearchCampaigns

Not allowed to add display only UserLists to search plus campaigns.

CannotAddDisplayOnlyListsToShoppingCampaigns

Not allowed to add display only UserLists to shopping campaigns.

CannotAddUserInterestsToSearchCampaigns

Not allowed to add User interests to search only campaigns.

CannotSetBidsOnCriterionTypeInSearchCampaigns

Not allowed to set bids for this criterion type in search campaigns

CannotAddUrlsToCriterionTypeForCampaignType

Final URLs, URL Templates and CustomParameters cannot be set for the criterion types of Gender, AgeRange, UserList, Placement, MobileApp, and MobileAppCategory in search campaigns and shopping campaigns.

InvalidCombinedAudience

Invalid combined audience criterion.

InvalidCustomAffinity

Invalid custom affinity criterion.

InvalidCustomIntent

Invalid custom intent criterion.

InvalidCustomAudience

Invalid custom audience criterion.

InvalidIpAddress

IP address is not valid.

InvalidIpFormat

IP format is not valid.

InvalidMobileApp

Mobile application is not valid.

InvalidMobileAppCategory

Mobile application category is not valid.

InvalidCriterionId

The CriterionId does not exist or is of the incorrect type.

CannotTargetCriterion

The Criterion is not allowed to be targeted.

CannotTargetObsoleteCriterion

The criterion is not allowed to be targeted as it is deprecated.

CriterionIdAndTypeMismatch

The CriterionId is not valid for the type.

InvalidProximityRadius

Distance for the radius for the proximity criterion is invalid.

InvalidProximityRadiusUnits

Units for the distance for the radius for the proximity criterion is invalid.

InvalidStreetaddressLength

Street address in the address is not valid.

InvalidCitynameLength

City name in the address is not valid.

InvalidRegioncodeLength

Region code in the address is not valid.

InvalidRegionnameLength

Region name in the address is not valid.

InvalidPostalcodeLength

Postal code in the address is not valid.

InvalidCountryCode

Country code in the address is not valid.

InvalidLatitude

Latitude for the GeoPoint is not valid.

InvalidLongitude

Longitude for the GeoPoint is not valid.

ProximityGeopointAndAddressBothCannotBeNull

The Proximity input is not valid. Both address and geoPoint cannot be null.

InvalidProximityAddress

The Proximity address cannot be geocoded to a valid lat/long.

InvalidUserDomainName

User domain name is not valid.

CriterionParameterTooLong

Length of serialized criterion parameter exceeded size limit.

AdScheduleTimeIntervalsOverlap

Time interval in the AdSchedule overlaps with another AdSchedule.

AdScheduleIntervalCannotSpanMultipleDays

AdSchedule time interval cannot span multiple days.

AdScheduleInvalidTimeInterval

AdSchedule time interval specified is invalid, endTime cannot be earlier than startTime.

AdScheduleExceededIntervalsPerDayLimit

The number of AdSchedule entries in a day exceeds the limit.

AdScheduleCriterionIdMismatchingFields

CriteriaId does not match the interval of the AdSchedule specified.

CannotBidModifyCriterionType

Cannot set bid modifier for this criterion type.

CannotBidModifyCriterionCampaignOptedOut

Cannot bid modify criterion, since it is opted out of the campaign.

CannotBidModifyNegativeCriterion

Cannot set bid modifier for a negative criterion.

BidModifierAlreadyExists

Bid Modifier already exists. Use SET operation to update.

FeedIdNotAllowed

Feed Id is not allowed in these Location Groups.

AccountIneligibleForCriteriaType

The account may not use the requested criteria type. For example, some accounts are restricted to keywords only.

CriteriaTypeInvalidForBiddingStrategy

The requested criteria type cannot be used with campaign or ad group bidding strategy.

CannotExcludeCriterion

The Criterion is not allowed to be excluded.

CannotRemoveCriterion

The criterion is not allowed to be removed. For example, we cannot remove any of the device criterion.

InvalidProductBiddingCategory

Bidding categories do not form a valid path in the Shopping bidding category taxonomy.

MissingShoppingSetting

ShoppingSetting must be added to the campaign before ProductScope criteria can be added.

InvalidMatchingFunction

Matching function is invalid.

LocationFilterNotAllowed

Filter parameters not allowed for location groups targeting.

InvalidFeedForLocationFilter

Feed not found, or the feed is not an enabled location feed.

LocationFilterInvalid

Given location filter parameter is invalid for location groups targeting.

CannotSetGeoTargetConstantsWithFeedItemSets

Cannot set geo target constants and feed item sets at the same time.

InvalidLocationGroupRadius

The location group radius is in the range but not at the valid increment.

InvalidLocationGroupRadiusUnit

The location group radius unit is invalid.

CannotAttachCriteriaAtCampaignAndAdgroup

Criteria type cannot be associated with a campaign and its ad group(s) simultaneously.

HotelLengthOfStayOverlapsWithExistingCriterion

Range represented by hotel length of stay’s min nights and max nights overlaps with an existing criterion.

HotelAdvanceBookingWindowOverlapsWithExistingCriterion

Range represented by hotel advance booking window’s min days and max days overlaps with an existing criterion.

FieldIncompatibleWithNegativeTargeting

The field is not allowed to be set when the negative field is set to true, e.g. we don’t allow bids in negative ad group or campaign criteria.

InvalidWebpageCondition

The combination of operand and operator in webpage condition is invalid.

InvalidWebpageConditionUrl

The URL of webpage condition is invalid.

WebpageConditionUrlCannotBeEmpty

The URL of webpage condition cannot be empty or contain white space.

WebpageConditionUrlUnsupportedProtocol

The URL of webpage condition contains an unsupported protocol.

WebpageConditionUrlCannotBeIpAddress

The URL of webpage condition cannot be an IP address.

WebpageConditionUrlDomainNotConsistentWithCampaignSetting

The domain of the URL is not consistent with the domain in campaign setting.

WebpageConditionUrlCannotBePublicSuffix

The URL of webpage condition cannot be a public suffix itself.

WebpageConditionUrlInvalidPublicSuffix

The URL of webpage condition has an invalid public suffix.

WebpageConditionUrlValueTrackValueNotSupported

Value track parameter is not supported in webpage condition URL.

WebpageCriterionUrlEqualsCanHaveOnlyOneCondition

Only one URL-EQUALS webpage condition is allowed in a webpage criterion and it cannot be combined with other conditions.

WebpageCriterionNotSupportedOnNonDsaAdGroup

A webpage criterion cannot be added to a non-DSA ad group.

CannotTargetUserListForSmartDisplayCampaigns

Cannot add positive user list criteria in Smart Display campaigns.

CannotTargetPlacementsForSearchCampaigns

Cannot add positive placement criterion types in search campaigns.

ListingScopeTooManyDimensionTypes

Listing scope contains too many dimension types.

ListingScopeTooManyInOperators

Listing scope has too many IN operators.

ListingScopeInOperatorNotSupported

Listing scope contains IN operator on an unsupported dimension type.

DuplicateListingDimensionType

There are dimensions with duplicate dimension type.

DuplicateListingDimensionValue

There are dimensions with duplicate dimension value.

CannotSetBidsOnListingGroupSubdivision

Listing group SUBDIVISION nodes cannot have bids.

InvalidListingGroupHierarchy

Ad group is invalid due to the listing groups it contains.

ListingGroupUnitCannotHaveChildren

Listing group unit cannot have children.

ListingGroupSubdivisionRequiresOthersCase

Subdivided listing groups must have an “others” case.

ListingGroupRequiresSameDimensionTypeAsSiblings

Dimension type of listing group must be the same as that of its siblings.

ListingGroupAlreadyExists

Listing group cannot be added to the ad group because it already exists.

ListingGroupDoesNotExist

Listing group referenced in the operation was not found in the ad group.

ListingGroupCannotBeRemoved

Recursive removal failed because listing group subdivision is being created or modified in this request.

InvalidListingGroupType

Listing group type is not allowed for specified ad group criterion type.

ListingGroupAddMayOnlyUseTempId

Listing group in an ADD operation specifies a non temporary criterion id.

ListingScopeTooLong

The combined length of dimension values of the Listing scope criterion is too long.

ListingScopeTooManyDimensions

Listing scope contains too many dimensions.

ListingGroupTooLong

The combined length of dimension values of the Listing group criterion is too long.

ListingGroupTreeTooDeep

Listing group tree is too deep.

InvalidListingDimension

Listing dimension is invalid (e.g. dimension contains illegal value, dimension type is represented with wrong class, etc). Listing dimension value can not contain “==” or “&+”.

InvalidListingDimensionType

Listing dimension type is either invalid for campaigns of this type or cannot be used in the current context. BIDDING_CATEGORY_Lx and PRODUCT_TYPE_Lx dimensions must be used in ascending order of their levels: L1, L2, L3, L4, L5… The levels must be specified sequentially and start from L1. Furthermore, an “others” Listing group cannot be subdivided with a dimension of the same type but of a higher level (“others” BIDDING_CATEGORY_L3 can be subdivided with BRAND but not with BIDDING_CATEGORY_L4).

AdvertiserNotOnAllowlistForCombinedAudienceOnDisplay

Customer is not on allowlist for composite audience in display campaigns.

CannotTargetRemovedCombinedAudience

Cannot target on a removed combined audience.

InvalidCombinedAudienceId

Combined audience ID is invalid.

CannotTargetRemovedCustomAudience

Can not target removed combined audience.

HotelCheckInDateRangeOverlapsWithExistingCriterion

Range represented by hotel check-in date’s start date and end date overlaps with an existing criterion.

HotelCheckInDateRangeStartDateTooEarly

Start date is earlier than earliest allowed value of yesterday UTC.

HotelCheckInDateRangeEndDateTooLate

End date later is than latest allowed day of 330 days in the future UTC.

HotelCheckInDateRangeReversed

Start date is after end date.

Implementations

impl CriterionError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of CriterionError.

pub fn from_i32(value: i32) -> Option<CriterionError>

Converts an i32 to a CriterionError, or None if value is not a valid variant.