Enum googapis::google::ads::googleads::v7::errors::extension_setting_error_enum::ExtensionSettingError

#[repr(i32)]
pub enum ExtensionSettingError {

    Unspecified,
    Unknown,
    ExtensionsRequired,
    FeedTypeExtensionTypeMismatch,
    InvalidFeedType,
    InvalidFeedTypeForCustomerExtensionSetting,
    CannotChangeFeedItemOnCreate,
    CannotUpdateNewlyCreatedExtension,
    NoExistingAdGroupExtensionSettingForType,
    NoExistingCampaignExtensionSettingForType,
    NoExistingCustomerExtensionSettingForType,
    AdGroupExtensionSettingAlreadyExists,
    CampaignExtensionSettingAlreadyExists,
    CustomerExtensionSettingAlreadyExists,
    AdGroupFeedAlreadyExistsForPlaceholderType,
    CampaignFeedAlreadyExistsForPlaceholderType,
    CustomerFeedAlreadyExistsForPlaceholderType,
    ValueOutOfRange,
    CannotSetFieldWithFinalUrls,
    FinalUrlsNotSet,
    InvalidPhoneNumber,
    PhoneNumberNotSupportedForCountry,
    CarrierSpecificShortNumberNotAllowed,
    PremiumRateNumberNotAllowed,
    DisallowedNumberType,
    InvalidDomesticPhoneNumberFormat,
    VanityPhoneNumberNotAllowed,
    InvalidCountryCode,
    InvalidCallConversionTypeId,
    CustomerNotInAllowlistForCalltracking,
    CalltrackingNotSupportedForCountry,
    InvalidAppId,
    QuotesInReviewExtensionSnippet,
    HyphensInReviewExtensionSnippet,
    ReviewExtensionSourceNotEligible,
    SourceNameInReviewExtensionText,
    MissingField,
    InconsistentCurrencyCodes,
    PriceExtensionHasDuplicatedHeaders,
    PriceItemHasDuplicatedHeaderAndDescription,
    PriceExtensionHasTooFewItems,
    PriceExtensionHasTooManyItems,
    UnsupportedValue,
    InvalidDevicePreference,
    InvalidScheduleEnd,
    DateTimeMustBeInAccountTimeZone,
    OverlappingSchedulesNotAllowed,
    ScheduleEndNotAfterStart,
    TooManySchedulesPerDay,
    DuplicateExtensionFeedItemEdit,
    InvalidSnippetsHeader,
    PhoneNumberNotSupportedWithCalltrackingForCountry,
    CampaignTargetingMismatch,
    CannotOperateOnRemovedFeed,
    ExtensionTypeRequired,
    IncompatibleUnderlyingMatchingFunction,
    StartDateAfterEndDate,
    InvalidPriceFormat,
    PromotionInvalidTime,
    PromotionCannotSetPercentDiscountAndMoneyDiscount,
    PromotionCannotSetPromotionCodeAndOrdersOverAmount,
    TooManyDecimalPlacesSpecified,
    InvalidLanguageCode,
    UnsupportedLanguage,
    CustomerConsentForCallRecordingRequired,
    ExtensionSettingUpdateIsANoop,
    DisallowedText,

}

Enum describing possible extension setting errors.

Variants

Unspecified

Enum unspecified.

Unknown

The received error code is not known in this version.

ExtensionsRequired

A platform restriction was provided without input extensions or existing extensions.

FeedTypeExtensionTypeMismatch

The provided feed type does not correspond to the provided extensions.

InvalidFeedType

The provided feed type cannot be used.

InvalidFeedTypeForCustomerExtensionSetting

The provided feed type cannot be used at the customer level.

CannotChangeFeedItemOnCreate

Cannot change a feed item field on a CREATE operation.

CannotUpdateNewlyCreatedExtension

Cannot update an extension that is not already in this setting.

NoExistingAdGroupExtensionSettingForType

There is no existing AdGroupExtensionSetting for this type.

NoExistingCampaignExtensionSettingForType

There is no existing CampaignExtensionSetting for this type.

NoExistingCustomerExtensionSettingForType

There is no existing CustomerExtensionSetting for this type.

AdGroupExtensionSettingAlreadyExists

The AdGroupExtensionSetting already exists. UPDATE should be used to modify the existing AdGroupExtensionSetting.

CampaignExtensionSettingAlreadyExists

The CampaignExtensionSetting already exists. UPDATE should be used to modify the existing CampaignExtensionSetting.

CustomerExtensionSettingAlreadyExists

The CustomerExtensionSetting already exists. UPDATE should be used to modify the existing CustomerExtensionSetting.

AdGroupFeedAlreadyExistsForPlaceholderType

An active ad group feed already exists for this place holder type.

CampaignFeedAlreadyExistsForPlaceholderType

An active campaign feed already exists for this place holder type.

CustomerFeedAlreadyExistsForPlaceholderType

An active customer feed already exists for this place holder type.

ValueOutOfRange

Value is not within the accepted range.

CannotSetFieldWithFinalUrls

Cannot simultaneously set specified field with final urls.

FinalUrlsNotSet

Must set field with final urls.

InvalidPhoneNumber

Phone number for a call extension is invalid.

PhoneNumberNotSupportedForCountry

Phone number for a call extension is not supported for the given country code.

CarrierSpecificShortNumberNotAllowed

A carrier specific number in short format is not allowed for call extensions.

PremiumRateNumberNotAllowed

Premium rate numbers are not allowed for call extensions.

DisallowedNumberType

Phone number type for a call extension is not allowed.

InvalidDomesticPhoneNumberFormat

Phone number for a call extension does not meet domestic format requirements.

VanityPhoneNumberNotAllowed

Vanity phone numbers (i.e. those including letters) are not allowed for call extensions.

InvalidCountryCode

Country code provided for a call extension is invalid.

InvalidCallConversionTypeId

Call conversion type id provided for a call extension is invalid.

CustomerNotInAllowlistForCalltracking

For a call extension, the customer is not on the allow-list for call tracking.

CalltrackingNotSupportedForCountry

Call tracking is not supported for the given country for a call extension.

InvalidAppId

App id provided for an app extension is invalid.

QuotesInReviewExtensionSnippet

Quotation marks present in the review text for a review extension.

HyphensInReviewExtensionSnippet

Hyphen character present in the review text for a review extension.

ReviewExtensionSourceNotEligible

A blocked review source name or url was provided for a review extension.

SourceNameInReviewExtensionText

Review source name should not be found in the review text.

MissingField

Field must be set.

InconsistentCurrencyCodes

Inconsistent currency codes.

PriceExtensionHasDuplicatedHeaders

Price extension cannot have duplicated headers.

PriceItemHasDuplicatedHeaderAndDescription

Price item cannot have duplicated header and description.

PriceExtensionHasTooFewItems

Price extension has too few items

PriceExtensionHasTooManyItems

Price extension has too many items

UnsupportedValue

The input value is not currently supported.

InvalidDevicePreference

Unknown or unsupported device preference.

InvalidScheduleEnd

Invalid feed item schedule end time (i.e., endHour = 24 and endMinute != 0).

DateTimeMustBeInAccountTimeZone

Date time zone does not match the account’s time zone.

OverlappingSchedulesNotAllowed

Overlapping feed item schedule times (e.g., 7-10AM and 8-11AM) are not allowed.

ScheduleEndNotAfterStart

Feed item schedule end time must be after start time.

TooManySchedulesPerDay

There are too many feed item schedules per day.

DuplicateExtensionFeedItemEdit

Cannot edit the same extension feed item more than once in the same request.

InvalidSnippetsHeader

Invalid structured snippet header.

PhoneNumberNotSupportedWithCalltrackingForCountry

Phone number with call tracking enabled is not supported for the specified country.

CampaignTargetingMismatch

The targeted adgroup must belong to the targeted campaign.

CannotOperateOnRemovedFeed

The feed used by the ExtensionSetting is removed and cannot be operated on. Remove the ExtensionSetting to allow a new one to be created using an active feed.

ExtensionTypeRequired

The ExtensionFeedItem type is required for this operation.

IncompatibleUnderlyingMatchingFunction

The matching function that links the extension feed to the customer, campaign, or ad group is not compatible with the ExtensionSetting services.

StartDateAfterEndDate

Start date must be before end date.

InvalidPriceFormat

Input price is not in a valid format.

PromotionInvalidTime

The promotion time is invalid.

PromotionCannotSetPercentDiscountAndMoneyDiscount

Cannot set both percent discount and money discount fields.

PromotionCannotSetPromotionCodeAndOrdersOverAmount

Cannot set both promotion code and orders over amount fields.

TooManyDecimalPlacesSpecified

This field has too many decimal places specified.

InvalidLanguageCode

The language code is not valid.

UnsupportedLanguage

The language is not supported.

CustomerConsentForCallRecordingRequired

Customer hasn’t consented for call recording, which is required for adding/updating call extensions. Please see https://support.google.com/google-ads/answer/7412639.

ExtensionSettingUpdateIsANoop

The UPDATE operation does not specify any fields other than the resource name in the update mask.

DisallowedText

The extension contains text which has been prohibited on policy grounds.

Implementations

impl ExtensionSettingError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ExtensionSettingError.

pub fn from_i32(value: i32) -> Option<ExtensionSettingError>

Converts an i32 to a ExtensionSettingError, or None if value is not a valid variant.