Enum googapis::google::ads::googleads::v7::errors::query_error_enum::QueryError

#[repr(i32)]
pub enum QueryError {

    Unspecified,
    Unknown,
    QueryError,
    BadEnumConstant,
    BadEscapeSequence,
    BadFieldName,
    BadLimitValue,
    BadNumber,
    BadOperator,
    BadParameterName,
    BadParameterValue,
    BadResourceTypeInFromClause,
    BadSymbol,
    BadValue,
    DateRangeTooWide,
    DateRangeTooNarrow,
    ExpectedAnd,
    ExpectedBy,
    ExpectedDimensionFieldInSelectClause,
    ExpectedFiltersOnDateRange,
    ExpectedFrom,
    ExpectedList,
    ExpectedReferencedFieldInSelectClause,
    ExpectedSelect,
    ExpectedSingleValue,
    ExpectedValueWithBetweenOperator,
    InvalidDateFormat,
    InvalidStringValue,
    InvalidValueWithBetweenOperator,
    InvalidValueWithDuringOperator,
    InvalidValueWithLikeOperator,
    OperatorFieldMismatch,
    ProhibitedEmptyListInCondition,
    ProhibitedEnumConstant,
    ProhibitedFieldCombinationInSelectClause,
    ProhibitedFieldInOrderByClause,
    ProhibitedFieldInSelectClause,
    ProhibitedFieldInWhereClause,
    ProhibitedResourceTypeInFromClause,
    ProhibitedResourceTypeInSelectClause,
    ProhibitedResourceTypeInWhereClause,
    ProhibitedMetricInSelectOrWhereClause,
    ProhibitedSegmentInSelectOrWhereClause,
    ProhibitedSegmentWithMetricInSelectOrWhereClause,
    LimitValueTooLow,
    ProhibitedNewlineInString,
    ProhibitedValueCombinationInList,
    ProhibitedValueCombinationWithBetweenOperator,
    StringNotTerminated,
    TooManySegments,
    UnexpectedEndOfQuery,
    UnexpectedFromClause,
    UnrecognizedField,
    UnexpectedInput,
    RequestedMetricsForManager,
    FilterHasTooManyValues,

}

Enum describing possible query errors.

Variants

Unspecified

Name unspecified.

Unknown

The received error code is not known in this version.

QueryError

Returned if all other query error reasons are not applicable.

BadEnumConstant

A condition used in the query references an invalid enum constant.

BadEscapeSequence

Query contains an invalid escape sequence.

BadFieldName

Field name is invalid.

BadLimitValue

Limit value is invalid (i.e. not a number)

BadNumber

Encountered number can not be parsed.

BadOperator

Invalid operator encountered.

BadParameterName

Parameter unknown or not supported.

BadParameterValue

Parameter have invalid value.

BadResourceTypeInFromClause

Invalid resource type was specified in the FROM clause.

BadSymbol

Non-ASCII symbol encountered outside of strings.

BadValue

Value is invalid.

DateRangeTooWide

Date filters fail to restrict date to a range smaller than 31 days. Applicable if the query is segmented by date.

DateRangeTooNarrow

Filters on date/week/month/quarter have a start date after end date.

ExpectedAnd

Expected AND between values with BETWEEN operator.

ExpectedBy

Expecting ORDER BY to have BY.

ExpectedDimensionFieldInSelectClause

There was no dimension field selected.

ExpectedFiltersOnDateRange

Missing filters on date related fields.

ExpectedFrom

Missing FROM clause.

ExpectedList

The operator used in the conditions requires the value to be a list.

ExpectedReferencedFieldInSelectClause

Fields used in WHERE or ORDER BY clauses are missing from the SELECT clause.

ExpectedSelect

SELECT is missing at the beginning of query.

ExpectedSingleValue

A list was passed as a value to a condition whose operator expects a single value.

ExpectedValueWithBetweenOperator

Missing one or both values with BETWEEN operator.

InvalidDateFormat

Invalid date format. Expected ‘YYYY-MM-DD’.

InvalidStringValue

Value passed was not a string when it should have been. I.e., it was a number or unquoted literal.

InvalidValueWithBetweenOperator

A String value passed to the BETWEEN operator does not parse as a date.

InvalidValueWithDuringOperator

The value passed to the DURING operator is not a Date range literal

InvalidValueWithLikeOperator

A non-string value was passed to the LIKE operator.

OperatorFieldMismatch

An operator was provided that is inapplicable to the field being filtered.

ProhibitedEmptyListInCondition

A Condition was found with an empty list.

ProhibitedEnumConstant

A condition used in the query references an unsupported enum constant.

ProhibitedFieldCombinationInSelectClause

Fields that are not allowed to be selected together were included in the SELECT clause.

ProhibitedFieldInOrderByClause

A field that is not orderable was included in the ORDER BY clause.

ProhibitedFieldInSelectClause

A field that is not selectable was included in the SELECT clause.

ProhibitedFieldInWhereClause

A field that is not filterable was included in the WHERE clause.

ProhibitedResourceTypeInFromClause

Resource type specified in the FROM clause is not supported by this service.

ProhibitedResourceTypeInSelectClause

A field that comes from an incompatible resource was included in the SELECT clause.

ProhibitedResourceTypeInWhereClause

A field that comes from an incompatible resource was included in the WHERE clause.

ProhibitedMetricInSelectOrWhereClause

A metric incompatible with the main resource or other selected segmenting resources was included in the SELECT or WHERE clause.

ProhibitedSegmentInSelectOrWhereClause

A segment incompatible with the main resource or other selected segmenting resources was included in the SELECT or WHERE clause.

ProhibitedSegmentWithMetricInSelectOrWhereClause

A segment in the SELECT clause is incompatible with a metric in the SELECT or WHERE clause.

LimitValueTooLow

The value passed to the limit clause is too low.

ProhibitedNewlineInString

Query has a string containing a newline character.

ProhibitedValueCombinationInList

List contains values of different types.

ProhibitedValueCombinationWithBetweenOperator

The values passed to the BETWEEN operator are not of the same type.

StringNotTerminated

Query contains unterminated string.

TooManySegments

Too many segments are specified in SELECT clause.

UnexpectedEndOfQuery

Query is incomplete and cannot be parsed.

UnexpectedFromClause

FROM clause cannot be specified in this query.

UnrecognizedField

Query contains one or more unrecognized fields.

UnexpectedInput

Query has an unexpected extra part.

RequestedMetricsForManager

Metrics cannot be requested for a manager account. To retrieve metrics, issue separate requests against each client account under the manager account.

FilterHasTooManyValues

The number of values (right-hand-side operands) in a filter exceeds the limit.

Implementations

impl QueryError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of QueryError.

pub fn from_i32(value: i32) -> Option<QueryError>

Converts an i32 to a QueryError, or None if value is not a valid variant.