Struct googapis::google::ads::googleads::v7::resources::AdGroupCriterion

pub struct AdGroupCriterion {

    pub resource_name: String,
    pub criterion_id: Option<i64>,
    pub display_name: String,
    pub status: i32,
    pub quality_info: Option<QualityInfo>,
    pub ad_group: Option<String>,
    pub type: i32,
    pub negative: Option<bool>,
    pub system_serving_status: i32,
    pub approval_status: i32,
    pub disapproval_reasons: Vec<String>,
    pub labels: Vec<String>,
    pub bid_modifier: Option<f64>,
    pub cpc_bid_micros: Option<i64>,
    pub cpm_bid_micros: Option<i64>,
    pub cpv_bid_micros: Option<i64>,
    pub percent_cpc_bid_micros: Option<i64>,
    pub effective_cpc_bid_micros: Option<i64>,
    pub effective_cpm_bid_micros: Option<i64>,
    pub effective_cpv_bid_micros: Option<i64>,
    pub effective_percent_cpc_bid_micros: Option<i64>,
    pub effective_cpc_bid_source: i32,
    pub effective_cpm_bid_source: i32,
    pub effective_cpv_bid_source: i32,
    pub effective_percent_cpc_bid_source: i32,
    pub position_estimates: Option<PositionEstimates>,
    pub final_urls: Vec<String>,
    pub final_mobile_urls: Vec<String>,
    pub final_url_suffix: Option<String>,
    pub tracking_url_template: Option<String>,
    pub url_custom_parameters: Vec<CustomParameter>,
    pub criterion: Option<Criterion>,

}

An ad group criterion.

Fields
resource_name: String

Immutable. The resource name of the ad group criterion. Ad group criterion resource names have the form:

customers/{customer_id}/adGroupCriteria/{ad_group_id}~{criterion_id}

criterion_id: Option<i64>

Output only. The ID of the criterion.

This field is ignored for mutates.

display_name: String

Output only. The display name of the criterion.

This field is ignored for mutates.

status: i32

The status of the criterion.

This is the status of the ad group criterion entity, set by the client. Note: UI reports may incorporate additional information that affects whether a criterion is eligible to run. In some cases a criterion that’s REMOVED in the API can still show as enabled in the UI. For example, campaigns by default show to users of all age ranges unless excluded. The UI will show each age range as “enabled”, since they’re eligible to see the ads; but AdGroupCriterion.status will show “removed”, since no positive criterion was added.

quality_info: Option<QualityInfo>

Output only. Information regarding the quality of the criterion.

ad_group: Option<String>

Immutable. The ad group to which the criterion belongs.

type: i32

Output only. The type of the criterion.

negative: Option<bool>

Immutable. Whether to target (false) or exclude (true) the criterion.

This field is immutable. To switch a criterion from positive to negative, remove then re-add it.

system_serving_status: i32

Output only. Serving status of the criterion.

approval_status: i32

Output only. Approval status of the criterion.

disapproval_reasons: Vec<String>

Output only. List of disapproval reasons of the criterion.

The different reasons for disapproving a criterion can be found here: https://support.google.com/adspolicy/answer/6008942

This field is read-only.

labels: Vec<String>

Output only. The resource names of labels attached to this ad group criterion.

bid_modifier: Option<f64>

The modifier for the bid when the criterion matches. The modifier must be in the range: 0.1 - 10.0. Most targetable criteria types support modifiers.

cpc_bid_micros: Option<i64>

The CPC (cost-per-click) bid.

cpm_bid_micros: Option<i64>

The CPM (cost-per-thousand viewable impressions) bid.

cpv_bid_micros: Option<i64>

The CPV (cost-per-view) bid.

percent_cpc_bid_micros: Option<i64>

The CPC bid amount, expressed as a fraction of the advertised price for some good or service. The valid range for the fraction is [0,1) and the value stored here is 1,000,000 * [fraction].

effective_cpc_bid_micros: Option<i64>

Output only. The effective CPC (cost-per-click) bid.

effective_cpm_bid_micros: Option<i64>

Output only. The effective CPM (cost-per-thousand viewable impressions) bid.

effective_cpv_bid_micros: Option<i64>

Output only. The effective CPV (cost-per-view) bid.

effective_percent_cpc_bid_micros: Option<i64>

Output only. The effective Percent CPC bid amount.

effective_cpc_bid_source: i32

Output only. Source of the effective CPC bid.

effective_cpm_bid_source: i32

Output only. Source of the effective CPM bid.

effective_cpv_bid_source: i32

Output only. Source of the effective CPV bid.

effective_percent_cpc_bid_source: i32

Output only. Source of the effective Percent CPC bid.

position_estimates: Option<PositionEstimates>

Output only. Estimates for criterion bids at various positions.

final_urls: Vec<String>

The list of possible final URLs after all cross-domain redirects for the ad.

final_mobile_urls: Vec<String>

The list of possible final mobile URLs after all cross-domain redirects.

final_url_suffix: Option<String>

URL template for appending params to final URL.

tracking_url_template: Option<String>

The URL template for constructing a tracking URL.

url_custom_parameters: Vec<CustomParameter>

The list of mappings used to substitute custom parameter tags in a tracking_url_template, final_urls, or mobile_final_urls.

criterion: Option<Criterion>

The ad group criterion.

Exactly one must be set.

Implementations

impl AdGroupCriterion

pub fn status(&self) -> AdGroupCriterionStatus

Returns the enum value of status, or the default if the field is set to an invalid enum value.

pub fn set_status(&mut self, value: AdGroupCriterionStatus)

Sets status to the provided enum value.

pub fn effective_cpc_bid_source(&self) -> BiddingSource

Returns the enum value of effective_cpc_bid_source, or the default if the field is set to an invalid enum value.

pub fn set_effective_cpc_bid_source(&mut self, value: BiddingSource)

Sets effective_cpc_bid_source to the provided enum value.

pub fn effective_cpm_bid_source(&self) -> BiddingSource

Returns the enum value of effective_cpm_bid_source, or the default if the field is set to an invalid enum value.

pub fn set_effective_cpm_bid_source(&mut self, value: BiddingSource)

Sets effective_cpm_bid_source to the provided enum value.

pub fn effective_cpv_bid_source(&self) -> BiddingSource

Returns the enum value of effective_cpv_bid_source, or the default if the field is set to an invalid enum value.

pub fn set_effective_cpv_bid_source(&mut self, value: BiddingSource)

Sets effective_cpv_bid_source to the provided enum value.

pub fn type(&self) -> CriterionType

Returns the enum value of type, or the default if the field is set to an invalid enum value.

pub fn set_type(&mut self, value: CriterionType)

Sets type to the provided enum value.

pub fn effective_percent_cpc_bid_source(&self) -> BiddingSource

Returns the enum value of effective_percent_cpc_bid_source, or the default if the field is set to an invalid enum value.

pub fn set_effective_percent_cpc_bid_source(&mut self, value: BiddingSource)

Sets effective_percent_cpc_bid_source to the provided enum value.

pub fn system_serving_status(&self) -> CriterionSystemServingStatus

Returns the enum value of system_serving_status, or the default if the field is set to an invalid enum value.

pub fn set_system_serving_status(&mut self, value: CriterionSystemServingStatus)

Sets system_serving_status to the provided enum value.

pub fn approval_status(&self) -> AdGroupCriterionApprovalStatus

Returns the enum value of approval_status, or the default if the field is set to an invalid enum value.

pub fn set_approval_status(&mut self, value: AdGroupCriterionApprovalStatus)

Sets approval_status to the provided enum value.

pub fn criterion_id(&self) -> i64

Returns the value of criterion_id, or the default value if criterion_id is unset.

pub fn ad_group(&self) -> &str

Returns the value of ad_group, or the default value if ad_group is unset.

pub fn negative(&self) -> bool

Returns the value of negative, or the default value if negative is unset.

pub fn bid_modifier(&self) -> f64

Returns the value of bid_modifier, or the default value if bid_modifier is unset.

pub fn cpc_bid_micros(&self) -> i64

Returns the value of cpc_bid_micros, or the default value if cpc_bid_micros is unset.

pub fn cpm_bid_micros(&self) -> i64

Returns the value of cpm_bid_micros, or the default value if cpm_bid_micros is unset.

pub fn cpv_bid_micros(&self) -> i64

Returns the value of cpv_bid_micros, or the default value if cpv_bid_micros is unset.

pub fn percent_cpc_bid_micros(&self) -> i64

Returns the value of percent_cpc_bid_micros, or the default value if percent_cpc_bid_micros is unset.

pub fn effective_cpc_bid_micros(&self) -> i64

Returns the value of effective_cpc_bid_micros, or the default value if effective_cpc_bid_micros is unset.

pub fn effective_cpm_bid_micros(&self) -> i64

Returns the value of effective_cpm_bid_micros, or the default value if effective_cpm_bid_micros is unset.

pub fn effective_cpv_bid_micros(&self) -> i64

Returns the value of effective_cpv_bid_micros, or the default value if effective_cpv_bid_micros is unset.

pub fn effective_percent_cpc_bid_micros(&self) -> i64

Returns the value of effective_percent_cpc_bid_micros, or the default value if effective_percent_cpc_bid_micros is unset.

pub fn final_url_suffix(&self) -> &str

Returns the value of final_url_suffix, or the default value if final_url_suffix is unset.

pub fn tracking_url_template(&self) -> &str

Returns the value of tracking_url_template, or the default value if tracking_url_template is unset.