Module googapis::google::ads::googleads::v7::services::apply_recommendation_operation

Nested message and enum types in ApplyRecommendationOperation.

Structs

CallExtensionParameters

Parameters to use when applying call extension recommendation.

CalloutExtensionParameters

Parameters to use when applying callout extension recommendation.

CampaignBudgetParameters

Parameters to use when applying a campaign budget recommendation.

KeywordParameters

Parameters to use when applying keyword recommendation.

MoveUnusedBudgetParameters

Parameters to use when applying move unused budget recommendation.

ResponsiveSearchAdParameters

Parameters to use when applying a responsive search ad recommendation.

SitelinkExtensionParameters

Parameters to use when applying sitelink extension recommendation.

TargetCpaOptInParameters

Parameters to use when applying Target CPA recommendation.

TargetRoasOptInParameters

Parameters to use when applying a Target ROAS opt-in recommendation.

TextAdParameters

Parameters to use when applying a text ad recommendation.

Enums

ApplyParameters

Parameters to use when applying the recommendation.