Struct googapis::google::ads::googleads::v8::common::Metrics

pub struct Metrics {

    pub absolute_top_impression_percentage: Option<f64>,
    pub active_view_cpm: Option<f64>,
    pub active_view_ctr: Option<f64>,
    pub active_view_impressions: Option<i64>,
    pub active_view_measurability: Option<f64>,
    pub active_view_measurable_cost_micros: Option<i64>,
    pub active_view_measurable_impressions: Option<i64>,
    pub active_view_viewability: Option<f64>,
    pub all_conversions_from_interactions_rate: Option<f64>,
    pub all_conversions_value: Option<f64>,
    pub all_conversions_value_by_conversion_date: f64,
    pub all_conversions: Option<f64>,
    pub all_conversions_by_conversion_date: f64,
    pub all_conversions_value_per_cost: Option<f64>,
    pub all_conversions_from_click_to_call: Option<f64>,
    pub all_conversions_from_directions: Option<f64>,
    pub all_conversions_from_interactions_value_per_interaction: Option<f64>,
    pub all_conversions_from_menu: Option<f64>,
    pub all_conversions_from_order: Option<f64>,
    pub all_conversions_from_other_engagement: Option<f64>,
    pub all_conversions_from_store_visit: Option<f64>,
    pub all_conversions_from_store_website: Option<f64>,
    pub average_cost: Option<f64>,
    pub average_cpc: Option<f64>,
    pub average_cpe: Option<f64>,
    pub average_cpm: Option<f64>,
    pub average_cpv: Option<f64>,
    pub average_page_views: Option<f64>,
    pub average_time_on_site: Option<f64>,
    pub benchmark_average_max_cpc: Option<f64>,
    pub benchmark_ctr: Option<f64>,
    pub bounce_rate: Option<f64>,
    pub clicks: Option<i64>,
    pub combined_clicks: Option<i64>,
    pub combined_clicks_per_query: Option<f64>,
    pub combined_queries: Option<i64>,
    pub content_budget_lost_impression_share: Option<f64>,
    pub content_impression_share: Option<f64>,
    pub conversion_last_received_request_date_time: Option<String>,
    pub conversion_last_conversion_date: Option<String>,
    pub content_rank_lost_impression_share: Option<f64>,
    pub conversions_from_interactions_rate: Option<f64>,
    pub conversions_value: Option<f64>,
    pub conversions_value_by_conversion_date: f64,
    pub conversions_value_per_cost: Option<f64>,
    pub conversions_from_interactions_value_per_interaction: Option<f64>,
    pub conversions: Option<f64>,
    pub conversions_by_conversion_date: f64,
    pub cost_micros: Option<i64>,
    pub cost_per_all_conversions: Option<f64>,
    pub cost_per_conversion: Option<f64>,
    pub cost_per_current_model_attributed_conversion: Option<f64>,
    pub cross_device_conversions: Option<f64>,
    pub ctr: Option<f64>,
    pub current_model_attributed_conversions: Option<f64>,
    pub current_model_attributed_conversions_from_interactions_rate: Option<f64>,
    pub current_model_attributed_conversions_from_interactions_value_per_interaction: Option<f64>,
    pub current_model_attributed_conversions_value: Option<f64>,
    pub current_model_attributed_conversions_value_per_cost: Option<f64>,
    pub engagement_rate: Option<f64>,
    pub engagements: Option<i64>,
    pub hotel_average_lead_value_micros: Option<f64>,
    pub hotel_price_difference_percentage: Option<f64>,
    pub hotel_eligible_impressions: Option<i64>,
    pub historical_creative_quality_score: i32,
    pub historical_landing_page_quality_score: i32,
    pub historical_quality_score: Option<i64>,
    pub historical_search_predicted_ctr: i32,
    pub gmail_forwards: Option<i64>,
    pub gmail_saves: Option<i64>,
    pub gmail_secondary_clicks: Option<i64>,
    pub impressions_from_store_reach: Option<i64>,
    pub impressions: Option<i64>,
    pub interaction_rate: Option<f64>,
    pub interactions: Option<i64>,
    pub interaction_event_types: Vec<i32>,
    pub invalid_click_rate: Option<f64>,
    pub invalid_clicks: Option<i64>,
    pub message_chats: Option<i64>,
    pub message_impressions: Option<i64>,
    pub message_chat_rate: Option<f64>,
    pub mobile_friendly_clicks_percentage: Option<f64>,
    pub optimization_score_uplift: Option<f64>,
    pub optimization_score_url: Option<String>,
    pub organic_clicks: Option<i64>,
    pub organic_clicks_per_query: Option<f64>,
    pub organic_impressions: Option<i64>,
    pub organic_impressions_per_query: Option<f64>,
    pub organic_queries: Option<i64>,
    pub percent_new_visitors: Option<f64>,
    pub phone_calls: Option<i64>,
    pub phone_impressions: Option<i64>,
    pub phone_through_rate: Option<f64>,
    pub relative_ctr: Option<f64>,
    pub search_absolute_top_impression_share: Option<f64>,
    pub search_budget_lost_absolute_top_impression_share: Option<f64>,
    pub search_budget_lost_impression_share: Option<f64>,
    pub search_budget_lost_top_impression_share: Option<f64>,
    pub search_click_share: Option<f64>,
    pub search_exact_match_impression_share: Option<f64>,
    pub search_impression_share: Option<f64>,
    pub search_rank_lost_absolute_top_impression_share: Option<f64>,
    pub search_rank_lost_impression_share: Option<f64>,
    pub search_rank_lost_top_impression_share: Option<f64>,
    pub search_top_impression_share: Option<f64>,
    pub speed_score: Option<i64>,
    pub top_impression_percentage: Option<f64>,
    pub valid_accelerated_mobile_pages_clicks_percentage: Option<f64>,
    pub value_per_all_conversions: Option<f64>,
    pub value_per_all_conversions_by_conversion_date: Option<f64>,
    pub value_per_conversion: Option<f64>,
    pub value_per_conversions_by_conversion_date: Option<f64>,
    pub value_per_current_model_attributed_conversion: Option<f64>,
    pub video_quartile_p100_rate: Option<f64>,
    pub video_quartile_p25_rate: Option<f64>,
    pub video_quartile_p50_rate: Option<f64>,
    pub video_quartile_p75_rate: Option<f64>,
    pub video_view_rate: Option<f64>,
    pub video_views: Option<i64>,
    pub view_through_conversions: Option<i64>,
    pub sk_ad_network_conversions: i64,

}

Metrics data.

Fields
absolute_top_impression_percentage: Option<f64>

The percent of your ad impressions that are shown as the very first ad above the organic search results.

active_view_cpm: Option<f64>

Average cost of viewable impressions (active_view_impressions).

active_view_ctr: Option<f64>

Active view measurable clicks divided by active view viewable impressions. This metric is reported only for display network.

active_view_impressions: Option<i64>

A measurement of how often your ad has become viewable on a Display Network site.

active_view_measurability: Option<f64>

The ratio of impressions that could be measured by Active View over the number of served impressions.

active_view_measurable_cost_micros: Option<i64>

The cost of the impressions you received that were measurable by Active View.

active_view_measurable_impressions: Option<i64>

The number of times your ads are appearing on placements in positions where they can be seen.

active_view_viewability: Option<f64>

The percentage of time when your ad appeared on an Active View enabled site (measurable impressions) and was viewable (viewable impressions).

all_conversions_from_interactions_rate: Option<f64>

All conversions from interactions (as oppose to view through conversions) divided by the number of ad interactions.

all_conversions_value: Option<f64>

The value of all conversions.

all_conversions_value_by_conversion_date: f64

The value of all conversions. When this column is selected with date, the values in date column means the conversion date. Details for the by_conversion_date columns are available at https://support.google.com/google-ads/answer/9549009.

all_conversions: Option<f64>

The total number of conversions. This includes all conversions regardless of the value of include_in_conversions_metric.

all_conversions_by_conversion_date: f64

The total number of conversions. This includes all conversions regardless of the value of include_in_conversions_metric. When this column is selected with date, the values in date column means the conversion date. Details for the by_conversion_date columns are available at https://support.google.com/google-ads/answer/9549009.

all_conversions_value_per_cost: Option<f64>

The value of all conversions divided by the total cost of ad interactions (such as clicks for text ads or views for video ads).

all_conversions_from_click_to_call: Option<f64>

The number of times people clicked the “Call” button to call a store during or after clicking an ad. This number doesn’t include whether or not calls were connected, or the duration of any calls. This metric applies to feed items only.

all_conversions_from_directions: Option<f64>

The number of times people clicked a “Get directions” button to navigate to a store after clicking an ad. This metric applies to feed items only.

all_conversions_from_interactions_value_per_interaction: Option<f64>

The value of all conversions from interactions divided by the total number of interactions.

all_conversions_from_menu: Option<f64>

The number of times people clicked a link to view a store’s menu after clicking an ad. This metric applies to feed items only.

all_conversions_from_order: Option<f64>

The number of times people placed an order at a store after clicking an ad. This metric applies to feed items only.

all_conversions_from_other_engagement: Option<f64>

The number of other conversions (for example, posting a review or saving a location for a store) that occurred after people clicked an ad. This metric applies to feed items only.

all_conversions_from_store_visit: Option<f64>

Estimated number of times people visited a store after clicking an ad. This metric applies to feed items only.

all_conversions_from_store_website: Option<f64>

The number of times that people were taken to a store’s URL after clicking an ad. This metric applies to feed items only.

average_cost: Option<f64>

The average amount you pay per interaction. This amount is the total cost of your ads divided by the total number of interactions.

average_cpc: Option<f64>

The total cost of all clicks divided by the total number of clicks received.

average_cpe: Option<f64>

The average amount that you’ve been charged for an ad engagement. This amount is the total cost of all ad engagements divided by the total number of ad engagements.

average_cpm: Option<f64>

Average cost-per-thousand impressions (CPM).

average_cpv: Option<f64>

The average amount you pay each time someone views your ad. The average CPV is defined by the total cost of all ad views divided by the number of views.

average_page_views: Option<f64>

Average number of pages viewed per session.

average_time_on_site: Option<f64>

Total duration of all sessions (in seconds) / number of sessions. Imported from Google Analytics.

benchmark_average_max_cpc: Option<f64>

An indication of how other advertisers are bidding on similar products.

benchmark_ctr: Option<f64>

An indication on how other advertisers’ Shopping ads for similar products are performing based on how often people who see their ad click on it.

bounce_rate: Option<f64>

Percentage of clicks where the user only visited a single page on your site. Imported from Google Analytics.

clicks: Option<i64>

The number of clicks.

combined_clicks: Option<i64>

The number of times your ad or your site’s listing in the unpaid results was clicked. See the help page at https://support.google.com/google-ads/answer/3097241 for details.

combined_clicks_per_query: Option<f64>

The number of times your ad or your site’s listing in the unpaid results was clicked (combined_clicks) divided by combined_queries. See the help page at https://support.google.com/google-ads/answer/3097241 for details.

combined_queries: Option<i64>

The number of searches that returned pages from your site in the unpaid results or showed one of your text ads. See the help page at https://support.google.com/google-ads/answer/3097241 for details.

content_budget_lost_impression_share: Option<f64>

The estimated percent of times that your ad was eligible to show on the Display Network but didn’t because your budget was too low. Note: Content budget lost impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

content_impression_share: Option<f64>

The impressions you’ve received on the Display Network divided by the estimated number of impressions you were eligible to receive. Note: Content impression share is reported in the range of 0.1 to 1. Any value below 0.1 is reported as 0.0999.

conversion_last_received_request_date_time: Option<String>

The last date/time a conversion tag for this conversion action successfully fired and was seen by Google Ads. This firing event may not have been the result of an attributable conversion (e.g. because the tag was fired from a browser that did not previously click an ad from an appropriate advertiser). The date/time is in the customer’s time zone.

conversion_last_conversion_date: Option<String>

The date of the most recent conversion for this conversion action. The date is in the customer’s time zone.

content_rank_lost_impression_share: Option<f64>

The estimated percentage of impressions on the Display Network that your ads didn’t receive due to poor Ad Rank. Note: Content rank lost impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

conversions_from_interactions_rate: Option<f64>

Conversions from interactions divided by the number of ad interactions (such as clicks for text ads or views for video ads). This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

conversions_value: Option<f64>

The value of conversions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

conversions_value_by_conversion_date: f64

The value of conversions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions. When this column is selected with date, the values in date column means the conversion date. Details for the by_conversion_date columns are available at https://support.google.com/google-ads/answer/9549009.

conversions_value_per_cost: Option<f64>

The value of conversions divided by the cost of ad interactions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

conversions_from_interactions_value_per_interaction: Option<f64>

The value of conversions from interactions divided by the number of ad interactions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

conversions: Option<f64>

The number of conversions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

conversions_by_conversion_date: f64

The number of conversions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions. When this column is selected with date, the values in date column means the conversion date. Details for the by_conversion_date columns are available at https://support.google.com/google-ads/answer/9549009.

cost_micros: Option<i64>

The sum of your cost-per-click (CPC) and cost-per-thousand impressions (CPM) costs during this period.

cost_per_all_conversions: Option<f64>

The cost of ad interactions divided by all conversions.

cost_per_conversion: Option<f64>

The cost of ad interactions divided by conversions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

cost_per_current_model_attributed_conversion: Option<f64>

The cost of ad interactions divided by current model attributed conversions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

cross_device_conversions: Option<f64>

Conversions from when a customer clicks on a Google Ads ad on one device, then converts on a different device or browser. Cross-device conversions are already included in all_conversions.

ctr: Option<f64>

The number of clicks your ad receives (Clicks) divided by the number of times your ad is shown (Impressions).

current_model_attributed_conversions: Option<f64>

Shows how your historic conversions data would look under the attribution model you’ve currently selected. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

current_model_attributed_conversions_from_interactions_rate: Option<f64>

Current model attributed conversions from interactions divided by the number of ad interactions (such as clicks for text ads or views for video ads). This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

current_model_attributed_conversions_from_interactions_value_per_interaction: Option<f64>

The value of current model attributed conversions from interactions divided by the number of ad interactions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

current_model_attributed_conversions_value: Option<f64>

The value of current model attributed conversions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

current_model_attributed_conversions_value_per_cost: Option<f64>

The value of current model attributed conversions divided by the cost of ad interactions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

engagement_rate: Option<f64>

How often people engage with your ad after it’s shown to them. This is the number of ad expansions divided by the number of times your ad is shown.

engagements: Option<i64>

The number of engagements. An engagement occurs when a viewer expands your Lightbox ad. Also, in the future, other ad types may support engagement metrics.

hotel_average_lead_value_micros: Option<f64>

Average lead value based on clicks.

hotel_price_difference_percentage: Option<f64>

The average price difference between the price offered by reporting hotel advertiser and the cheapest price offered by the competing advertiser.

hotel_eligible_impressions: Option<i64>

The number of impressions that hotel partners could have had given their feed performance.

historical_creative_quality_score: i32

The creative historical quality score.

historical_landing_page_quality_score: i32

The quality of historical landing page experience.

historical_quality_score: Option<i64>

The historical quality score.

historical_search_predicted_ctr: i32

The historical search predicted click through rate (CTR).

gmail_forwards: Option<i64>

The number of times the ad was forwarded to someone else as a message.

gmail_saves: Option<i64>

The number of times someone has saved your Gmail ad to their inbox as a message.

gmail_secondary_clicks: Option<i64>

The number of clicks to the landing page on the expanded state of Gmail ads.

impressions_from_store_reach: Option<i64>

The number of times a store’s location-based ad was shown. This metric applies to feed items only.

impressions: Option<i64>

Count of how often your ad has appeared on a search results page or website on the Google Network.

interaction_rate: Option<f64>

How often people interact with your ad after it is shown to them. This is the number of interactions divided by the number of times your ad is shown.

interactions: Option<i64>

The number of interactions. An interaction is the main user action associated with an ad format-clicks for text and shopping ads, views for video ads, and so on.

interaction_event_types: Vec<i32>

The types of payable and free interactions.

invalid_click_rate: Option<f64>

The percentage of clicks filtered out of your total number of clicks (filtered + non-filtered clicks) during the reporting period.

invalid_clicks: Option<i64>

Number of clicks Google considers illegitimate and doesn’t charge you for.

message_chats: Option<i64>

Number of message chats initiated for Click To Message impressions that were message tracking eligible.

message_impressions: Option<i64>

Number of Click To Message impressions that were message tracking eligible.

message_chat_rate: Option<f64>

Number of message chats initiated (message_chats) divided by the number of message impressions (message_impressions). Rate at which a user initiates a message chat from an ad impression with a messaging option and message tracking enabled. Note that this rate can be more than 1.0 for a given message impression.

mobile_friendly_clicks_percentage: Option<f64>

The percentage of mobile clicks that go to a mobile-friendly page.

optimization_score_uplift: Option<f64>

Total optimization score uplift of all recommendations.

optimization_score_url: Option<String>

URL for the optimization score page in the Google Ads web interface. This metric can be selected from customer or campaign, and can be segmented by segments.recommendation_type. For example, SELECT metrics.optimization_score_url, segments.recommendation_type FROM customer will return a URL for each unique (customer, recommendation_type) combination.

organic_clicks: Option<i64>

The number of times someone clicked your site’s listing in the unpaid results for a particular query. See the help page at https://support.google.com/google-ads/answer/3097241 for details.

organic_clicks_per_query: Option<f64>

The number of times someone clicked your site’s listing in the unpaid results (organic_clicks) divided by the total number of searches that returned pages from your site (organic_queries). See the help page at https://support.google.com/google-ads/answer/3097241 for details.

organic_impressions: Option<i64>

The number of listings for your site in the unpaid search results. See the help page at https://support.google.com/google-ads/answer/3097241 for details.

organic_impressions_per_query: Option<f64>

The number of times a page from your site was listed in the unpaid search results (organic_impressions) divided by the number of searches returning your site’s listing in the unpaid results (organic_queries). See the help page at https://support.google.com/google-ads/answer/3097241 for details.

organic_queries: Option<i64>

The total number of searches that returned your site’s listing in the unpaid results. See the help page at https://support.google.com/google-ads/answer/3097241 for details.

percent_new_visitors: Option<f64>

Percentage of first-time sessions (from people who had never visited your site before). Imported from Google Analytics.

phone_calls: Option<i64>

Number of offline phone calls.

phone_impressions: Option<i64>

Number of offline phone impressions.

phone_through_rate: Option<f64>

Number of phone calls received (phone_calls) divided by the number of times your phone number is shown (phone_impressions).

relative_ctr: Option<f64>

Your clickthrough rate (Ctr) divided by the average clickthrough rate of all advertisers on the websites that show your ads. Measures how your ads perform on Display Network sites compared to other ads on the same sites.

search_absolute_top_impression_share: Option<f64>

The percentage of the customer’s Shopping or Search ad impressions that are shown in the most prominent Shopping position. See https://support.google.com/google-ads/answer/7501826 for details. Any value below 0.1 is reported as 0.0999.

search_budget_lost_absolute_top_impression_share: Option<f64>

The number estimating how often your ad wasn’t the very first ad above the organic search results due to a low budget. Note: Search budget lost absolute top impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

search_budget_lost_impression_share: Option<f64>

The estimated percent of times that your ad was eligible to show on the Search Network but didn’t because your budget was too low. Note: Search budget lost impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

search_budget_lost_top_impression_share: Option<f64>

The number estimating how often your ad didn’t show anywhere above the organic search results due to a low budget. Note: Search budget lost top impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

search_click_share: Option<f64>

The number of clicks you’ve received on the Search Network divided by the estimated number of clicks you were eligible to receive. Note: Search click share is reported in the range of 0.1 to 1. Any value below 0.1 is reported as 0.0999.

search_exact_match_impression_share: Option<f64>

The impressions you’ve received divided by the estimated number of impressions you were eligible to receive on the Search Network for search terms that matched your keywords exactly (or were close variants of your keyword), regardless of your keyword match types. Note: Search exact match impression share is reported in the range of 0.1 to 1. Any value below 0.1 is reported as 0.0999.

search_impression_share: Option<f64>

The impressions you’ve received on the Search Network divided by the estimated number of impressions you were eligible to receive. Note: Search impression share is reported in the range of 0.1 to 1. Any value below 0.1 is reported as 0.0999.

search_rank_lost_absolute_top_impression_share: Option<f64>

The number estimating how often your ad wasn’t the very first ad above the organic search results due to poor Ad Rank. Note: Search rank lost absolute top impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

search_rank_lost_impression_share: Option<f64>

The estimated percentage of impressions on the Search Network that your ads didn’t receive due to poor Ad Rank. Note: Search rank lost impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

search_rank_lost_top_impression_share: Option<f64>

The number estimating how often your ad didn’t show anywhere above the organic search results due to poor Ad Rank. Note: Search rank lost top impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

search_top_impression_share: Option<f64>

The impressions you’ve received in the top location (anywhere above the organic search results) compared to the estimated number of impressions you were eligible to receive in the top location. Note: Search top impression share is reported in the range of 0.1 to 1. Any value below 0.1 is reported as 0.0999.

speed_score: Option<i64>

A measure of how quickly your page loads after clicks on your mobile ads. The score is a range from 1 to 10, 10 being the fastest.

top_impression_percentage: Option<f64>

The percent of your ad impressions that are shown anywhere above the organic search results.

valid_accelerated_mobile_pages_clicks_percentage: Option<f64>

The percentage of ad clicks to Accelerated Mobile Pages (AMP) landing pages that reach a valid AMP page.

value_per_all_conversions: Option<f64>

The value of all conversions divided by the number of all conversions.

value_per_all_conversions_by_conversion_date: Option<f64>

The value of all conversions divided by the number of all conversions. When this column is selected with date, the values in date column means the conversion date. Details for the by_conversion_date columns are available at https://support.google.com/google-ads/answer/9549009.

value_per_conversion: Option<f64>

The value of conversions divided by the number of conversions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

value_per_conversions_by_conversion_date: Option<f64>

The value of conversions divided by the number of conversions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions. When this column is selected with date, the values in date column means the conversion date. Details for the by_conversion_date columns are available at https://support.google.com/google-ads/answer/9549009.

value_per_current_model_attributed_conversion: Option<f64>

The value of current model attributed conversions divided by the number of the conversions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

video_quartile_p100_rate: Option<f64>

Percentage of impressions where the viewer watched all of your video.

video_quartile_p25_rate: Option<f64>

Percentage of impressions where the viewer watched 25% of your video.

video_quartile_p50_rate: Option<f64>

Percentage of impressions where the viewer watched 50% of your video.

video_quartile_p75_rate: Option<f64>

Percentage of impressions where the viewer watched 75% of your video.

video_view_rate: Option<f64>

The number of views your TrueView video ad receives divided by its number of impressions, including thumbnail impressions for TrueView in-display ads.

video_views: Option<i64>

The number of times your video ads were viewed.

view_through_conversions: Option<i64>

The total number of view-through conversions. These happen when a customer sees an image or rich media ad, then later completes a conversion on your site without interacting with (e.g., clicking on) another ad.

sk_ad_network_conversions: i64

The number of iOS Store Kit Ad Network conversions.

Implementations

impl Metrics

pub fn historical_creative_quality_score(&self) -> QualityScoreBucket

Returns the enum value of historical_creative_quality_score, or the default if the field is set to an invalid enum value.

pub fn set_historical_creative_quality_score(
&mut self,
value: QualityScoreBucket
)

Sets historical_creative_quality_score to the provided enum value.

pub fn historical_landing_page_quality_score(&self) -> QualityScoreBucket

Returns the enum value of historical_landing_page_quality_score, or the default if the field is set to an invalid enum value.

pub fn set_historical_landing_page_quality_score(
&mut self,
value: QualityScoreBucket
)

Sets historical_landing_page_quality_score to the provided enum value.

pub fn historical_search_predicted_ctr(&self) -> QualityScoreBucket

Returns the enum value of historical_search_predicted_ctr, or the default if the field is set to an invalid enum value.

pub fn set_historical_search_predicted_ctr(&mut self, value: QualityScoreBucket)

Sets historical_search_predicted_ctr to the provided enum value.

pub fn interaction_event_types(
&self
) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<InteractionEventType>>

Returns an iterator which yields the valid enum values contained in interaction_event_types.

pub fn push_interaction_event_types(&mut self, value: InteractionEventType)

Appends the provided enum value to interaction_event_types.

pub fn clicks(&self) -> i64

Returns the value of clicks, or the default value if clicks is unset.

pub fn video_quartile_p100_rate(&self) -> f64

Returns the value of video_quartile_p100_rate, or the default value if video_quartile_p100_rate is unset.

pub fn video_quartile_p25_rate(&self) -> f64

Returns the value of video_quartile_p25_rate, or the default value if video_quartile_p25_rate is unset.

pub fn video_quartile_p50_rate(&self) -> f64

Returns the value of video_quartile_p50_rate, or the default value if video_quartile_p50_rate is unset.

pub fn video_quartile_p75_rate(&self) -> f64

Returns the value of video_quartile_p75_rate, or the default value if video_quartile_p75_rate is unset.

pub fn search_absolute_top_impression_share(&self) -> f64

Returns the value of search_absolute_top_impression_share, or the default value if search_absolute_top_impression_share is unset.

pub fn search_budget_lost_absolute_top_impression_share(&self) -> f64

Returns the value of search_budget_lost_absolute_top_impression_share, or the default value if search_budget_lost_absolute_top_impression_share is unset.

pub fn search_budget_lost_impression_share(&self) -> f64

Returns the value of search_budget_lost_impression_share, or the default value if search_budget_lost_impression_share is unset.

pub fn search_budget_lost_top_impression_share(&self) -> f64

Returns the value of search_budget_lost_top_impression_share, or the default value if search_budget_lost_top_impression_share is unset.

pub fn search_click_share(&self) -> f64

Returns the value of search_click_share, or the default value if search_click_share is unset.

pub fn search_exact_match_impression_share(&self) -> f64

Returns the value of search_exact_match_impression_share, or the default value if search_exact_match_impression_share is unset.

pub fn search_impression_share(&self) -> f64

Returns the value of search_impression_share, or the default value if search_impression_share is unset.

pub fn search_rank_lost_absolute_top_impression_share(&self) -> f64

Returns the value of search_rank_lost_absolute_top_impression_share, or the default value if search_rank_lost_absolute_top_impression_share is unset.

pub fn search_rank_lost_impression_share(&self) -> f64

Returns the value of search_rank_lost_impression_share, or the default value if search_rank_lost_impression_share is unset.

pub fn search_rank_lost_top_impression_share(&self) -> f64

Returns the value of search_rank_lost_top_impression_share, or the default value if search_rank_lost_top_impression_share is unset.

pub fn search_top_impression_share(&self) -> f64

Returns the value of search_top_impression_share, or the default value if search_top_impression_share is unset.

pub fn speed_score(&self) -> i64

Returns the value of speed_score, or the default value if speed_score is unset.

pub fn top_impression_percentage(&self) -> f64

Returns the value of top_impression_percentage, or the default value if top_impression_percentage is unset.

pub fn valid_accelerated_mobile_pages_clicks_percentage(&self) -> f64

Returns the value of valid_accelerated_mobile_pages_clicks_percentage, or the default value if valid_accelerated_mobile_pages_clicks_percentage is unset.

pub fn value_per_all_conversions(&self) -> f64

Returns the value of value_per_all_conversions, or the default value if value_per_all_conversions is unset.

pub fn value_per_conversion(&self) -> f64

Returns the value of value_per_conversion, or the default value if value_per_conversion is unset.

pub fn value_per_current_model_attributed_conversion(&self) -> f64

Returns the value of value_per_current_model_attributed_conversion, or the default value if value_per_current_model_attributed_conversion is unset.

pub fn video_view_rate(&self) -> f64

Returns the value of video_view_rate, or the default value if video_view_rate is unset.

pub fn video_views(&self) -> i64

Returns the value of video_views, or the default value if video_views is unset.

pub fn view_through_conversions(&self) -> i64

Returns the value of view_through_conversions, or the default value if view_through_conversions is unset.

pub fn combined_clicks(&self) -> i64

Returns the value of combined_clicks, or the default value if combined_clicks is unset.

pub fn combined_clicks_per_query(&self) -> f64

Returns the value of combined_clicks_per_query, or the default value if combined_clicks_per_query is unset.

pub fn combined_queries(&self) -> i64

Returns the value of combined_queries, or the default value if combined_queries is unset.

pub fn content_budget_lost_impression_share(&self) -> f64

Returns the value of content_budget_lost_impression_share, or the default value if content_budget_lost_impression_share is unset.

pub fn content_impression_share(&self) -> f64

Returns the value of content_impression_share, or the default value if content_impression_share is unset.

pub fn conversion_last_received_request_date_time(&self) -> &str

Returns the value of conversion_last_received_request_date_time, or the default value if conversion_last_received_request_date_time is unset.

pub fn conversion_last_conversion_date(&self) -> &str

Returns the value of conversion_last_conversion_date, or the default value if conversion_last_conversion_date is unset.

pub fn content_rank_lost_impression_share(&self) -> f64

Returns the value of content_rank_lost_impression_share, or the default value if content_rank_lost_impression_share is unset.

pub fn conversions_from_interactions_rate(&self) -> f64

Returns the value of conversions_from_interactions_rate, or the default value if conversions_from_interactions_rate is unset.

pub fn conversions_value(&self) -> f64

Returns the value of conversions_value, or the default value if conversions_value is unset.

pub fn conversions_value_per_cost(&self) -> f64

Returns the value of conversions_value_per_cost, or the default value if conversions_value_per_cost is unset.

pub fn conversions_from_interactions_value_per_interaction(&self) -> f64

Returns the value of conversions_from_interactions_value_per_interaction, or the default value if conversions_from_interactions_value_per_interaction is unset.

pub fn conversions(&self) -> f64

Returns the value of conversions, or the default value if conversions is unset.

pub fn cost_micros(&self) -> i64

Returns the value of cost_micros, or the default value if cost_micros is unset.

pub fn cost_per_all_conversions(&self) -> f64

Returns the value of cost_per_all_conversions, or the default value if cost_per_all_conversions is unset.

pub fn cost_per_conversion(&self) -> f64

Returns the value of cost_per_conversion, or the default value if cost_per_conversion is unset.

pub fn cost_per_current_model_attributed_conversion(&self) -> f64

Returns the value of cost_per_current_model_attributed_conversion, or the default value if cost_per_current_model_attributed_conversion is unset.

pub fn cross_device_conversions(&self) -> f64

Returns the value of cross_device_conversions, or the default value if cross_device_conversions is unset.

pub fn ctr(&self) -> f64

Returns the value of ctr, or the default value if ctr is unset.

pub fn current_model_attributed_conversions(&self) -> f64

Returns the value of current_model_attributed_conversions, or the default value if current_model_attributed_conversions is unset.

pub fn current_model_attributed_conversions_from_interactions_rate(&self) -> f64

Returns the value of current_model_attributed_conversions_from_interactions_rate, or the default value if current_model_attributed_conversions_from_interactions_rate is unset.

pub fn current_model_attributed_conversions_from_interactions_value_per_interaction(
&self
) -> f64

Returns the value of current_model_attributed_conversions_from_interactions_value_per_interaction, or the default value if current_model_attributed_conversions_from_interactions_value_per_interaction is unset.

pub fn current_model_attributed_conversions_value(&self) -> f64

Returns the value of current_model_attributed_conversions_value, or the default value if current_model_attributed_conversions_value is unset.

pub fn current_model_attributed_conversions_value_per_cost(&self) -> f64

Returns the value of current_model_attributed_conversions_value_per_cost, or the default value if current_model_attributed_conversions_value_per_cost is unset.

pub fn engagement_rate(&self) -> f64

Returns the value of engagement_rate, or the default value if engagement_rate is unset.

pub fn engagements(&self) -> i64

Returns the value of engagements, or the default value if engagements is unset.

pub fn absolute_top_impression_percentage(&self) -> f64

Returns the value of absolute_top_impression_percentage, or the default value if absolute_top_impression_percentage is unset.

pub fn active_view_cpm(&self) -> f64

Returns the value of active_view_cpm, or the default value if active_view_cpm is unset.

pub fn active_view_ctr(&self) -> f64

Returns the value of active_view_ctr, or the default value if active_view_ctr is unset.

pub fn active_view_impressions(&self) -> i64

Returns the value of active_view_impressions, or the default value if active_view_impressions is unset.

pub fn active_view_measurability(&self) -> f64

Returns the value of active_view_measurability, or the default value if active_view_measurability is unset.

pub fn active_view_measurable_cost_micros(&self) -> i64

Returns the value of active_view_measurable_cost_micros, or the default value if active_view_measurable_cost_micros is unset.

pub fn active_view_measurable_impressions(&self) -> i64

Returns the value of active_view_measurable_impressions, or the default value if active_view_measurable_impressions is unset.

pub fn active_view_viewability(&self) -> f64

Returns the value of active_view_viewability, or the default value if active_view_viewability is unset.

pub fn all_conversions_from_interactions_rate(&self) -> f64

Returns the value of all_conversions_from_interactions_rate, or the default value if all_conversions_from_interactions_rate is unset.

pub fn all_conversions_value(&self) -> f64

Returns the value of all_conversions_value, or the default value if all_conversions_value is unset.

pub fn all_conversions(&self) -> f64

Returns the value of all_conversions, or the default value if all_conversions is unset.

pub fn all_conversions_value_per_cost(&self) -> f64

Returns the value of all_conversions_value_per_cost, or the default value if all_conversions_value_per_cost is unset.

pub fn all_conversions_from_click_to_call(&self) -> f64

Returns the value of all_conversions_from_click_to_call, or the default value if all_conversions_from_click_to_call is unset.

pub fn all_conversions_from_directions(&self) -> f64

Returns the value of all_conversions_from_directions, or the default value if all_conversions_from_directions is unset.

pub fn all_conversions_from_interactions_value_per_interaction(&self) -> f64

Returns the value of all_conversions_from_interactions_value_per_interaction, or the default value if all_conversions_from_interactions_value_per_interaction is unset.

pub fn all_conversions_from_menu(&self) -> f64

Returns the value of all_conversions_from_menu, or the default value if all_conversions_from_menu is unset.

pub fn all_conversions_from_order(&self) -> f64

Returns the value of all_conversions_from_order, or the default value if all_conversions_from_order is unset.

pub fn all_conversions_from_other_engagement(&self) -> f64

Returns the value of all_conversions_from_other_engagement, or the default value if all_conversions_from_other_engagement is unset.

pub fn all_conversions_from_store_visit(&self) -> f64

Returns the value of all_conversions_from_store_visit, or the default value if all_conversions_from_store_visit is unset.

pub fn all_conversions_from_store_website(&self) -> f64

Returns the value of all_conversions_from_store_website, or the default value if all_conversions_from_store_website is unset.

pub fn average_cost(&self) -> f64

Returns the value of average_cost, or the default value if average_cost is unset.

pub fn average_cpc(&self) -> f64

Returns the value of average_cpc, or the default value if average_cpc is unset.

pub fn average_cpe(&self) -> f64

Returns the value of average_cpe, or the default value if average_cpe is unset.

pub fn average_cpm(&self) -> f64

Returns the value of average_cpm, or the default value if average_cpm is unset.

pub fn average_cpv(&self) -> f64

Returns the value of average_cpv, or the default value if average_cpv is unset.

pub fn average_page_views(&self) -> f64

Returns the value of average_page_views, or the default value if average_page_views is unset.

pub fn average_time_on_site(&self) -> f64

Returns the value of average_time_on_site, or the default value if average_time_on_site is unset.

pub fn benchmark_average_max_cpc(&self) -> f64

Returns the value of benchmark_average_max_cpc, or the default value if benchmark_average_max_cpc is unset.

pub fn benchmark_ctr(&self) -> f64

Returns the value of benchmark_ctr, or the default value if benchmark_ctr is unset.

pub fn bounce_rate(&self) -> f64

Returns the value of bounce_rate, or the default value if bounce_rate is unset.

pub fn hotel_average_lead_value_micros(&self) -> f64

Returns the value of hotel_average_lead_value_micros, or the default value if hotel_average_lead_value_micros is unset.

pub fn hotel_price_difference_percentage(&self) -> f64

Returns the value of hotel_price_difference_percentage, or the default value if hotel_price_difference_percentage is unset.

pub fn hotel_eligible_impressions(&self) -> i64

Returns the value of hotel_eligible_impressions, or the default value if hotel_eligible_impressions is unset.

pub fn historical_quality_score(&self) -> i64

Returns the value of historical_quality_score, or the default value if historical_quality_score is unset.

pub fn gmail_forwards(&self) -> i64

Returns the value of gmail_forwards, or the default value if gmail_forwards is unset.

pub fn gmail_saves(&self) -> i64

Returns the value of gmail_saves, or the default value if gmail_saves is unset.

pub fn gmail_secondary_clicks(&self) -> i64

Returns the value of gmail_secondary_clicks, or the default value if gmail_secondary_clicks is unset.

pub fn impressions_from_store_reach(&self) -> i64

Returns the value of impressions_from_store_reach, or the default value if impressions_from_store_reach is unset.

pub fn impressions(&self) -> i64

Returns the value of impressions, or the default value if impressions is unset.

pub fn interaction_rate(&self) -> f64

Returns the value of interaction_rate, or the default value if interaction_rate is unset.

pub fn interactions(&self) -> i64

Returns the value of interactions, or the default value if interactions is unset.

pub fn invalid_click_rate(&self) -> f64

Returns the value of invalid_click_rate, or the default value if invalid_click_rate is unset.

pub fn invalid_clicks(&self) -> i64

Returns the value of invalid_clicks, or the default value if invalid_clicks is unset.

pub fn message_chats(&self) -> i64

Returns the value of message_chats, or the default value if message_chats is unset.

pub fn message_impressions(&self) -> i64

Returns the value of message_impressions, or the default value if message_impressions is unset.

pub fn message_chat_rate(&self) -> f64

Returns the value of message_chat_rate, or the default value if message_chat_rate is unset.

pub fn mobile_friendly_clicks_percentage(&self) -> f64

Returns the value of mobile_friendly_clicks_percentage, or the default value if mobile_friendly_clicks_percentage is unset.

pub fn organic_clicks(&self) -> i64

Returns the value of organic_clicks, or the default value if organic_clicks is unset.

pub fn organic_clicks_per_query(&self) -> f64

Returns the value of organic_clicks_per_query, or the default value if organic_clicks_per_query is unset.

pub fn organic_impressions(&self) -> i64

Returns the value of organic_impressions, or the default value if organic_impressions is unset.

pub fn organic_impressions_per_query(&self) -> f64

Returns the value of organic_impressions_per_query, or the default value if organic_impressions_per_query is unset.

pub fn organic_queries(&self) -> i64

Returns the value of organic_queries, or the default value if organic_queries is unset.

pub fn percent_new_visitors(&self) -> f64

Returns the value of percent_new_visitors, or the default value if percent_new_visitors is unset.

pub fn phone_calls(&self) -> i64

Returns the value of phone_calls, or the default value if phone_calls is unset.

pub fn phone_impressions(&self) -> i64

Returns the value of phone_impressions, or the default value if phone_impressions is unset.

pub fn phone_through_rate(&self) -> f64

Returns the value of phone_through_rate, or the default value if phone_through_rate is unset.

pub fn relative_ctr(&self) -> f64

Returns the value of relative_ctr, or the default value if relative_ctr is unset.

pub fn value_per_all_conversions_by_conversion_date(&self) -> f64

Returns the value of value_per_all_conversions_by_conversion_date, or the default value if value_per_all_conversions_by_conversion_date is unset.

pub fn value_per_conversions_by_conversion_date(&self) -> f64

Returns the value of value_per_conversions_by_conversion_date, or the default value if value_per_conversions_by_conversion_date is unset.

pub fn optimization_score_uplift(&self) -> f64

Returns the value of optimization_score_uplift, or the default value if optimization_score_uplift is unset.

pub fn optimization_score_url(&self) -> &str

Returns the value of optimization_score_url, or the default value if optimization_score_url is unset.