Module googapis::google::ads::googleads::v8::enums

Modules

access_invitation_status_enum
Nested message and enum types in AccessInvitationStatusEnum.

access_reason_enum
Nested message and enum types in AccessReasonEnum.

access_role_enum
Nested message and enum types in AccessRoleEnum.

account_budget_proposal_status_enum
Nested message and enum types in AccountBudgetProposalStatusEnum.

account_budget_proposal_type_enum
Nested message and enum types in AccountBudgetProposalTypeEnum.

account_budget_status_enum
Nested message and enum types in AccountBudgetStatusEnum.

account_link_status_enum
Nested message and enum types in AccountLinkStatusEnum.

ad_customizer_placeholder_field_enum
Nested message and enum types in AdCustomizerPlaceholderFieldEnum.

ad_destination_type_enum
Nested message and enum types in AdDestinationTypeEnum.

ad_group_ad_rotation_mode_enum
Nested message and enum types in AdGroupAdRotationModeEnum.

ad_group_ad_status_enum
Nested message and enum types in AdGroupAdStatusEnum.

ad_group_criterion_approval_status_enum
Nested message and enum types in AdGroupCriterionApprovalStatusEnum.

ad_group_criterion_status_enum
Nested message and enum types in AdGroupCriterionStatusEnum.

ad_group_status_enum
Nested message and enum types in AdGroupStatusEnum.

ad_group_type_enum
Nested message and enum types in AdGroupTypeEnum.

ad_network_type_enum
Nested message and enum types in AdNetworkTypeEnum.

ad_serving_optimization_status_enum
Nested message and enum types in AdServingOptimizationStatusEnum.

ad_strength_enum
Nested message and enum types in AdStrengthEnum.

ad_type_enum
Nested message and enum types in AdTypeEnum.

advertising_channel_sub_type_enum
Nested message and enum types in AdvertisingChannelSubTypeEnum.

advertising_channel_type_enum
Nested message and enum types in AdvertisingChannelTypeEnum.

affiliate_location_feed_relationship_type_enum
Nested message and enum types in AffiliateLocationFeedRelationshipTypeEnum.

affiliate_location_placeholder_field_enum
Nested message and enum types in AffiliateLocationPlaceholderFieldEnum.

age_range_type_enum
Nested message and enum types in AgeRangeTypeEnum.

app_campaign_app_store_enum
Nested message and enum types in AppCampaignAppStoreEnum.

app_campaign_bidding_strategy_goal_type_enum
Nested message and enum types in AppCampaignBiddingStrategyGoalTypeEnum.

app_payment_model_type_enum
Nested message and enum types in AppPaymentModelTypeEnum.

app_placeholder_field_enum
Nested message and enum types in AppPlaceholderFieldEnum.

app_store_enum
Nested message and enum types in AppStoreEnum.

app_url_operating_system_type_enum
Nested message and enum types in AppUrlOperatingSystemTypeEnum.

asset_field_type_enum
Nested message and enum types in AssetFieldTypeEnum.

asset_link_status_enum
Nested message and enum types in AssetLinkStatusEnum.

asset_performance_label_enum
Nested message and enum types in AssetPerformanceLabelEnum.

asset_type_enum
Nested message and enum types in AssetTypeEnum.

attribution_model_enum
Nested message and enum types in AttributionModelEnum.

batch_job_status_enum
Nested message and enum types in BatchJobStatusEnum.

bid_modifier_source_enum
Nested message and enum types in BidModifierSourceEnum.

bidding_source_enum
Nested message and enum types in BiddingSourceEnum.

bidding_strategy_status_enum
Nested message and enum types in BiddingStrategyStatusEnum.

bidding_strategy_type_enum
Nested message and enum types in BiddingStrategyTypeEnum.

billing_setup_status_enum
Nested message and enum types in BillingSetupStatusEnum.

brand_safety_suitability_enum
Nested message and enum types in BrandSafetySuitabilityEnum.

budget_campaign_association_status_enum
Nested message and enum types in BudgetCampaignAssociationStatusEnum.

budget_delivery_method_enum
Nested message and enum types in BudgetDeliveryMethodEnum.

budget_period_enum
Nested message and enum types in BudgetPeriodEnum.

budget_status_enum
Nested message and enum types in BudgetStatusEnum.

budget_type_enum
Nested message and enum types in BudgetTypeEnum.

call_conversion_reporting_state_enum
Nested message and enum types in CallConversionReportingStateEnum.

call_placeholder_field_enum
Nested message and enum types in CallPlaceholderFieldEnum.

call_tracking_display_location_enum
Nested message and enum types in CallTrackingDisplayLocationEnum.

call_type_enum
Nested message and enum types in CallTypeEnum.

callout_placeholder_field_enum
Nested message and enum types in CalloutPlaceholderFieldEnum.

campaign_criterion_status_enum
Nested message and enum types in CampaignCriterionStatusEnum.

campaign_draft_status_enum
Nested message and enum types in CampaignDraftStatusEnum.

campaign_experiment_status_enum
Nested message and enum types in CampaignExperimentStatusEnum.

campaign_experiment_traffic_split_type_enum
Nested message and enum types in CampaignExperimentTrafficSplitTypeEnum.

campaign_experiment_type_enum
Nested message and enum types in CampaignExperimentTypeEnum.

campaign_serving_status_enum
Nested message and enum types in CampaignServingStatusEnum.

campaign_shared_set_status_enum
Nested message and enum types in CampaignSharedSetStatusEnum.

campaign_status_enum
Nested message and enum types in CampaignStatusEnum.

change_client_type_enum
Nested message and enum types in ChangeClientTypeEnum.

change_event_resource_type_enum
Nested message and enum types in ChangeEventResourceTypeEnum.

change_status_operation_enum
Nested message and enum types in ChangeStatusOperationEnum.

change_status_resource_type_enum
Nested message and enum types in ChangeStatusResourceTypeEnum.

click_type_enum
Nested message and enum types in ClickTypeEnum.

combined_audience_status_enum
Nested message and enum types in CombinedAudienceStatusEnum.

content_label_type_enum
Nested message and enum types in ContentLabelTypeEnum.

conversion_action_category_enum
Nested message and enum types in ConversionActionCategoryEnum.

conversion_action_counting_type_enum
Nested message and enum types in ConversionActionCountingTypeEnum.

conversion_action_status_enum
Nested message and enum types in ConversionActionStatusEnum.

conversion_action_type_enum
Nested message and enum types in ConversionActionTypeEnum.

conversion_adjustment_type_enum
Nested message and enum types in ConversionAdjustmentTypeEnum.

conversion_attribution_event_type_enum
Nested message and enum types in ConversionAttributionEventTypeEnum.

conversion_custom_variable_status_enum
Nested message and enum types in ConversionCustomVariableStatusEnum.

conversion_lag_bucket_enum
Nested message and enum types in ConversionLagBucketEnum.

conversion_or_adjustment_lag_bucket_enum
Nested message and enum types in ConversionOrAdjustmentLagBucketEnum.

criterion_category_channel_availability_mode_enum
Nested message and enum types in CriterionCategoryChannelAvailabilityModeEnum.

criterion_category_locale_availability_mode_enum
Nested message and enum types in CriterionCategoryLocaleAvailabilityModeEnum.

criterion_system_serving_status_enum
Nested message and enum types in CriterionSystemServingStatusEnum.

criterion_type_enum
Nested message and enum types in CriterionTypeEnum.

custom_audience_member_type_enum
Nested message and enum types in CustomAudienceMemberTypeEnum.

custom_audience_status_enum
Nested message and enum types in CustomAudienceStatusEnum.

custom_audience_type_enum
Nested message and enum types in CustomAudienceTypeEnum.

custom_interest_member_type_enum
Nested message and enum types in CustomInterestMemberTypeEnum.

custom_interest_status_enum
Nested message and enum types in CustomInterestStatusEnum.

custom_interest_type_enum
Nested message and enum types in CustomInterestTypeEnum.

custom_placeholder_field_enum
Nested message and enum types in CustomPlaceholderFieldEnum.

customer_match_upload_key_type_enum
Nested message and enum types in CustomerMatchUploadKeyTypeEnum.

customer_pay_per_conversion_eligibility_failure_reason_enum
Nested message and enum types in CustomerPayPerConversionEligibilityFailureReasonEnum.

data_driven_model_status_enum
Nested message and enum types in DataDrivenModelStatusEnum.

day_of_week_enum
Nested message and enum types in DayOfWeekEnum.

device_enum
Nested message and enum types in DeviceEnum.

display_ad_format_setting_enum
Nested message and enum types in DisplayAdFormatSettingEnum.

display_upload_product_type_enum
Nested message and enum types in DisplayUploadProductTypeEnum.

distance_bucket_enum
Nested message and enum types in DistanceBucketEnum.

dsa_page_feed_criterion_field_enum
Nested message and enum types in DsaPageFeedCriterionFieldEnum.

education_placeholder_field_enum
Nested message and enum types in EducationPlaceholderFieldEnum.

extension_setting_device_enum
Nested message and enum types in ExtensionSettingDeviceEnum.

extension_type_enum
Nested message and enum types in ExtensionTypeEnum.

external_conversion_source_enum
Nested message and enum types in ExternalConversionSourceEnum.

feed_attribute_type_enum
Nested message and enum types in FeedAttributeTypeEnum.

feed_item_quality_approval_status_enum
Nested message and enum types in FeedItemQualityApprovalStatusEnum.

feed_item_quality_disapproval_reason_enum
Nested message and enum types in FeedItemQualityDisapprovalReasonEnum.

feed_item_set_status_enum
Nested message and enum types in FeedItemSetStatusEnum.

feed_item_set_string_filter_type_enum
Nested message and enum types in FeedItemSetStringFilterTypeEnum.

feed_item_status_enum
Nested message and enum types in FeedItemStatusEnum.

feed_item_target_device_enum
Nested message and enum types in FeedItemTargetDeviceEnum.

feed_item_target_status_enum
Nested message and enum types in FeedItemTargetStatusEnum.

feed_item_target_type_enum
Nested message and enum types in FeedItemTargetTypeEnum.

feed_item_validation_status_enum
Nested message and enum types in FeedItemValidationStatusEnum.

feed_link_status_enum
Nested message and enum types in FeedLinkStatusEnum.

feed_mapping_criterion_type_enum
Nested message and enum types in FeedMappingCriterionTypeEnum.

feed_mapping_status_enum
Nested message and enum types in FeedMappingStatusEnum.

feed_origin_enum
Nested message and enum types in FeedOriginEnum.

feed_status_enum
Nested message and enum types in FeedStatusEnum.

flight_placeholder_field_enum
Nested message and enum types in FlightPlaceholderFieldEnum.

frequency_cap_event_type_enum
Nested message and enum types in FrequencyCapEventTypeEnum.

frequency_cap_level_enum
Nested message and enum types in FrequencyCapLevelEnum.

frequency_cap_time_unit_enum
Nested message and enum types in FrequencyCapTimeUnitEnum.

gender_type_enum
Nested message and enum types in GenderTypeEnum.

geo_target_constant_status_enum
Nested message and enum types in GeoTargetConstantStatusEnum.

geo_targeting_restriction_enum
Nested message and enum types in GeoTargetingRestrictionEnum.

geo_targeting_type_enum
Nested message and enum types in GeoTargetingTypeEnum.

google_ads_field_category_enum
Nested message and enum types in GoogleAdsFieldCategoryEnum.

google_ads_field_data_type_enum
Nested message and enum types in GoogleAdsFieldDataTypeEnum.

google_voice_call_status_enum
Nested message and enum types in GoogleVoiceCallStatusEnum.

hotel_date_selection_type_enum
Nested message and enum types in HotelDateSelectionTypeEnum.

hotel_placeholder_field_enum
Nested message and enum types in HotelPlaceholderFieldEnum.

hotel_price_bucket_enum
Nested message and enum types in HotelPriceBucketEnum.

hotel_rate_type_enum
Nested message and enum types in HotelRateTypeEnum.

image_placeholder_field_enum
Nested message and enum types in ImagePlaceholderFieldEnum.

income_range_type_enum
Nested message and enum types in IncomeRangeTypeEnum.

interaction_event_type_enum
Nested message and enum types in InteractionEventTypeEnum.

interaction_type_enum
Nested message and enum types in InteractionTypeEnum.

invoice_type_enum
Nested message and enum types in InvoiceTypeEnum.

job_placeholder_field_enum
Nested message and enum types in JobPlaceholderFieldEnum.

keyword_match_type_enum
Nested message and enum types in KeywordMatchTypeEnum.

keyword_plan_aggregate_metric_type_enum
Nested message and enum types in KeywordPlanAggregateMetricTypeEnum.

keyword_plan_competition_level_enum
Nested message and enum types in KeywordPlanCompetitionLevelEnum.

keyword_plan_concept_group_type_enum
Nested message and enum types in KeywordPlanConceptGroupTypeEnum.

keyword_plan_forecast_interval_enum
Nested message and enum types in KeywordPlanForecastIntervalEnum.

keyword_plan_keyword_annotation_enum
Nested message and enum types in KeywordPlanKeywordAnnotationEnum.

keyword_plan_network_enum
Nested message and enum types in KeywordPlanNetworkEnum.

label_status_enum
Nested message and enum types in LabelStatusEnum.

lead_form_call_to_action_type_enum
Nested message and enum types in LeadFormCallToActionTypeEnum.

lead_form_desired_intent_enum
Nested message and enum types in LeadFormDesiredIntentEnum.

lead_form_field_user_input_type_enum
Nested message and enum types in LeadFormFieldUserInputTypeEnum.

lead_form_post_submit_call_to_action_type_enum
Nested message and enum types in LeadFormPostSubmitCallToActionTypeEnum.

legacy_app_install_ad_app_store_enum
Nested message and enum types in LegacyAppInstallAdAppStoreEnum.

linked_account_type_enum
Nested message and enum types in LinkedAccountTypeEnum.

listing_group_type_enum
Nested message and enum types in ListingGroupTypeEnum.

local_placeholder_field_enum
Nested message and enum types in LocalPlaceholderFieldEnum.

location_extension_targeting_criterion_field_enum
Nested message and enum types in LocationExtensionTargetingCriterionFieldEnum.

location_group_radius_units_enum
Nested message and enum types in LocationGroupRadiusUnitsEnum.

location_placeholder_field_enum
Nested message and enum types in LocationPlaceholderFieldEnum.

location_source_type_enum
Nested message and enum types in LocationSourceTypeEnum.

manager_link_status_enum
Nested message and enum types in ManagerLinkStatusEnum.

matching_function_context_type_enum
Nested message and enum types in MatchingFunctionContextTypeEnum.

matching_function_operator_enum
Nested message and enum types in MatchingFunctionOperatorEnum.

media_type_enum
Nested message and enum types in MediaTypeEnum.

merchant_center_link_status_enum
Nested message and enum types in MerchantCenterLinkStatusEnum.

message_placeholder_field_enum
Nested message and enum types in MessagePlaceholderFieldEnum.

mime_type_enum
Nested message and enum types in MimeTypeEnum.

minute_of_hour_enum
Nested message and enum types in MinuteOfHourEnum.

mobile_app_vendor_enum
Nested message and enum types in MobileAppVendorEnum.

mobile_device_type_enum
Nested message and enum types in MobileDeviceTypeEnum.

month_of_year_enum
Nested message and enum types in MonthOfYearEnum.

negative_geo_target_type_enum
Nested message and enum types in NegativeGeoTargetTypeEnum.

offline_user_data_job_failure_reason_enum
Nested message and enum types in OfflineUserDataJobFailureReasonEnum.

offline_user_data_job_status_enum
Nested message and enum types in OfflineUserDataJobStatusEnum.

offline_user_data_job_type_enum
Nested message and enum types in OfflineUserDataJobTypeEnum.

operating_system_version_operator_type_enum
Nested message and enum types in OperatingSystemVersionOperatorTypeEnum.

optimization_goal_type_enum
Nested message and enum types in OptimizationGoalTypeEnum.

parental_status_type_enum
Nested message and enum types in ParentalStatusTypeEnum.

payment_mode_enum
Nested message and enum types in PaymentModeEnum.

placeholder_type_enum
Nested message and enum types in PlaceholderTypeEnum.

placement_type_enum
Nested message and enum types in PlacementTypeEnum.

policy_approval_status_enum
Nested message and enum types in PolicyApprovalStatusEnum.

policy_review_status_enum
Nested message and enum types in PolicyReviewStatusEnum.

policy_topic_entry_type_enum
Nested message and enum types in PolicyTopicEntryTypeEnum.

policy_topic_evidence_destination_mismatch_url_type_enum
Nested message and enum types in PolicyTopicEvidenceDestinationMismatchUrlTypeEnum.

policy_topic_evidence_destination_not_working_device_enum
Nested message and enum types in PolicyTopicEvidenceDestinationNotWorkingDeviceEnum.

policy_topic_evidence_destination_not_working_dns_error_type_enum
Nested message and enum types in PolicyTopicEvidenceDestinationNotWorkingDnsErrorTypeEnum.

positive_geo_target_type_enum
Nested message and enum types in PositiveGeoTargetTypeEnum.

preferred_content_type_enum
Nested message and enum types in PreferredContentTypeEnum.

price_extension_price_qualifier_enum
Nested message and enum types in PriceExtensionPriceQualifierEnum.

price_extension_price_unit_enum
Nested message and enum types in PriceExtensionPriceUnitEnum.

price_extension_type_enum
Nested message and enum types in PriceExtensionTypeEnum.

price_placeholder_field_enum
Nested message and enum types in PricePlaceholderFieldEnum.

product_bidding_category_level_enum
Nested message and enum types in ProductBiddingCategoryLevelEnum.

product_bidding_category_status_enum
Nested message and enum types in ProductBiddingCategoryStatusEnum.

product_channel_enum
Nested message and enum types in ProductChannelEnum.

product_channel_exclusivity_enum
Nested message and enum types in ProductChannelExclusivityEnum.

product_condition_enum
Nested message and enum types in ProductConditionEnum.

product_custom_attribute_index_enum
Nested message and enum types in ProductCustomAttributeIndexEnum.

product_type_level_enum
Nested message and enum types in ProductTypeLevelEnum.

promotion_extension_discount_modifier_enum
Nested message and enum types in PromotionExtensionDiscountModifierEnum.

promotion_extension_occasion_enum
Nested message and enum types in PromotionExtensionOccasionEnum.

promotion_placeholder_field_enum
Nested message and enum types in PromotionPlaceholderFieldEnum.

proximity_radius_units_enum
Nested message and enum types in ProximityRadiusUnitsEnum.

quality_score_bucket_enum
Nested message and enum types in QualityScoreBucketEnum.

reach_plan_ad_length_enum
Nested message and enum types in ReachPlanAdLengthEnum.

reach_plan_age_range_enum
Nested message and enum types in ReachPlanAgeRangeEnum.

reach_plan_network_enum
Nested message and enum types in ReachPlanNetworkEnum.

real_estate_placeholder_field_enum
Nested message and enum types in RealEstatePlaceholderFieldEnum.

recommendation_type_enum
Nested message and enum types in RecommendationTypeEnum.

resource_change_operation_enum
Nested message and enum types in ResourceChangeOperationEnum.

resource_limit_type_enum
Nested message and enum types in ResourceLimitTypeEnum.

response_content_type_enum
Nested message and enum types in ResponseContentTypeEnum.

search_engine_results_page_type_enum
Nested message and enum types in SearchEngineResultsPageTypeEnum.

search_term_match_type_enum
Nested message and enum types in SearchTermMatchTypeEnum.

search_term_targeting_status_enum
Nested message and enum types in SearchTermTargetingStatusEnum.

served_asset_field_type_enum
Nested message and enum types in ServedAssetFieldTypeEnum.

shared_set_status_enum
Nested message and enum types in SharedSetStatusEnum.

shared_set_type_enum
Nested message and enum types in SharedSetTypeEnum.

simulation_modification_method_enum
Nested message and enum types in SimulationModificationMethodEnum.

simulation_type_enum
Nested message and enum types in SimulationTypeEnum.

sitelink_placeholder_field_enum
Nested message and enum types in SitelinkPlaceholderFieldEnum.

slot_enum
Nested message and enum types in SlotEnum.

spending_limit_type_enum
Nested message and enum types in SpendingLimitTypeEnum.

structured_snippet_placeholder_field_enum
Nested message and enum types in StructuredSnippetPlaceholderFieldEnum.

summary_row_setting_enum
Nested message and enum types in SummaryRowSettingEnum.

system_managed_resource_source_enum
Nested message and enum types in SystemManagedResourceSourceEnum.

target_cpa_opt_in_recommendation_goal_enum
Nested message and enum types in TargetCpaOptInRecommendationGoalEnum.

target_impression_share_location_enum
Nested message and enum types in TargetImpressionShareLocationEnum.

targeting_dimension_enum
Nested message and enum types in TargetingDimensionEnum.

time_type_enum
Nested message and enum types in TimeTypeEnum.

tracking_code_page_format_enum
Nested message and enum types in TrackingCodePageFormatEnum.

tracking_code_type_enum
Nested message and enum types in TrackingCodeTypeEnum.

travel_placeholder_field_enum
Nested message and enum types in TravelPlaceholderFieldEnum.

user_identifier_source_enum
Nested message and enum types in UserIdentifierSourceEnum.

user_interest_taxonomy_type_enum
Nested message and enum types in UserInterestTaxonomyTypeEnum.

user_list_access_status_enum
Nested message and enum types in UserListAccessStatusEnum.

user_list_closing_reason_enum
Nested message and enum types in UserListClosingReasonEnum.

user_list_combined_rule_operator_enum
Nested message and enum types in UserListCombinedRuleOperatorEnum.

user_list_crm_data_source_type_enum
Nested message and enum types in UserListCrmDataSourceTypeEnum.

user_list_date_rule_item_operator_enum
Nested message and enum types in UserListDateRuleItemOperatorEnum.

user_list_logical_rule_operator_enum
Nested message and enum types in UserListLogicalRuleOperatorEnum.

user_list_membership_status_enum
Nested message and enum types in UserListMembershipStatusEnum.

user_list_number_rule_item_operator_enum
Nested message and enum types in UserListNumberRuleItemOperatorEnum.

user_list_prepopulation_status_enum
Nested message and enum types in UserListPrepopulationStatusEnum.

user_list_rule_type_enum
Nested message and enum types in UserListRuleTypeEnum.

user_list_size_range_enum
Nested message and enum types in UserListSizeRangeEnum.

user_list_string_rule_item_operator_enum
Nested message and enum types in UserListStringRuleItemOperatorEnum.

user_list_type_enum
Nested message and enum types in UserListTypeEnum.

vanity_pharma_display_url_mode_enum
Nested message and enum types in VanityPharmaDisplayUrlModeEnum.

vanity_pharma_text_enum
Nested message and enum types in VanityPharmaTextEnum.

webpage_condition_operand_enum
Nested message and enum types in WebpageConditionOperandEnum.

webpage_condition_operator_enum
Nested message and enum types in WebpageConditionOperatorEnum.

Structs

AccessInvitationStatusEnum
Container for enum for identifying the status of access invitation

AccessReasonEnum
Indicates the way the resource such as user list is related to a user.

AccessRoleEnum
Container for enum describing possible access role for user.

AccountBudgetProposalStatusEnum
Message describing AccountBudgetProposal statuses.

AccountBudgetProposalTypeEnum
Message describing AccountBudgetProposal types.

AccountBudgetStatusEnum
Message describing AccountBudget statuses.

AccountLinkStatusEnum
Container for enum describing possible statuses of an account link.

AdCustomizerPlaceholderFieldEnum
Values for Ad Customizer placeholder fields.

AdDestinationTypeEnum
Container for enumeration of Google Ads destination types.

AdGroupAdRotationModeEnum
Container for enum describing possible ad rotation modes of ads within an ad group.

AdGroupAdStatusEnum
Container for enum describing possible statuses of an AdGroupAd.

AdGroupCriterionApprovalStatusEnum
Container for enum describing possible AdGroupCriterion approval statuses.

AdGroupCriterionStatusEnum
Message describing AdGroupCriterion statuses.

AdGroupStatusEnum
Container for enum describing possible statuses of an ad group.

AdGroupTypeEnum
Defines types of an ad group, specific to a particular campaign channel type. This type drives validations that restrict which entities can be added to the ad group.

AdNetworkTypeEnum
Container for enumeration of Google Ads network types.

AdServingOptimizationStatusEnum
Possible ad serving statuses of a campaign.

AdStrengthEnum
Container for enum describing possible ad strengths.

AdTypeEnum
Container for enum describing possible types of an ad.

AdvertisingChannelSubTypeEnum
An immutable specialization of an Advertising Channel.

AdvertisingChannelTypeEnum
The channel type a campaign may target to serve on.

AffiliateLocationFeedRelationshipTypeEnum
Container for enum describing possible values for a relationship type for an affiliate location feed.

AffiliateLocationPlaceholderFieldEnum
Values for Affiliate Location placeholder fields.

AgeRangeTypeEnum
Container for enum describing the type of demographic age ranges.

AppCampaignAppStoreEnum
The application store that distributes mobile applications.

AppCampaignBiddingStrategyGoalTypeEnum
Container for enum describing goal towards which the bidding strategy of an app campaign should optimize for.

AppPaymentModelTypeEnum
Represents a criterion for targeting paid apps.

AppPlaceholderFieldEnum
Values for App placeholder fields.

AppStoreEnum
Container for enum describing app store type in an app extension.

AppUrlOperatingSystemTypeEnum
The possible OS types for a deeplink AppUrl.

AssetFieldTypeEnum
Container for enum describing the possible placements of an asset.

AssetLinkStatusEnum
Container for enum describing possible statuses of an asset link.

AssetPerformanceLabelEnum
Container for enum describing the performance label of an asset.

AssetTypeEnum
Container for enum describing the types of asset.

AttributionModelEnum
Container for enum representing the attribution model that describes how to distribute credit for a particular conversion across potentially many prior interactions.

BatchJobStatusEnum
Container for enum describing possible batch job statuses.

BidModifierSourceEnum
Container for enum describing possible bid modifier sources.

BiddingSourceEnum
Container for enum describing possible bidding sources.

BiddingStrategyStatusEnum
Message describing BiddingStrategy statuses.

BiddingStrategyTypeEnum
Container for enum describing possible bidding strategy types.

BillingSetupStatusEnum
Message describing BillingSetup statuses.

BrandSafetySuitabilityEnum
Container for enum with 3-Tier brand safety suitability control.

BudgetCampaignAssociationStatusEnum
Message describing the status of the association between the Budget and the Campaign.

BudgetDeliveryMethodEnum
Message describing Budget delivery methods. A delivery method determines the rate at which the Budget is spent.

BudgetPeriodEnum
Message describing Budget period.

BudgetStatusEnum
Message describing a Budget status

BudgetTypeEnum
Describes Budget types.

CallConversionReportingStateEnum
Container for enum describing possible data types for call conversion reporting state.

CallPlaceholderFieldEnum
Values for Call placeholder fields.

CallTrackingDisplayLocationEnum
Container for enum describing possible call tracking display locations.

CallTypeEnum
Container for enum describing possible types of property from where the call was made.

CalloutPlaceholderFieldEnum
Values for Callout placeholder fields.

CampaignCriterionStatusEnum
Message describing CampaignCriterion statuses.

CampaignDraftStatusEnum
Container for enum describing possible statuses of a campaign draft.

CampaignExperimentStatusEnum
Container for enum describing possible statuses of a campaign experiment.

CampaignExperimentTrafficSplitTypeEnum
Container for enum describing campaign experiment traffic split type.

CampaignExperimentTypeEnum
Container for enum describing campaign experiment type.

CampaignServingStatusEnum
Message describing Campaign serving statuses.

CampaignSharedSetStatusEnum
Container for enum describing types of campaign shared set statuses.

CampaignStatusEnum
Container for enum describing possible statuses of a campaign.

ChangeClientTypeEnum
Container for enum describing the sources that the change event resource was made through.

ChangeEventResourceTypeEnum
Container for enum describing supported resource types for the ChangeEvent resource.

ChangeStatusOperationEnum
Container for enum describing operations for the ChangeStatus resource.

ChangeStatusResourceTypeEnum
Container for enum describing supported resource types for the ChangeStatus resource.

ClickTypeEnum
Container for enumeration of Google Ads click types.

CombinedAudienceStatusEnum
The status of combined audience.

ContentLabelTypeEnum
Container for enum describing content label types in ContentLabel.

ConversionActionCategoryEnum
Container for enum describing the category of conversions that are associated with a ConversionAction.

ConversionActionCountingTypeEnum
Container for enum describing the conversion deduplication mode for conversion optimizer.

ConversionActionStatusEnum
Container for enum describing possible statuses of a conversion action.

ConversionActionTypeEnum
Container for enum describing possible types of a conversion action.

ConversionAdjustmentTypeEnum
Container for enum describing conversion adjustment types.

ConversionAttributionEventTypeEnum
Container for enum indicating the event type the conversion is attributed to.

ConversionCustomVariableStatusEnum
Container for enum describing possible statuses of a conversion custom variable.

ConversionLagBucketEnum
Container for enum representing the number of days between impression and conversion.

ConversionOrAdjustmentLagBucketEnum
Container for enum representing the number of days between the impression and the conversion or between the impression and adjustments to the conversion.

CriterionCategoryChannelAvailabilityModeEnum
Describes channel availability mode for a criterion availability - whether the availability is meant to include all advertising channels, or a particular channel with all its channel subtypes, or a channel with a certain subset of channel subtypes.

CriterionCategoryLocaleAvailabilityModeEnum
Describes locale availability mode for a criterion availability - whether it’s available globally, or a particular country with all languages, or a particular language with all countries, or a country-language pair.

CriterionSystemServingStatusEnum
Container for enum describing possible criterion system serving statuses.

CriterionTypeEnum
The possible types of a criterion.

CustomAudienceMemberTypeEnum
The type of custom audience member.

CustomAudienceStatusEnum
The status of custom audience.

CustomAudienceTypeEnum
The types of custom audience.

CustomInterestMemberTypeEnum
The types of custom interest member, either KEYWORD or URL.

CustomInterestStatusEnum
The status of custom interest.

CustomInterestTypeEnum
The types of custom interest.

CustomPlaceholderFieldEnum
Values for Custom placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

CustomerMatchUploadKeyTypeEnum
Indicates what type of data are the user list’s members matched from.

CustomerPayPerConversionEligibilityFailureReasonEnum
Container for enum describing reasons why a customer is not eligible to use PaymentMode.CONVERSIONS.

DataDrivenModelStatusEnum
Container for enum indicating data driven model status.

DayOfWeekEnum
Container for enumeration of days of the week, e.g., “Monday”.

DeviceEnum
Container for enumeration of Google Ads devices available for targeting.

DisplayAdFormatSettingEnum
Container for display ad format settings.

DisplayUploadProductTypeEnum
Container for display upload product types. Product types that have the word “DYNAMIC” in them must be associated with a campaign that has a dynamic remarketing feed. See https://support.google.com/google-ads/answer/6053288 for more info about dynamic remarketing. Other product types are regarded as “static” and do not have this requirement.

DistanceBucketEnum
Container for distance buckets of a user’s distance from an advertiser’s location extension.

DsaPageFeedCriterionFieldEnum
Values for Dynamic Search Ad Page Feed criterion fields.

EducationPlaceholderFieldEnum
Values for Education placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

ExtensionSettingDeviceEnum
Container for enum describing extension setting device types.

ExtensionTypeEnum
Container for enum describing possible data types for an extension in an extension setting.

ExternalConversionSourceEnum
Container for enum describing the external conversion source that is associated with a ConversionAction.

FeedAttributeTypeEnum
Container for enum describing possible data types for a feed attribute.

FeedItemQualityApprovalStatusEnum
Container for enum describing possible quality evaluation approval statuses of a feed item.

FeedItemQualityDisapprovalReasonEnum
Container for enum describing possible quality evaluation disapproval reasons of a feed item.

FeedItemSetStatusEnum
Container for enum describing possible statuses of a feed item set.

FeedItemSetStringFilterTypeEnum
The type of string matching to be used for a dynamic FeedItemSet filter.

FeedItemStatusEnum
Container for enum describing possible statuses of a feed item.

FeedItemTargetDeviceEnum
Container for enum describing possible data types for a feed item target device.

FeedItemTargetStatusEnum
Container for enum describing possible statuses of a feed item target.

FeedItemTargetTypeEnum
Container for enum describing possible types of a feed item target.

FeedItemValidationStatusEnum
Container for enum describing possible validation statuses of a feed item.

FeedLinkStatusEnum
Container for an enum describing possible statuses of a feed link.

FeedMappingCriterionTypeEnum
Container for enum describing possible criterion types for a feed mapping.

FeedMappingStatusEnum
Container for enum describing possible statuses of a feed mapping.

FeedOriginEnum
Container for enum describing possible values for a feed origin.

FeedStatusEnum
Container for enum describing possible statuses of a feed.

FlightPlaceholderFieldEnum
Values for Flight placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

FrequencyCapEventTypeEnum
Container for enum describing the type of event that the cap applies to.

FrequencyCapLevelEnum
Container for enum describing the level on which the cap is to be applied.

FrequencyCapTimeUnitEnum
Container for enum describing the unit of time the cap is defined at.

GenderTypeEnum
Container for enum describing the type of demographic genders.

GeoTargetConstantStatusEnum
Container for describing the status of a geo target constant.

GeoTargetingRestrictionEnum
Message describing feed item geo targeting restriction.

GeoTargetingTypeEnum
Container for enum describing possible geo targeting types.

GoogleAdsFieldCategoryEnum
Container for enum that determines if the described artifact is a resource or a field, and if it is a field, when it segments search queries.

GoogleAdsFieldDataTypeEnum
Container holding the various data types.

GoogleVoiceCallStatusEnum
Container for enum describing possible statuses of a google voice call.

HotelDateSelectionTypeEnum
Container for enum describing possible hotel date selection types

HotelPlaceholderFieldEnum
Values for Hotel placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

HotelPriceBucketEnum
Container for enum describing hotel price bucket for a hotel itinerary.

HotelRateTypeEnum
Container for enum describing possible hotel rate types.

ImagePlaceholderFieldEnum
Values for Advertiser Provided Image placeholder fields.

IncomeRangeTypeEnum
Container for enum describing the type of demographic income ranges.

InteractionEventTypeEnum
Container for enum describing types of payable and free interactions.

InteractionTypeEnum
Container for enum describing possible interaction types.

InvoiceTypeEnum
Container for enum describing the type of invoices.

JobPlaceholderFieldEnum
Values for Job placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

KeywordMatchTypeEnum
Message describing Keyword match types.

KeywordPlanAggregateMetricTypeEnum
The enumeration of keyword plan aggregate metric types.

KeywordPlanCompetitionLevelEnum
Container for enumeration of keyword competition levels. The competition level indicates how competitive ad placement is for a keyword and is determined by the number of advertisers bidding on that keyword relative to all keywords across Google. The competition level can depend on the location and Search Network targeting options you’ve selected.

KeywordPlanConceptGroupTypeEnum
Container for enumeration of keyword plan concept group types.

KeywordPlanForecastIntervalEnum
Container for enumeration of forecast intervals.

KeywordPlanKeywordAnnotationEnum
Container for enumeration of keyword plan keyword annotations.

KeywordPlanNetworkEnum
Container for enumeration of keyword plan forecastable network types.

LabelStatusEnum
Container for enum describing possible status of a label.

LeadFormCallToActionTypeEnum
Describes the type of call-to-action phrases in a lead form.

LeadFormDesiredIntentEnum
Describes the desired level of intent of generated leads.

LeadFormFieldUserInputTypeEnum
Describes the input type of a lead form field.

LeadFormPostSubmitCallToActionTypeEnum
Describes the type of post-submit call-to-action phrases for a lead form.

LegacyAppInstallAdAppStoreEnum
Container for enum describing app store type in a legacy app install ad.

LinkedAccountTypeEnum
Container for enum describing different types of Linked accounts.

ListingGroupTypeEnum
Container for enum describing the type of the listing group.

LocalPlaceholderFieldEnum
Values for Local placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

LocationExtensionTargetingCriterionFieldEnum
Values for Location Extension Targeting criterion fields.

LocationGroupRadiusUnitsEnum
Container for enum describing unit of radius in location group.

LocationPlaceholderFieldEnum
Values for Location placeholder fields.

LocationSourceTypeEnum
Used to distinguish the location source type.

ManagerLinkStatusEnum
Container for enum describing possible status of a manager and client link.

MatchingFunctionContextTypeEnum
Container for context types for an operand in a matching function.

MatchingFunctionOperatorEnum
Container for enum describing matching function operator.

MediaTypeEnum
Container for enum describing the types of media.

MerchantCenterLinkStatusEnum
Container for enum describing possible statuses of a Google Merchant Center link.

MessagePlaceholderFieldEnum
Values for Message placeholder fields.

MimeTypeEnum
Container for enum describing the mime types.

MinuteOfHourEnum
Container for enumeration of quarter-hours.

MobileAppVendorEnum
Container for enum describing different types of mobile app vendors.

MobileDeviceTypeEnum
Container for enum describing the types of mobile device.

MonthOfYearEnum
Container for enumeration of months of the year, e.g., “January”.

NegativeGeoTargetTypeEnum
Container for enum describing possible negative geo target types.

OfflineUserDataJobFailureReasonEnum
Container for enum describing reasons why an offline user data job failed to be processed.

OfflineUserDataJobStatusEnum
Container for enum describing status of an offline user data job.

OfflineUserDataJobTypeEnum
Container for enum describing types of an offline user data job.

OperatingSystemVersionOperatorTypeEnum
Container for enum describing the type of OS operators.

OptimizationGoalTypeEnum
Container for enum describing the type of optimization goal.

ParentalStatusTypeEnum
Container for enum describing the type of demographic parental statuses.

PaymentModeEnum
Container for enum describing possible payment modes.

PlaceholderTypeEnum
Container for enum describing possible placeholder types for a feed mapping.

PlacementTypeEnum
Container for enum describing possible placement types.

PolicyApprovalStatusEnum
Container for enum describing possible policy approval statuses.

PolicyReviewStatusEnum
Container for enum describing possible policy review statuses.

PolicyTopicEntryTypeEnum
Container for enum describing possible policy topic entry types.

PolicyTopicEvidenceDestinationMismatchUrlTypeEnum
Container for enum describing possible policy topic evidence destination mismatch url types.

PolicyTopicEvidenceDestinationNotWorkingDeviceEnum
Container for enum describing possible policy topic evidence destination not working devices.

PolicyTopicEvidenceDestinationNotWorkingDnsErrorTypeEnum
Container for enum describing possible policy topic evidence destination not working DNS error types.

PositiveGeoTargetTypeEnum
Container for enum describing possible positive geo target types.

PreferredContentTypeEnum
Container for enumeration of preferred content criterion type.

PriceExtensionPriceQualifierEnum
Container for enum describing a price extension price qualifier.

PriceExtensionPriceUnitEnum
Container for enum describing price extension price unit.

PriceExtensionTypeEnum
Container for enum describing types for a price extension.

PricePlaceholderFieldEnum
Values for Price placeholder fields.

ProductBiddingCategoryLevelEnum
Level of a product bidding category.

ProductBiddingCategoryStatusEnum
Status of the product bidding category.

ProductChannelEnum
Locality of a product offer.

ProductChannelExclusivityEnum
Availability of a product offer.

ProductConditionEnum
Condition of a product offer.

ProductCustomAttributeIndexEnum
Container for enum describing the index of the product custom attribute.

ProductTypeLevelEnum
Level of the type of a product offer.

PromotionExtensionDiscountModifierEnum
Container for enum describing possible a promotion extension discount modifier.

PromotionExtensionOccasionEnum
Container for enum describing a promotion extension occasion. For more information about the occasions please check: https://support.google.com/google-ads/answer/7367521

PromotionPlaceholderFieldEnum
Values for Promotion placeholder fields.

ProximityRadiusUnitsEnum
Container for enum describing unit of radius in proximity.

QualityScoreBucketEnum
The relative performance compared to other advertisers.

ReachPlanAdLengthEnum
Message describing length of a plannable video ad.

ReachPlanAgeRangeEnum
Message describing plannable age ranges.

ReachPlanNetworkEnum
Container for enum describing plannable networks.

RealEstatePlaceholderFieldEnum
Values for Real Estate placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

RecommendationTypeEnum
Container for enum describing types of recommendations.

ResourceChangeOperationEnum
Container for enum describing resource change operations in the ChangeEvent resource.

ResourceLimitTypeEnum
Container for enum describing possible resource limit types.

ResponseContentTypeEnum
Container for possible response content types.

SearchEngineResultsPageTypeEnum
The type of the search engine results page.

SearchTermMatchTypeEnum
Container for enum describing match types for a keyword triggering an ad.

SearchTermTargetingStatusEnum
Container for enum indicating whether a search term is one of your targeted or excluded keywords.

ServedAssetFieldTypeEnum
Container for enum describing possible asset field types.

SharedSetStatusEnum
Container for enum describing types of shared set statuses.

SharedSetTypeEnum
Container for enum describing types of shared sets.

SimulationModificationMethodEnum
Container for enum describing the method by which a simulation modifies a field.

SimulationTypeEnum
Container for enum describing the field a simulation modifies.

SitelinkPlaceholderFieldEnum
Values for Sitelink placeholder fields.

SlotEnum
Container for enumeration of possible positions of the Ad.

SpendingLimitTypeEnum
Message describing spending limit types.

StructuredSnippetPlaceholderFieldEnum
Values for Structured Snippet placeholder fields.

SummaryRowSettingEnum
Indicates summary row setting in request parameter.

SystemManagedResourceSourceEnum
Container for enum describing possible system managed entity sources.

TargetCpaOptInRecommendationGoalEnum
Container for enum describing goals for TargetCpaOptIn recommendation.

TargetImpressionShareLocationEnum
Container for enum describing where on the first search results page the automated bidding system should target impressions for the TargetImpressionShare bidding strategy.

TargetingDimensionEnum
The dimensions that can be targeted.

TimeTypeEnum
Message describing time types.

TrackingCodePageFormatEnum
Container for enum describing the format of the web page where the tracking tag and snippet will be installed.

TrackingCodeTypeEnum
Container for enum describing the type of the generated tag snippets for tracking conversions.

TravelPlaceholderFieldEnum
Values for Travel placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

UserIdentifierSourceEnum
Container for enum describing the source of the user identifier for offline Store Sales third party uploads.

UserInterestTaxonomyTypeEnum
Message describing a UserInterestTaxonomyType.

UserListAccessStatusEnum
Indicates if this client still has access to the list.

UserListClosingReasonEnum
Indicates the reason why the userlist was closed. This enum is only used when a list is auto-closed by the system.

UserListCombinedRuleOperatorEnum
Logical operator connecting two rules.

UserListCrmDataSourceTypeEnum
Indicates source of Crm upload data.

UserListDateRuleItemOperatorEnum
Supported rule operator for date type.

UserListLogicalRuleOperatorEnum
The logical operator of the rule.

UserListMembershipStatusEnum
Membership status of this user list. Indicates whether a user list is open or active. Only open user lists can accumulate more users and can be used for targeting.

UserListNumberRuleItemOperatorEnum
Supported rule operator for number type.

UserListPrepopulationStatusEnum
Indicates status of prepopulation based on the rule.

UserListRuleTypeEnum
Rule based user list rule type.

UserListSizeRangeEnum
Size range in terms of number of users of a UserList.

UserListStringRuleItemOperatorEnum
Supported rule operator for string type.

UserListTypeEnum
The user list types.

VanityPharmaDisplayUrlModeEnum
The display mode for vanity pharma URLs.

VanityPharmaTextEnum
The text that will be displayed in display URL of the text ad when website description is the selected display mode for vanity pharma URLs.

WebpageConditionOperandEnum
Container for enum describing webpage condition operand in webpage criterion.

WebpageConditionOperatorEnum
Container for enum describing webpage condition operator in webpage criterion.