Enum googapis::google::ads::googleads::v8::enums::lead_form_field_user_input_type_enum::LeadFormFieldUserInputType

#[repr(i32)]
pub enum LeadFormFieldUserInputType {

    Unspecified,
    Unknown,
    FullName,
    Email,
    PhoneNumber,
    PostalCode,
    City,
    Region,
    Country,
    WorkEmail,
    CompanyName,
    WorkPhone,
    JobTitle,
    FirstName,
    LastName,
    VehicleModel,
    VehicleType,
    PreferredDealership,
    VehiclePurchaseTimeline,
    VehicleOwnership,
    VehiclePaymentType,
    VehicleCondition,
    CompanySize,
    AnnualSales,
    YearsInBusiness,
    JobDepartment,
    JobRole,
    EducationProgram,
    EducationCourse,
    Product,
    Service,
    Offer,
    Category,
    PreferredContactMethod,
    PreferredLocation,
    PreferredContactTime,
    PurchaseTimeline,
    YearsOfExperience,
    JobIndustry,
    LevelOfEducation,
    PropertyType,
    RealtorHelpGoal,
    PropertyCommunity,
    PriceRange,
    NumberOfBedrooms,
    FurnishedProperty,
    PetsAllowedProperty,
    NextPlannedPurchase,
    EventSignupInterest,
    PreferredShoppingPlaces,
    FavoriteBrand,
    TransportationCommercialLicenseType,
    EventBookingInterest,
    DestinationCountry,
    DestinationCity,
    DepartureCountry,
    DepartureCity,
    DepartureDate,
    ReturnDate,
    NumberOfTravelers,
    TravelBudget,
    TravelAccommodation,

}

Enum describing the input type of a lead form field.

Variants

Unspecified

Not specified.

Unknown

Used for return value only. Represents value unknown in this version.

FullName

The user will be asked to fill in their given and family name. This field cannot be set at the same time as GIVEN_NAME or FAMILY_NAME.

Email

The user will be asked to fill in their email address.

PhoneNumber

The user will be asked to fill in their phone number.

PostalCode

The user will be asked to fill in their zip code.

City

The user will be asked to fill in their city.

Region

The user will be asked to fill in their region part of the address (e.g. state for US, province for Canada).

Country

The user will be asked to fill in their country.

WorkEmail

The user will be asked to fill in their work email address.

CompanyName

The user will be asked to fill in their company name.

WorkPhone

The user will be asked to fill in their work phone.

JobTitle

The user will be asked to fill in their job title.

FirstName

The user will be asked to fill in their first name. This field can not be set at the same time as FULL_NAME.

LastName

The user will be asked to fill in their last name. This field can not be set at the same time as FULL_NAME.

VehicleModel

Question: “Which model are you interested in?” Category: “Auto”

VehicleType

Question: “Which type of vehicle are you interested in?” Category: “Auto”

PreferredDealership

Question: “What is your preferred dealership?” Category: “Auto”

VehiclePurchaseTimeline

Question: “When do you plan on purchasing a vehicle?” Category: “Auto”

VehicleOwnership

Question: “Do you own a vehicle?” Category: “Auto”

VehiclePaymentType

Question: “What vehicle ownership option are you interested in?” Category: “Auto”

VehicleCondition

Question: “What type of vehicle condition are you interested in?” Category: “Auto”

CompanySize

Question: “What size is your company?” Category: “Business”

AnnualSales

Question: “What is your annual sales volume?” Category: “Business”

YearsInBusiness

Question: “How many years have you been in business?” Category: “Business”

JobDepartment

Question: “What is your job department?” Category: “Business”

JobRole

Question: “What is your job role?” Category: “Business”

EducationProgram

Question: “Which program are you interested in?” Category: “Education”

EducationCourse

Question: “Which course are you interested in?” Category: “Education”

Product

Question: “Which product are you interested in?” Category: “General”

Service

Question: “Which service are you interested in?” Category: “General”

Offer

Question: “Which offer are you interested in?” Category: “General”

Category

Question: “Which category are you interested in?” Category: “General”

PreferredContactMethod

Question: “What is your preferred method of contact?” Category: “General”

PreferredLocation

Question: “What is your preferred location?” Category: “General”

PreferredContactTime

Question: “What is the best time to contact you?” Category: “General”

PurchaseTimeline

Question: “When are you looking to make a purchase?” Category: “General”

YearsOfExperience

Question: “How many years of work experience do you have?” Category: “Jobs”

JobIndustry

Question: “What industry do you work in?” Category: “Jobs”

LevelOfEducation

Question: “What is your highest level of education?” Category: “Jobs”

PropertyType

Question: “What type of property are you looking for?” Category: “Real Estate”

RealtorHelpGoal

Question: “What do you need a realtor’s help with?” Category: “Real Estate”

PropertyCommunity

Question: “What neighborhood are you interested in?” Category: “Real Estate”

PriceRange

Question: “What price range are you looking for?” Category: “Real Estate”

NumberOfBedrooms

Question: “How many bedrooms are you looking for?” Category: “Real Estate”

FurnishedProperty

Question: “Are you looking for a fully furnished property?” Category: “Real Estate”

PetsAllowedProperty

Question: “Are you looking for properties that allow pets?” Category: “Real Estate”

NextPlannedPurchase

Question: “What is the next product you plan to purchase?” Category: “Retail”

EventSignupInterest

Question: “Would you like to sign up for an event?” Category: “Retail”

PreferredShoppingPlaces

Question: “Where are you interested in shopping?” Category: “Retail”

FavoriteBrand

Question: “What is your favorite brand?” Category: “Retail”

TransportationCommercialLicenseType

Question: “Which type of valid commercial license do you have?” Category: “Transportation”

EventBookingInterest

Question: “Interested in booking an event?” Category: “Travel”

DestinationCountry

Question: “What is your destination country?” Category: “Travel”

DestinationCity

Question: “What is your destination city?” Category: “Travel”

DepartureCountry

Question: “What is your departure country?” Category: “Travel”

DepartureCity

Question: “What is your departure city?” Category: “Travel”

DepartureDate

Question: “What is your departure date?” Category: “Travel”

ReturnDate

Question: “What is your return date?” Category: “Travel”

NumberOfTravelers

Question: “How many people are you traveling with?” Category: “Travel”

TravelBudget

Question: “What is your travel budget?” Category: “Travel”

TravelAccommodation

Question: “Where do you want to stay during your travel?” Category: “Travel”

Implementations

impl LeadFormFieldUserInputType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of LeadFormFieldUserInputType.

pub fn from_i32(value: i32) -> Option<LeadFormFieldUserInputType>

Converts an i32 to a LeadFormFieldUserInputType, or None if value is not a valid variant.