Enum googapis::google::ads::googleads::v8::enums::price_placeholder_field_enum::PricePlaceholderField

#[repr(i32)]
pub enum PricePlaceholderField {

    Unspecified,
    Unknown,
    Type,
    PriceQualifier,
    TrackingTemplate,
    Language,
    FinalUrlSuffix,
    Item1Header,
    Item1Description,
    Item1Price,
    Item1Unit,
    Item1FinalUrls,
    Item1FinalMobileUrls,
    Item2Header,
    Item2Description,
    Item2Price,
    Item2Unit,
    Item2FinalUrls,
    Item2FinalMobileUrls,
    Item3Header,
    Item3Description,
    Item3Price,
    Item3Unit,
    Item3FinalUrls,
    Item3FinalMobileUrls,
    Item4Header,
    Item4Description,
    Item4Price,
    Item4Unit,
    Item4FinalUrls,
    Item4FinalMobileUrls,
    Item5Header,
    Item5Description,
    Item5Price,
    Item5Unit,
    Item5FinalUrls,
    Item5FinalMobileUrls,
    Item6Header,
    Item6Description,
    Item6Price,
    Item6Unit,
    Item6FinalUrls,
    Item6FinalMobileUrls,
    Item7Header,
    Item7Description,
    Item7Price,
    Item7Unit,
    Item7FinalUrls,
    Item7FinalMobileUrls,
    Item8Header,
    Item8Description,
    Item8Price,
    Item8Unit,
    Item8FinalUrls,
    Item8FinalMobileUrls,

}

Possible values for Price placeholder fields.

Variants

Unspecified

Not specified.

Unknown

Used for return value only. Represents value unknown in this version.

Type

Data Type: STRING. The type of your price feed. Must match one of the predefined price feed type exactly.

PriceQualifier

Data Type: STRING. The qualifier of each price. Must match one of the predefined price qualifiers exactly.

TrackingTemplate

Data Type: URL. Tracking template for the price feed when using Upgraded URLs.

Language

Data Type: STRING. Language of the price feed. Must match one of the available available locale codes exactly.

FinalUrlSuffix

Data Type: STRING. Final URL suffix for the price feed when using parallel tracking.

Item1Header

Data Type: STRING. The header of item 1 of the table.

Item1Description

Data Type: STRING. The description of item 1 of the table.

Item1Price

Data Type: MONEY. The price (money with currency) of item 1 of the table, e.g., 30 USD. The currency must match one of the available currencies.

Item1Unit

Data Type: STRING. The price unit of item 1 of the table. Must match one of the predefined price units.

Item1FinalUrls

Data Type: URL_LIST. The final URLs of item 1 of the table when using Upgraded URLs.

Item1FinalMobileUrls

Data Type: URL_LIST. The final mobile URLs of item 1 of the table when using Upgraded URLs.

Item2Header

Data Type: STRING. The header of item 2 of the table.

Item2Description

Data Type: STRING. The description of item 2 of the table.

Item2Price

Data Type: MONEY. The price (money with currency) of item 2 of the table, e.g., 30 USD. The currency must match one of the available currencies.

Item2Unit

Data Type: STRING. The price unit of item 2 of the table. Must match one of the predefined price units.

Item2FinalUrls

Data Type: URL_LIST. The final URLs of item 2 of the table when using Upgraded URLs.

Item2FinalMobileUrls

Data Type: URL_LIST. The final mobile URLs of item 2 of the table when using Upgraded URLs.

Item3Header

Data Type: STRING. The header of item 3 of the table.

Item3Description

Data Type: STRING. The description of item 3 of the table.

Item3Price

Data Type: MONEY. The price (money with currency) of item 3 of the table, e.g., 30 USD. The currency must match one of the available currencies.

Item3Unit

Data Type: STRING. The price unit of item 3 of the table. Must match one of the predefined price units.

Item3FinalUrls

Data Type: URL_LIST. The final URLs of item 3 of the table when using Upgraded URLs.

Item3FinalMobileUrls

Data Type: URL_LIST. The final mobile URLs of item 3 of the table when using Upgraded URLs.

Item4Header

Data Type: STRING. The header of item 4 of the table.

Item4Description

Data Type: STRING. The description of item 4 of the table.

Item4Price

Data Type: MONEY. The price (money with currency) of item 4 of the table, e.g., 30 USD. The currency must match one of the available currencies.

Item4Unit

Data Type: STRING. The price unit of item 4 of the table. Must match one of the predefined price units.

Item4FinalUrls

Data Type: URL_LIST. The final URLs of item 4 of the table when using Upgraded URLs.

Item4FinalMobileUrls

Data Type: URL_LIST. The final mobile URLs of item 4 of the table when using Upgraded URLs.

Item5Header

Data Type: STRING. The header of item 5 of the table.

Item5Description

Data Type: STRING. The description of item 5 of the table.

Item5Price

Data Type: MONEY. The price (money with currency) of item 5 of the table, e.g., 30 USD. The currency must match one of the available currencies.

Item5Unit

Data Type: STRING. The price unit of item 5 of the table. Must match one of the predefined price units.

Item5FinalUrls

Data Type: URL_LIST. The final URLs of item 5 of the table when using Upgraded URLs.

Item5FinalMobileUrls

Data Type: URL_LIST. The final mobile URLs of item 5 of the table when using Upgraded URLs.

Item6Header

Data Type: STRING. The header of item 6 of the table.

Item6Description

Data Type: STRING. The description of item 6 of the table.

Item6Price

Data Type: MONEY. The price (money with currency) of item 6 of the table, e.g., 30 USD. The currency must match one of the available currencies.

Item6Unit

Data Type: STRING. The price unit of item 6 of the table. Must match one of the predefined price units.

Item6FinalUrls

Data Type: URL_LIST. The final URLs of item 6 of the table when using Upgraded URLs.

Item6FinalMobileUrls

Data Type: URL_LIST. The final mobile URLs of item 6 of the table when using Upgraded URLs.

Item7Header

Data Type: STRING. The header of item 7 of the table.

Item7Description

Data Type: STRING. The description of item 7 of the table.

Item7Price

Data Type: MONEY. The price (money with currency) of item 7 of the table, e.g., 30 USD. The currency must match one of the available currencies.

Item7Unit

Data Type: STRING. The price unit of item 7 of the table. Must match one of the predefined price units.

Item7FinalUrls

Data Type: URL_LIST. The final URLs of item 7 of the table when using Upgraded URLs.

Item7FinalMobileUrls

Data Type: URL_LIST. The final mobile URLs of item 7 of the table when using Upgraded URLs.

Item8Header

Data Type: STRING. The header of item 8 of the table.

Item8Description

Data Type: STRING. The description of item 8 of the table.

Item8Price

Data Type: MONEY. The price (money with currency) of item 8 of the table, e.g., 30 USD. The currency must match one of the available currencies.

Item8Unit

Data Type: STRING. The price unit of item 8 of the table. Must match one of the predefined price units.

Item8FinalUrls

Data Type: URL_LIST. The final URLs of item 8 of the table when using Upgraded URLs.

Item8FinalMobileUrls

Data Type: URL_LIST. The final mobile URLs of item 8 of the table when using Upgraded URLs.

Implementations

impl PricePlaceholderField

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of PricePlaceholderField.

pub fn from_i32(value: i32) -> Option<PricePlaceholderField>

Converts an i32 to a PricePlaceholderField, or None if value is not a valid variant.