Enum googapis::google::ads::googleads::v8::enums::resource_limit_type_enum::ResourceLimitType

#[repr(i32)]
pub enum ResourceLimitType {

    Unspecified,
    Unknown,
    CampaignsPerCustomer,
    BaseCampaignsPerCustomer,
    ExperimentCampaignsPerCustomer,
    HotelCampaignsPerCustomer,
    SmartShoppingCampaignsPerCustomer,
    AdGroupsPerCampaign,
    AdGroupsPerShoppingCampaign,
    AdGroupsPerHotelCampaign,
    ReportingAdGroupsPerLocalCampaign,
    ReportingAdGroupsPerAppCampaign,
    ManagedAdGroupsPerSmartCampaign,
    AdGroupCriteriaPerCustomer,
    BaseAdGroupCriteriaPerCustomer,
    ExperimentAdGroupCriteriaPerCustomer,
    AdGroupCriteriaPerCampaign,
    CampaignCriteriaPerCustomer,
    BaseCampaignCriteriaPerCustomer,
    ExperimentCampaignCriteriaPerCustomer,
    WebpageCriteriaPerCustomer,
    BaseWebpageCriteriaPerCustomer,
    ExperimentWebpageCriteriaPerCustomer,
    CombinedAudienceCriteriaPerAdGroup,
    CustomerNegativePlacementCriteriaPerCustomer,
    CustomerNegativeYoutubeChannelCriteriaPerCustomer,
    CriteriaPerAdGroup,
    ListingGroupsPerAdGroup,
    ExplicitlySharedBudgetsPerCustomer,
    ImplicitlySharedBudgetsPerCustomer,
    CombinedAudienceCriteriaPerCampaign,
    NegativeKeywordsPerCampaign,
    NegativePlacementsPerCampaign,
    GeoTargetsPerCampaign,
    NegativeIpBlocksPerCampaign,
    ProximitiesPerCampaign,
    ListingScopesPerShoppingCampaign,
    ListingScopesPerNonShoppingCampaign,
    NegativeKeywordsPerSharedSet,
    NegativePlacementsPerSharedSet,
    SharedSetsPerCustomerForTypeDefault,
    SharedSetsPerCustomerForNegativePlacementListLower,
    HotelAdvanceBookingWindowBidModifiersPerAdGroup,
    BiddingStrategiesPerCustomer,
    BasicUserListsPerCustomer,
    LogicalUserListsPerCustomer,
    BaseAdGroupAdsPerCustomer,
    ExperimentAdGroupAdsPerCustomer,
    AdGroupAdsPerCampaign,
    TextAndOtherAdsPerAdGroup,
    ImageAdsPerAdGroup,
    ShoppingSmartAdsPerAdGroup,
    ResponsiveSearchAdsPerAdGroup,
    AppAdsPerAdGroup,
    AppEngagementAdsPerAdGroup,
    LocalAdsPerAdGroup,
    VideoAdsPerAdGroup,
    LeadFormAssetLinksPerCampaign,
    VersionsPerAd,
    UserFeedsPerCustomer,
    SystemFeedsPerCustomer,
    FeedAttributesPerFeed,
    FeedItemsPerCustomer,
    CampaignFeedsPerCustomer,
    BaseCampaignFeedsPerCustomer,
    ExperimentCampaignFeedsPerCustomer,
    AdGroupFeedsPerCustomer,
    BaseAdGroupFeedsPerCustomer,
    ExperimentAdGroupFeedsPerCustomer,
    AdGroupFeedsPerCampaign,
    FeedItemSetsPerCustomer,
    FeedItemsPerFeedItemSet,
    CampaignExperimentsPerCustomer,
    ExperimentArmsPerVideoExperiment,
    OwnedLabelsPerCustomer,
    LabelsPerCampaign,
    LabelsPerAdGroup,
    LabelsPerAdGroupAd,
    LabelsPerAdGroupCriterion,
    TargetCustomersPerLabel,
    KeywordPlansPerUserPerCustomer,
    KeywordPlanAdGroupKeywordsPerKeywordPlan,
    KeywordPlanAdGroupsPerKeywordPlan,
    KeywordPlanNegativeKeywordsPerKeywordPlan,
    KeywordPlanCampaignsPerKeywordPlan,
    ConversionActionsPerCustomer,
    BatchJobOperationsPerJob,
    BatchJobsPerCustomer,
    HotelCheckInDateRangeBidModifiersPerAdGroup,

}

Resource limit type.

Variants

Unspecified

No value has been specified.

Unknown

Used for return value only. Represents an unclassified operation unknown in this version.

CampaignsPerCustomer

Number of ENABLED and PAUSED campaigns per customer.

BaseCampaignsPerCustomer

Number of ENABLED and PAUSED base campaigns per customer.

ExperimentCampaignsPerCustomer

Number of ENABLED and PAUSED experiment campaigns per customer.

HotelCampaignsPerCustomer

Number of ENABLED and PAUSED Hotel campaigns per customer.

SmartShoppingCampaignsPerCustomer

Number of ENABLED and PAUSED Smart Shopping campaigns per customer.

AdGroupsPerCampaign

Number of ENABLED ad groups per campaign.

AdGroupsPerShoppingCampaign

Number of ENABLED ad groups per Shopping campaign.

AdGroupsPerHotelCampaign

Number of ENABLED ad groups per Hotel campaign.

ReportingAdGroupsPerLocalCampaign

Number of ENABLED reporting ad groups per local campaign.

ReportingAdGroupsPerAppCampaign

Number of ENABLED reporting ad groups per App campaign. It includes app campaign and app campaign for engagement.

ManagedAdGroupsPerSmartCampaign

Number of ENABLED managed ad groups per smart campaign.

AdGroupCriteriaPerCustomer

Number of ENABLED ad group criteria per customer. An ad group criterion is considered as ENABLED if:

it’s not REMOVED

its ad group is not REMOVED

its campaign is not REMOVED.

BaseAdGroupCriteriaPerCustomer

Number of ad group criteria across all base campaigns for a customer.

ExperimentAdGroupCriteriaPerCustomer

Number of ad group criteria across all experiment campaigns for a customer.

AdGroupCriteriaPerCampaign

Number of ENABLED ad group criteria per campaign. An ad group criterion is considered as ENABLED if:

it’s not REMOVED

its ad group is not REMOVED.

CampaignCriteriaPerCustomer

Number of ENABLED campaign criteria per customer.

BaseCampaignCriteriaPerCustomer

Number of ENABLED campaign criteria across all base campaigns for a customer.

ExperimentCampaignCriteriaPerCustomer

Number of ENABLED campaign criteria across all experiment campaigns for a customer.

WebpageCriteriaPerCustomer

Number of ENABLED webpage criteria per customer, including campaign level and ad group level.

BaseWebpageCriteriaPerCustomer

Number of ENABLED webpage criteria across all base campaigns for a customer.

ExperimentWebpageCriteriaPerCustomer

Meximum number of ENABLED webpage criteria across all experiment campaigns for a customer.

CombinedAudienceCriteriaPerAdGroup

Number of combined audience criteria per ad group.

CustomerNegativePlacementCriteriaPerCustomer

Limit for placement criterion type group in customer negative criterion.

CustomerNegativeYoutubeChannelCriteriaPerCustomer

Limit for YouTube TV channels in customer negative criterion.

CriteriaPerAdGroup

Number of ENABLED criteria per ad group.

ListingGroupsPerAdGroup

Number of listing group criteria per ad group.

ExplicitlySharedBudgetsPerCustomer

Number of ENABLED explicitly shared budgets per customer.

ImplicitlySharedBudgetsPerCustomer

Number of ENABLED implicitly shared budgets per customer.

CombinedAudienceCriteriaPerCampaign

Number of combined audience criteria per campaign.

NegativeKeywordsPerCampaign

Number of negative keywords per campaign.

NegativePlacementsPerCampaign

Number of excluded campaign criteria in placement dimension, e.g. placement, mobile application, YouTube channel, etc. The API criterion type is NOT limited to placement only, and this does not include exclusions at the ad group or other levels.

GeoTargetsPerCampaign

Number of geo targets per campaign.

NegativeIpBlocksPerCampaign

Number of negative IP blocks per campaign.

ProximitiesPerCampaign

Number of proximity targets per campaign.

ListingScopesPerShoppingCampaign

Number of listing scopes per Shopping campaign.

ListingScopesPerNonShoppingCampaign

Number of listing scopes per non-Shopping campaign.

NegativeKeywordsPerSharedSet

Number of criteria per negative keyword shared set.

NegativePlacementsPerSharedSet

Number of criteria per negative placement shared set.

SharedSetsPerCustomerForTypeDefault

Default number of shared sets allowed per type per customer.

SharedSetsPerCustomerForNegativePlacementListLower

Number of shared sets of negative placement list type for a manager customer.

HotelAdvanceBookingWindowBidModifiersPerAdGroup

Number of hotel_advance_booking_window bid modifiers per ad group.

BiddingStrategiesPerCustomer

Number of ENABLED shared bidding strategies per customer.

BasicUserListsPerCustomer

Number of open basic user lists per customer.

LogicalUserListsPerCustomer

Number of open logical user lists per customer.

BaseAdGroupAdsPerCustomer

Number of ENABLED and PAUSED ad group ads across all base campaigns for a customer.

ExperimentAdGroupAdsPerCustomer

Number of ENABLED and PAUSED ad group ads across all experiment campaigns for a customer.

AdGroupAdsPerCampaign

Number of ENABLED and PAUSED ad group ads per campaign.

TextAndOtherAdsPerAdGroup

Number of ENABLED ads per ad group that do not fall in to other buckets. Includes text and many other types.

ImageAdsPerAdGroup

Number of ENABLED image ads per ad group.

ShoppingSmartAdsPerAdGroup

Number of ENABLED shopping smart ads per ad group.

ResponsiveSearchAdsPerAdGroup

Number of ENABLED responsive search ads per ad group.

AppAdsPerAdGroup

Number of ENABLED app ads per ad group.

AppEngagementAdsPerAdGroup

Number of ENABLED app engagement ads per ad group.

LocalAdsPerAdGroup

Number of ENABLED local ads per ad group.

VideoAdsPerAdGroup

Number of ENABLED video ads per ad group.

LeadFormAssetLinksPerCampaign

Number of ENABLED lead form asset links per campaign.

VersionsPerAd

Number of versions per ad.

UserFeedsPerCustomer

Number of ENABLED user feeds per customer.

SystemFeedsPerCustomer

Number of ENABLED system feeds per customer.

FeedAttributesPerFeed

Number of feed attributes per feed.

FeedItemsPerCustomer

Number of ENABLED feed items per customer.

CampaignFeedsPerCustomer

Number of ENABLED campaign feeds per customer.

BaseCampaignFeedsPerCustomer

Number of ENABLED campaign feeds across all base campaigns for a customer.

ExperimentCampaignFeedsPerCustomer

Number of ENABLED campaign feeds across all experiment campaigns for a customer.

AdGroupFeedsPerCustomer

Number of ENABLED ad group feeds per customer.

BaseAdGroupFeedsPerCustomer

Number of ENABLED ad group feeds across all base campaigns for a customer.

ExperimentAdGroupFeedsPerCustomer

Number of ENABLED ad group feeds across all experiment campaigns for a customer.

AdGroupFeedsPerCampaign

Number of ENABLED ad group feeds per campaign.

FeedItemSetsPerCustomer

Number of ENABLED feed items per customer.

FeedItemsPerFeedItemSet

Number of feed items per feed item set.

CampaignExperimentsPerCustomer

Number of ENABLED campaign experiments per customer.

ExperimentArmsPerVideoExperiment

Number of video experiment arms per experiment.

OwnedLabelsPerCustomer

Number of owned labels per customer.

LabelsPerCampaign

Number of applied labels per campaign.

LabelsPerAdGroup

Number of applied labels per ad group.

LabelsPerAdGroupAd

Number of applied labels per ad group ad.

LabelsPerAdGroupCriterion

Number of applied labels per ad group criterion.

TargetCustomersPerLabel

Number of customers with a single label applied.

KeywordPlansPerUserPerCustomer

Number of ENABLED keyword plans per user per customer. The limit is applied per <user, customer> pair because by default a plan is private to a user of a customer. Each user of a customer has his or her own independent limit.

KeywordPlanAdGroupKeywordsPerKeywordPlan

Number of keyword plan ad group keywords per keyword plan.

KeywordPlanAdGroupsPerKeywordPlan

Number of keyword plan ad groups per keyword plan.

KeywordPlanNegativeKeywordsPerKeywordPlan

Number of keyword plan negative keywords (both campaign and ad group) per keyword plan.

KeywordPlanCampaignsPerKeywordPlan

Number of keyword plan campaigns per keyword plan.

ConversionActionsPerCustomer

Number of ENABLED conversion actions per customer.

BatchJobOperationsPerJob

Number of operations in a single batch job.

BatchJobsPerCustomer

Number of PENDING or ENABLED batch jobs per customer.

HotelCheckInDateRangeBidModifiersPerAdGroup

Number of hotel check-in date range bid modifiers per ad agroup.

Implementations

impl ResourceLimitType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ResourceLimitType.

pub fn from_i32(value: i32) -> Option<ResourceLimitType>

Converts an i32 to a ResourceLimitType, or None if value is not a valid variant.