Module googapis::google::ads::googleads::v8::errors

Modules

access_invitation_error_enum
Nested message and enum types in AccessInvitationErrorEnum.

account_budget_proposal_error_enum
Nested message and enum types in AccountBudgetProposalErrorEnum.

account_link_error_enum
Nested message and enum types in AccountLinkErrorEnum.

ad_customizer_error_enum
Nested message and enum types in AdCustomizerErrorEnum.

ad_error_enum
Nested message and enum types in AdErrorEnum.

ad_group_ad_error_enum
Nested message and enum types in AdGroupAdErrorEnum.

ad_group_bid_modifier_error_enum
Nested message and enum types in AdGroupBidModifierErrorEnum.

ad_group_criterion_error_enum
Nested message and enum types in AdGroupCriterionErrorEnum.

ad_group_error_enum
Nested message and enum types in AdGroupErrorEnum.

ad_group_feed_error_enum
Nested message and enum types in AdGroupFeedErrorEnum.

ad_parameter_error_enum
Nested message and enum types in AdParameterErrorEnum.

ad_sharing_error_enum
Nested message and enum types in AdSharingErrorEnum.

adx_error_enum
Nested message and enum types in AdxErrorEnum.

asset_error_enum
Nested message and enum types in AssetErrorEnum.

asset_link_error_enum
Nested message and enum types in AssetLinkErrorEnum.

authentication_error_enum
Nested message and enum types in AuthenticationErrorEnum.

authorization_error_enum
Nested message and enum types in AuthorizationErrorEnum.

batch_job_error_enum
Nested message and enum types in BatchJobErrorEnum.

bidding_error_enum
Nested message and enum types in BiddingErrorEnum.

bidding_strategy_error_enum
Nested message and enum types in BiddingStrategyErrorEnum.

billing_setup_error_enum
Nested message and enum types in BillingSetupErrorEnum.

campaign_budget_error_enum
Nested message and enum types in CampaignBudgetErrorEnum.

campaign_criterion_error_enum
Nested message and enum types in CampaignCriterionErrorEnum.

campaign_draft_error_enum
Nested message and enum types in CampaignDraftErrorEnum.

campaign_error_enum
Nested message and enum types in CampaignErrorEnum.

campaign_experiment_error_enum
Nested message and enum types in CampaignExperimentErrorEnum.

campaign_feed_error_enum
Nested message and enum types in CampaignFeedErrorEnum.

campaign_shared_set_error_enum
Nested message and enum types in CampaignSharedSetErrorEnum.

change_event_error_enum
Nested message and enum types in ChangeEventErrorEnum.

change_status_error_enum
Nested message and enum types in ChangeStatusErrorEnum.

collection_size_error_enum
Nested message and enum types in CollectionSizeErrorEnum.

context_error_enum
Nested message and enum types in ContextErrorEnum.

conversion_action_error_enum
Nested message and enum types in ConversionActionErrorEnum.

conversion_adjustment_upload_error_enum
Nested message and enum types in ConversionAdjustmentUploadErrorEnum.

conversion_custom_variable_error_enum
Nested message and enum types in ConversionCustomVariableErrorEnum.

conversion_upload_error_enum
Nested message and enum types in ConversionUploadErrorEnum.

country_code_error_enum
Nested message and enum types in CountryCodeErrorEnum.

criterion_error_enum
Nested message and enum types in CriterionErrorEnum.

currency_code_error_enum
Nested message and enum types in CurrencyCodeErrorEnum.

custom_audience_error_enum
Nested message and enum types in CustomAudienceErrorEnum.

custom_interest_error_enum
Nested message and enum types in CustomInterestErrorEnum.

customer_client_link_error_enum
Nested message and enum types in CustomerClientLinkErrorEnum.

customer_error_enum
Nested message and enum types in CustomerErrorEnum.

customer_feed_error_enum
Nested message and enum types in CustomerFeedErrorEnum.

customer_manager_link_error_enum
Nested message and enum types in CustomerManagerLinkErrorEnum.

customer_user_access_error_enum
Nested message and enum types in CustomerUserAccessErrorEnum.

database_error_enum
Nested message and enum types in DatabaseErrorEnum.

date_error_enum
Nested message and enum types in DateErrorEnum.

date_range_error_enum
Nested message and enum types in DateRangeErrorEnum.

distinct_error_enum
Nested message and enum types in DistinctErrorEnum.

enum_error_enum
Nested message and enum types in EnumErrorEnum.

error_code
Nested message and enum types in ErrorCode.

error_location
Nested message and enum types in ErrorLocation.

extension_feed_item_error_enum
Nested message and enum types in ExtensionFeedItemErrorEnum.

extension_setting_error_enum
Nested message and enum types in ExtensionSettingErrorEnum.

feed_attribute_reference_error_enum
Nested message and enum types in FeedAttributeReferenceErrorEnum.

feed_error_enum
Nested message and enum types in FeedErrorEnum.

feed_item_error_enum
Nested message and enum types in FeedItemErrorEnum.

feed_item_set_error_enum
Nested message and enum types in FeedItemSetErrorEnum.

feed_item_set_link_error_enum
Nested message and enum types in FeedItemSetLinkErrorEnum.

feed_item_target_error_enum
Nested message and enum types in FeedItemTargetErrorEnum.

feed_item_validation_error_enum
Nested message and enum types in FeedItemValidationErrorEnum.

feed_mapping_error_enum
Nested message and enum types in FeedMappingErrorEnum.

field_error_enum
Nested message and enum types in FieldErrorEnum.

field_mask_error_enum
Nested message and enum types in FieldMaskErrorEnum.

function_error_enum
Nested message and enum types in FunctionErrorEnum.

function_parsing_error_enum
Nested message and enum types in FunctionParsingErrorEnum.

geo_target_constant_suggestion_error_enum
Nested message and enum types in GeoTargetConstantSuggestionErrorEnum.

header_error_enum
Nested message and enum types in HeaderErrorEnum.

id_error_enum
Nested message and enum types in IdErrorEnum.

image_error_enum
Nested message and enum types in ImageErrorEnum.

internal_error_enum
Nested message and enum types in InternalErrorEnum.

invoice_error_enum
Nested message and enum types in InvoiceErrorEnum.

keyword_plan_ad_group_error_enum
Nested message and enum types in KeywordPlanAdGroupErrorEnum.

keyword_plan_ad_group_keyword_error_enum
Nested message and enum types in KeywordPlanAdGroupKeywordErrorEnum.

keyword_plan_campaign_error_enum
Nested message and enum types in KeywordPlanCampaignErrorEnum.

keyword_plan_campaign_keyword_error_enum
Nested message and enum types in KeywordPlanCampaignKeywordErrorEnum.

keyword_plan_error_enum
Nested message and enum types in KeywordPlanErrorEnum.

keyword_plan_idea_error_enum
Nested message and enum types in KeywordPlanIdeaErrorEnum.

label_error_enum
Nested message and enum types in LabelErrorEnum.

language_code_error_enum
Nested message and enum types in LanguageCodeErrorEnum.

list_operation_error_enum
Nested message and enum types in ListOperationErrorEnum.

manager_link_error_enum
Nested message and enum types in ManagerLinkErrorEnum.

media_bundle_error_enum
Nested message and enum types in MediaBundleErrorEnum.

media_file_error_enum
Nested message and enum types in MediaFileErrorEnum.

media_upload_error_enum
Nested message and enum types in MediaUploadErrorEnum.

multiplier_error_enum
Nested message and enum types in MultiplierErrorEnum.

mutate_error_enum
Nested message and enum types in MutateErrorEnum.

new_resource_creation_error_enum
Nested message and enum types in NewResourceCreationErrorEnum.

not_allowlisted_error_enum
Nested message and enum types in NotAllowlistedErrorEnum.

not_empty_error_enum
Nested message and enum types in NotEmptyErrorEnum.

null_error_enum
Nested message and enum types in NullErrorEnum.

offline_user_data_job_error_enum
Nested message and enum types in OfflineUserDataJobErrorEnum.

operation_access_denied_error_enum
Nested message and enum types in OperationAccessDeniedErrorEnum.

operator_error_enum
Nested message and enum types in OperatorErrorEnum.

partial_failure_error_enum
Nested message and enum types in PartialFailureErrorEnum.

payments_account_error_enum
Nested message and enum types in PaymentsAccountErrorEnum.

policy_finding_error_enum
Nested message and enum types in PolicyFindingErrorEnum.

policy_validation_parameter_error_enum
Nested message and enum types in PolicyValidationParameterErrorEnum.

policy_violation_error_enum
Nested message and enum types in PolicyViolationErrorEnum.

query_error_enum
Nested message and enum types in QueryErrorEnum.

quota_error_details
Nested message and enum types in QuotaErrorDetails.

quota_error_enum
Nested message and enum types in QuotaErrorEnum.

range_error_enum
Nested message and enum types in RangeErrorEnum.

reach_plan_error_enum
Nested message and enum types in ReachPlanErrorEnum.

recommendation_error_enum
Nested message and enum types in RecommendationErrorEnum.

region_code_error_enum
Nested message and enum types in RegionCodeErrorEnum.

request_error_enum
Nested message and enum types in RequestErrorEnum.

resource_access_denied_error_enum
Nested message and enum types in ResourceAccessDeniedErrorEnum.

resource_count_limit_exceeded_error_enum
Nested message and enum types in ResourceCountLimitExceededErrorEnum.

setting_error_enum
Nested message and enum types in SettingErrorEnum.

shared_criterion_error_enum
Nested message and enum types in SharedCriterionErrorEnum.

shared_set_error_enum
Nested message and enum types in SharedSetErrorEnum.

size_limit_error_enum
Nested message and enum types in SizeLimitErrorEnum.

string_format_error_enum
Nested message and enum types in StringFormatErrorEnum.

string_length_error_enum
Nested message and enum types in StringLengthErrorEnum.

third_party_app_analytics_link_error_enum
Nested message and enum types in ThirdPartyAppAnalyticsLinkErrorEnum.

time_zone_error_enum
Nested message and enum types in TimeZoneErrorEnum.

url_field_error_enum
Nested message and enum types in UrlFieldErrorEnum.

user_data_error_enum
Nested message and enum types in UserDataErrorEnum.

user_list_error_enum
Nested message and enum types in UserListErrorEnum.

youtube_video_registration_error_enum
Nested message and enum types in YoutubeVideoRegistrationErrorEnum.

Structs

AccessInvitationErrorEnum
Container for enum describing possible AccessInvitation errors.

AccountBudgetProposalErrorEnum
Container for enum describing possible account budget proposal errors.

AccountLinkErrorEnum
Container for enum describing possible account link errors.

AdCustomizerErrorEnum
Container for enum describing possible ad customizer errors.

AdErrorEnum
Container for enum describing possible ad errors.

AdGroupAdErrorEnum
Container for enum describing possible ad group ad errors.

AdGroupBidModifierErrorEnum
Container for enum describing possible ad group bid modifier errors.

AdGroupCriterionErrorEnum
Container for enum describing possible ad group criterion errors.

AdGroupErrorEnum
Container for enum describing possible ad group errors.

AdGroupFeedErrorEnum
Container for enum describing possible ad group feed errors.

AdParameterErrorEnum
Container for enum describing possible ad parameter errors.

AdSharingErrorEnum
Container for enum describing possible ad sharing errors.

AdxErrorEnum
Container for enum describing possible adx errors.

AssetErrorEnum
Container for enum describing possible asset errors.

AssetLinkErrorEnum
Container for enum describing possible asset link errors.

AuthenticationErrorEnum
Container for enum describing possible authentication errors.

AuthorizationErrorEnum
Container for enum describing possible authorization errors.

BatchJobErrorEnum
Container for enum describing possible batch job errors.

BiddingErrorEnum
Container for enum describing possible bidding errors.

BiddingStrategyErrorEnum
Container for enum describing possible bidding strategy errors.

BillingSetupErrorEnum
Container for enum describing possible billing setup errors.

CampaignBudgetErrorEnum
Container for enum describing possible campaign budget errors.

CampaignCriterionErrorEnum
Container for enum describing possible campaign criterion errors.

CampaignDraftErrorEnum
Container for enum describing possible campaign draft errors.

CampaignErrorEnum
Container for enum describing possible campaign errors.

CampaignExperimentErrorEnum
Container for enum describing possible campaign experiment errors.

CampaignFeedErrorEnum
Container for enum describing possible campaign feed errors.

CampaignSharedSetErrorEnum
Container for enum describing possible campaign shared set errors.

ChangeEventErrorEnum
Container for enum describing possible change event errors.

ChangeStatusErrorEnum
Container for enum describing possible change status errors.

CollectionSizeErrorEnum
Container for enum describing possible collection size errors.

ContextErrorEnum
Container for enum describing possible context errors.

ConversionActionErrorEnum
Container for enum describing possible conversion action errors.

ConversionAdjustmentUploadErrorEnum
Container for enum describing possible conversion adjustment upload errors.

ConversionCustomVariableErrorEnum
Container for enum describing possible conversion custom variable errors.

ConversionUploadErrorEnum
Container for enum describing possible conversion upload errors.

CountryCodeErrorEnum
Container for enum describing country code errors.

CriterionErrorEnum
Container for enum describing possible criterion errors.

CurrencyCodeErrorEnum
Container for enum describing possible currency code errors.

CustomAudienceErrorEnum
Container for enum describing possible custom audience errors.

CustomInterestErrorEnum
Container for enum describing possible custom interest errors.

CustomerClientLinkErrorEnum
Container for enum describing possible CustomeClientLink errors.

CustomerErrorEnum
Container for enum describing possible customer errors.

CustomerFeedErrorEnum
Container for enum describing possible customer feed errors.

CustomerManagerLinkErrorEnum
Container for enum describing possible CustomerManagerLink errors.

CustomerUserAccessErrorEnum
Container for enum describing possible CustomerUserAccess errors.

DatabaseErrorEnum
Container for enum describing possible database errors.

DateErrorEnum
Container for enum describing possible date errors.

DateRangeErrorEnum
Container for enum describing possible date range errors.

DistinctErrorEnum
Container for enum describing possible distinct errors.

EnumErrorEnum
Container for enum describing possible enum errors.

ErrorCode
The error reason represented by type and enum.

ErrorDetails
Additional error details.

ErrorLocation
Describes the part of the request proto that caused the error.

ExtensionFeedItemErrorEnum
Container for enum describing possible extension feed item error.

ExtensionSettingErrorEnum
Container for enum describing validation errors of extension settings.

FeedAttributeReferenceErrorEnum
Container for enum describing possible feed attribute reference errors.

FeedErrorEnum
Container for enum describing possible feed errors.

FeedItemErrorEnum
Container for enum describing possible feed item errors.

FeedItemSetErrorEnum
Container for enum describing possible feed item set errors.

FeedItemSetLinkErrorEnum
Container for enum describing possible feed item set link errors.

FeedItemTargetErrorEnum
Container for enum describing possible feed item target errors.

FeedItemValidationErrorEnum
Container for enum describing possible validation errors of a feed item.

FeedMappingErrorEnum
Container for enum describing possible feed item errors.

FieldErrorEnum
Container for enum describing possible field errors.

FieldMaskErrorEnum
Container for enum describing possible field mask errors.

FunctionErrorEnum
Container for enum describing possible function errors.

FunctionParsingErrorEnum
Container for enum describing possible function parsing errors.

GeoTargetConstantSuggestionErrorEnum
Container for enum describing possible geo target constant suggestion errors.

GoogleAdsError
GoogleAds-specific error.

GoogleAdsFailure
Describes how a GoogleAds API call failed. It’s returned inside google.rpc.Status.details when a call fails.

HeaderErrorEnum
Container for enum describing possible header errors.

IdErrorEnum
Container for enum describing possible id errors.

ImageErrorEnum
Container for enum describing possible image errors.

InternalErrorEnum
Container for enum describing possible internal errors.

InvoiceErrorEnum
Container for enum describing possible invoice errors.

KeywordPlanAdGroupErrorEnum
Container for enum describing possible errors from applying a keyword plan ad group.

KeywordPlanAdGroupKeywordErrorEnum
Container for enum describing possible errors from applying an ad group keyword or a campaign keyword from a keyword plan.

KeywordPlanCampaignErrorEnum
Container for enum describing possible errors from applying a keyword plan campaign.

KeywordPlanCampaignKeywordErrorEnum
Container for enum describing possible errors from applying a keyword plan campaign keyword.

KeywordPlanErrorEnum
Container for enum describing possible errors from applying a keyword plan resource (keyword plan, keyword plan campaign, keyword plan ad group or keyword plan keyword) or KeywordPlanService RPC.

KeywordPlanIdeaErrorEnum
Container for enum describing possible errors from KeywordPlanIdeaService.

LabelErrorEnum
Container for enum describing possible label errors.

LanguageCodeErrorEnum
Container for enum describing language code errors.

ListOperationErrorEnum
Container for enum describing possible list operation errors.

ManagerLinkErrorEnum
Container for enum describing possible ManagerLink errors.

MediaBundleErrorEnum
Container for enum describing possible media bundle errors.

MediaFileErrorEnum
Container for enum describing possible media file errors.

MediaUploadErrorEnum
Container for enum describing possible media uploading errors.

MultiplierErrorEnum
Container for enum describing possible multiplier errors.

MutateErrorEnum
Container for enum describing possible mutate errors.

NewResourceCreationErrorEnum
Container for enum describing possible new resource creation errors.

NotAllowlistedErrorEnum
Container for enum describing possible not allowlisted errors.

NotEmptyErrorEnum
Container for enum describing possible not empty errors.

NullErrorEnum
Container for enum describing possible null errors.

OfflineUserDataJobErrorEnum
Container for enum describing possible offline user data job errors.

OperationAccessDeniedErrorEnum
Container for enum describing possible operation access denied errors.

OperatorErrorEnum
Container for enum describing possible operator errors.

PartialFailureErrorEnum
Container for enum describing possible partial failure errors.

PaymentsAccountErrorEnum
Container for enum describing possible errors in payments account service.

PolicyFindingDetails
Error returned as part of a mutate response. This error indicates one or more policy findings in the fields of a resource.

PolicyFindingErrorEnum
Container for enum describing possible policy finding errors.

PolicyValidationParameterErrorEnum
Container for enum describing possible policy validation parameter errors.

PolicyViolationDetails
Error returned as part of a mutate response. This error indicates single policy violation by some text in one of the fields.

PolicyViolationErrorEnum
Container for enum describing possible policy violation errors.

QueryErrorEnum
Container for enum describing possible query errors.

QuotaErrorDetails
Additional quota error details when there is QuotaError.

QuotaErrorEnum
Container for enum describing possible quota errors.

RangeErrorEnum
Container for enum describing possible range errors.

ReachPlanErrorEnum
Container for enum describing possible errors returned from the ReachPlanService.

RecommendationErrorEnum
Container for enum describing possible errors from applying a recommendation.

RegionCodeErrorEnum
Container for enum describing possible region code errors.

RequestErrorEnum
Container for enum describing possible request errors.

ResourceAccessDeniedErrorEnum
Container for enum describing possible resource access denied errors.

ResourceCountDetails
Error details returned when an resource count limit was exceeded.

ResourceCountLimitExceededErrorEnum
Container for enum describing possible resource count limit exceeded errors.

SettingErrorEnum
Container for enum describing possible setting errors.

SharedCriterionErrorEnum
Container for enum describing possible shared criterion errors.

SharedSetErrorEnum
Container for enum describing possible shared set errors.

SizeLimitErrorEnum
Container for enum describing possible size limit errors.

StringFormatErrorEnum
Container for enum describing possible string format errors.

StringLengthErrorEnum
Container for enum describing possible string length errors.

ThirdPartyAppAnalyticsLinkErrorEnum
Container for enum describing possible third party app analytics link errors.

TimeZoneErrorEnum
Container for enum describing possible time zone errors.

UrlFieldErrorEnum
Container for enum describing possible url field errors.

UserDataErrorEnum
Container for enum describing possible user data errors.

UserListErrorEnum
Container for enum describing possible user list errors.

YoutubeVideoRegistrationErrorEnum
Container for enum describing YouTube video registration errors.