Enum googapis::google::ads::googleads::v8::errors::campaign_error_enum::CampaignError

#[repr(i32)]
pub enum CampaignError {

    Unspecified,
    Unknown,
    CannotTargetContentNetwork,
    CannotTargetSearchNetwork,
    CannotTargetSearchNetworkWithoutGoogleSearch,
    CannotTargetGoogleSearchForCpmCampaign,
    CampaignMustTargetAtLeastOneNetwork,
    CannotTargetPartnerSearchNetwork,
    CannotTargetContentNetworkOnlyWithCriteriaLevelBiddingStrategy,
    CampaignDurationMustContainAllRunnableTrials,
    CannotModifyForTrialCampaign,
    DuplicateCampaignName,
    IncompatibleCampaignField,
    InvalidCampaignName,
    InvalidAdServingOptimizationStatus,
    InvalidTrackingUrl,
    CannotSetBothTrackingUrlTemplateAndTrackingSetting,
    MaxImpressionsNotInRange,
    TimeUnitNotSupported,
    InvalidOperationIfServingStatusHasEnded,
    BudgetCannotBeShared,
    CampaignCannotUseSharedBudget,
    CannotChangeBudgetOnCampaignWithTrials,
    CampaignLabelDoesNotExist,
    CampaignLabelAlreadyExists,
    MissingShoppingSetting,
    InvalidShoppingSalesCountry,
    AdvertisingChannelTypeNotAvailableForAccountType,
    InvalidAdvertisingChannelSubType,
    AtLeastOneConversionMustBeSelected,
    CannotSetAdRotationMode,
    CannotModifyStartDateIfAlreadyStarted,
    CannotSetDateToPast,
    MissingHotelCustomerLink,
    InvalidHotelCustomerLink,
    MissingHotelSetting,
    CannotUseSharedCampaignBudgetWhilePartOfCampaignGroup,
    AppNotFound,
    ShoppingEnableLocalNotSupportedForCampaignType,
    MerchantNotAllowedForComparisonListingAds,
    InsufficientAppInstallsCount,
    SensitiveCategoryApp,
    HecAgreementRequired,
    NotCompatibleWithViewThroughConversionOptimization,
    InvalidExcludedParentAssetFieldType,

}

Enum describing possible campaign errors.

Variants

Unspecified

Enum unspecified.

Unknown

The received error code is not known in this version.

CannotTargetContentNetwork

Cannot target content network.

CannotTargetSearchNetwork

Cannot target search network.

CannotTargetSearchNetworkWithoutGoogleSearch

Cannot cover search network without google search network.

CannotTargetGoogleSearchForCpmCampaign

Cannot target Google Search network for a CPM campaign.

CampaignMustTargetAtLeastOneNetwork

Must target at least one network.

CannotTargetPartnerSearchNetwork

Only some Google partners are allowed to target partner search network.

CannotTargetContentNetworkOnlyWithCriteriaLevelBiddingStrategy

Cannot target content network only as campaign has criteria-level bidding strategy.

CampaignDurationMustContainAllRunnableTrials

Cannot modify the start or end date such that the campaign duration would not contain the durations of all runnable trials.

CannotModifyForTrialCampaign

Cannot modify dates, budget or status of a trial campaign.

DuplicateCampaignName

Trying to modify the name of an active or paused campaign, where the name is already assigned to another active or paused campaign.

IncompatibleCampaignField

Two fields are in conflicting modes.

InvalidCampaignName

Campaign name cannot be used.

InvalidAdServingOptimizationStatus

Given status is invalid.

InvalidTrackingUrl

Error in the campaign level tracking URL.

CannotSetBothTrackingUrlTemplateAndTrackingSetting

Cannot set both tracking URL template and tracking setting. A user has to clear legacy tracking setting in order to add tracking URL template.

MaxImpressionsNotInRange

The maximum number of impressions for Frequency Cap should be an integer greater than 0.

TimeUnitNotSupported

Only the Day, Week and Month time units are supported.

InvalidOperationIfServingStatusHasEnded

Operation not allowed on a campaign whose serving status has ended

BudgetCannotBeShared

This budget is exclusively linked to a Campaign that is using experiments so it cannot be shared.

CampaignCannotUseSharedBudget

Campaigns using experiments cannot use a shared budget.

CannotChangeBudgetOnCampaignWithTrials

A different budget cannot be assigned to a campaign when there are running or scheduled trials.

CampaignLabelDoesNotExist

No link found between the campaign and the label.

CampaignLabelAlreadyExists

The label has already been attached to the campaign.

MissingShoppingSetting

A ShoppingSetting was not found when creating a shopping campaign.

InvalidShoppingSalesCountry

The country in shopping setting is not an allowed country.

AdvertisingChannelTypeNotAvailableForAccountType

The requested channel type is not available according to the customer’s account setting.

InvalidAdvertisingChannelSubType

The AdvertisingChannelSubType is not a valid subtype of the primary channel type.

AtLeastOneConversionMustBeSelected

At least one conversion must be selected.

CannotSetAdRotationMode

Setting ad rotation mode for a campaign is not allowed. Ad rotation mode at campaign is deprecated.

CannotModifyStartDateIfAlreadyStarted

Trying to change start date on a campaign that has started.

CannotSetDateToPast

Trying to modify a date into the past.

MissingHotelCustomerLink

Hotel center id in the hotel setting does not match any customer links.

InvalidHotelCustomerLink

Hotel center id in the hotel setting must match an active customer link.

MissingHotelSetting

Hotel setting was not found when creating a hotel ads campaign.

CannotUseSharedCampaignBudgetWhilePartOfCampaignGroup

A Campaign cannot use shared campaign budgets and be part of a campaign group.

AppNotFound

The app ID was not found.

ShoppingEnableLocalNotSupportedForCampaignType

Campaign.shopping_setting.enable_local is not supported for the specified campaign type.

MerchantNotAllowedForComparisonListingAds

The merchant does not support the creation of campaigns for Shopping Comparison Listing Ads.

InsufficientAppInstallsCount

The App campaign for engagement cannot be created because there aren’t enough installs.

SensitiveCategoryApp

The App campaign for engagement cannot be created because the app is sensitive.

HecAgreementRequired

Customers with Housing, Employment, or Credit ads must accept updated personalized ads policy to continue creating campaigns.

NotCompatibleWithViewThroughConversionOptimization

The field is not compatible with view through conversion optimization.

InvalidExcludedParentAssetFieldType

The field type cannot be excluded because an active campaign-asset link of this type exists.

Implementations

impl CampaignError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of CampaignError.

pub fn from_i32(value: i32) -> Option<CampaignError>

Converts an i32 to a CampaignError, or None if value is not a valid variant.