Module googapis::google::ads::googleads::v8::resources::feed

Nested message and enum types in Feed.

Modules

places_location_feed_data
Nested message and enum types in PlacesLocationFeedData.

Structs

AffiliateLocationFeedData
Data used to configure an affiliate location feed populated with the specified chains.

PlacesLocationFeedData
Data used to configure a location feed populated from Google My Business Locations.

Enums

SystemFeedGenerationData
The system data for the Feed. This data specifies information for generating the feed items of the system generated feed.