Struct googapis::google::ads::googleads::v8::services::GoogleAdsRow

pub struct GoogleAdsRow {

    pub account_budget: Option<AccountBudget>,
    pub account_budget_proposal: Option<AccountBudgetProposal>,
    pub account_link: Option<AccountLink>,
    pub ad_group: Option<AdGroup>,
    pub ad_group_ad: Option<AdGroupAd>,
    pub ad_group_ad_asset_view: Option<AdGroupAdAssetView>,
    pub ad_group_ad_label: Option<AdGroupAdLabel>,
    pub ad_group_asset: Option<AdGroupAsset>,
    pub ad_group_audience_view: Option<AdGroupAudienceView>,
    pub ad_group_bid_modifier: Option<AdGroupBidModifier>,
    pub ad_group_criterion: Option<AdGroupCriterion>,
    pub ad_group_criterion_label: Option<AdGroupCriterionLabel>,
    pub ad_group_criterion_simulation: Option<AdGroupCriterionSimulation>,
    pub ad_group_extension_setting: Option<AdGroupExtensionSetting>,
    pub ad_group_feed: Option<AdGroupFeed>,
    pub ad_group_label: Option<AdGroupLabel>,
    pub ad_group_simulation: Option<AdGroupSimulation>,
    pub ad_parameter: Option<AdParameter>,
    pub age_range_view: Option<AgeRangeView>,
    pub ad_schedule_view: Option<AdScheduleView>,
    pub domain_category: Option<DomainCategory>,
    pub asset: Option<Asset>,
    pub asset_field_type_view: Option<AssetFieldTypeView>,
    pub batch_job: Option<BatchJob>,
    pub bidding_strategy: Option<BiddingStrategy>,
    pub bidding_strategy_simulation: Option<BiddingStrategySimulation>,
    pub billing_setup: Option<BillingSetup>,
    pub call_view: Option<CallView>,
    pub campaign_budget: Option<CampaignBudget>,
    pub campaign: Option<Campaign>,
    pub campaign_asset: Option<CampaignAsset>,
    pub campaign_audience_view: Option<CampaignAudienceView>,
    pub campaign_bid_modifier: Option<CampaignBidModifier>,
    pub campaign_criterion: Option<CampaignCriterion>,
    pub campaign_criterion_simulation: Option<CampaignCriterionSimulation>,
    pub campaign_draft: Option<CampaignDraft>,
    pub campaign_experiment: Option<CampaignExperiment>,
    pub campaign_extension_setting: Option<CampaignExtensionSetting>,
    pub campaign_feed: Option<CampaignFeed>,
    pub campaign_label: Option<CampaignLabel>,
    pub campaign_shared_set: Option<CampaignSharedSet>,
    pub campaign_simulation: Option<CampaignSimulation>,
    pub carrier_constant: Option<CarrierConstant>,
    pub change_event: Option<ChangeEvent>,
    pub change_status: Option<ChangeStatus>,
    pub combined_audience: Option<CombinedAudience>,
    pub conversion_action: Option<ConversionAction>,
    pub conversion_custom_variable: Option<ConversionCustomVariable>,
    pub click_view: Option<ClickView>,
    pub currency_constant: Option<CurrencyConstant>,
    pub custom_audience: Option<CustomAudience>,
    pub custom_interest: Option<CustomInterest>,
    pub customer: Option<Customer>,
    pub customer_asset: Option<CustomerAsset>,
    pub accessible_bidding_strategy: Option<AccessibleBiddingStrategy>,
    pub customer_manager_link: Option<CustomerManagerLink>,
    pub customer_client_link: Option<CustomerClientLink>,
    pub customer_client: Option<CustomerClient>,
    pub customer_extension_setting: Option<CustomerExtensionSetting>,
    pub customer_feed: Option<CustomerFeed>,
    pub customer_label: Option<CustomerLabel>,
    pub customer_negative_criterion: Option<CustomerNegativeCriterion>,
    pub customer_user_access: Option<CustomerUserAccess>,
    pub customer_user_access_invitation: Option<CustomerUserAccessInvitation>,
    pub detail_placement_view: Option<DetailPlacementView>,
    pub detailed_demographic: Option<DetailedDemographic>,
    pub display_keyword_view: Option<DisplayKeywordView>,
    pub distance_view: Option<DistanceView>,
    pub dynamic_search_ads_search_term_view: Option<DynamicSearchAdsSearchTermView>,
    pub expanded_landing_page_view: Option<ExpandedLandingPageView>,
    pub extension_feed_item: Option<ExtensionFeedItem>,
    pub feed: Option<Feed>,
    pub feed_item: Option<FeedItem>,
    pub feed_item_set: Option<FeedItemSet>,
    pub feed_item_set_link: Option<FeedItemSetLink>,
    pub feed_item_target: Option<FeedItemTarget>,
    pub feed_mapping: Option<FeedMapping>,
    pub feed_placeholder_view: Option<FeedPlaceholderView>,
    pub gender_view: Option<GenderView>,
    pub geo_target_constant: Option<GeoTargetConstant>,
    pub geographic_view: Option<GeographicView>,
    pub group_placement_view: Option<GroupPlacementView>,
    pub hotel_group_view: Option<HotelGroupView>,
    pub hotel_performance_view: Option<HotelPerformanceView>,
    pub income_range_view: Option<IncomeRangeView>,
    pub keyword_view: Option<KeywordView>,
    pub keyword_plan: Option<KeywordPlan>,
    pub keyword_plan_campaign: Option<KeywordPlanCampaign>,
    pub keyword_plan_campaign_keyword: Option<KeywordPlanCampaignKeyword>,
    pub keyword_plan_ad_group: Option<KeywordPlanAdGroup>,
    pub keyword_plan_ad_group_keyword: Option<KeywordPlanAdGroupKeyword>,
    pub keyword_theme_constant: Option<KeywordThemeConstant>,
    pub label: Option<Label>,
    pub landing_page_view: Option<LandingPageView>,
    pub language_constant: Option<LanguageConstant>,
    pub location_view: Option<LocationView>,
    pub managed_placement_view: Option<ManagedPlacementView>,
    pub media_file: Option<MediaFile>,
    pub mobile_app_category_constant: Option<MobileAppCategoryConstant>,
    pub mobile_device_constant: Option<MobileDeviceConstant>,
    pub offline_user_data_job: Option<OfflineUserDataJob>,
    pub operating_system_version_constant: Option<OperatingSystemVersionConstant>,
    pub paid_organic_search_term_view: Option<PaidOrganicSearchTermView>,
    pub parental_status_view: Option<ParentalStatusView>,
    pub product_bidding_category_constant: Option<ProductBiddingCategoryConstant>,
    pub product_group_view: Option<ProductGroupView>,
    pub recommendation: Option<Recommendation>,
    pub search_term_view: Option<SearchTermView>,
    pub shared_criterion: Option<SharedCriterion>,
    pub shared_set: Option<SharedSet>,
    pub smart_campaign_setting: Option<SmartCampaignSetting>,
    pub shopping_performance_view: Option<ShoppingPerformanceView>,
    pub smart_campaign_search_term_view: Option<SmartCampaignSearchTermView>,
    pub third_party_app_analytics_link: Option<ThirdPartyAppAnalyticsLink>,
    pub topic_view: Option<TopicView>,
    pub user_interest: Option<UserInterest>,
    pub life_event: Option<LifeEvent>,
    pub user_list: Option<UserList>,
    pub user_location_view: Option<UserLocationView>,
    pub remarketing_action: Option<RemarketingAction>,
    pub topic_constant: Option<TopicConstant>,
    pub video: Option<Video>,
    pub webpage_view: Option<WebpageView>,
    pub metrics: Option<Metrics>,
    pub segments: Option<Segments>,

}

A returned row from the query.

Fields
account_budget: Option<AccountBudget>

The account budget in the query.

account_budget_proposal: Option<AccountBudgetProposal>

The account budget proposal referenced in the query.

account_link: Option<AccountLink>

The AccountLink referenced in the query.

ad_group: Option<AdGroup>

The ad group referenced in the query.

ad_group_ad: Option<AdGroupAd>

The ad referenced in the query.

ad_group_ad_asset_view: Option<AdGroupAdAssetView>

The ad group ad asset view in the query.

ad_group_ad_label: Option<AdGroupAdLabel>

The ad group ad label referenced in the query.

ad_group_asset: Option<AdGroupAsset>

The ad group asset referenced in the query.

ad_group_audience_view: Option<AdGroupAudienceView>

The ad group audience view referenced in the query.

ad_group_bid_modifier: Option<AdGroupBidModifier>

The bid modifier referenced in the query.

ad_group_criterion: Option<AdGroupCriterion>

The criterion referenced in the query.

ad_group_criterion_label: Option<AdGroupCriterionLabel>

The ad group criterion label referenced in the query.

ad_group_criterion_simulation: Option<AdGroupCriterionSimulation>

The ad group criterion simulation referenced in the query.

ad_group_extension_setting: Option<AdGroupExtensionSetting>

The ad group extension setting referenced in the query.

ad_group_feed: Option<AdGroupFeed>

The ad group feed referenced in the query.

ad_group_label: Option<AdGroupLabel>

The ad group label referenced in the query.

ad_group_simulation: Option<AdGroupSimulation>

The ad group simulation referenced in the query.

ad_parameter: Option<AdParameter>

The ad parameter referenced in the query.

age_range_view: Option<AgeRangeView>

The age range view referenced in the query.

ad_schedule_view: Option<AdScheduleView>

The ad schedule view referenced in the query.

domain_category: Option<DomainCategory>

The domain category referenced in the query.

asset: Option<Asset>

The asset referenced in the query.

asset_field_type_view: Option<AssetFieldTypeView>

The asset field type view referenced in the query.

batch_job: Option<BatchJob>

The batch job referenced in the query.

bidding_strategy: Option<BiddingStrategy>

The bidding strategy referenced in the query.

bidding_strategy_simulation: Option<BiddingStrategySimulation>

The bidding strategy simulation referenced in the query.

billing_setup: Option<BillingSetup>

The billing setup referenced in the query.

call_view: Option<CallView>

The call view referenced in the query.

campaign_budget: Option<CampaignBudget>

The campaign budget referenced in the query.

campaign: Option<Campaign>

The campaign referenced in the query.

campaign_asset: Option<CampaignAsset>

The campaign asset referenced in the query.

campaign_audience_view: Option<CampaignAudienceView>

The campaign audience view referenced in the query.

campaign_bid_modifier: Option<CampaignBidModifier>

The campaign bid modifier referenced in the query.

campaign_criterion: Option<CampaignCriterion>

The campaign criterion referenced in the query.

campaign_criterion_simulation: Option<CampaignCriterionSimulation>

The campaign criterion simulation referenced in the query.

campaign_draft: Option<CampaignDraft>

The campaign draft referenced in the query.

campaign_experiment: Option<CampaignExperiment>

The campaign experiment referenced in the query.

campaign_extension_setting: Option<CampaignExtensionSetting>

The campaign extension setting referenced in the query.

campaign_feed: Option<CampaignFeed>

The campaign feed referenced in the query.

campaign_label: Option<CampaignLabel>

The campaign label referenced in the query.

campaign_shared_set: Option<CampaignSharedSet>

Campaign Shared Set referenced in AWQL query.

campaign_simulation: Option<CampaignSimulation>

The campaign simulation referenced in the query.

carrier_constant: Option<CarrierConstant>

The carrier constant referenced in the query.

change_event: Option<ChangeEvent>

The ChangeEvent referenced in the query.

change_status: Option<ChangeStatus>

The ChangeStatus referenced in the query.

combined_audience: Option<CombinedAudience>

The CombinedAudience referenced in the query.

conversion_action: Option<ConversionAction>

The conversion action referenced in the query.

conversion_custom_variable: Option<ConversionCustomVariable>

The conversion custom variable referenced in the query.

click_view: Option<ClickView>

The ClickView referenced in the query.

currency_constant: Option<CurrencyConstant>

The currency constant referenced in the query.

custom_audience: Option<CustomAudience>

The CustomAudience referenced in the query.

custom_interest: Option<CustomInterest>

The CustomInterest referenced in the query.

customer: Option<Customer>

The customer referenced in the query.

customer_asset: Option<CustomerAsset>

The customer asset referenced in the query.

accessible_bidding_strategy: Option<AccessibleBiddingStrategy>

The accessible bidding strategy referenced in the query.

customer_manager_link: Option<CustomerManagerLink>

The CustomerManagerLink referenced in the query.

customer_client_link: Option<CustomerClientLink>

The CustomerClientLink referenced in the query.

customer_client: Option<CustomerClient>

The CustomerClient referenced in the query.

customer_extension_setting: Option<CustomerExtensionSetting>

The customer extension setting referenced in the query.

customer_feed: Option<CustomerFeed>

The customer feed referenced in the query.

customer_label: Option<CustomerLabel>

The customer label referenced in the query.

customer_negative_criterion: Option<CustomerNegativeCriterion>

The customer negative criterion referenced in the query.

customer_user_access: Option<CustomerUserAccess>

The CustomerUserAccess referenced in the query.

customer_user_access_invitation: Option<CustomerUserAccessInvitation>

The CustomerUserAccessInvitation referenced in the query.

detail_placement_view: Option<DetailPlacementView>

The detail placement view referenced in the query.

detailed_demographic: Option<DetailedDemographic>

The detailed demographic referenced in the query.

display_keyword_view: Option<DisplayKeywordView>

The display keyword view referenced in the query.

distance_view: Option<DistanceView>

The distance view referenced in the query.

dynamic_search_ads_search_term_view: Option<DynamicSearchAdsSearchTermView>

The dynamic search ads search term view referenced in the query.

expanded_landing_page_view: Option<ExpandedLandingPageView>

The expanded landing page view referenced in the query.

extension_feed_item: Option<ExtensionFeedItem>

The extension feed item referenced in the query.

feed: Option<Feed>

The feed referenced in the query.

feed_item: Option<FeedItem>

The feed item referenced in the query.

feed_item_set: Option<FeedItemSet>

The feed item set referenced in the query.

feed_item_set_link: Option<FeedItemSetLink>

The feed item set link referenced in the query.

feed_item_target: Option<FeedItemTarget>

The feed item target referenced in the query.

feed_mapping: Option<FeedMapping>

The feed mapping referenced in the query.

feed_placeholder_view: Option<FeedPlaceholderView>

The feed placeholder view referenced in the query.

gender_view: Option<GenderView>

The gender view referenced in the query.

geo_target_constant: Option<GeoTargetConstant>

The geo target constant referenced in the query.

geographic_view: Option<GeographicView>

The geographic view referenced in the query.

group_placement_view: Option<GroupPlacementView>

The group placement view referenced in the query.

hotel_group_view: Option<HotelGroupView>

The hotel group view referenced in the query.

hotel_performance_view: Option<HotelPerformanceView>

The hotel performance view referenced in the query.

income_range_view: Option<IncomeRangeView>

The income range view referenced in the query.

keyword_view: Option<KeywordView>

The keyword view referenced in the query.

keyword_plan: Option<KeywordPlan>

The keyword plan referenced in the query.

keyword_plan_campaign: Option<KeywordPlanCampaign>

The keyword plan campaign referenced in the query.

keyword_plan_campaign_keyword: Option<KeywordPlanCampaignKeyword>

The keyword plan campaign keyword referenced in the query.

keyword_plan_ad_group: Option<KeywordPlanAdGroup>

The keyword plan ad group referenced in the query.

keyword_plan_ad_group_keyword: Option<KeywordPlanAdGroupKeyword>

The keyword plan ad group referenced in the query.

keyword_theme_constant: Option<KeywordThemeConstant>

The keyword theme constant referenced in the query.

label: Option<Label>

The label referenced in the query.

landing_page_view: Option<LandingPageView>

The landing page view referenced in the query.

language_constant: Option<LanguageConstant>

The language constant referenced in the query.

location_view: Option<LocationView>

The location view referenced in the query.

managed_placement_view: Option<ManagedPlacementView>

The managed placement view referenced in the query.

media_file: Option<MediaFile>

The media file referenced in the query.

mobile_app_category_constant: Option<MobileAppCategoryConstant>

The mobile app category constant referenced in the query.

mobile_device_constant: Option<MobileDeviceConstant>

The mobile device constant referenced in the query.

offline_user_data_job: Option<OfflineUserDataJob>

The offline user data job referenced in the query.

operating_system_version_constant: Option<OperatingSystemVersionConstant>

The operating system version constant referenced in the query.

paid_organic_search_term_view: Option<PaidOrganicSearchTermView>

The paid organic search term view referenced in the query.

parental_status_view: Option<ParentalStatusView>

The parental status view referenced in the query.

product_bidding_category_constant: Option<ProductBiddingCategoryConstant>

The Product Bidding Category referenced in the query.

product_group_view: Option<ProductGroupView>

The product group view referenced in the query.

recommendation: Option<Recommendation>

The recommendation referenced in the query.

search_term_view: Option<SearchTermView>

The search term view referenced in the query.

shared_criterion: Option<SharedCriterion>

The shared set referenced in the query.

shared_set: Option<SharedSet>

The shared set referenced in the query.

smart_campaign_setting: Option<SmartCampaignSetting>

The Smart campaign setting referenced in the query.

shopping_performance_view: Option<ShoppingPerformanceView>

The shopping performance view referenced in the query.

smart_campaign_search_term_view: Option<SmartCampaignSearchTermView>

The Smart campaign search term view referenced in the query.

third_party_app_analytics_link: Option<ThirdPartyAppAnalyticsLink>

The AccountLink referenced in the query.

topic_view: Option<TopicView>

The topic view referenced in the query.

user_interest: Option<UserInterest>

The user interest referenced in the query.

life_event: Option<LifeEvent>

The life event referenced in the query.

user_list: Option<UserList>

The user list referenced in the query.

user_location_view: Option<UserLocationView>

The user location view referenced in the query.

remarketing_action: Option<RemarketingAction>

The remarketing action referenced in the query.

topic_constant: Option<TopicConstant>

The topic constant referenced in the query.

video: Option<Video>

The video referenced in the query.

webpage_view: Option<WebpageView>

The webpage view referenced in the query.

metrics: Option<Metrics>

The metrics.

segments: Option<Segments>

The segments.