Module googapis::google::ads::googleads::v9::common::policy_topic_evidence

Nested message and enum types in PolicyTopicEvidence.

Modules

destination_not_working

Nested message and enum types in DestinationNotWorking.

Structs

DestinationMismatch

Evidence of mismatches between the URLs of a resource.

DestinationNotWorking

Evidence details when the destination is returning an HTTP error code or isn’t functional in all locations for commonly used devices.

DestinationTextList

A list of strings found in a destination page that caused a policy finding.

TextList

A list of fragments of text that violated a policy.

WebsiteList

A list of websites that caused a policy finding. Used for ONE_WEBSITE_PER_AD_GROUP policy topic, for example. In case there are more than five websites, only the top five (those that appear in resources the most) will be listed here.

Enums

Value

Specific evidence information depending on the evidence type.