Struct googapis::google::ads::googleads::v9::common::Segments

pub struct Segments {

    pub ad_destination_type: i32,
    pub ad_network_type: i32,
    pub budget_campaign_association_status: Option<BudgetCampaignAssociationStatus>,
    pub click_type: i32,
    pub conversion_action: Option<String>,
    pub conversion_action_category: i32,
    pub conversion_action_name: Option<String>,
    pub conversion_adjustment: Option<bool>,
    pub conversion_attribution_event_type: i32,
    pub conversion_lag_bucket: i32,
    pub conversion_or_adjustment_lag_bucket: i32,
    pub date: Option<String>,
    pub day_of_week: i32,
    pub device: i32,
    pub external_conversion_source: i32,
    pub geo_target_airport: Option<String>,
    pub geo_target_canton: Option<String>,
    pub geo_target_city: Option<String>,
    pub geo_target_country: Option<String>,
    pub geo_target_county: Option<String>,
    pub geo_target_district: Option<String>,
    pub geo_target_metro: Option<String>,
    pub geo_target_most_specific_location: Option<String>,
    pub geo_target_postal_code: Option<String>,
    pub geo_target_province: Option<String>,
    pub geo_target_region: Option<String>,
    pub geo_target_state: Option<String>,
    pub hotel_booking_window_days: Option<i64>,
    pub hotel_center_id: Option<i64>,
    pub hotel_check_in_date: Option<String>,
    pub hotel_check_in_day_of_week: i32,
    pub hotel_city: Option<String>,
    pub hotel_class: Option<i32>,
    pub hotel_country: Option<String>,
    pub hotel_date_selection_type: i32,
    pub hotel_length_of_stay: Option<i32>,
    pub hotel_rate_rule_id: Option<String>,
    pub hotel_rate_type: i32,
    pub hotel_price_bucket: i32,
    pub hotel_state: Option<String>,
    pub hour: Option<i32>,
    pub interaction_on_this_extension: Option<bool>,
    pub keyword: Option<Keyword>,
    pub month: Option<String>,
    pub month_of_year: i32,
    pub partner_hotel_id: Option<String>,
    pub placeholder_type: i32,
    pub product_aggregator_id: Option<i64>,
    pub product_bidding_category_level1: Option<String>,
    pub product_bidding_category_level2: Option<String>,
    pub product_bidding_category_level3: Option<String>,
    pub product_bidding_category_level4: Option<String>,
    pub product_bidding_category_level5: Option<String>,
    pub product_brand: Option<String>,
    pub product_channel: i32,
    pub product_channel_exclusivity: i32,
    pub product_condition: i32,
    pub product_country: Option<String>,
    pub product_custom_attribute0: Option<String>,
    pub product_custom_attribute1: Option<String>,
    pub product_custom_attribute2: Option<String>,
    pub product_custom_attribute3: Option<String>,
    pub product_custom_attribute4: Option<String>,
    pub product_item_id: Option<String>,
    pub product_language: Option<String>,
    pub product_merchant_id: Option<i64>,
    pub product_store_id: Option<String>,
    pub product_title: Option<String>,
    pub product_type_l1: Option<String>,
    pub product_type_l2: Option<String>,
    pub product_type_l3: Option<String>,
    pub product_type_l4: Option<String>,
    pub product_type_l5: Option<String>,
    pub quarter: Option<String>,
    pub recommendation_type: i32,
    pub search_engine_results_page_type: i32,
    pub search_term_match_type: i32,
    pub slot: i32,
    pub conversion_value_rule_primary_dimension: i32,
    pub webpage: Option<String>,
    pub week: Option<String>,
    pub year: Option<i32>,
    pub sk_ad_network_conversion_value: Option<i64>,
    pub asset_interaction_target: Option<AssetInteractionTarget>,

}

Segment only fields.

Fields
ad_destination_type: i32

Ad Destination type.

ad_network_type: i32

Ad network type.

budget_campaign_association_status: Option<BudgetCampaignAssociationStatus>

Budget campaign association status.

click_type: i32

Click type.

conversion_action: Option<String>

Resource name of the conversion action.

conversion_action_category: i32

Conversion action category.

conversion_action_name: Option<String>

Conversion action name.

conversion_adjustment: Option<bool>

This segments your conversion columns by the original conversion and conversion value vs. the delta if conversions were adjusted. False row has the data as originally stated; While true row has the delta between data now and the data as originally stated. Summing the two together results post-adjustment data.

conversion_attribution_event_type: i32

Conversion attribution event type.

conversion_lag_bucket: i32

An enum value representing the number of days between the impression and the conversion.

conversion_or_adjustment_lag_bucket: i32

An enum value representing the number of days between the impression and the conversion or between the impression and adjustments to the conversion.

date: Option<String>

Date to which metrics apply. yyyy-MM-dd format, e.g., 2018-04-17.

day_of_week: i32

Day of the week, e.g., MONDAY.

device: i32

Device to which metrics apply.

external_conversion_source: i32

External conversion source.

geo_target_airport: Option<String>

Resource name of the geo target constant that represents an airport.

geo_target_canton: Option<String>

Resource name of the geo target constant that represents a canton.

geo_target_city: Option<String>

Resource name of the geo target constant that represents a city.

geo_target_country: Option<String>

Resource name of the geo target constant that represents a country.

geo_target_county: Option<String>

Resource name of the geo target constant that represents a county.

geo_target_district: Option<String>

Resource name of the geo target constant that represents a district.

geo_target_metro: Option<String>

Resource name of the geo target constant that represents a metro.

geo_target_most_specific_location: Option<String>

Resource name of the geo target constant that represents the most specific location.

geo_target_postal_code: Option<String>

Resource name of the geo target constant that represents a postal code.

geo_target_province: Option<String>

Resource name of the geo target constant that represents a province.

geo_target_region: Option<String>

Resource name of the geo target constant that represents a region.

geo_target_state: Option<String>

Resource name of the geo target constant that represents a state.

hotel_booking_window_days: Option<i64>

Hotel booking window in days.

hotel_center_id: Option<i64>

Hotel center ID.

hotel_check_in_date: Option<String>

Hotel check-in date. Formatted as yyyy-MM-dd.

hotel_check_in_day_of_week: i32

Hotel check-in day of week.

hotel_city: Option<String>

Hotel city.

hotel_class: Option<i32>

Hotel class.

hotel_country: Option<String>

Hotel country.

hotel_date_selection_type: i32

Hotel date selection type.

hotel_length_of_stay: Option<i32>

Hotel length of stay.

hotel_rate_rule_id: Option<String>

Hotel rate rule ID.

hotel_rate_type: i32

Hotel rate type.

hotel_price_bucket: i32

Hotel price bucket.

hotel_state: Option<String>

Hotel state.

hour: Option<i32>

Hour of day as a number between 0 and 23, inclusive.

interaction_on_this_extension: Option<bool>

Only used with feed item metrics. Indicates whether the interaction metrics occurred on the feed item itself or a different extension or ad unit.

keyword: Option<Keyword>

Keyword criterion.

month: Option<String>

Month as represented by the date of the first day of a month. Formatted as yyyy-MM-dd.

month_of_year: i32

Month of the year, e.g., January.

partner_hotel_id: Option<String>

Partner hotel ID.

placeholder_type: i32

Placeholder type. This is only used with feed item metrics.

product_aggregator_id: Option<i64>

Aggregator ID of the product.

product_bidding_category_level1: Option<String>

Bidding category (level 1) of the product.

product_bidding_category_level2: Option<String>

Bidding category (level 2) of the product.

product_bidding_category_level3: Option<String>

Bidding category (level 3) of the product.

product_bidding_category_level4: Option<String>

Bidding category (level 4) of the product.

product_bidding_category_level5: Option<String>

Bidding category (level 5) of the product.

product_brand: Option<String>

Brand of the product.

product_channel: i32

Channel of the product.

product_channel_exclusivity: i32

Channel exclusivity of the product.

product_condition: i32

Condition of the product.

product_country: Option<String>

Resource name of the geo target constant for the country of sale of the product.

product_custom_attribute0: Option<String>

Custom attribute 0 of the product.

product_custom_attribute1: Option<String>

Custom attribute 1 of the product.

product_custom_attribute2: Option<String>

Custom attribute 2 of the product.

product_custom_attribute3: Option<String>

Custom attribute 3 of the product.

product_custom_attribute4: Option<String>

Custom attribute 4 of the product.

product_item_id: Option<String>

Item ID of the product.

product_language: Option<String>

Resource name of the language constant for the language of the product.

product_merchant_id: Option<i64>

Merchant ID of the product.

product_store_id: Option<String>

Store ID of the product.

product_title: Option<String>

Title of the product.

product_type_l1: Option<String>

Type (level 1) of the product.

product_type_l2: Option<String>

Type (level 2) of the product.

product_type_l3: Option<String>

Type (level 3) of the product.

product_type_l4: Option<String>

Type (level 4) of the product.

product_type_l5: Option<String>

Type (level 5) of the product.

quarter: Option<String>

Quarter as represented by the date of the first day of a quarter. Uses the calendar year for quarters, e.g., the second quarter of 2018 starts on 2018-04-01. Formatted as yyyy-MM-dd.

recommendation_type: i32

Recommendation type.

search_engine_results_page_type: i32

Type of the search engine results page.

search_term_match_type: i32

Match type of the keyword that triggered the ad, including variants.

slot: i32

Position of the ad.

conversion_value_rule_primary_dimension: i32

Primary dimension of applied conversion value rules. NO_RULE_APPLIED shows the total recorded value of conversions that do not have a value rule applied. ORIGINAL shows the original value of conversions to which a value rule has been applied. GEO_LOCATION, DEVICE, AUDIENCE show the net adjustment after value rules were applied.

webpage: Option<String>

Resource name of the ad group criterion that represents webpage criterion.

week: Option<String>

Week as defined as Monday through Sunday, and represented by the date of Monday. Formatted as yyyy-MM-dd.

year: Option<i32>

Year, formatted as yyyy.

sk_ad_network_conversion_value: Option<i64>

iOS Store Kit Ad Network conversion value. Null value means this segment is not applicable, e.g. non-iOS campaign.

asset_interaction_target: Option<AssetInteractionTarget>

Only used with CustomerAsset, CampaignAsset and AdGroupAsset metrics. Indicates whether the interaction metrics occurred on the asset itself or a different asset or ad unit. Interactions (e.g. clicks) are counted across all the parts of the served ad (e.g. Ad itself and other components like Sitelinks) when they are served together. When interaction_on_this_asset is true, it means the interactions are on this specific asset and when interaction_on_this_asset is false, it means the interactions is not on this specific asset but on other parts of the served ad this asset is served with.

Implementations

impl Segments

pub fn device(&self) -> Device

Returns the enum value of device, or the default if the field is set to an invalid enum value.

pub fn set_device(&mut self, value: Device)

Sets device to the provided enum value.

pub fn conversion_attribution_event_type(
&self
) -> ConversionAttributionEventType

Returns the enum value of conversion_attribution_event_type, or the default if the field is set to an invalid enum value.

pub fn set_conversion_attribution_event_type(
&mut self,
value: ConversionAttributionEventType
)

Sets conversion_attribution_event_type to the provided enum value.

pub fn ad_network_type(&self) -> AdNetworkType

Returns the enum value of ad_network_type, or the default if the field is set to an invalid enum value.

pub fn set_ad_network_type(&mut self, value: AdNetworkType)

Sets ad_network_type to the provided enum value.

pub fn day_of_week(&self) -> DayOfWeek

Returns the enum value of day_of_week, or the default if the field is set to an invalid enum value.

pub fn set_day_of_week(&mut self, value: DayOfWeek)

Sets day_of_week to the provided enum value.

pub fn hotel_check_in_day_of_week(&self) -> DayOfWeek

Returns the enum value of hotel_check_in_day_of_week, or the default if the field is set to an invalid enum value.

pub fn set_hotel_check_in_day_of_week(&mut self, value: DayOfWeek)

Sets hotel_check_in_day_of_week to the provided enum value.

pub fn hotel_date_selection_type(&self) -> HotelDateSelectionType

Returns the enum value of hotel_date_selection_type, or the default if the field is set to an invalid enum value.

pub fn set_hotel_date_selection_type(&mut self, value: HotelDateSelectionType)

Sets hotel_date_selection_type to the provided enum value.

pub fn month_of_year(&self) -> MonthOfYear

Returns the enum value of month_of_year, or the default if the field is set to an invalid enum value.

pub fn set_month_of_year(&mut self, value: MonthOfYear)

Sets month_of_year to the provided enum value.

pub fn placeholder_type(&self) -> PlaceholderType

Returns the enum value of placeholder_type, or the default if the field is set to an invalid enum value.

pub fn set_placeholder_type(&mut self, value: PlaceholderType)

Sets placeholder_type to the provided enum value.

pub fn search_term_match_type(&self) -> SearchTermMatchType

Returns the enum value of search_term_match_type, or the default if the field is set to an invalid enum value.

pub fn set_search_term_match_type(&mut self, value: SearchTermMatchType)

Sets search_term_match_type to the provided enum value.

pub fn slot(&self) -> Slot

Returns the enum value of slot, or the default if the field is set to an invalid enum value.

pub fn set_slot(&mut self, value: Slot)

Sets slot to the provided enum value.

pub fn click_type(&self) -> ClickType

Returns the enum value of click_type, or the default if the field is set to an invalid enum value.

pub fn set_click_type(&mut self, value: ClickType)

Sets click_type to the provided enum value.

pub fn product_channel(&self) -> ProductChannel

Returns the enum value of product_channel, or the default if the field is set to an invalid enum value.

pub fn set_product_channel(&mut self, value: ProductChannel)

Sets product_channel to the provided enum value.

pub fn product_channel_exclusivity(&self) -> ProductChannelExclusivity

Returns the enum value of product_channel_exclusivity, or the default if the field is set to an invalid enum value.

pub fn set_product_channel_exclusivity(
&mut self,
value: ProductChannelExclusivity
)

Sets product_channel_exclusivity to the provided enum value.

pub fn product_condition(&self) -> ProductCondition

Returns the enum value of product_condition, or the default if the field is set to an invalid enum value.

pub fn set_product_condition(&mut self, value: ProductCondition)

Sets product_condition to the provided enum value.

pub fn conversion_lag_bucket(&self) -> ConversionLagBucket

Returns the enum value of conversion_lag_bucket, or the default if the field is set to an invalid enum value.

pub fn set_conversion_lag_bucket(&mut self, value: ConversionLagBucket)

Sets conversion_lag_bucket to the provided enum value.

pub fn conversion_or_adjustment_lag_bucket(
&self
) -> ConversionOrAdjustmentLagBucket

Returns the enum value of conversion_or_adjustment_lag_bucket, or the default if the field is set to an invalid enum value.

pub fn set_conversion_or_adjustment_lag_bucket(
&mut self,
value: ConversionOrAdjustmentLagBucket
)

Sets conversion_or_adjustment_lag_bucket to the provided enum value.

pub fn conversion_action_category(&self) -> ConversionActionCategory

Returns the enum value of conversion_action_category, or the default if the field is set to an invalid enum value.

pub fn set_conversion_action_category(
&mut self,
value: ConversionActionCategory
)

Sets conversion_action_category to the provided enum value.

pub fn external_conversion_source(&self) -> ExternalConversionSource

Returns the enum value of external_conversion_source, or the default if the field is set to an invalid enum value.

pub fn set_external_conversion_source(
&mut self,
value: ExternalConversionSource
)

Sets external_conversion_source to the provided enum value.

pub fn search_engine_results_page_type(&self) -> SearchEngineResultsPageType

Returns the enum value of search_engine_results_page_type, or the default if the field is set to an invalid enum value.

pub fn set_search_engine_results_page_type(
&mut self,
value: SearchEngineResultsPageType
)

Sets search_engine_results_page_type to the provided enum value.

pub fn hotel_rate_type(&self) -> HotelRateType

Returns the enum value of hotel_rate_type, or the default if the field is set to an invalid enum value.

pub fn set_hotel_rate_type(&mut self, value: HotelRateType)

Sets hotel_rate_type to the provided enum value.

pub fn hotel_price_bucket(&self) -> HotelPriceBucket

Returns the enum value of hotel_price_bucket, or the default if the field is set to an invalid enum value.

pub fn set_hotel_price_bucket(&mut self, value: HotelPriceBucket)

Sets hotel_price_bucket to the provided enum value.

pub fn date(&self) -> &str

Returns the value of date, or the default value if date is unset.

pub fn hotel_center_id(&self) -> i64

Returns the value of hotel_center_id, or the default value if hotel_center_id is unset.

pub fn hotel_check_in_date(&self) -> &str

Returns the value of hotel_check_in_date, or the default value if hotel_check_in_date is unset.

pub fn hotel_city(&self) -> &str

Returns the value of hotel_city, or the default value if hotel_city is unset.

pub fn hotel_class(&self) -> i32

Returns the value of hotel_class, or the default value if hotel_class is unset.

pub fn hotel_country(&self) -> &str

Returns the value of hotel_country, or the default value if hotel_country is unset.

pub fn hotel_length_of_stay(&self) -> i32

Returns the value of hotel_length_of_stay, or the default value if hotel_length_of_stay is unset.

pub fn hotel_rate_rule_id(&self) -> &str

Returns the value of hotel_rate_rule_id, or the default value if hotel_rate_rule_id is unset.

pub fn hotel_state(&self) -> &str

Returns the value of hotel_state, or the default value if hotel_state is unset.

pub fn hour(&self) -> i32

Returns the value of hour, or the default value if hour is unset.

pub fn interaction_on_this_extension(&self) -> bool

Returns the value of interaction_on_this_extension, or the default value if interaction_on_this_extension is unset.

pub fn month(&self) -> &str

Returns the value of month, or the default value if month is unset.

pub fn partner_hotel_id(&self) -> &str

Returns the value of partner_hotel_id, or the default value if partner_hotel_id is unset.

pub fn product_bidding_category_level1(&self) -> &str

Returns the value of product_bidding_category_level1, or the default value if product_bidding_category_level1 is unset.

pub fn product_bidding_category_level2(&self) -> &str

Returns the value of product_bidding_category_level2, or the default value if product_bidding_category_level2 is unset.

pub fn product_bidding_category_level3(&self) -> &str

Returns the value of product_bidding_category_level3, or the default value if product_bidding_category_level3 is unset.

pub fn product_bidding_category_level4(&self) -> &str

Returns the value of product_bidding_category_level4, or the default value if product_bidding_category_level4 is unset.

pub fn product_bidding_category_level5(&self) -> &str

Returns the value of product_bidding_category_level5, or the default value if product_bidding_category_level5 is unset.

pub fn product_brand(&self) -> &str

Returns the value of product_brand, or the default value if product_brand is unset.

pub fn product_country(&self) -> &str

Returns the value of product_country, or the default value if product_country is unset.

pub fn product_custom_attribute0(&self) -> &str

Returns the value of product_custom_attribute0, or the default value if product_custom_attribute0 is unset.

pub fn product_custom_attribute1(&self) -> &str

Returns the value of product_custom_attribute1, or the default value if product_custom_attribute1 is unset.

pub fn product_custom_attribute2(&self) -> &str

Returns the value of product_custom_attribute2, or the default value if product_custom_attribute2 is unset.

pub fn product_custom_attribute3(&self) -> &str

Returns the value of product_custom_attribute3, or the default value if product_custom_attribute3 is unset.

pub fn product_custom_attribute4(&self) -> &str

Returns the value of product_custom_attribute4, or the default value if product_custom_attribute4 is unset.

pub fn product_item_id(&self) -> &str

Returns the value of product_item_id, or the default value if product_item_id is unset.

pub fn product_language(&self) -> &str

Returns the value of product_language, or the default value if product_language is unset.

pub fn product_store_id(&self) -> &str

Returns the value of product_store_id, or the default value if product_store_id is unset.

pub fn product_title(&self) -> &str

Returns the value of product_title, or the default value if product_title is unset.

pub fn product_type_l1(&self) -> &str

Returns the value of product_type_l1, or the default value if product_type_l1 is unset.

pub fn product_type_l2(&self) -> &str

Returns the value of product_type_l2, or the default value if product_type_l2 is unset.

pub fn product_type_l3(&self) -> &str

Returns the value of product_type_l3, or the default value if product_type_l3 is unset.

pub fn product_type_l4(&self) -> &str

Returns the value of product_type_l4, or the default value if product_type_l4 is unset.

pub fn product_type_l5(&self) -> &str

Returns the value of product_type_l5, or the default value if product_type_l5 is unset.

pub fn conversion_action(&self) -> &str

Returns the value of conversion_action, or the default value if conversion_action is unset.

pub fn conversion_action_name(&self) -> &str

Returns the value of conversion_action_name, or the default value if conversion_action_name is unset.

pub fn conversion_adjustment(&self) -> bool

Returns the value of conversion_adjustment, or the default value if conversion_adjustment is unset.

pub fn geo_target_airport(&self) -> &str

Returns the value of geo_target_airport, or the default value if geo_target_airport is unset.

pub fn geo_target_canton(&self) -> &str

Returns the value of geo_target_canton, or the default value if geo_target_canton is unset.

pub fn geo_target_city(&self) -> &str

Returns the value of geo_target_city, or the default value if geo_target_city is unset.

pub fn geo_target_country(&self) -> &str

Returns the value of geo_target_country, or the default value if geo_target_country is unset.

pub fn geo_target_county(&self) -> &str

Returns the value of geo_target_county, or the default value if geo_target_county is unset.

pub fn geo_target_district(&self) -> &str

Returns the value of geo_target_district, or the default value if geo_target_district is unset.

pub fn geo_target_metro(&self) -> &str

Returns the value of geo_target_metro, or the default value if geo_target_metro is unset.

pub fn geo_target_most_specific_location(&self) -> &str

Returns the value of geo_target_most_specific_location, or the default value if geo_target_most_specific_location is unset.

pub fn geo_target_postal_code(&self) -> &str

Returns the value of geo_target_postal_code, or the default value if geo_target_postal_code is unset.

pub fn geo_target_province(&self) -> &str

Returns the value of geo_target_province, or the default value if geo_target_province is unset.

pub fn geo_target_region(&self) -> &str

Returns the value of geo_target_region, or the default value if geo_target_region is unset.

pub fn geo_target_state(&self) -> &str

Returns the value of geo_target_state, or the default value if geo_target_state is unset.

pub fn quarter(&self) -> &str

Returns the value of quarter, or the default value if quarter is unset.

pub fn webpage(&self) -> &str

Returns the value of webpage, or the default value if webpage is unset.

pub fn week(&self) -> &str

Returns the value of week, or the default value if week is unset.

pub fn year(&self) -> i32

Returns the value of year, or the default value if year is unset.

pub fn product_aggregator_id(&self) -> i64

Returns the value of product_aggregator_id, or the default value if product_aggregator_id is unset.

pub fn product_merchant_id(&self) -> i64

Returns the value of product_merchant_id, or the default value if product_merchant_id is unset.

pub fn hotel_booking_window_days(&self) -> i64

Returns the value of hotel_booking_window_days, or the default value if hotel_booking_window_days is unset.

pub fn ad_destination_type(&self) -> AdDestinationType

Returns the enum value of ad_destination_type, or the default if the field is set to an invalid enum value.

pub fn set_ad_destination_type(&mut self, value: AdDestinationType)

Sets ad_destination_type to the provided enum value.

pub fn sk_ad_network_conversion_value(&self) -> i64

Returns the value of sk_ad_network_conversion_value, or the default value if sk_ad_network_conversion_value is unset.

pub fn conversion_value_rule_primary_dimension(
&self
) -> ConversionValueRulePrimaryDimension

Returns the enum value of conversion_value_rule_primary_dimension, or the default if the field is set to an invalid enum value.

pub fn set_conversion_value_rule_primary_dimension(
&mut self,
value: ConversionValueRulePrimaryDimension
)

Sets conversion_value_rule_primary_dimension to the provided enum value.

pub fn recommendation_type(&self) -> RecommendationType

Returns the enum value of recommendation_type, or the default if the field is set to an invalid enum value.

pub fn set_recommendation_type(&mut self, value: RecommendationType)

Sets recommendation_type to the provided enum value.