Enum googapis::google::ads::googleads::v9::enums::ad_type_enum::AdType

#[repr(i32)]
pub enum AdType {

    Unspecified,
    Unknown,
    TextAd,
    ExpandedTextAd,
    ExpandedDynamicSearchAd,
    HotelAd,
    ShoppingSmartAd,
    ShoppingProductAd,
    VideoAd,
    GmailAd,
    ImageAd,
    ResponsiveSearchAd,
    LegacyResponsiveDisplayAd,
    AppAd,
    LegacyAppInstallAd,
    ResponsiveDisplayAd,
    LocalAd,
    Html5UploadAd,
    DynamicHtml5Ad,
    AppEngagementAd,
    ShoppingComparisonListingAd,
    VideoBumperAd,
    VideoNonSkippableInStreamAd,
    VideoOutstreamAd,
    VideoTrueviewDiscoveryAd,
    VideoTrueviewInStreamAd,
    VideoResponsiveAd,
    SmartCampaignAd,
    CallAd,
    AppPreRegistrationAd,

}

The possible types of an ad.

Variants

Unspecified

No value has been specified.

Unknown

The received value is not known in this version.

This is a response-only value.

TextAd

The ad is a text ad.

ExpandedTextAd

The ad is an expanded text ad.

ExpandedDynamicSearchAd

The ad is an expanded dynamic search ad.

HotelAd

The ad is a hotel ad.

ShoppingSmartAd

The ad is a Smart Shopping ad.

ShoppingProductAd

The ad is a standard Shopping ad.

VideoAd

The ad is a video ad.

GmailAd

This ad is a Gmail ad.

ImageAd

This ad is an Image ad.

ResponsiveSearchAd

The ad is a responsive search ad.

LegacyResponsiveDisplayAd

The ad is a legacy responsive display ad.

AppAd

The ad is an app ad.

LegacyAppInstallAd

The ad is a legacy app install ad.

ResponsiveDisplayAd

The ad is a responsive display ad.

LocalAd

The ad is a local ad.

Html5UploadAd

The ad is a display upload ad with the HTML5_UPLOAD_AD product type.

DynamicHtml5Ad

The ad is a display upload ad with one of the DYNAMIC_HTML5_* product types.

AppEngagementAd

The ad is an app engagement ad.

ShoppingComparisonListingAd

The ad is a Shopping Comparison Listing ad.

VideoBumperAd

Video bumper ad.

VideoNonSkippableInStreamAd

Video non-skippable in-stream ad.

VideoOutstreamAd

Video outstream ad.

VideoTrueviewDiscoveryAd

Video TrueView in-display ad.

VideoTrueviewInStreamAd

Video TrueView in-stream ad.

VideoResponsiveAd

Video responsive ad.

SmartCampaignAd

Smart campaign ad.

CallAd

Call ad.

AppPreRegistrationAd

Universal app pre-registration ad.

Implementations

impl AdType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of AdType.

pub fn from_i32(value: i32) -> Option<AdType>

Converts an i32 to a AdType, or None if value is not a valid variant.