Enum googapis::google::ads::googleads::v9::enums::click_type_enum::ClickType

#[repr(i32)]
pub enum ClickType {

    Unspecified,
    Unknown,
    AppDeeplink,
    Breadcrumbs,
    BroadbandPlan,
    CallTracking,
    Calls,
    ClickOnEngagementAd,
    GetDirections,
    LocationExpansion,
    LocationFormatCall,
    LocationFormatDirections,
    LocationFormatImage,
    LocationFormatLandingPage,
    LocationFormatMap,
    LocationFormatStoreInfo,
    LocationFormatText,
    MobileCallTracking,
    OfferPrints,
    Other,
    ProductExtensionClicks,
    ProductListingAdClicks,
    Sitelinks,
    StoreLocator,
    UrlClicks,
    VideoAppStoreClicks,
    VideoCallToActionClicks,
    VideoCardActionHeadlineClicks,
    VideoEndCapClicks,
    VideoWebsiteClicks,
    VisualSitelinks,
    WirelessPlan,
    ProductListingAdLocal,
    ProductListingAdMultichannelLocal,
    ProductListingAdMultichannelOnline,
    ProductListingAdsCoupon,
    ProductListingAdTransactable,
    ProductAdAppDeeplink,
    ShowcaseAdCategoryLink,
    ShowcaseAdLocalStorefrontLink,
    ShowcaseAdOnlineProductLink,
    ShowcaseAdLocalProductLink,
    PromotionExtension,
    SwipeableGalleryAdHeadline,
    SwipeableGalleryAdSwipes,
    SwipeableGalleryAdSeeMore,
    SwipeableGalleryAdSitelinkOne,
    SwipeableGalleryAdSitelinkTwo,
    SwipeableGalleryAdSitelinkThree,
    SwipeableGalleryAdSitelinkFour,
    SwipeableGalleryAdSitelinkFive,
    HotelPrice,
    PriceExtension,
    HotelBookOnGoogleRoomSelection,
    ShoppingComparisonListing,

}

Enumerates Google Ads click types.

Variants

Unspecified

Not specified.

Unknown

The value is unknown in this version.

AppDeeplink

App engagement ad deep link.

Breadcrumbs

Breadcrumbs.

BroadbandPlan

Broadband Plan.

CallTracking

Manually dialed phone calls.

Calls

Phone calls.

ClickOnEngagementAd

Click on engagement ad.

GetDirections

Driving direction.

LocationExpansion

Get location details.

LocationFormatCall

Call.

LocationFormatDirections

Directions.

LocationFormatImage

Image(s).

LocationFormatLandingPage

Go to landing page.

LocationFormatMap

Map.

LocationFormatStoreInfo

Go to store info.

LocationFormatText

Text.

MobileCallTracking

Mobile phone calls.

OfferPrints

Print offer.

Other

Other.

ProductExtensionClicks

Product plusbox offer.

ProductListingAdClicks

Shopping - Product - Online.

Sitelinks

Sitelink.

StoreLocator

Show nearby locations.

UrlClicks

Headline.

VideoAppStoreClicks

App store.

VideoCallToActionClicks

Call-to-Action overlay.

VideoCardActionHeadlineClicks

Cards.

VideoEndCapClicks

End cap.

VideoWebsiteClicks

Website.

VisualSitelinks

Visual Sitelinks.

WirelessPlan

Wireless Plan.

ProductListingAdLocal

Shopping - Product - Local.

ProductListingAdMultichannelLocal

Shopping - Product - MultiChannel Local.

ProductListingAdMultichannelOnline

Shopping - Product - MultiChannel Online.

ProductListingAdsCoupon

Shopping - Product - Coupon.

ProductListingAdTransactable

Shopping - Product - Sell on Google.

ProductAdAppDeeplink

Shopping - Product - App engagement ad deep link.

ShowcaseAdCategoryLink

Shopping - Showcase - Category.

ShowcaseAdLocalStorefrontLink

Shopping - Showcase - Local storefront.

ShowcaseAdOnlineProductLink

Shopping - Showcase - Online product.

ShowcaseAdLocalProductLink

Shopping - Showcase - Local product.

PromotionExtension

Promotion Extension.

SwipeableGalleryAdHeadline

Ad Headline.

SwipeableGalleryAdSwipes

Swipes.

SwipeableGalleryAdSeeMore

See More.

SwipeableGalleryAdSitelinkOne

Sitelink 1.

SwipeableGalleryAdSitelinkTwo

Sitelink 2.

SwipeableGalleryAdSitelinkThree

Sitelink 3.

SwipeableGalleryAdSitelinkFour

Sitelink 4.

SwipeableGalleryAdSitelinkFive

Sitelink 5.

HotelPrice

Hotel price.

PriceExtension

Price Extension.

HotelBookOnGoogleRoomSelection

Book on Google hotel room selection.

ShoppingComparisonListing

Shopping - Comparison Listing.

Implementations

impl ClickType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ClickType.

pub fn from_i32(value: i32) -> Option<ClickType>

Converts an i32 to a ClickType, or None if value is not a valid variant.