Module googapis::google::ads::googleads::v9::resources::accessible_bidding_strategy

Nested message and enum types in AccessibleBiddingStrategy.

Structs

MaximizeConversionValue

An automated bidding strategy to help get the most conversion value for your campaigns while spending your budget.

MaximizeConversions

An automated bidding strategy to help get the most conversions for your campaigns while spending your budget.

TargetCpa

An automated bid strategy that sets bids to help get as many conversions as possible at the target cost-per-acquisition (CPA) you set.

TargetImpressionShare

An automated bidding strategy that sets bids so that a certain percentage of search ads are shown at the top of the first page (or other targeted location).

TargetRoas

An automated bidding strategy that helps you maximize revenue while averaging a specific target return on ad spend (ROAS).

TargetSpend

An automated bid strategy that sets your bids to help get as many clicks as possible within your budget.

Enums

Scheme

The bidding scheme.