Module googapis::google::ads::googleads::v9::resources::billing_setup

Nested message and enum types in BillingSetup.

Structs

PaymentsAccountInfo

Container of payments account information for this billing.

Enums

EndTime

When the billing setup ends / ended. This is either FOREVER or the start time of the next scheduled billing setup.

StartTime

When creating a new billing setup, this is when the setup should take effect. NOW is the only acceptable start time if the customer doesn’t have any approved setups.