Module googapis::google::ads::googleads::v9::resources::campaign

Nested message and enum types in Campaign.

Structs

AppCampaignSetting

Campaign-level settings for App Campaigns.

DynamicSearchAdsSetting

The setting for controlling Dynamic Search Ads (DSA).

GeoTargetTypeSetting

Represents a collection of settings related to ads geotargeting.

HotelSettingInfo

Campaign-level settings for hotel ads.

LocalCampaignSetting

Campaign setting for local campaigns.

NetworkSettings

The network settings for the campaign.

OptimizationGoalSetting

Optimization goal setting for this campaign, which includes a set of optimization goal types.

SelectiveOptimization

Selective optimization setting for this campaign, which includes a set of conversion actions to optimize this campaign towards.

ShoppingSetting

The setting for Shopping campaigns. Defines the universe of products that can be advertised by the campaign, and how this campaign interacts with other Shopping campaigns.

TrackingSetting

Campaign-level settings for tracking information.

VanityPharma

Describes how unbranded pharma ads will be displayed.

Enums

CampaignBiddingStrategy

The bidding strategy for the campaign.