Enum googapis::google::ads::googleads::v9::services::mutate_operation_response::Response

pub enum Response {

    AdGroupAdLabelResult(MutateAdGroupAdLabelResult),
    AdGroupAdResult(MutateAdGroupAdResult),
    AdGroupAssetResult(MutateAdGroupAssetResult),
    AdGroupBidModifierResult(MutateAdGroupBidModifierResult),
    AdGroupCriterionCustomizerResult(MutateAdGroupCriterionCustomizerResult),
    AdGroupCriterionLabelResult(MutateAdGroupCriterionLabelResult),
    AdGroupCriterionResult(MutateAdGroupCriterionResult),
    AdGroupCustomizerResult(MutateAdGroupCustomizerResult),
    AdGroupExtensionSettingResult(MutateAdGroupExtensionSettingResult),
    AdGroupFeedResult(MutateAdGroupFeedResult),
    AdGroupLabelResult(MutateAdGroupLabelResult),
    AdGroupResult(MutateAdGroupResult),
    AdParameterResult(MutateAdParameterResult),
    AdResult(MutateAdResult),
    AssetResult(MutateAssetResult),
    AssetGroupAssetResult(MutateAssetGroupAssetResult),
    AssetGroupListingGroupFilterResult(MutateAssetGroupListingGroupFilterResult),
    AssetGroupResult(MutateAssetGroupResult),
    AssetSetAssetResult(MutateAssetSetAssetResult),
    AssetSetResult(MutateAssetSetResult),
    BiddingDataExclusionResult(MutateBiddingDataExclusionsResult),
    BiddingSeasonalityAdjustmentResult(MutateBiddingSeasonalityAdjustmentsResult),
    BiddingStrategyResult(MutateBiddingStrategyResult),
    CampaignAssetResult(MutateCampaignAssetResult),
    CampaignAssetSetResult(MutateCampaignAssetSetResult),
    CampaignBidModifierResult(MutateCampaignBidModifierResult),
    CampaignBudgetResult(MutateCampaignBudgetResult),
    CampaignConversionGoalResult(MutateCampaignConversionGoalResult),
    CampaignCriterionResult(MutateCampaignCriterionResult),
    CampaignCustomizerResult(MutateCampaignCustomizerResult),
    CampaignDraftResult(MutateCampaignDraftResult),
    CampaignExperimentResult(MutateCampaignExperimentResult),
    CampaignExtensionSettingResult(MutateCampaignExtensionSettingResult),
    CampaignFeedResult(MutateCampaignFeedResult),
    CampaignLabelResult(MutateCampaignLabelResult),
    CampaignResult(MutateCampaignResult),
    CampaignSharedSetResult(MutateCampaignSharedSetResult),
    ConversionActionResult(MutateConversionActionResult),
    ConversionCustomVariableResult(MutateConversionCustomVariableResult),
    ConversionGoalCampaignConfigResult(MutateConversionGoalCampaignConfigResult),
    ConversionValueRuleResult(MutateConversionValueRuleResult),
    ConversionValueRuleSetResult(MutateConversionValueRuleSetResult),
    CustomConversionGoalResult(MutateCustomConversionGoalResult),
    CustomerAssetResult(MutateCustomerAssetResult),
    CustomerConversionGoalResult(MutateCustomerConversionGoalResult),
    CustomerCustomizerResult(MutateCustomerCustomizerResult),
    CustomerExtensionSettingResult(MutateCustomerExtensionSettingResult),
    CustomerFeedResult(MutateCustomerFeedResult),
    CustomerLabelResult(MutateCustomerLabelResult),
    CustomerNegativeCriterionResult(MutateCustomerNegativeCriteriaResult),
    CustomerResult(MutateCustomerResult),
    CustomizerAttributeResult(MutateCustomizerAttributeResult),
    ExtensionFeedItemResult(MutateExtensionFeedItemResult),
    FeedItemResult(MutateFeedItemResult),
    FeedItemSetResult(MutateFeedItemSetResult),
    FeedItemSetLinkResult(MutateFeedItemSetLinkResult),
    FeedItemTargetResult(MutateFeedItemTargetResult),
    FeedMappingResult(MutateFeedMappingResult),
    FeedResult(MutateFeedResult),
    KeywordPlanAdGroupResult(MutateKeywordPlanAdGroupResult),
    KeywordPlanCampaignResult(MutateKeywordPlanCampaignResult),
    KeywordPlanAdGroupKeywordResult(MutateKeywordPlanAdGroupKeywordResult),
    KeywordPlanCampaignKeywordResult(MutateKeywordPlanCampaignKeywordResult),
    KeywordPlanResult(MutateKeywordPlansResult),
    LabelResult(MutateLabelResult),
    MediaFileResult(MutateMediaFileResult),
    RemarketingActionResult(MutateRemarketingActionResult),
    SharedCriterionResult(MutateSharedCriterionResult),
    SharedSetResult(MutateSharedSetResult),
    SmartCampaignSettingResult(MutateSmartCampaignSettingResult),
    UserListResult(MutateUserListResult),

}

The mutate response.

Variants

AdGroupAdLabelResult(MutateAdGroupAdLabelResult)

The result for the ad group ad label mutate.

Tuple Fields of AdGroupAdLabelResult

0: MutateAdGroupAdLabelResult

AdGroupAdResult(MutateAdGroupAdResult)

The result for the ad group ad mutate.

Tuple Fields of AdGroupAdResult

0: MutateAdGroupAdResult

AdGroupAssetResult(MutateAdGroupAssetResult)

The result for the ad group asset mutate.

Tuple Fields of AdGroupAssetResult

0: MutateAdGroupAssetResult

AdGroupBidModifierResult(MutateAdGroupBidModifierResult)

The result for the ad group bid modifier mutate.

Tuple Fields of AdGroupBidModifierResult

0: MutateAdGroupBidModifierResult

AdGroupCriterionCustomizerResult(MutateAdGroupCriterionCustomizerResult)

The result for the ad group criterion customizer mutate.

Tuple Fields of AdGroupCriterionCustomizerResult

0: MutateAdGroupCriterionCustomizerResult

AdGroupCriterionLabelResult(MutateAdGroupCriterionLabelResult)

The result for the ad group criterion label mutate.

Tuple Fields of AdGroupCriterionLabelResult

0: MutateAdGroupCriterionLabelResult

AdGroupCriterionResult(MutateAdGroupCriterionResult)

The result for the ad group criterion mutate.

Tuple Fields of AdGroupCriterionResult

0: MutateAdGroupCriterionResult

AdGroupCustomizerResult(MutateAdGroupCustomizerResult)

The result for the ad group customizer mutate.

Tuple Fields of AdGroupCustomizerResult

0: MutateAdGroupCustomizerResult

AdGroupExtensionSettingResult(MutateAdGroupExtensionSettingResult)

The result for the ad group extension setting mutate.

Tuple Fields of AdGroupExtensionSettingResult

0: MutateAdGroupExtensionSettingResult

AdGroupFeedResult(MutateAdGroupFeedResult)

The result for the ad group feed mutate.

Tuple Fields of AdGroupFeedResult

0: MutateAdGroupFeedResult

AdGroupLabelResult(MutateAdGroupLabelResult)

The result for the ad group label mutate.

Tuple Fields of AdGroupLabelResult

0: MutateAdGroupLabelResult

AdGroupResult(MutateAdGroupResult)

The result for the ad group mutate.

Tuple Fields of AdGroupResult

0: MutateAdGroupResult

AdParameterResult(MutateAdParameterResult)

The result for the ad parameter mutate.

Tuple Fields of AdParameterResult

0: MutateAdParameterResult

AdResult(MutateAdResult)

The result for the ad mutate.

Tuple Fields of AdResult

0: MutateAdResult

AssetResult(MutateAssetResult)

The result for the asset mutate.

Tuple Fields of AssetResult

0: MutateAssetResult

AssetGroupAssetResult(MutateAssetGroupAssetResult)

The result for the asset group asset mutate.

Tuple Fields of AssetGroupAssetResult

0: MutateAssetGroupAssetResult

AssetGroupListingGroupFilterResult(MutateAssetGroupListingGroupFilterResult)

The result for the asset group listing group filter mutate.

Tuple Fields of AssetGroupListingGroupFilterResult

0: MutateAssetGroupListingGroupFilterResult

AssetGroupResult(MutateAssetGroupResult)

The result for the asset group mutate.

Tuple Fields of AssetGroupResult

0: MutateAssetGroupResult

AssetSetAssetResult(MutateAssetSetAssetResult)

The result for the asset set asset mutate.

Tuple Fields of AssetSetAssetResult

0: MutateAssetSetAssetResult

AssetSetResult(MutateAssetSetResult)

The result for the asset set mutate.

Tuple Fields of AssetSetResult

0: MutateAssetSetResult

BiddingDataExclusionResult(MutateBiddingDataExclusionsResult)

The result for the bidding data exclusion mutate.

Tuple Fields of BiddingDataExclusionResult

0: MutateBiddingDataExclusionsResult

BiddingSeasonalityAdjustmentResult(MutateBiddingSeasonalityAdjustmentsResult)

The result for the bidding seasonality adjustment mutate.

Tuple Fields of BiddingSeasonalityAdjustmentResult

0: MutateBiddingSeasonalityAdjustmentsResult

BiddingStrategyResult(MutateBiddingStrategyResult)

The result for the bidding strategy mutate.

Tuple Fields of BiddingStrategyResult

0: MutateBiddingStrategyResult

CampaignAssetResult(MutateCampaignAssetResult)

The result for the campaign asset mutate.

Tuple Fields of CampaignAssetResult

0: MutateCampaignAssetResult

CampaignAssetSetResult(MutateCampaignAssetSetResult)

The result for the campaign asset set mutate.

Tuple Fields of CampaignAssetSetResult

0: MutateCampaignAssetSetResult

CampaignBidModifierResult(MutateCampaignBidModifierResult)

The result for the campaign bid modifier mutate.

Tuple Fields of CampaignBidModifierResult

0: MutateCampaignBidModifierResult

CampaignBudgetResult(MutateCampaignBudgetResult)

The result for the campaign budget mutate.

Tuple Fields of CampaignBudgetResult

0: MutateCampaignBudgetResult

CampaignConversionGoalResult(MutateCampaignConversionGoalResult)

The result for the campaign conversion goal mutate.

Tuple Fields of CampaignConversionGoalResult

0: MutateCampaignConversionGoalResult

CampaignCriterionResult(MutateCampaignCriterionResult)

The result for the campaign criterion mutate.

Tuple Fields of CampaignCriterionResult

0: MutateCampaignCriterionResult

CampaignCustomizerResult(MutateCampaignCustomizerResult)

The result for the campaign customizer mutate.

Tuple Fields of CampaignCustomizerResult

0: MutateCampaignCustomizerResult

CampaignDraftResult(MutateCampaignDraftResult)

The result for the campaign draft mutate.

Tuple Fields of CampaignDraftResult

0: MutateCampaignDraftResult

CampaignExperimentResult(MutateCampaignExperimentResult)

The result for the campaign experiment mutate.

Tuple Fields of CampaignExperimentResult

0: MutateCampaignExperimentResult

CampaignExtensionSettingResult(MutateCampaignExtensionSettingResult)

The result for the campaign extension setting mutate.

Tuple Fields of CampaignExtensionSettingResult

0: MutateCampaignExtensionSettingResult

CampaignFeedResult(MutateCampaignFeedResult)

The result for the campaign feed mutate.

Tuple Fields of CampaignFeedResult

0: MutateCampaignFeedResult

CampaignLabelResult(MutateCampaignLabelResult)

The result for the campaign label mutate.

Tuple Fields of CampaignLabelResult

0: MutateCampaignLabelResult

CampaignResult(MutateCampaignResult)

The result for the campaign mutate.

Tuple Fields of CampaignResult

0: MutateCampaignResult

CampaignSharedSetResult(MutateCampaignSharedSetResult)

The result for the campaign shared set mutate.

Tuple Fields of CampaignSharedSetResult

0: MutateCampaignSharedSetResult

ConversionActionResult(MutateConversionActionResult)

The result for the conversion action mutate.

Tuple Fields of ConversionActionResult

0: MutateConversionActionResult

ConversionCustomVariableResult(MutateConversionCustomVariableResult)

The result for the conversion custom variable mutate.

Tuple Fields of ConversionCustomVariableResult

0: MutateConversionCustomVariableResult

ConversionGoalCampaignConfigResult(MutateConversionGoalCampaignConfigResult)

The result for the conversion goal campaign config mutate.

Tuple Fields of ConversionGoalCampaignConfigResult

0: MutateConversionGoalCampaignConfigResult

ConversionValueRuleResult(MutateConversionValueRuleResult)

The result for the conversion value rule mutate.

Tuple Fields of ConversionValueRuleResult

0: MutateConversionValueRuleResult

ConversionValueRuleSetResult(MutateConversionValueRuleSetResult)

The result for the conversion value rule set mutate.

Tuple Fields of ConversionValueRuleSetResult

0: MutateConversionValueRuleSetResult

CustomConversionGoalResult(MutateCustomConversionGoalResult)

The result for the custom conversion goal mutate.

Tuple Fields of CustomConversionGoalResult

0: MutateCustomConversionGoalResult

CustomerAssetResult(MutateCustomerAssetResult)

The result for the customer asset mutate.

Tuple Fields of CustomerAssetResult

0: MutateCustomerAssetResult

CustomerConversionGoalResult(MutateCustomerConversionGoalResult)

The result for the customer conversion goal mutate.

Tuple Fields of CustomerConversionGoalResult

0: MutateCustomerConversionGoalResult

CustomerCustomizerResult(MutateCustomerCustomizerResult)

The result for the customer customizer mutate.

Tuple Fields of CustomerCustomizerResult

0: MutateCustomerCustomizerResult

CustomerExtensionSettingResult(MutateCustomerExtensionSettingResult)

The result for the customer extension setting mutate.

Tuple Fields of CustomerExtensionSettingResult

0: MutateCustomerExtensionSettingResult

CustomerFeedResult(MutateCustomerFeedResult)

The result for the customer feed mutate.

Tuple Fields of CustomerFeedResult

0: MutateCustomerFeedResult

CustomerLabelResult(MutateCustomerLabelResult)

The result for the customer label mutate.

Tuple Fields of CustomerLabelResult

0: MutateCustomerLabelResult

CustomerNegativeCriterionResult(MutateCustomerNegativeCriteriaResult)

The result for the customer negative criterion mutate.

Tuple Fields of CustomerNegativeCriterionResult

0: MutateCustomerNegativeCriteriaResult

CustomerResult(MutateCustomerResult)

The result for the customer mutate.

Tuple Fields of CustomerResult

0: MutateCustomerResult

CustomizerAttributeResult(MutateCustomizerAttributeResult)

The result for the customizer attribute mutate.

Tuple Fields of CustomizerAttributeResult

0: MutateCustomizerAttributeResult

ExtensionFeedItemResult(MutateExtensionFeedItemResult)

The result for the extension feed item mutate.

Tuple Fields of ExtensionFeedItemResult

0: MutateExtensionFeedItemResult

FeedItemResult(MutateFeedItemResult)

The result for the feed item mutate.

Tuple Fields of FeedItemResult

0: MutateFeedItemResult

FeedItemSetResult(MutateFeedItemSetResult)

The result for the feed item set mutate.

Tuple Fields of FeedItemSetResult

0: MutateFeedItemSetResult

FeedItemSetLinkResult(MutateFeedItemSetLinkResult)

The result for the feed item set link mutate.

Tuple Fields of FeedItemSetLinkResult

0: MutateFeedItemSetLinkResult

FeedItemTargetResult(MutateFeedItemTargetResult)

The result for the feed item target mutate.

Tuple Fields of FeedItemTargetResult

0: MutateFeedItemTargetResult

FeedMappingResult(MutateFeedMappingResult)

The result for the feed mapping mutate.

Tuple Fields of FeedMappingResult

0: MutateFeedMappingResult

FeedResult(MutateFeedResult)

The result for the feed mutate.

Tuple Fields of FeedResult

0: MutateFeedResult

KeywordPlanAdGroupResult(MutateKeywordPlanAdGroupResult)

The result for the keyword plan ad group mutate.

Tuple Fields of KeywordPlanAdGroupResult

0: MutateKeywordPlanAdGroupResult

KeywordPlanCampaignResult(MutateKeywordPlanCampaignResult)

The result for the keyword plan campaign mutate.

Tuple Fields of KeywordPlanCampaignResult

0: MutateKeywordPlanCampaignResult

KeywordPlanAdGroupKeywordResult(MutateKeywordPlanAdGroupKeywordResult)

The result for the keyword plan ad group keyword mutate.

Tuple Fields of KeywordPlanAdGroupKeywordResult

0: MutateKeywordPlanAdGroupKeywordResult

KeywordPlanCampaignKeywordResult(MutateKeywordPlanCampaignKeywordResult)

The result for the keyword plan campaign keyword mutate.

Tuple Fields of KeywordPlanCampaignKeywordResult

0: MutateKeywordPlanCampaignKeywordResult

KeywordPlanResult(MutateKeywordPlansResult)

The result for the keyword plan mutate.

Tuple Fields of KeywordPlanResult

0: MutateKeywordPlansResult

LabelResult(MutateLabelResult)

The result for the label mutate.

Tuple Fields of LabelResult

0: MutateLabelResult

MediaFileResult(MutateMediaFileResult)

The result for the media file mutate.

Tuple Fields of MediaFileResult

0: MutateMediaFileResult

RemarketingActionResult(MutateRemarketingActionResult)

The result for the remarketing action mutate.

Tuple Fields of RemarketingActionResult

0: MutateRemarketingActionResult

SharedCriterionResult(MutateSharedCriterionResult)

The result for the shared criterion mutate.

Tuple Fields of SharedCriterionResult

0: MutateSharedCriterionResult

SharedSetResult(MutateSharedSetResult)

The result for the shared set mutate.

Tuple Fields of SharedSetResult

0: MutateSharedSetResult

SmartCampaignSettingResult(MutateSmartCampaignSettingResult)

The result for the Smart campaign setting mutate.

Tuple Fields of SmartCampaignSettingResult

0: MutateSmartCampaignSettingResult

UserListResult(MutateUserListResult)

The result for the user list mutate.

Tuple Fields of UserListResult

0: MutateUserListResult

Implementations

impl Response

pub fn encode<B>(&self, buf: &mut B) where
B: BufMut,

pub fn merge<B>(
field: &mut Option<Response>,
tag: u32,
wire_type: WireType,
buf: &mut B,
ctx: DecodeContext
) -> Result<(), DecodeError> where
B: Buf,

pub fn encoded_len(&self) -> usize