Module googapis::google::analytics::admin::v1alpha

Modules

analytics_admin_service_client
Generated client implementations.

change_history_change
Nested message and enum types in ChangeHistoryChange.

custom_dimension
Nested message and enum types in CustomDimension.

custom_metric
Nested message and enum types in CustomMetric.

Structs

Account
A resource message representing a Google Analytics account.

AccountSummary
A virtual resource representing an overview of an account and all its child GA4 properties.

AndroidAppDataStream
A resource message representing a Google Analytics Android app stream.

ArchiveCustomDimensionRequest
Request message for ArchiveCustomDimension RPC.

ArchiveCustomMetricRequest
Request message for ArchiveCustomMetric RPC.

AuditUserLink
Read-only resource used to summarize a principal’s effective roles.

AuditUserLinksRequest
Request message for AuditUserLinks RPC.

AuditUserLinksResponse
Response message for AuditUserLinks RPC.

BatchCreateUserLinksRequest
Request message for BatchCreateUserLinks RPC.

BatchCreateUserLinksResponse
Response message for BatchCreateUserLinks RPC.

BatchDeleteUserLinksRequest
Request message for BatchDeleteUserLinks RPC.

BatchGetUserLinksRequest
Request message for BatchGetUserLinks RPC.

BatchGetUserLinksResponse
Response message for BatchGetUserLinks RPC.

BatchUpdateUserLinksRequest
Request message for BatchUpdateUserLinks RPC.

BatchUpdateUserLinksResponse
Response message for BatchUpdateUserLinks RPC.

ChangeHistoryChange
A description of a change to a single Google Analytics resource.

ChangeHistoryEvent
A set of changes within a Google Analytics account or its child properties that resulted from the same cause. Common causes would be updates made in the Google Analytics UI, changes from customer support, or automatic Google Analytics system changes.

ConversionEvent
A conversion event in a Google Analytics property.

CreateConversionEventRequest
Request message for CreateConversionEvent RPC

CreateCustomDimensionRequest
Request message for CreateCustomDimension RPC.

CreateCustomMetricRequest
Request message for CreateCustomMetric RPC.

CreateFirebaseLinkRequest
Request message for CreateFirebaseLink RPC

CreateGoogleAdsLinkRequest
Request message for CreateGoogleAdsLink RPC

CreateMeasurementProtocolSecretRequest
Request message for CreateMeasurementProtocolSecret RPC

CreatePropertyRequest
Request message for CreateProperty RPC.

CreateUserLinkRequest
Request message for CreateUserLink RPC.

CreateWebDataStreamRequest
Request message for CreateWebDataStream RPC.

CustomDimension
A definition for a CustomDimension.

CustomMetric
A definition for a custom metric.

DataSharingSettings
A resource message representing data sharing settings of a Google Analytics account.

DeleteAccountRequest
Request message for DeleteAccount RPC.

DeleteAndroidAppDataStreamRequest
Request message for DeleteAndroidAppDataStream RPC.

DeleteConversionEventRequest
Request message for DeleteConversionEvent RPC

DeleteFirebaseLinkRequest
Request message for DeleteFirebaseLink RPC

DeleteGoogleAdsLinkRequest
Request message for DeleteGoogleAdsLink RPC.

DeleteIosAppDataStreamRequest
Request message for DeleteIosAppDataStream RPC.

DeleteMeasurementProtocolSecretRequest
Request message for DeleteMeasurementProtocolSecret RPC

DeletePropertyRequest
Request message for DeleteProperty RPC.

DeleteUserLinkRequest
Request message for DeleteUserLink RPC.

DeleteWebDataStreamRequest
Request message for DeleteWebDataStream RPC.

EnhancedMeasurementSettings
Singleton resource under a WebDataStream, configuring measurement of additional site interactions and content.

FirebaseLink
A link between an GA4 property and a Firebase project.

GetAccountRequest
Request message for GetAccount RPC.

GetAndroidAppDataStreamRequest
Request message for GetAndroidAppDataStream RPC.

GetConversionEventRequest
Request message for GetConversionEvent RPC

GetCustomDimensionRequest
Request message for GetCustomDimension RPC.

GetCustomMetricRequest
Request message for GetCustomMetric RPC.

GetDataSharingSettingsRequest
Request message for GetDataSharingSettings RPC.

GetEnhancedMeasurementSettingsRequest
Request message for GetEnhancedMeasurementSettings RPC.

GetGlobalSiteTagRequest
Request message for GetGlobalSiteTag RPC.

GetGoogleSignalsSettingsRequest
Request message for GetGoogleSignalsSettings RPC

GetIosAppDataStreamRequest
Request message for GetIosAppDataStream RPC.

GetMeasurementProtocolSecretRequest
Request message for GetMeasurementProtocolSecret RPC.

GetPropertyRequest
Request message for GetProperty RPC.

GetUserLinkRequest
Request message for GetUserLink RPC.

GetWebDataStreamRequest
Request message for GetWebDataStream RPC.

GlobalSiteTag
Read-only resource with the tag for sending data from a website to a WebDataStream.

GoogleAdsLink
A link between an GA4 property and a Google Ads account.

GoogleSignalsSettings
Settings values for Google Signals. This is a singleton resource.

IosAppDataStream
A resource message representing a Google Analytics IOS app stream.

ListAccountSummariesRequest
Request message for ListAccountSummaries RPC.

ListAccountSummariesResponse
Response message for ListAccountSummaries RPC.

ListAccountsRequest
Request message for ListAccounts RPC.

ListAccountsResponse
Request message for ListAccounts RPC.

ListAndroidAppDataStreamsRequest
Request message for ListAndroidAppDataStreams RPC.

ListAndroidAppDataStreamsResponse
Request message for ListAndroidDataStreams RPC.

ListConversionEventsRequest
Request message for ListConversionEvents RPC

ListConversionEventsResponse
Response message for ListConversionEvents RPC.

ListCustomDimensionsRequest
Request message for ListCustomDimensions RPC.

ListCustomDimensionsResponse
Response message for ListCustomDimensions RPC.

ListCustomMetricsRequest
Request message for ListCustomMetrics RPC.

ListCustomMetricsResponse
Response message for ListCustomMetrics RPC.

ListFirebaseLinksRequest
Request message for ListFirebaseLinks RPC

ListFirebaseLinksResponse
Response message for ListFirebaseLinks RPC

ListGoogleAdsLinksRequest
Request message for ListGoogleAdsLinks RPC.

ListGoogleAdsLinksResponse
Response message for ListGoogleAdsLinks RPC.

ListIosAppDataStreamsRequest
Request message for ListIosAppDataStreams RPC.

ListIosAppDataStreamsResponse
Request message for ListIosAppDataStreams RPC.

ListMeasurementProtocolSecretsRequest
Request message for ListMeasurementProtocolSecret RPC

ListMeasurementProtocolSecretsResponse
Response message for ListMeasurementProtocolSecret RPC

ListPropertiesRequest
Request message for ListProperties RPC.

ListPropertiesResponse
Response message for ListProperties RPC.

ListUserLinksRequest
Request message for ListUserLinks RPC.

ListUserLinksResponse
Response message for ListUserLinks RPC.

ListWebDataStreamsRequest
Request message for ListWebDataStreams RPC.

ListWebDataStreamsResponse
Request message for ListWebDataStreams RPC.

MeasurementProtocolSecret
A secret value used for sending hits to Measurement Protocol.

Property
A resource message representing a Google Analytics GA4 property.

PropertySummary
A virtual resource representing metadata for an GA4 property.

ProvisionAccountTicketRequest
Request message for ProvisionAccountTicket RPC.

ProvisionAccountTicketResponse
Response message for ProvisionAccountTicket RPC.

SearchChangeHistoryEventsRequest
Request message for SearchChangeHistoryEvents RPC.

SearchChangeHistoryEventsResponse
Response message for SearchAccounts RPC.

UpdateAccountRequest
Request message for UpdateAccount RPC.

UpdateAndroidAppDataStreamRequest
Request message for UpdateAndroidAppDataStream RPC.

UpdateCustomDimensionRequest
Request message for UpdateCustomDimension RPC.

UpdateCustomMetricRequest
Request message for UpdateCustomMetric RPC.

UpdateEnhancedMeasurementSettingsRequest
Request message for UpdateEnhancedMeasurementSettings RPC.

UpdateFirebaseLinkRequest
Request message for UpdateFirebaseLink RPC

UpdateGoogleAdsLinkRequest
Request message for UpdateGoogleAdsLink RPC

UpdateGoogleSignalsSettingsRequest
Request message for UpdateGoogleSignalsSettings RPC

UpdateIosAppDataStreamRequest
Request message for UpdateIosAppDataStream RPC.

UpdateMeasurementProtocolSecretRequest
Request message for UpdateMeasurementProtocolSecret RPC

UpdatePropertyRequest
Request message for UpdateProperty RPC.

UpdateUserLinkRequest
Request message for UpdateUserLink RPC.

UpdateWebDataStreamRequest
Request message for UpdateWebDataStream RPC.

UserLink
A resource message representing a user’s permissions on an Account or Property resource.

WebDataStream
A resource message representing a Google Analytics web stream.

Enums

ActionType
Types of actions that may change a resource.

ActorType
Different kinds of actors that can make changes to Google Analytics resources.

ChangeHistoryResourceType
Types of resources whose changes may be returned from change history.

GoogleSignalsConsent
Consent field of the Google Signals settings (i.e., whether the user has consented to the Google Signals terms of service.)

GoogleSignalsState
Status of the Google Signals settings (i.e., whether this feature has been enabled for the property).

IndustryCategory
The category selected for this property, used for industry benchmarking.

MaximumUserAccess
Maximum access settings that Firebase user receive on the linked Analytics property.