Module googapis::google::api::expr::conformance::v1alpha1

Modules

conformance_service_client

Generated client implementations.

eval_request

Nested message and enum types in EvalRequest.

issue_details

Nested message and enum types in IssueDetails.

Structs

CheckRequest

Request message for the Check method.

CheckResponse

Response message for the Check method.

EvalRequest

Request message for the Eval method.

EvalResponse

Response message for the Eval method.

IssueDetails

Warnings or errors in service execution are represented by [google.rpc.Status][google.rpc.Status] messages, with the following message in the details field.

ParseRequest

Request message for the Parse method.

ParseResponse

Response message for the Parse method.