Module googapis::google::api::expr::v1alpha1::decl

Nested message and enum types in Decl.

Modules

function_decl

Nested message and enum types in FunctionDecl.

Structs

FunctionDecl

Function declaration specifies one or more overloads which indicate the function’s parameter types and return type.

IdentDecl

Identifier declaration which specifies its type and optional Expr value.

Enums

DeclKind

Required. The declaration kind.