Module googapis::google::api::expr::v1beta1

Modules

decl

Nested message and enum types in Decl.

eval_state

Nested message and enum types in EvalState.

expr

Nested message and enum types in Expr.

expr_value

Nested message and enum types in ExprValue.

literal

Nested message and enum types in Literal.

map_value

Nested message and enum types in MapValue.

value

Nested message and enum types in Value.

Structs

Decl

A declaration.

DeclType

The declared type of a variable.

EnumValue

An enum value.

ErrorSet

A set of errors.

EvalState

The state of an evaluation.

Expr

An abstract representation of a common expression.

ExprValue

The value of an evaluated expression.

FunctionDecl

A function declaration.

IdRef

A reference to an expression id.

IdentDecl

An identifier declaration.

ListValue

A list.

Literal

Represents a primitive literal.

MapValue

A map.

ParsedExpr

An expression together with source information as returned by the parser.

SourceInfo

Source information collected at parse time.

SourcePosition

A specific position in source.

UnknownSet

A set of expressions for which the value is unknown.

Value

Represents a CEL value.