Module googapis::google::api::serviceusage::v1beta1

Modules

get_service_identity_response

Nested message and enum types in GetServiceIdentityResponse.

import_admin_overrides_request

Nested message and enum types in ImportAdminOverridesRequest.

import_consumer_overrides_request

Nested message and enum types in ImportConsumerOverridesRequest.

service_usage_client

Generated client implementations.

Structs

AdminQuotaPolicy

Quota policy created by quota administrator.

BatchCreateAdminOverridesResponse

Response message for BatchCreateAdminOverrides

BatchCreateConsumerOverridesResponse

Response message for BatchCreateConsumerOverrides

BatchEnableServicesRequest

Request message for the BatchEnableServices method.

ConsumerQuotaLimit

Consumer quota settings for a quota limit.

ConsumerQuotaMetric

Consumer quota settings for a quota metric.

CreateAdminOverrideRequest

Request message for CreateAdminOverride.

CreateAdminQuotaPolicyMetadata

Metadata message that provides information such as progress, partial failures, and similar information on each GetOperation call of LRO returned by CreateAdminQuotaPolicy.

CreateConsumerOverrideRequest

Request message for CreateConsumerOverride.

DeleteAdminOverrideRequest

Request message for DeleteAdminOverride.

DeleteAdminQuotaPolicyMetadata

Metadata message that provides information such as progress, partial failures, and similar information on each GetOperation call of LRO returned by DeleteAdminQuotaPolicy.

DeleteConsumerOverrideRequest

Request message for DeleteConsumerOverride.

DisableServiceRequest

Request message for the DisableService method.

EnableServiceRequest

Request message for the EnableService method.

GenerateServiceIdentityRequest

Request message for generating service identity.

GetConsumerQuotaLimitRequest

Request message for GetConsumerQuotaLimit

GetConsumerQuotaMetricRequest

Request message for GetConsumerQuotaMetric

GetServiceIdentityMetadata

Metadata for the GetServiceIdentity method.

GetServiceIdentityResponse

Response message for getting service identity.

GetServiceRequest

Request message for the GetService method.

ImportAdminOverridesMetadata

Metadata message that provides information such as progress, partial failures, and similar information on each GetOperation call of LRO returned by ImportAdminOverrides.

ImportAdminOverridesRequest

Request message for ImportAdminOverrides

ImportAdminOverridesResponse

Response message for ImportAdminOverrides

ImportAdminQuotaPoliciesMetadata

Metadata message that provides information such as progress, partial failures, and similar information on each GetOperation call of LRO returned by ImportAdminQuotaPolicies.

ImportAdminQuotaPoliciesResponse

Response message for ImportAdminQuotaPolicies

ImportConsumerOverridesMetadata

Metadata message that provides information such as progress, partial failures, and similar information on each GetOperation call of LRO returned by ImportConsumerOverrides.

ImportConsumerOverridesRequest

Request message for ImportConsumerOverrides

ImportConsumerOverridesResponse

Response message for ImportConsumerOverrides

ListAdminOverridesRequest

Request message for ListAdminOverrides

ListAdminOverridesResponse

Response message for ListAdminOverrides.

ListConsumerOverridesRequest

Request message for ListConsumerOverrides

ListConsumerOverridesResponse

Response message for ListConsumerOverrides.

ListConsumerQuotaMetricsRequest

Request message for ListConsumerQuotaMetrics

ListConsumerQuotaMetricsResponse

Response message for ListConsumerQuotaMetrics

ListServicesRequest

Request message for the ListServices method.

ListServicesResponse

Response message for the ListServices method.

OperationMetadata

The operation metadata returned for the batchend services operation.

OverrideInlineSource

Import data embedded in the request message

QuotaBucket

A quota bucket is a quota provisioning unit for a specific set of dimensions.

QuotaOverride

A quota override

Service

A service that is available for use by the consumer.

ServiceConfig

The configuration of the service.

ServiceIdentity

Service identity for a service. This is the identity that service producer should use to access consumer resources.

UpdateAdminOverrideRequest

Request message for UpdateAdminOverride.

UpdateAdminQuotaPolicyMetadata

Metadata message that provides information such as progress, partial failures, and similar information on each GetOperation call of LRO returned by UpdateAdminQuotaPolicy.

UpdateConsumerOverrideRequest

Request message for UpdateConsumerOverride.

Enums

QuotaSafetyCheck

Enumerations of quota safety checks.

QuotaView

Selected view of quota. Can be used to request more detailed quota information when retrieving quota metrics and limits.

State

Whether or not a service has been enabled for use by a consumer.