Module googapis::google::area120::tables::v1alpha1

Modules

tables_service_client
Generated client implementations.

Structs

BatchCreateRowsRequest
Request message for TablesService.BatchCreateRows.

BatchCreateRowsResponse
Response message for TablesService.BatchCreateRows.

BatchDeleteRowsRequest
Request message for TablesService.BatchDeleteRows

BatchUpdateRowsRequest
Request message for TablesService.BatchUpdateRows.

BatchUpdateRowsResponse
Response message for TablesService.BatchUpdateRows.

ColumnDescription
Details on a column in the table.

CreateRowRequest
Request message for TablesService.CreateRow.

DeleteRowRequest
Request message for TablesService.DeleteRow

GetRowRequest
Request message for TablesService.GetRow.

GetTableRequest
Request message for TablesService.GetTable.

GetWorkspaceRequest
Request message for TablesService.GetWorkspace.

LabeledItem
A single item in a labeled column.

ListRowsRequest
Request message for TablesService.ListRows.

ListRowsResponse
Response message for TablesService.ListRows.

ListTablesRequest
Request message for TablesService.ListTables.

ListTablesResponse
Response message for TablesService.ListTables.

ListWorkspacesRequest
Request message for TablesService.ListWorkspaces.

ListWorkspacesResponse
Response message for TablesService.ListWorkspaces.

LookupDetails
Details about a lookup column whose value comes from the associated relationship.

RelationshipDetails
Details about a relationship column.

Row
A single row in a table.

Table
A single table.

UpdateRowRequest
Request message for TablesService.UpdateRow.

Workspace
A single workspace.

Enums

View
Column identifier used for the values in the row.