Module googapis::google::bigtable::v2::mutation

Nested message and enum types in Mutation.

Structs

DeleteFromColumn
A Mutation which deletes cells from the specified column, optionally restricting the deletions to a given timestamp range.

DeleteFromFamily
A Mutation which deletes all cells from the specified column family.

DeleteFromRow
A Mutation which deletes all cells from the containing row.

SetCell
A Mutation which sets the value of the specified cell.

Enums

Mutation
Which of the possible Mutation types to apply.