Module googapis::google::bigtable::v2::row_filter

Nested message and enum types in RowFilter.

Structs

Chain

A RowFilter which sends rows through several RowFilters in sequence.

Condition

A RowFilter which evaluates one of two possible RowFilters, depending on whether or not a predicate RowFilter outputs any cells from the input row.

Interleave

A RowFilter which sends each row to each of several component RowFilters and interleaves the results.

Enums

Filter

Which of the possible RowFilter types to apply. If none are set, this RowFilter returns all cells in the input row.