Module googapis::google::chromeos::moblab::v1beta1

Modules

build

Nested message and enum types in Build.

build_service_client

Generated client implementations.

Structs

Build

Resource that represents a build for the given build target, model, milestone and build version.

BuildArtifact

Resource that represents a build artifact stored in Google Cloud Storage for the given build target, model, build version and bucket.

BuildTarget

Resource that represents a build target.

CheckBuildStageStatusRequest

Request message for checking if the build artifact is staged.

CheckBuildStageStatusResponse

Response message for checking the stage status of a build artifact.

FindMostStableBuildRequest

Request message for finding the most stable build.

FindMostStableBuildResponse

Response message for finding the most stable build.

ListBuildTargetsRequest

Request message for listing build targets.

ListBuildTargetsResponse

Response message for listing build targets.

ListBuildsRequest

Request message for listing builds.

ListBuildsResponse

Response message for listing builds.

ListModelsRequest

Request message for listing models.

ListModelsResponse

Response message for listing models.

Milestone

Resource that represents a chrome OS milestone.

Model

Resource that represents a model. Each model belongs to a build target. For non-unified build, the model name is the same as its build target name.

StageBuildMetadata

Metadata message for staging a build artifact.

StageBuildRequest

Request message for staging a build artifact.

StageBuildResponse

Response message for staging a build artifact.