Module googapis::google::cloud

Modules

accessapproval

aiplatform

apigateway

apigeeconnect

asset

assuredworkloads

audit

automl

bigquery

billing

binaryauthorization

channel

clouddms

common

compute

contactcenterinsights

datacatalog

datafusion

datalabeling

dataproc

dataqna

datastream

deploy

dialogflow

documentai

domains

essentialcontacts

eventarc

filestore

functions

gaming

gkebackup

gkeconnect

gkehub

gsuiteaddons

iap

identitytoolkit

ids

integrations

iot

kms

language

lifesciences

location

managedidentities

mediatranslation

memcache

metastore

ml

networkconnectivity

networkmanagement

networksecurity

networkservices

notebooks

orchestration

orgpolicy

osconfig

oslogin

phishingprotection

policytroubleshooter

privatecatalog

pubsublite

recaptchaenterprise

recommendationengine

recommender

redis

resourcemanager

resourcesettings

retail

runtimeconfig

saasaccelerator

scheduler

secretmanager

secrets

security

securitycenter

servicedirectory

shell

speech

sql

support

talent

tasks

texttospeech

tpu

translation

video

videointelligence

vision

vmmigration

vpcaccess

webrisk

websecurityscanner

workflows

Enums

OperationResponseMapping

An enum to be used to mark the essential (for polling) fields in an API-specific Operation object. A custom Operation object may contain many different fields, but only few of them are essential to conduct a successful polling process.