Module googapis::google::cloud::accessapproval::v1

Modules

access_approval_client
Generated client implementations.

access_reason
Nested message and enum types in AccessReason.

approval_request
Nested message and enum types in ApprovalRequest.

Structs

AccessApprovalSettings
Settings on a Project/Folder/Organization related to Access Approval.

AccessLocations
Home office and physical location of the principal.

AccessReason

ApprovalRequest
A request for the customer to approve access to a resource.

ApproveApprovalRequestMessage
Request to approve an ApprovalRequest.

ApproveDecision
A decision that has been made to approve access to a resource.

DeleteAccessApprovalSettingsMessage
Request to delete access approval settings.

DismissApprovalRequestMessage
Request to dismiss an approval request.

DismissDecision
A decision that has been made to dismiss an approval request.

EnrolledService
Represents the enrollment of a cloud resource into a specific service.

GetAccessApprovalSettingsMessage
Request to get access approval settings.

GetApprovalRequestMessage
Request to get an approval request.

ListApprovalRequestsMessage
Request to list approval requests.

ListApprovalRequestsResponse
Response to listing of ApprovalRequest objects.

ResourceProperties
The properties associated with the resource of the request.

UpdateAccessApprovalSettingsMessage
Request to update access approval settings.

Enums

EnrollmentLevel
Represents the type of enrollment for a given service to Access Approval.