Struct googapis::google::cloud::aiplatform::v1beta1::job_service_client::JobServiceClient

pub struct JobServiceClient<T> { /* fields omitted */ }

A service for creating and managing Vertex AI’s jobs.

Implementations

impl<T> JobServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + Sync + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> JobServiceClient<InterceptedService<T, F>> where
F: FnMut(Request<()>) -> Result<Request<()>, Status>,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_custom_job(
&mut self,
request: impl IntoRequest<CreateCustomJobRequest>
) -> Result<Response<CustomJob>, Status>

Creates a CustomJob. A created CustomJob right away will be attempted to be run.

pub async fn get_custom_job(
&mut self,
request: impl IntoRequest<GetCustomJobRequest>
) -> Result<Response<CustomJob>, Status>

Gets a CustomJob.

pub async fn list_custom_jobs(
&mut self,
request: impl IntoRequest<ListCustomJobsRequest>
) -> Result<Response<ListCustomJobsResponse>, Status>

Lists CustomJobs in a Location.

pub async fn delete_custom_job(
&mut self,
request: impl IntoRequest<DeleteCustomJobRequest>
) -> Result<Response<Operation>, Status>

Deletes a CustomJob.

pub async fn cancel_custom_job(
&mut self,
request: impl IntoRequest<CancelCustomJobRequest>
) -> Result<Response<()>, Status>

Cancels a CustomJob. Starts asynchronous cancellation on the CustomJob. The server makes a best effort to cancel the job, but success is not guaranteed. Clients can use [JobService.GetCustomJob][google.cloud.aiplatform.v1beta1.JobService.GetCustomJob] or other methods to check whether the cancellation succeeded or whether the job completed despite cancellation. On successful cancellation, the CustomJob is not deleted; instead it becomes a job with a [CustomJob.error][google.cloud.aiplatform.v1beta1.CustomJob.error] value with a [google.rpc.Status.code][google.rpc.Status.code] of 1, corresponding to Code.CANCELLED, and [CustomJob.state][google.cloud.aiplatform.v1beta1.CustomJob.state] is set to CANCELLED.

pub async fn create_data_labeling_job(
&mut self,
request: impl IntoRequest<CreateDataLabelingJobRequest>
) -> Result<Response<DataLabelingJob>, Status>

Creates a DataLabelingJob.

pub async fn get_data_labeling_job(
&mut self,
request: impl IntoRequest<GetDataLabelingJobRequest>
) -> Result<Response<DataLabelingJob>, Status>

Gets a DataLabelingJob.

pub async fn list_data_labeling_jobs(
&mut self,
request: impl IntoRequest<ListDataLabelingJobsRequest>
) -> Result<Response<ListDataLabelingJobsResponse>, Status>

Lists DataLabelingJobs in a Location.

pub async fn delete_data_labeling_job(
&mut self,
request: impl IntoRequest<DeleteDataLabelingJobRequest>
) -> Result<Response<Operation>, Status>

Deletes a DataLabelingJob.

pub async fn cancel_data_labeling_job(
&mut self,
request: impl IntoRequest<CancelDataLabelingJobRequest>
) -> Result<Response<()>, Status>

Cancels a DataLabelingJob. Success of cancellation is not guaranteed.

pub async fn create_hyperparameter_tuning_job(
&mut self,
request: impl IntoRequest<CreateHyperparameterTuningJobRequest>
) -> Result<Response<HyperparameterTuningJob>, Status>

Creates a HyperparameterTuningJob

pub async fn get_hyperparameter_tuning_job(
&mut self,
request: impl IntoRequest<GetHyperparameterTuningJobRequest>
) -> Result<Response<HyperparameterTuningJob>, Status>

Gets a HyperparameterTuningJob

pub async fn list_hyperparameter_tuning_jobs(
&mut self,
request: impl IntoRequest<ListHyperparameterTuningJobsRequest>
) -> Result<Response<ListHyperparameterTuningJobsResponse>, Status>

Lists HyperparameterTuningJobs in a Location.

pub async fn delete_hyperparameter_tuning_job(
&mut self,
request: impl IntoRequest<DeleteHyperparameterTuningJobRequest>
) -> Result<Response<Operation>, Status>

Deletes a HyperparameterTuningJob.

pub async fn cancel_hyperparameter_tuning_job(
&mut self,
request: impl IntoRequest<CancelHyperparameterTuningJobRequest>
) -> Result<Response<()>, Status>

Cancels a HyperparameterTuningJob. Starts asynchronous cancellation on the HyperparameterTuningJob. The server makes a best effort to cancel the job, but success is not guaranteed. Clients can use [JobService.GetHyperparameterTuningJob][google.cloud.aiplatform.v1beta1.JobService.GetHyperparameterTuningJob] or other methods to check whether the cancellation succeeded or whether the job completed despite cancellation. On successful cancellation, the HyperparameterTuningJob is not deleted; instead it becomes a job with a [HyperparameterTuningJob.error][google.cloud.aiplatform.v1beta1.HyperparameterTuningJob.error] value with a [google.rpc.Status.code][google.rpc.Status.code] of 1, corresponding to Code.CANCELLED, and [HyperparameterTuningJob.state][google.cloud.aiplatform.v1beta1.HyperparameterTuningJob.state] is set to CANCELLED.

pub async fn create_batch_prediction_job(
&mut self,
request: impl IntoRequest<CreateBatchPredictionJobRequest>
) -> Result<Response<BatchPredictionJob>, Status>

Creates a BatchPredictionJob. A BatchPredictionJob once created will right away be attempted to start.

pub async fn get_batch_prediction_job(
&mut self,
request: impl IntoRequest<GetBatchPredictionJobRequest>
) -> Result<Response<BatchPredictionJob>, Status>

Gets a BatchPredictionJob

pub async fn list_batch_prediction_jobs(
&mut self,
request: impl IntoRequest<ListBatchPredictionJobsRequest>
) -> Result<Response<ListBatchPredictionJobsResponse>, Status>

Lists BatchPredictionJobs in a Location.

pub async fn delete_batch_prediction_job(
&mut self,
request: impl IntoRequest<DeleteBatchPredictionJobRequest>
) -> Result<Response<Operation>, Status>

Deletes a BatchPredictionJob. Can only be called on jobs that already finished.

pub async fn cancel_batch_prediction_job(
&mut self,
request: impl IntoRequest<CancelBatchPredictionJobRequest>
) -> Result<Response<()>, Status>

Cancels a BatchPredictionJob.

Starts asynchronous cancellation on the BatchPredictionJob. The server makes the best effort to cancel the job, but success is not guaranteed. Clients can use [JobService.GetBatchPredictionJob][google.cloud.aiplatform.v1beta1.JobService.GetBatchPredictionJob] or other methods to check whether the cancellation succeeded or whether the job completed despite cancellation. On a successful cancellation, the BatchPredictionJob is not deleted;instead its [BatchPredictionJob.state][google.cloud.aiplatform.v1beta1.BatchPredictionJob.state] is set to CANCELLED. Any files already outputted by the job are not deleted.

pub async fn create_model_deployment_monitoring_job(
&mut self,
request: impl IntoRequest<CreateModelDeploymentMonitoringJobRequest>
) -> Result<Response<ModelDeploymentMonitoringJob>, Status>

Creates a ModelDeploymentMonitoringJob. It will run periodically on a configured interval.

pub async fn search_model_deployment_monitoring_stats_anomalies(
&mut self,
request: impl IntoRequest<SearchModelDeploymentMonitoringStatsAnomaliesRequest>
) -> Result<Response<SearchModelDeploymentMonitoringStatsAnomaliesResponse>, Status>

Searches Model Monitoring Statistics generated within a given time window.

pub async fn get_model_deployment_monitoring_job(
&mut self,
request: impl IntoRequest<GetModelDeploymentMonitoringJobRequest>
) -> Result<Response<ModelDeploymentMonitoringJob>, Status>

Gets a ModelDeploymentMonitoringJob.

pub async fn list_model_deployment_monitoring_jobs(
&mut self,
request: impl IntoRequest<ListModelDeploymentMonitoringJobsRequest>
) -> Result<Response<ListModelDeploymentMonitoringJobsResponse>, Status>

Lists ModelDeploymentMonitoringJobs in a Location.

pub async fn update_model_deployment_monitoring_job(
&mut self,
request: impl IntoRequest<UpdateModelDeploymentMonitoringJobRequest>
) -> Result<Response<Operation>, Status>

Updates a ModelDeploymentMonitoringJob.

pub async fn delete_model_deployment_monitoring_job(
&mut self,
request: impl IntoRequest<DeleteModelDeploymentMonitoringJobRequest>
) -> Result<Response<Operation>, Status>

Deletes a ModelDeploymentMonitoringJob.

pub async fn pause_model_deployment_monitoring_job(
&mut self,
request: impl IntoRequest<PauseModelDeploymentMonitoringJobRequest>
) -> Result<Response<()>, Status>

Pauses a ModelDeploymentMonitoringJob. If the job is running, the server makes a best effort to cancel the job. Will mark [ModelDeploymentMonitoringJob.state][google.cloud.aiplatform.v1beta1.ModelDeploymentMonitoringJob.state] to ‘PAUSED’.

pub async fn resume_model_deployment_monitoring_job(
&mut self,
request: impl IntoRequest<ResumeModelDeploymentMonitoringJobRequest>
) -> Result<Response<()>, Status>

Resumes a paused ModelDeploymentMonitoringJob. It will start to run from next scheduled time. A deleted ModelDeploymentMonitoringJob can’t be resumed.