Struct googapis::google::cloud::aiplatform::v1beta1::metadata_service_client::MetadataServiceClient

pub struct MetadataServiceClient<T> { /* fields omitted */ }

Service for reading and writing metadata entries.

Implementations

impl<T> MetadataServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + Sync + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> MetadataServiceClient<InterceptedService<T, F>> where
F: FnMut(Request<()>) -> Result<Request<()>, Status>,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_metadata_store(
&mut self,
request: impl IntoRequest<CreateMetadataStoreRequest>
) -> Result<Response<Operation>, Status>

Initializes a MetadataStore, including allocation of resources.

pub async fn get_metadata_store(
&mut self,
request: impl IntoRequest<GetMetadataStoreRequest>
) -> Result<Response<MetadataStore>, Status>

Retrieves a specific MetadataStore.

pub async fn list_metadata_stores(
&mut self,
request: impl IntoRequest<ListMetadataStoresRequest>
) -> Result<Response<ListMetadataStoresResponse>, Status>

Lists MetadataStores for a Location.

pub async fn delete_metadata_store(
&mut self,
request: impl IntoRequest<DeleteMetadataStoreRequest>
) -> Result<Response<Operation>, Status>

Deletes a single MetadataStore.

pub async fn create_artifact(
&mut self,
request: impl IntoRequest<CreateArtifactRequest>
) -> Result<Response<Artifact>, Status>

Creates an Artifact associated with a MetadataStore.

pub async fn get_artifact(
&mut self,
request: impl IntoRequest<GetArtifactRequest>
) -> Result<Response<Artifact>, Status>

Retrieves a specific Artifact.

pub async fn list_artifacts(
&mut self,
request: impl IntoRequest<ListArtifactsRequest>
) -> Result<Response<ListArtifactsResponse>, Status>

Lists Artifacts in the MetadataStore.

pub async fn update_artifact(
&mut self,
request: impl IntoRequest<UpdateArtifactRequest>
) -> Result<Response<Artifact>, Status>

Updates a stored Artifact.

pub async fn delete_artifact(
&mut self,
request: impl IntoRequest<DeleteArtifactRequest>
) -> Result<Response<Operation>, Status>

Deletes an Artifact.

pub async fn purge_artifacts(
&mut self,
request: impl IntoRequest<PurgeArtifactsRequest>
) -> Result<Response<Operation>, Status>

Purges Artifacts.

pub async fn create_context(
&mut self,
request: impl IntoRequest<CreateContextRequest>
) -> Result<Response<Context>, Status>

Creates a Context associated with a MetadataStore.

pub async fn get_context(
&mut self,
request: impl IntoRequest<GetContextRequest>
) -> Result<Response<Context>, Status>

Retrieves a specific Context.

pub async fn list_contexts(
&mut self,
request: impl IntoRequest<ListContextsRequest>
) -> Result<Response<ListContextsResponse>, Status>

Lists Contexts on the MetadataStore.

pub async fn update_context(
&mut self,
request: impl IntoRequest<UpdateContextRequest>
) -> Result<Response<Context>, Status>

Updates a stored Context.

pub async fn delete_context(
&mut self,
request: impl IntoRequest<DeleteContextRequest>
) -> Result<Response<Operation>, Status>

Deletes a stored Context.

pub async fn purge_contexts(
&mut self,
request: impl IntoRequest<PurgeContextsRequest>
) -> Result<Response<Operation>, Status>

Purges Contexts.

pub async fn add_context_artifacts_and_executions(
&mut self,
request: impl IntoRequest<AddContextArtifactsAndExecutionsRequest>
) -> Result<Response<AddContextArtifactsAndExecutionsResponse>, Status>

Adds a set of Artifacts and Executions to a Context. If any of the Artifacts or Executions have already been added to a Context, they are simply skipped.

pub async fn add_context_children(
&mut self,
request: impl IntoRequest<AddContextChildrenRequest>
) -> Result<Response<AddContextChildrenResponse>, Status>

Adds a set of Contexts as children to a parent Context. If any of the child Contexts have already been added to the parent Context, they are simply skipped. If this call would create a cycle or cause any Context to have more than 10 parents, the request will fail with an INVALID_ARGUMENT error.

pub async fn query_context_lineage_subgraph(
&mut self,
request: impl IntoRequest<QueryContextLineageSubgraphRequest>
) -> Result<Response<LineageSubgraph>, Status>

Retrieves Artifacts and Executions within the specified Context, connected by Event edges and returned as a LineageSubgraph.

pub async fn create_execution(
&mut self,
request: impl IntoRequest<CreateExecutionRequest>
) -> Result<Response<Execution>, Status>

Creates an Execution associated with a MetadataStore.

pub async fn get_execution(
&mut self,
request: impl IntoRequest<GetExecutionRequest>
) -> Result<Response<Execution>, Status>

Retrieves a specific Execution.

pub async fn list_executions(
&mut self,
request: impl IntoRequest<ListExecutionsRequest>
) -> Result<Response<ListExecutionsResponse>, Status>

Lists Executions in the MetadataStore.

pub async fn update_execution(
&mut self,
request: impl IntoRequest<UpdateExecutionRequest>
) -> Result<Response<Execution>, Status>

Updates a stored Execution.

pub async fn delete_execution(
&mut self,
request: impl IntoRequest<DeleteExecutionRequest>
) -> Result<Response<Operation>, Status>

Deletes an Execution.

pub async fn purge_executions(
&mut self,
request: impl IntoRequest<PurgeExecutionsRequest>
) -> Result<Response<Operation>, Status>

Purges Executions.

pub async fn add_execution_events(
&mut self,
request: impl IntoRequest<AddExecutionEventsRequest>
) -> Result<Response<AddExecutionEventsResponse>, Status>

Adds Events to the specified Execution. An Event indicates whether an Artifact was used as an input or output for an Execution. If an Event already exists between the Execution and the Artifact, the Event is skipped.

pub async fn query_execution_inputs_and_outputs(
&mut self,
request: impl IntoRequest<QueryExecutionInputsAndOutputsRequest>
) -> Result<Response<LineageSubgraph>, Status>

Obtains the set of input and output Artifacts for this Execution, in the form of LineageSubgraph that also contains the Execution and connecting Events.

pub async fn create_metadata_schema(
&mut self,
request: impl IntoRequest<CreateMetadataSchemaRequest>
) -> Result<Response<MetadataSchema>, Status>

Creates a MetadataSchema.

pub async fn get_metadata_schema(
&mut self,
request: impl IntoRequest<GetMetadataSchemaRequest>
) -> Result<Response<MetadataSchema>, Status>

Retrieves a specific MetadataSchema.

pub async fn list_metadata_schemas(
&mut self,
request: impl IntoRequest<ListMetadataSchemasRequest>
) -> Result<Response<ListMetadataSchemasResponse>, Status>

Lists MetadataSchemas.

pub async fn query_artifact_lineage_subgraph(
&mut self,
request: impl IntoRequest<QueryArtifactLineageSubgraphRequest>
) -> Result<Response<LineageSubgraph>, Status>

Retrieves lineage of an Artifact represented through Artifacts and Executions connected by Event edges and returned as a LineageSubgraph.