Struct googapis::google::cloud::aiplatform::v1beta1::tensorboard_service_client::TensorboardServiceClient

pub struct TensorboardServiceClient<T> { /* fields omitted */ }

TensorboardService

Implementations

impl<T> TensorboardServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> TensorboardServiceClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_tensorboard(
&mut self,
request: impl IntoRequest<CreateTensorboardRequest>
) -> Result<Response<Operation>, Status>

Creates a Tensorboard.

pub async fn get_tensorboard(
&mut self,
request: impl IntoRequest<GetTensorboardRequest>
) -> Result<Response<Tensorboard>, Status>

Gets a Tensorboard.

pub async fn update_tensorboard(
&mut self,
request: impl IntoRequest<UpdateTensorboardRequest>
) -> Result<Response<Operation>, Status>

Updates a Tensorboard.

pub async fn list_tensorboards(
&mut self,
request: impl IntoRequest<ListTensorboardsRequest>
) -> Result<Response<ListTensorboardsResponse>, Status>

Lists Tensorboards in a Location.

pub async fn delete_tensorboard(
&mut self,
request: impl IntoRequest<DeleteTensorboardRequest>
) -> Result<Response<Operation>, Status>

Deletes a Tensorboard.

pub async fn create_tensorboard_experiment(
&mut self,
request: impl IntoRequest<CreateTensorboardExperimentRequest>
) -> Result<Response<TensorboardExperiment>, Status>

Creates a TensorboardExperiment.

pub async fn get_tensorboard_experiment(
&mut self,
request: impl IntoRequest<GetTensorboardExperimentRequest>
) -> Result<Response<TensorboardExperiment>, Status>

Gets a TensorboardExperiment.

pub async fn update_tensorboard_experiment(
&mut self,
request: impl IntoRequest<UpdateTensorboardExperimentRequest>
) -> Result<Response<TensorboardExperiment>, Status>

Updates a TensorboardExperiment.

pub async fn list_tensorboard_experiments(
&mut self,
request: impl IntoRequest<ListTensorboardExperimentsRequest>
) -> Result<Response<ListTensorboardExperimentsResponse>, Status>

Lists TensorboardExperiments in a Location.

pub async fn delete_tensorboard_experiment(
&mut self,
request: impl IntoRequest<DeleteTensorboardExperimentRequest>
) -> Result<Response<Operation>, Status>

Deletes a TensorboardExperiment.

pub async fn create_tensorboard_run(
&mut self,
request: impl IntoRequest<CreateTensorboardRunRequest>
) -> Result<Response<TensorboardRun>, Status>

Creates a TensorboardRun.

pub async fn batch_create_tensorboard_runs(
&mut self,
request: impl IntoRequest<BatchCreateTensorboardRunsRequest>
) -> Result<Response<BatchCreateTensorboardRunsResponse>, Status>

Batch create TensorboardRuns.

pub async fn get_tensorboard_run(
&mut self,
request: impl IntoRequest<GetTensorboardRunRequest>
) -> Result<Response<TensorboardRun>, Status>

Gets a TensorboardRun.

pub async fn update_tensorboard_run(
&mut self,
request: impl IntoRequest<UpdateTensorboardRunRequest>
) -> Result<Response<TensorboardRun>, Status>

Updates a TensorboardRun.

pub async fn list_tensorboard_runs(
&mut self,
request: impl IntoRequest<ListTensorboardRunsRequest>
) -> Result<Response<ListTensorboardRunsResponse>, Status>

Lists TensorboardRuns in a Location.

pub async fn delete_tensorboard_run(
&mut self,
request: impl IntoRequest<DeleteTensorboardRunRequest>
) -> Result<Response<Operation>, Status>

Deletes a TensorboardRun.

pub async fn batch_create_tensorboard_time_series(
&mut self,
request: impl IntoRequest<BatchCreateTensorboardTimeSeriesRequest>
) -> Result<Response<BatchCreateTensorboardTimeSeriesResponse>, Status>

Batch create TensorboardTimeSeries that belong to a TensorboardExperiment.

pub async fn create_tensorboard_time_series(
&mut self,
request: impl IntoRequest<CreateTensorboardTimeSeriesRequest>
) -> Result<Response<TensorboardTimeSeries>, Status>

Creates a TensorboardTimeSeries.

pub async fn get_tensorboard_time_series(
&mut self,
request: impl IntoRequest<GetTensorboardTimeSeriesRequest>
) -> Result<Response<TensorboardTimeSeries>, Status>

Gets a TensorboardTimeSeries.

pub async fn update_tensorboard_time_series(
&mut self,
request: impl IntoRequest<UpdateTensorboardTimeSeriesRequest>
) -> Result<Response<TensorboardTimeSeries>, Status>

Updates a TensorboardTimeSeries.

pub async fn list_tensorboard_time_series(
&mut self,
request: impl IntoRequest<ListTensorboardTimeSeriesRequest>
) -> Result<Response<ListTensorboardTimeSeriesResponse>, Status>

Lists TensorboardTimeSeries in a Location.

pub async fn delete_tensorboard_time_series(
&mut self,
request: impl IntoRequest<DeleteTensorboardTimeSeriesRequest>
) -> Result<Response<Operation>, Status>

Deletes a TensorboardTimeSeries.

pub async fn batch_read_tensorboard_time_series_data(
&mut self,
request: impl IntoRequest<BatchReadTensorboardTimeSeriesDataRequest>
) -> Result<Response<BatchReadTensorboardTimeSeriesDataResponse>, Status>

Reads multiple TensorboardTimeSeries’ data. The data point number limit is 1000 for scalars, 100 for tensors and blob references. If the number of data points stored is less than the limit, all data will be returned. Otherwise, that limit number of data points will be randomly selected from this time series and returned.

pub async fn read_tensorboard_time_series_data(
&mut self,
request: impl IntoRequest<ReadTensorboardTimeSeriesDataRequest>
) -> Result<Response<ReadTensorboardTimeSeriesDataResponse>, Status>

Reads a TensorboardTimeSeries’ data. By default, if the number of data points stored is less than 1000, all data will be returned. Otherwise, 1000 data points will be randomly selected from this time series and returned. This value can be changed by changing max_data_points, which can’t be greater than 10k.

pub async fn read_tensorboard_blob_data(
&mut self,
request: impl IntoRequest<ReadTensorboardBlobDataRequest>
) -> Result<Response<Streaming<ReadTensorboardBlobDataResponse>>, Status>

Gets bytes of TensorboardBlobs. This is to allow reading blob data stored in consumer project’s Cloud Storage bucket without users having to obtain Cloud Storage access permission.

pub async fn write_tensorboard_experiment_data(
&mut self,
request: impl IntoRequest<WriteTensorboardExperimentDataRequest>
) -> Result<Response<WriteTensorboardExperimentDataResponse>, Status>

Write time series data points of multiple TensorboardTimeSeries in multiple TensorboardRun’s. If any data fail to be ingested, an error will be returned.

pub async fn write_tensorboard_run_data(
&mut self,
request: impl IntoRequest<WriteTensorboardRunDataRequest>
) -> Result<Response<WriteTensorboardRunDataResponse>, Status>

Write time series data points into multiple TensorboardTimeSeries under a TensorboardRun. If any data fail to be ingested, an error will be returned.

pub async fn export_tensorboard_time_series_data(
&mut self,
request: impl IntoRequest<ExportTensorboardTimeSeriesDataRequest>
) -> Result<Response<ExportTensorboardTimeSeriesDataResponse>, Status>

Exports a TensorboardTimeSeries’ data. Data is returned in paginated responses.