Struct googapis::google::cloud::asset::v1::asset_service_client::AssetServiceClient

pub struct AssetServiceClient<T> { /* fields omitted */ }

Asset service definition.

Implementations

impl<T> AssetServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + Sync + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> AssetServiceClient<InterceptedService<T, F>> where
F: FnMut(Request<()>) -> Result<Request<()>, Status>,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn export_assets(
&mut self,
request: impl IntoRequest<ExportAssetsRequest>
) -> Result<Response<Operation>, Status>

Exports assets with time and resource types to a given Cloud Storage location/BigQuery table. For Cloud Storage location destinations, the output format is newline-delimited JSON. Each line represents a [google.cloud.asset.v1.Asset][google.cloud.asset.v1.Asset] in the JSON format; for BigQuery table destinations, the output table stores the fields in asset proto as columns. This API implements the [google.longrunning.Operation][google.longrunning.Operation] API , which allows you to keep track of the export. We recommend intervals of at least 2 seconds with exponential retry to poll the export operation result. For regular-size resource parent, the export operation usually finishes within 5 minutes.

pub async fn list_assets(
&mut self,
request: impl IntoRequest<ListAssetsRequest>
) -> Result<Response<ListAssetsResponse>, Status>

Lists assets with time and resource types and returns paged results in response.

pub async fn batch_get_assets_history(
&mut self,
request: impl IntoRequest<BatchGetAssetsHistoryRequest>
) -> Result<Response<BatchGetAssetsHistoryResponse>, Status>

Batch gets the update history of assets that overlap a time window. For IAM_POLICY content, this API outputs history when the asset and its attached IAM POLICY both exist. This can create gaps in the output history. Otherwise, this API outputs history with asset in both non-delete or deleted status. If a specified asset does not exist, this API returns an INVALID_ARGUMENT error.

pub async fn create_feed(
&mut self,
request: impl IntoRequest<CreateFeedRequest>
) -> Result<Response<Feed>, Status>

Creates a feed in a parent project/folder/organization to listen to its asset updates.

pub async fn get_feed(
&mut self,
request: impl IntoRequest<GetFeedRequest>
) -> Result<Response<Feed>, Status>

Gets details about an asset feed.

pub async fn list_feeds(
&mut self,
request: impl IntoRequest<ListFeedsRequest>
) -> Result<Response<ListFeedsResponse>, Status>

Lists all asset feeds in a parent project/folder/organization.

pub async fn update_feed(
&mut self,
request: impl IntoRequest<UpdateFeedRequest>
) -> Result<Response<Feed>, Status>

Updates an asset feed configuration.

pub async fn delete_feed(
&mut self,
request: impl IntoRequest<DeleteFeedRequest>
) -> Result<Response<()>, Status>

Deletes an asset feed.

pub async fn search_all_resources(
&mut self,
request: impl IntoRequest<SearchAllResourcesRequest>
) -> Result<Response<SearchAllResourcesResponse>, Status>

Searches all Cloud resources within the specified scope, such as a project, folder, or organization. The caller must be granted the cloudasset.assets.searchAllResources permission on the desired scope, otherwise the request will be rejected.

pub async fn search_all_iam_policies(
&mut self,
request: impl IntoRequest<SearchAllIamPoliciesRequest>
) -> Result<Response<SearchAllIamPoliciesResponse>, Status>

Searches all IAM policies within the specified scope, such as a project, folder, or organization. The caller must be granted the cloudasset.assets.searchAllIamPolicies permission on the desired scope, otherwise the request will be rejected.

pub async fn analyze_iam_policy(
&mut self,
request: impl IntoRequest<AnalyzeIamPolicyRequest>
) -> Result<Response<AnalyzeIamPolicyResponse>, Status>

Analyzes IAM policies to answer which identities have what accesses on which resources.

pub async fn analyze_iam_policy_longrunning(
&mut self,
request: impl IntoRequest<AnalyzeIamPolicyLongrunningRequest>
) -> Result<Response<Operation>, Status>

Analyzes IAM policies asynchronously to answer which identities have what accesses on which resources, and writes the analysis results to a Google Cloud Storage or a BigQuery destination. For Cloud Storage destination, the output format is the JSON format that represents a [AnalyzeIamPolicyResponse][google.cloud.asset.v1.AnalyzeIamPolicyResponse]. This method implements the [google.longrunning.Operation][google.longrunning.Operation], which allows you to track the operation status. We recommend intervals of at least 2 seconds with exponential backoff retry to poll the operation result. The metadata contains the metadata for the long-running operation.

pub async fn analyze_move(
&mut self,
request: impl IntoRequest<AnalyzeMoveRequest>
) -> Result<Response<AnalyzeMoveResponse>, Status>

Analyze moving a resource to a specified destination without kicking off the actual move. The analysis is best effort depending on the user’s permissions of viewing different hierarchical policies and configurations. The policies and configuration are subject to change before the actual resource migration takes place.