Module googapis::google::cloud::asset::v1::iam_policy_analysis_query

Nested message and enum types in IamPolicyAnalysisQuery.

Modules

condition_context

Nested message and enum types in ConditionContext.

Structs

AccessSelector

Specifies roles and/or permissions to analyze, to determine both the identities possessing them and the resources they control. If multiple values are specified, results will include roles or permissions matching any of them. The total number of roles and permissions should be equal or less than 10.

ConditionContext

The IAM conditions context.

IdentitySelector

Specifies an identity for which to determine resource access, based on roles assigned either directly to them or to the groups they belong to, directly or indirectly.

Options

Contains query options.

ResourceSelector

Specifies the resource to analyze for access policies, which may be set directly on the resource, or on ancestors such as organizations, folders or projects.