Module googapis::google::cloud::bigquery::logging::v1

Modules

audit_data

Nested message and enum types in AuditData.

big_query_acl

Nested message and enum types in BigQueryAcl.

job_configuration

Nested message and enum types in JobConfiguration.

job_statistics

Nested message and enum types in JobStatistics.

Structs

AuditData

BigQuery request and response messages for audit log. Note: Table.schema has been deprecated in favor of Table.schemaJson. Table.schema may continue to be present in your logs during this transition.

BigQueryAcl

An access control list.

Dataset

BigQuery dataset information. See the [Dataset](/bigquery/docs/reference/v2/datasets) API resource for more details on individual fields.

DatasetInfo

User-provided metadata for a dataset.

DatasetInsertRequest

Dataset insert request.

DatasetInsertResponse

Dataset insert response.

DatasetListRequest

Dataset list request.

DatasetName

The fully-qualified name for a dataset.

DatasetUpdateRequest

Dataset update request.

DatasetUpdateResponse

Dataset update response.

EncryptionInfo

Describes encryption properties for a table or a job

Job

Describes a job.

JobCompletedEvent

Query job completed event.

JobConfiguration

Job configuration information. See the [Jobs](/bigquery/docs/reference/v2/jobs) API resource for more details on individual fields.

JobGetQueryResultsRequest

Job getQueryResults request.

JobGetQueryResultsResponse

Job getQueryResults response.

JobInsertRequest

Job insert request.

JobInsertResponse

Job insert response.

JobName

The fully-qualified name for a job.

JobQueryDoneResponse

Job getQueryDone response.

JobQueryRequest

Job query request.

JobQueryResponse

Job query response.

JobStatistics

Job statistics that may change after a job starts.

JobStatus

Running state of a job.

Table

Describes a BigQuery table. See the [Table](/bigquery/docs/reference/v2/tables) API resource for more details on individual fields. Note: Table.schema has been deprecated in favor of Table.schemaJson. Table.schema may continue to be present in your logs during this transition.

TableDataListRequest

Table data-list request.

TableDataReadEvent

Table data read event. Only present for tables, not views, and is only included in the log record for the project that owns the table.

TableDefinition

Describes an external data source used in a query.

TableInfo

User-provided metadata for a table.

TableInsertRequest

Table insert request.

TableInsertResponse

Table insert response.

TableName

The fully-qualified name for a table.

TableUpdateRequest

Table update request.

TableUpdateResponse

Table update response.

TableViewDefinition

Describes a virtual table defined by a SQL query.