Module googapis::google::cloud::bigquery::migration::v2alpha

Modules

migration_service_client
Generated client implementations.

migration_subtask
Nested message and enum types in MigrationSubtask.

migration_task
Nested message and enum types in MigrationTask.

migration_workflow
Nested message and enum types in MigrationWorkflow.

typed_value
Nested message and enum types in TypedValue.

Structs

CreateMigrationWorkflowRequest
Request to create a migration workflow resource.

DeleteMigrationWorkflowRequest
A request to delete a previously created migration workflow.

ErrorDetail
Provides details for errors, e.g. issues that where encountered when processing a subtask.

ErrorLocation
Holds information about where the error is located.

GetMigrationSubtaskRequest
A request to get a previously created migration subtasks.

GetMigrationWorkflowRequest
A request to get a previously created migration workflow.

ListMigrationSubtasksRequest
A request to list previously created migration subtasks.

ListMigrationSubtasksResponse
Response object for a ListMigrationSubtasks call.

ListMigrationWorkflowsRequest
A request to list previously created migration workflows.

ListMigrationWorkflowsResponse
Response object for a ListMigrationWorkflows call.

MigrationSubtask
A subtask for a migration which carries details about the configuration of the subtask. The content of the details should not matter to the end user, but is a contract between the subtask creator and subtask worker.

MigrationTask
A single task for a migration which has details about the configuration of the task.

MigrationWorkflow
A migration workflow which specifies what needs to be done for an EDW migration.

Point
A single data point in a time series.

ResourceErrorDetail
Provides details for errors and the corresponding resources.

StartMigrationWorkflowRequest
A request to start a previously created migration workflow.

TimeInterval
A time interval extending just after a start time through an end time. If the start time is the same as the end time, then the interval represents a single point in time.

TimeSeries
The metrics object for a SubTask.

TypedValue
A single strongly-typed value.