Module googapis::google::cloud::bigquery::reservation::v1beta1

Modules

assignment

Nested message and enum types in Assignment.

capacity_commitment

Nested message and enum types in CapacityCommitment.

reservation_service_client

Generated client implementations.

Structs

Assignment

A Assignment allows a project to submit jobs of a certain type using slots from the specified reservation.

BiReservation

Represents a BI Reservation.

CapacityCommitment

Capacity commitment is a way to purchase compute capacity for BigQuery jobs (in the form of slots) with some committed period of usage. Annual commitments renew by default. Commitments can be removed after their commitment end time passes.

CreateAssignmentRequest

The request for [ReservationService.CreateAssignment][google.cloud.bigquery.reservation.v1beta1.ReservationService.CreateAssignment]. Note: “bigquery.reservationAssignments.create” permission is required on the related assignee.

CreateCapacityCommitmentRequest

The request for [ReservationService.CreateCapacityCommitment][google.cloud.bigquery.reservation.v1beta1.ReservationService.CreateCapacityCommitment].

CreateReservationRequest

The request for [ReservationService.CreateReservation][google.cloud.bigquery.reservation.v1beta1.ReservationService.CreateReservation].

DeleteAssignmentRequest

The request for [ReservationService.DeleteAssignment][google.cloud.bigquery.reservation.v1beta1.ReservationService.DeleteAssignment]. Note: “bigquery.reservationAssignments.delete” permission is required on the related assignee.

DeleteCapacityCommitmentRequest

The request for [ReservationService.DeleteCapacityCommitment][google.cloud.bigquery.reservation.v1beta1.ReservationService.DeleteCapacityCommitment].

DeleteReservationRequest

The request for [ReservationService.DeleteReservation][google.cloud.bigquery.reservation.v1beta1.ReservationService.DeleteReservation].

GetBiReservationRequest

A request to get a singleton BI reservation.

GetCapacityCommitmentRequest

The request for [ReservationService.GetCapacityCommitment][google.cloud.bigquery.reservation.v1beta1.ReservationService.GetCapacityCommitment].

GetReservationRequest

The request for [ReservationService.GetReservation][google.cloud.bigquery.reservation.v1beta1.ReservationService.GetReservation].

ListAssignmentsRequest

The request for [ReservationService.ListAssignments][google.cloud.bigquery.reservation.v1beta1.ReservationService.ListAssignments].

ListAssignmentsResponse

The response for [ReservationService.ListAssignments][google.cloud.bigquery.reservation.v1beta1.ReservationService.ListAssignments].

ListCapacityCommitmentsRequest

The request for [ReservationService.ListCapacityCommitments][google.cloud.bigquery.reservation.v1beta1.ReservationService.ListCapacityCommitments].

ListCapacityCommitmentsResponse

The response for [ReservationService.ListCapacityCommitments][google.cloud.bigquery.reservation.v1beta1.ReservationService.ListCapacityCommitments].

ListReservationsRequest

The request for [ReservationService.ListReservations][google.cloud.bigquery.reservation.v1beta1.ReservationService.ListReservations].

ListReservationsResponse

The response for [ReservationService.ListReservations][google.cloud.bigquery.reservation.v1beta1.ReservationService.ListReservations].

MergeCapacityCommitmentsRequest

The request for [ReservationService.MergeCapacityCommitments][google.cloud.bigquery.reservation.v1beta1.ReservationService.MergeCapacityCommitments].

MoveAssignmentRequest

The request for [ReservationService.MoveAssignment][google.cloud.bigquery.reservation.v1beta1.ReservationService.MoveAssignment].

Reservation

A reservation is a mechanism used to guarantee slots to users.

SearchAssignmentsRequest

The request for [ReservationService.SearchAssignments][google.cloud.bigquery.reservation.v1beta1.ReservationService.SearchAssignments]. Note: “bigquery.reservationAssignments.search” permission is required on the related assignee.

SearchAssignmentsResponse

The response for [ReservationService.SearchAssignments][google.cloud.bigquery.reservation.v1beta1.ReservationService.SearchAssignments].

SplitCapacityCommitmentRequest

The request for [ReservationService.SplitCapacityCommitment][google.cloud.bigquery.reservation.v1beta1.ReservationService.SplitCapacityCommitment].

SplitCapacityCommitmentResponse

The response for [ReservationService.SplitCapacityCommitment][google.cloud.bigquery.reservation.v1beta1.ReservationService.SplitCapacityCommitment].

UpdateBiReservationRequest

A request to update a BI reservation.

UpdateCapacityCommitmentRequest

The request for [ReservationService.UpdateCapacityCommitment][google.cloud.bigquery.reservation.v1beta1.ReservationService.UpdateCapacityCommitment].

UpdateReservationRequest

The request for [ReservationService.UpdateReservation][google.cloud.bigquery.reservation.v1beta1.ReservationService.UpdateReservation].