Module googapis::google::cloud::bigquery::storage::v1

Modules

arrow_serialization_options
Nested message and enum types in ArrowSerializationOptions.

big_query_read_client
Generated client implementations.

read_rows_response
Nested message and enum types in ReadRowsResponse.

read_session
Nested message and enum types in ReadSession.

stream_stats
Nested message and enum types in StreamStats.

Structs

ArrowRecordBatch
Arrow RecordBatch.

ArrowSchema
Arrow schema as specified in https://arrow.apache.org/docs/python/api/datatypes.html and serialized to bytes using IPC: https://arrow.apache.org/docs/format/Columnar.html#serialization-and-interprocess-communication-ipc

ArrowSerializationOptions
Contains options specific to Arrow Serialization.

AvroRows
Avro rows.

AvroSchema
Avro schema.

CreateReadSessionRequest
Request message for CreateReadSession.

ReadRowsRequest
Request message for ReadRows.

ReadRowsResponse
Response from calling ReadRows may include row data, progress and throttling information.

ReadSession
Information about the ReadSession.

ReadStream
Information about a single stream that gets data out of the storage system. Most of the information about ReadStream instances is aggregated, making ReadStream lightweight.

SplitReadStreamRequest
Request message for SplitReadStream.

SplitReadStreamResponse
Response message for SplitReadStream.

StreamStats
Estimated stream statistics for a given Stream.

ThrottleState
Information on if the current connection is being throttled.

Enums

DataFormat
Data format for input or output data.