Struct googapis::google::cloud::bigquery::v2::model::training_run::TrainingOptions

pub struct TrainingOptions {

    pub max_iterations: i64,
    pub loss_type: i32,
    pub learn_rate: f64,
    pub l1_regularization: Option<f64>,
    pub l2_regularization: Option<f64>,
    pub min_relative_progress: Option<f64>,
    pub warm_start: Option<bool>,
    pub early_stop: Option<bool>,
    pub input_label_columns: Vec<String>,
    pub data_split_method: i32,
    pub data_split_eval_fraction: f64,
    pub data_split_column: String,
    pub learn_rate_strategy: i32,
    pub initial_learn_rate: f64,
    pub label_class_weights: HashMap<String, f64>,
    pub user_column: String,
    pub item_column: String,
    pub distance_type: i32,
    pub num_clusters: i64,
    pub model_uri: String,
    pub optimization_strategy: i32,
    pub hidden_units: Vec<i64>,
    pub batch_size: i64,
    pub dropout: Option<f64>,
    pub max_tree_depth: i64,
    pub subsample: f64,
    pub min_split_loss: Option<f64>,
    pub num_factors: i64,
    pub feedback_type: i32,
    pub wals_alpha: Option<f64>,
    pub kmeans_initialization_method: i32,
    pub kmeans_initialization_column: String,
    pub time_series_timestamp_column: String,
    pub time_series_data_column: String,
    pub auto_arima: bool,
    pub non_seasonal_order: Option<ArimaOrder>,
    pub data_frequency: i32,
    pub include_drift: bool,
    pub holiday_region: i32,
    pub time_series_id_column: String,
    pub time_series_id_columns: Vec<String>,
    pub horizon: i64,
    pub preserve_input_structs: bool,
    pub auto_arima_max_order: i64,
    pub decompose_time_series: Option<bool>,
    pub clean_spikes_and_dips: Option<bool>,
    pub adjust_step_changes: Option<bool>,

}

Options used in model training.

Fields
max_iterations: i64

The maximum number of iterations in training. Used only for iterative training algorithms.

loss_type: i32

Type of loss function used during training run.

learn_rate: f64

Learning rate in training. Used only for iterative training algorithms.

l1_regularization: Option<f64>

L1 regularization coefficient.

l2_regularization: Option<f64>

L2 regularization coefficient.

min_relative_progress: Option<f64>

When early_stop is true, stops training when accuracy improvement is less than ‘min_relative_progress’. Used only for iterative training algorithms.

warm_start: Option<bool>

Whether to train a model from the last checkpoint.

early_stop: Option<bool>

Whether to stop early when the loss doesn’t improve significantly any more (compared to min_relative_progress). Used only for iterative training algorithms.

input_label_columns: Vec<String>

Name of input label columns in training data.

data_split_method: i32

The data split type for training and evaluation, e.g. RANDOM.

data_split_eval_fraction: f64

The fraction of evaluation data over the whole input data. The rest of data will be used as training data. The format should be double. Accurate to two decimal places. Default value is 0.2.

data_split_column: String

The column to split data with. This column won’t be used as a feature.

When data_split_method is CUSTOM, the corresponding column should be boolean. The rows with true value tag are eval data, and the false are training data.

When data_split_method is SEQ, the first DATA_SPLIT_EVAL_FRACTION rows (from smallest to largest) in the corresponding column are used as training data, and the rest are eval data. It respects the order in Orderable data types: https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#data-type-properties

learn_rate_strategy: i32

The strategy to determine learn rate for the current iteration.

initial_learn_rate: f64

Specifies the initial learning rate for the line search learn rate strategy.

label_class_weights: HashMap<String, f64>

Weights associated with each label class, for rebalancing the training data. Only applicable for classification models.

user_column: String

User column specified for matrix factorization models.

item_column: String

Item column specified for matrix factorization models.

distance_type: i32

Distance type for clustering models.

num_clusters: i64

Number of clusters for clustering models.

model_uri: String

Google Cloud Storage URI from which the model was imported. Only applicable for imported models.

optimization_strategy: i32

Optimization strategy for training linear regression models.

hidden_units: Vec<i64>

Hidden units for dnn models.

batch_size: i64

Batch size for dnn models.

dropout: Option<f64>

Dropout probability for dnn models.

max_tree_depth: i64

Maximum depth of a tree for boosted tree models.

subsample: f64

Subsample fraction of the training data to grow tree to prevent overfitting for boosted tree models.

min_split_loss: Option<f64>

Minimum split loss for boosted tree models.

num_factors: i64

Num factors specified for matrix factorization models.

feedback_type: i32

Feedback type that specifies which algorithm to run for matrix factorization.

wals_alpha: Option<f64>

Hyperparameter for matrix factoration when implicit feedback type is specified.

kmeans_initialization_method: i32

The method used to initialize the centroids for kmeans algorithm.

kmeans_initialization_column: String

The column used to provide the initial centroids for kmeans algorithm when kmeans_initialization_method is CUSTOM.

time_series_timestamp_column: String

Column to be designated as time series timestamp for ARIMA model.

time_series_data_column: String

Column to be designated as time series data for ARIMA model.

auto_arima: bool

Whether to enable auto ARIMA or not.

non_seasonal_order: Option<ArimaOrder>

A specification of the non-seasonal part of the ARIMA model: the three components (p, d, q) are the AR order, the degree of differencing, and the MA order.

data_frequency: i32

The data frequency of a time series.

include_drift: bool

Include drift when fitting an ARIMA model.

holiday_region: i32

The geographical region based on which the holidays are considered in time series modeling. If a valid value is specified, then holiday effects modeling is enabled.

time_series_id_column: String

The time series id column that was used during ARIMA model training.

time_series_id_columns: Vec<String>

The time series id columns that were used during ARIMA model training.

horizon: i64

The number of periods ahead that need to be forecasted.

preserve_input_structs: bool

Whether to preserve the input structs in output feature names. Suppose there is a struct A with field b. When false (default), the output feature name is A_b. When true, the output feature name is A.b.

auto_arima_max_order: i64

The max value of non-seasonal p and q.

decompose_time_series: Option<bool>

If true, perform decompose time series and save the results.

clean_spikes_and_dips: Option<bool>

If true, clean spikes and dips in the input time series.

adjust_step_changes: Option<bool>

If true, detect step changes and make data adjustment in the input time series.

Implementations

impl TrainingOptions

pub fn loss_type(&self) -> LossType

Returns the enum value of loss_type, or the default if the field is set to an invalid enum value.

pub fn set_loss_type(&mut self, value: LossType)

Sets loss_type to the provided enum value.

pub fn data_split_method(&self) -> DataSplitMethod

Returns the enum value of data_split_method, or the default if the field is set to an invalid enum value.

pub fn set_data_split_method(&mut self, value: DataSplitMethod)

Sets data_split_method to the provided enum value.

pub fn learn_rate_strategy(&self) -> LearnRateStrategy

Returns the enum value of learn_rate_strategy, or the default if the field is set to an invalid enum value.

pub fn set_learn_rate_strategy(&mut self, value: LearnRateStrategy)

Sets learn_rate_strategy to the provided enum value.

pub fn distance_type(&self) -> DistanceType

Returns the enum value of distance_type, or the default if the field is set to an invalid enum value.

pub fn set_distance_type(&mut self, value: DistanceType)

Sets distance_type to the provided enum value.

pub fn optimization_strategy(&self) -> OptimizationStrategy

Returns the enum value of optimization_strategy, or the default if the field is set to an invalid enum value.

pub fn set_optimization_strategy(&mut self, value: OptimizationStrategy)

Sets optimization_strategy to the provided enum value.

pub fn feedback_type(&self) -> FeedbackType

Returns the enum value of feedback_type, or the default if the field is set to an invalid enum value.

pub fn set_feedback_type(&mut self, value: FeedbackType)

Sets feedback_type to the provided enum value.

pub fn kmeans_initialization_method(&self) -> KmeansInitializationMethod

Returns the enum value of kmeans_initialization_method, or the default if the field is set to an invalid enum value.

pub fn set_kmeans_initialization_method(
&mut self,
value: KmeansInitializationMethod
)

Sets kmeans_initialization_method to the provided enum value.

pub fn data_frequency(&self) -> DataFrequency

Returns the enum value of data_frequency, or the default if the field is set to an invalid enum value.

pub fn set_data_frequency(&mut self, value: DataFrequency)

Sets data_frequency to the provided enum value.

pub fn holiday_region(&self) -> HolidayRegion

Returns the enum value of holiday_region, or the default if the field is set to an invalid enum value.

pub fn set_holiday_region(&mut self, value: HolidayRegion)

Sets holiday_region to the provided enum value.