Struct googapis::google::cloud::clouddms::v1::CloudSqlSettings

pub struct CloudSqlSettings {

    pub database_version: i32,
    pub user_labels: HashMap<String, String>,
    pub tier: String,
    pub storage_auto_resize_limit: Option<i64>,
    pub activation_policy: i32,
    pub ip_config: Option<SqlIpConfig>,
    pub auto_storage_increase: Option<bool>,
    pub database_flags: HashMap<String, String>,
    pub data_disk_type: i32,
    pub data_disk_size_gb: Option<i64>,
    pub zone: String,
    pub source_id: String,
    pub root_password: String,
    pub root_password_set: bool,
    pub collation: String,

}

Settings for creating a Cloud SQL database instance.

Fields
database_version: i32

The database engine type and version.

user_labels: HashMap<String, String>

The resource labels for a Cloud SQL instance to use to annotate any related underlying resources such as Compute Engine VMs. An object containing a list of “key”: “value” pairs.

Example: { "name": "wrench", "mass": "18kg", "count": "3" }.

tier: String

The tier (or machine type) for this instance, for example: db-n1-standard-1 (MySQL instances) or db-custom-1-3840 (PostgreSQL instances). For more information, see Cloud SQL Instance Settings.

storage_auto_resize_limit: Option<i64>

The maximum size to which storage capacity can be automatically increased. The default value is 0, which specifies that there is no limit.

activation_policy: i32

The activation policy specifies when the instance is activated; it is applicable only when the instance state is ‘RUNNABLE’. Valid values:

‘ALWAYS’: The instance is on, and remains so even in the absence of connection requests.

NEVER: The instance is off; it is not activated, even if a connection request arrives.

ip_config: Option<SqlIpConfig>

The settings for IP Management. This allows to enable or disable the instance IP and manage which external networks can connect to the instance. The IPv4 address cannot be disabled.

auto_storage_increase: Option<bool>

[default: ON] If you enable this setting, Cloud SQL checks your available storage every 30 seconds. If the available storage falls below a threshold size, Cloud SQL automatically adds additional storage capacity. If the available storage repeatedly falls below the threshold size, Cloud SQL continues to add storage until it reaches the maximum of 30 TB.

database_flags: HashMap<String, String>

The database flags passed to the Cloud SQL instance at startup. An object containing a list of “key”: value pairs. Example: { “name”: “wrench”, “mass”: “1.3kg”, “count”: “3” }.

data_disk_type: i32

The type of storage: PD_SSD (default) or PD_HDD.

data_disk_size_gb: Option<i64>

The storage capacity available to the database, in GB. The minimum (and default) size is 10GB.

zone: String

The Google Cloud Platform zone where your Cloud SQL datdabse instance is located.

source_id: String

The Database Migration Service source connection profile ID, in the format: projects/my_project_name/locations/us-central1/connectionProfiles/connection_profile_ID

root_password: String

Input only. Initial root password.

root_password_set: bool

Output only. Indicates If this connection profile root password is stored.

collation: String

The Cloud SQL default instance level collation.

Implementations

impl CloudSqlSettings

pub fn database_version(&self) -> SqlDatabaseVersion

Returns the enum value of database_version, or the default if the field is set to an invalid enum value.

pub fn set_database_version(&mut self, value: SqlDatabaseVersion)

Sets database_version to the provided enum value.

pub fn activation_policy(&self) -> SqlActivationPolicy

Returns the enum value of activation_policy, or the default if the field is set to an invalid enum value.

pub fn set_activation_policy(&mut self, value: SqlActivationPolicy)

Sets activation_policy to the provided enum value.

pub fn data_disk_type(&self) -> SqlDataDiskType

Returns the enum value of data_disk_type, or the default if the field is set to an invalid enum value.

pub fn set_data_disk_type(&mut self, value: SqlDataDiskType)

Sets data_disk_type to the provided enum value.