Enum googapis::google::cloud::compute::v1::quota::Metric

#[repr(i32)]
pub enum Metric {

    UndefinedMetric,
    A2Cpus,
    AffinityGroups,
    Autoscalers,
    BackendBuckets,
    BackendServices,
    C2dCpus,
    C2Cpus,
    C3Cpus,
    Commitments,
    CommittedA2Cpus,
    CommittedC2dCpus,
    CommittedC2Cpus,
    CommittedC3Cpus,
    CommittedCpus,
    CommittedE2Cpus,
    CommittedLicenses,
    CommittedLocalSsdTotalGb,
    CommittedMemoryOptimizedCpus,
    CommittedN2aCpus,
    CommittedN2dCpus,
    CommittedN2Cpus,
    CommittedNvidiaA100Gpus,
    CommittedNvidiaK80Gpus,
    CommittedNvidiaP100Gpus,
    CommittedNvidiaP4Gpus,
    CommittedNvidiaT4Gpus,
    CommittedNvidiaV100Gpus,
    CommittedT2dCpus,
    Cpus,
    CpusAllRegions,
    DisksTotalGb,
    E2Cpus,
    ExternalNetworkLbForwardingRules,
    ExternalProtocolForwardingRules,
    ExternalVpnGateways,
    Firewalls,
    ForwardingRules,
    GlobalInternalAddresses,
    GpusAllRegions,
    HealthChecks,
    Images,
    Instances,
    InstanceGroups,
    InstanceGroupManagers,
    InstanceTemplates,
    Interconnects,
    InterconnectAttachmentsPerRegion,
    InterconnectAttachmentsTotalMbps,
    InterconnectTotalGbps,
    InternalAddresses,
    InternalTrafficDirectorForwardingRules,
    InPlaceSnapshots,
    InUseAddresses,
    InUseBackupSchedules,
    InUseSnapshotSchedules,
    LocalSsdTotalGb,
    M1Cpus,
    M2Cpus,
    MachineImages,
    N2aCpus,
    N2dCpus,
    N2Cpus,
    Networks,
    NetworkEndpointGroups,
    NetworkFirewallPolicies,
    NodeGroups,
    NodeTemplates,
    NvidiaA100Gpus,
    NvidiaK80Gpus,
    NvidiaP100Gpus,
    NvidiaP100VwsGpus,
    NvidiaP4Gpus,
    NvidiaP4VwsGpus,
    NvidiaT4Gpus,
    NvidiaT4VwsGpus,
    NvidiaV100Gpus,
    PacketMirrorings,
    PdExtremeTotalProvisionedIops,
    PreemptibleCpus,
    PreemptibleLocalSsdGb,
    PreemptibleNvidiaA100Gpus,
    PreemptibleNvidiaK80Gpus,
    PreemptibleNvidiaP100Gpus,
    PreemptibleNvidiaP100VwsGpus,
    PreemptibleNvidiaP4Gpus,
    PreemptibleNvidiaP4VwsGpus,
    PreemptibleNvidiaT4Gpus,
    PreemptibleNvidiaT4VwsGpus,
    PreemptibleNvidiaV100Gpus,
    PscIlbConsumerForwardingRulesPerProducerNetwork,
    PscInternalLbForwardingRules,
    PublicAdvertisedPrefixes,
    PublicDelegatedPrefixes,
    RegionalAutoscalers,
    RegionalInstanceGroupManagers,
    Reservations,
    ResourcePolicies,
    Routers,
    Routes,
    SecurityPolicies,
    SecurityPoliciesPerRegion,
    SecurityPolicyCevalRules,
    SecurityPolicyRules,
    SecurityPolicyRulesPerRegion,
    ServiceAttachments,
    Snapshots,
    SsdTotalGb,
    SslCertificates,
    StaticAddresses,
    StaticByoipAddresses,
    Subnetworks,
    T2dCpus,
    TargetHttpsProxies,
    TargetHttpProxies,
    TargetInstances,
    TargetPools,
    TargetSslProxies,
    TargetTcpProxies,
    TargetVpnGateways,
    UrlMaps,
    VpnGateways,
    VpnTunnels,
    XpnServiceProjects,

}

[Output Only] Name of the quota metric.

Variants

UndefinedMetric

A value indicating that the enum field is not set.

A2Cpus

AffinityGroups

Autoscalers

BackendBuckets

BackendServices

C2dCpus

C2Cpus

C3Cpus

Commitments

CommittedA2Cpus

CommittedC2dCpus

CommittedC2Cpus

CommittedC3Cpus

CommittedCpus

CommittedE2Cpus

CommittedLicenses

CommittedLocalSsdTotalGb

CommittedMemoryOptimizedCpus

CommittedN2aCpus

CommittedN2dCpus

CommittedN2Cpus

CommittedNvidiaA100Gpus

CommittedNvidiaK80Gpus

CommittedNvidiaP100Gpus

CommittedNvidiaP4Gpus

CommittedNvidiaT4Gpus

CommittedNvidiaV100Gpus

CommittedT2dCpus

Cpus

Guest CPUs

CpusAllRegions

DisksTotalGb

E2Cpus

ExternalNetworkLbForwardingRules

ExternalProtocolForwardingRules

ExternalVpnGateways

Firewalls

ForwardingRules

GlobalInternalAddresses

GpusAllRegions

HealthChecks

Images

Instances

InstanceGroups

InstanceGroupManagers

InstanceTemplates

Interconnects

InterconnectAttachmentsPerRegion

InterconnectAttachmentsTotalMbps

InterconnectTotalGbps

InternalAddresses

InternalTrafficDirectorForwardingRules

InPlaceSnapshots

InUseAddresses

InUseBackupSchedules

InUseSnapshotSchedules

LocalSsdTotalGb

M1Cpus

M2Cpus

MachineImages

N2aCpus

N2dCpus

N2Cpus

Networks

NetworkEndpointGroups

NetworkFirewallPolicies

NodeGroups

NodeTemplates

NvidiaA100Gpus

NvidiaK80Gpus

NvidiaP100Gpus

NvidiaP100VwsGpus

NvidiaP4Gpus

NvidiaP4VwsGpus

NvidiaT4Gpus

NvidiaT4VwsGpus

NvidiaV100Gpus

PacketMirrorings

PdExtremeTotalProvisionedIops

PreemptibleCpus

PreemptibleLocalSsdGb

PreemptibleNvidiaA100Gpus

PreemptibleNvidiaK80Gpus

PreemptibleNvidiaP100Gpus

PreemptibleNvidiaP100VwsGpus

PreemptibleNvidiaP4Gpus

PreemptibleNvidiaP4VwsGpus

PreemptibleNvidiaT4Gpus

PreemptibleNvidiaT4VwsGpus

PreemptibleNvidiaV100Gpus

PscIlbConsumerForwardingRulesPerProducerNetwork

PscInternalLbForwardingRules

PublicAdvertisedPrefixes

PublicDelegatedPrefixes

RegionalAutoscalers

RegionalInstanceGroupManagers

Reservations

ResourcePolicies

Routers

Routes

SecurityPolicies

SecurityPoliciesPerRegion

SecurityPolicyCevalRules

SecurityPolicyRules

SecurityPolicyRulesPerRegion

ServiceAttachments

Snapshots

The total number of snapshots allowed for a single project.

SsdTotalGb

SslCertificates

StaticAddresses

StaticByoipAddresses

Subnetworks

T2dCpus

TargetHttpsProxies

TargetHttpProxies

TargetInstances

TargetPools

TargetSslProxies

TargetTcpProxies

TargetVpnGateways

UrlMaps

VpnGateways

VpnTunnels

XpnServiceProjects

Implementations

impl Metric

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Metric.

pub fn from_i32(value: i32) -> Option<Metric>

Converts an i32 to a Metric, or None if value is not a valid variant.