Struct googapis::google::cloud::compute::v1::BackendService

pub struct BackendService {

    pub affinity_cookie_ttl_sec: Option<i32>,
    pub backends: Vec<Backend>,
    pub cdn_policy: Option<BackendServiceCdnPolicy>,
    pub circuit_breakers: Option<CircuitBreakers>,
    pub connection_draining: Option<ConnectionDraining>,
    pub consistent_hash: Option<ConsistentHashLoadBalancerSettings>,
    pub creation_timestamp: Option<String>,
    pub custom_request_headers: Vec<String>,
    pub custom_response_headers: Vec<String>,
    pub description: Option<String>,
    pub enable_c_d_n: Option<bool>,
    pub failover_policy: Option<BackendServiceFailoverPolicy>,
    pub fingerprint: Option<String>,
    pub health_checks: Vec<String>,
    pub iap: Option<BackendServiceIap>,
    pub id: Option<u64>,
    pub kind: Option<String>,
    pub load_balancing_scheme: Option<String>,
    pub locality_lb_policy: Option<String>,
    pub log_config: Option<BackendServiceLogConfig>,
    pub max_stream_duration: Option<Duration>,
    pub name: Option<String>,
    pub network: Option<String>,
    pub outlier_detection: Option<OutlierDetection>,
    pub port: Option<i32>,
    pub port_name: Option<String>,
    pub protocol: Option<String>,
    pub region: Option<String>,
    pub security_policy: Option<String>,
    pub security_settings: Option<SecuritySettings>,
    pub self_link: Option<String>,
    pub session_affinity: Option<String>,
    pub subsetting: Option<Subsetting>,
    pub timeout_sec: Option<i32>,

}

Represents a Backend Service resource. A backend service defines how Google Cloud load balancers distribute traffic. The backend service configuration contains a set of values, such as the protocol used to connect to backends, various distribution and session settings, health checks, and timeouts. These settings provide fine-grained control over how your load balancer behaves. Most of the settings have default values that allow for easy configuration if you need to get started quickly. Backend services in Google Compute Engine can be either regionally or globally scoped. * [Global](https://cloud.google.com/compute/docs/reference/rest/v1/backendServices) * [Regional](https://cloud.google.com/compute/docs/reference/rest/v1/regionBackendServices) For more information, see Backend Services.

Fields
affinity_cookie_ttl_sec: Option<i32>

Lifetime of cookies in seconds. This setting is applicable to external and internal HTTP(S) load balancers and Traffic Director and requires GENERATED_COOKIE or HTTP_COOKIE session affinity. If set to 0, the cookie is non-persistent and lasts only until the end of the browser session (or equivalent). The maximum allowed value is one day (86,400). Not supported when the backend service is referenced by a URL map that is bound to target gRPC proxy that has validateForProxyless field set to true.

backends: Vec<Backend>

The list of backends that serve this BackendService.

cdn_policy: Option<BackendServiceCdnPolicy>

Cloud CDN configuration for this BackendService. Only available for specified load balancer types.

circuit_breakers: Option<CircuitBreakers>connection_draining: Option<ConnectionDraining>consistent_hash: Option<ConsistentHashLoadBalancerSettings>

Consistent Hash-based load balancing can be used to provide soft session affinity based on HTTP headers, cookies or other properties. This load balancing policy is applicable only for HTTP connections. The affinity to a particular destination host will be lost when one or more hosts are added/removed from the destination service. This field specifies parameters that control consistent hashing. This field is only applicable when localityLbPolicy is set to MAGLEV or RING_HASH. This field is applicable to either: - A regional backend service with the service_protocol set to HTTP, HTTPS, or HTTP2, and load_balancing_scheme set to INTERNAL_MANAGED. - A global backend service with the load_balancing_scheme set to INTERNAL_SELF_MANAGED. Not supported when the backend service is referenced by a URL map that is bound to target gRPC proxy that has validateForProxyless field set to true.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

custom_request_headers: Vec<String>

Headers that the load balancer adds to proxied requests. See Creating custom headers.

custom_response_headers: Vec<String>

Headers that the load balancer adds to proxied responses. See Creating custom headers.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

enable_c_d_n: Option<bool>

If true, enables Cloud CDN for the backend service of an external HTTP(S) load balancer.

failover_policy: Option<BackendServiceFailoverPolicy>

Requires at least one backend instance group to be defined as a backup (failover) backend. For load balancers that have configurable failover: Internal TCP/UDP Load Balancing and external TCP/UDP Load Balancing.

fingerprint: Option<String>

Fingerprint of this resource. A hash of the contents stored in this object. This field is used in optimistic locking. This field will be ignored when inserting a BackendService. An up-to-date fingerprint must be provided in order to update the BackendService, otherwise the request will fail with error 412 conditionNotMet. To see the latest fingerprint, make a get() request to retrieve a BackendService.

health_checks: Vec<String>

The list of URLs to the healthChecks, httpHealthChecks (legacy), or httpsHealthChecks (legacy) resource for health checking this backend service. Not all backend services support legacy health checks. See Load balancer guide. Currently, at most one health check can be specified for each backend service. Backend services with instance group or zonal NEG backends must have a health check. Backend services with internet or serverless NEG backends must not have a health check.

iap: Option<BackendServiceIap>

The configurations for Identity-Aware Proxy on this resource. Not available for Internal TCP/UDP Load Balancing and Network Load Balancing.

id: Option<u64>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

kind: Option<String>

[Output Only] Type of resource. Always compute#backendService for backend services.

load_balancing_scheme: Option<String>

Specifies the load balancer type. A backend service created for one type of load balancer cannot be used with another. For more information, refer to Choosing a load balancer. Check the LoadBalancingScheme enum for the list of possible values.

locality_lb_policy: Option<String>

The load balancing algorithm used within the scope of the locality. The possible values are: - ROUND_ROBIN: This is a simple policy in which each healthy backend is selected in round robin order. This is the default. - LEAST_REQUEST: An O(1) algorithm which selects two random healthy hosts and picks the host which has fewer active requests. - RING_HASH: The ring/modulo hash load balancer implements consistent hashing to backends. The algorithm has the property that the addition/removal of a host from a set of N hosts only affects 1/N of the requests. - RANDOM: The load balancer selects a random healthy host. - ORIGINAL_DESTINATION: Backend host is selected based on the client connection metadata, i.e., connections are opened to the same address as the destination address of the incoming connection before the connection was redirected to the load balancer. - MAGLEV: used as a drop in replacement for the ring hash load balancer. Maglev is not as stable as ring hash but has faster table lookup build times and host selection times. For more information about Maglev, see https://ai.google/research/pubs/pub44824 This field is applicable to either: - A regional backend service with the service_protocol set to HTTP, HTTPS, or HTTP2, and load_balancing_scheme set to INTERNAL_MANAGED. - A global backend service with the load_balancing_scheme set to INTERNAL_SELF_MANAGED. If sessionAffinity is not NONE, and this field is not set to MAGLEV or RING_HASH, session affinity settings will not take effect. Only the default ROUND_ROBIN policy is supported when the backend service is referenced by a URL map that is bound to target gRPC proxy that has validateForProxyless field set to true. Check the LocalityLbPolicy enum for the list of possible values.

log_config: Option<BackendServiceLogConfig>

This field denotes the logging options for the load balancer traffic served by this backend service. If logging is enabled, logs will be exported to Stackdriver.

max_stream_duration: Option<Duration>

Specifies the default maximum duration (timeout) for streams to this service. Duration is computed from the beginning of the stream until the response has been completely processed, including all retries. A stream that does not complete in this duration is closed. If not specified, there will be no timeout limit, i.e. the maximum duration is infinite. This value can be overridden in the PathMatcher configuration of the UrlMap that references this backend service. This field is only allowed when the loadBalancingScheme of the backend service is INTERNAL_SELF_MANAGED.

name: Option<String>

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression \[a-z]([-a-z0-9]*[a-z0-9\])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

network: Option<String>

The URL of the network to which this backend service belongs. This field can only be specified when the load balancing scheme is set to INTERNAL.

outlier_detection: Option<OutlierDetection>

Settings controlling the eviction of unhealthy hosts from the load balancing pool for the backend service. If not set, this feature is considered disabled. This field is applicable to either: - A regional backend service with the service_protocol set to HTTP, HTTPS, or HTTP2, and load_balancing_scheme set to INTERNAL_MANAGED. - A global backend service with the load_balancing_scheme set to INTERNAL_SELF_MANAGED. Not supported when the backend service is referenced by a URL map that is bound to target gRPC proxy that has validateForProxyless field set to true.

port: Option<i32>

Deprecated in favor of portName. The TCP port to connect on the backend. The default value is 80. For Internal TCP/UDP Load Balancing and Network Load Balancing, omit port.

port_name: Option<String>

A named port on a backend instance group representing the port for communication to the backend VMs in that group. The named port must be defined on each backend instance group. This parameter has no meaning if the backends are NEGs. For Internal TCP/UDP Load Balancing and Network Load Balancing, omit port_name.

protocol: Option<String>

The protocol this BackendService uses to communicate with backends. Possible values are HTTP, HTTPS, HTTP2, TCP, SSL, UDP or GRPC. depending on the chosen load balancer or Traffic Director configuration. Refer to the documentation for the load balancers or for Traffic Director for more information. Must be set to GRPC when the backend service is referenced by a URL map that is bound to target gRPC proxy. Check the Protocol enum for the list of possible values.

region: Option<String>

[Output Only] URL of the region where the regional backend service resides. This field is not applicable to global backend services. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.

security_policy: Option<String>

[Output Only] The resource URL for the security policy associated with this backend service.

security_settings: Option<SecuritySettings>

This field specifies the security policy that applies to this backend service. This field is applicable to either: - A regional backend service with the service_protocol set to HTTP, HTTPS, or HTTP2, and load_balancing_scheme set to INTERNAL_MANAGED. - A global backend service with the load_balancing_scheme set to INTERNAL_SELF_MANAGED.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

session_affinity: Option<String>

Type of session affinity to use. The default is NONE. For a detailed description of session affinity options, see: Session affinity. Not supported when the backend service is referenced by a URL map that is bound to target gRPC proxy that has validateForProxyless field set to true. Check the SessionAffinity enum for the list of possible values.

subsetting: Option<Subsetting>timeout_sec: Option<i32>

Not supported when the backend service is referenced by a URL map that is bound to target gRPC proxy that has validateForProxyless field set to true. Instead, use maxStreamDuration.

Implementations

impl BackendService

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn port(&self) -> i32

Returns the value of port, or the default value if port is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn port_name(&self) -> &str

Returns the value of port_name, or the default value if port_name is unset.

pub fn timeout_sec(&self) -> i32

Returns the value of timeout_sec, or the default value if timeout_sec is unset.

pub fn protocol(&self) -> &str

Returns the value of protocol, or the default value if protocol is unset.

pub fn locality_lb_policy(&self) -> &str

Returns the value of locality_lb_policy, or the default value if locality_lb_policy is unset.

pub fn region(&self) -> &str

Returns the value of region, or the default value if region is unset.

pub fn security_policy(&self) -> &str

Returns the value of security_policy, or the default value if security_policy is unset.

pub fn network(&self) -> &str

Returns the value of network, or the default value if network is unset.

pub fn fingerprint(&self) -> &str

Returns the value of fingerprint, or the default value if fingerprint is unset.

pub fn enable_c_d_n(&self) -> bool

Returns the value of enable_c_d_n, or the default value if enable_c_d_n is unset.

pub fn load_balancing_scheme(&self) -> &str

Returns the value of load_balancing_scheme, or the default value if load_balancing_scheme is unset.

pub fn affinity_cookie_ttl_sec(&self) -> i32

Returns the value of affinity_cookie_ttl_sec, or the default value if affinity_cookie_ttl_sec is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.

pub fn session_affinity(&self) -> &str

Returns the value of session_affinity, or the default value if session_affinity is unset.