Struct googapis::google::cloud::compute::v1::RouterNat

pub struct RouterNat {

    pub drain_nat_ips: Vec<String>,
    pub enable_endpoint_independent_mapping: Option<bool>,
    pub icmp_idle_timeout_sec: Option<i32>,
    pub log_config: Option<RouterNatLogConfig>,
    pub min_ports_per_vm: Option<i32>,
    pub name: Option<String>,
    pub nat_ip_allocate_option: Option<String>,
    pub nat_ips: Vec<String>,
    pub rules: Vec<RouterNatRule>,
    pub source_subnetwork_ip_ranges_to_nat: Option<String>,
    pub subnetworks: Vec<RouterNatSubnetworkToNat>,
    pub tcp_established_idle_timeout_sec: Option<i32>,
    pub tcp_time_wait_timeout_sec: Option<i32>,
    pub tcp_transitory_idle_timeout_sec: Option<i32>,
    pub udp_idle_timeout_sec: Option<i32>,

}

Represents a Nat resource. It enables the VMs within the specified subnetworks to access Internet without external IP addresses. It specifies a list of subnetworks (and the ranges within) that want to use NAT. Customers can also provide the external IPs that would be used for NAT. GCP would auto-allocate ephemeral IPs if no external IPs are provided.

Fields
drain_nat_ips: Vec<String>

A list of URLs of the IP resources to be drained. These IPs must be valid static external IPs that have been assigned to the NAT. These IPs should be used for updating/patching a NAT only.

enable_endpoint_independent_mapping: Option<bool>icmp_idle_timeout_sec: Option<i32>

Timeout (in seconds) for ICMP connections. Defaults to 30s if not set.

log_config: Option<RouterNatLogConfig>

Configure logging on this NAT.

min_ports_per_vm: Option<i32>

Minimum number of ports allocated to a VM from this NAT config. If not set, a default number of ports is allocated to a VM. This is rounded up to the nearest power of 2. For example, if the value of this field is 50, at least 64 ports are allocated to a VM.

name: Option<String>

Unique name of this Nat service. The name must be 1-63 characters long and comply with RFC1035.

nat_ip_allocate_option: Option<String>

Specify the NatIpAllocateOption, which can take one of the following values: - MANUAL_ONLY: Uses only Nat IP addresses provided by customers. When there are not enough specified Nat IPs, the Nat service fails for new VMs. - AUTO_ONLY: Nat IPs are allocated by Google Cloud Platform; customers can’t specify any Nat IPs. When choosing AUTO_ONLY, then nat_ip should be empty. Check the NatIpAllocateOption enum for the list of possible values.

nat_ips: Vec<String>

A list of URLs of the IP resources used for this Nat service. These IP addresses must be valid static external IP addresses assigned to the project.

rules: Vec<RouterNatRule>

A list of rules associated with this NAT.

source_subnetwork_ip_ranges_to_nat: Option<String>

Specify the Nat option, which can take one of the following values: - ALL_SUBNETWORKS_ALL_IP_RANGES: All of the IP ranges in every Subnetwork are allowed to Nat. - ALL_SUBNETWORKS_ALL_PRIMARY_IP_RANGES: All of the primary IP ranges in every Subnetwork are allowed to Nat. - LIST_OF_SUBNETWORKS: A list of Subnetworks are allowed to Nat (specified in the field subnetwork below) The default is SUBNETWORK_IP_RANGE_TO_NAT_OPTION_UNSPECIFIED. Note that if this field contains ALL_SUBNETWORKS_ALL_IP_RANGES or ALL_SUBNETWORKS_ALL_PRIMARY_IP_RANGES, then there should not be any other Router.Nat section in any Router for this network in this region. Check the SourceSubnetworkIpRangesToNat enum for the list of possible values.

subnetworks: Vec<RouterNatSubnetworkToNat>

A list of Subnetwork resources whose traffic should be translated by NAT Gateway. It is used only when LIST_OF_SUBNETWORKS is selected for the SubnetworkIpRangeToNatOption above.

tcp_established_idle_timeout_sec: Option<i32>

Timeout (in seconds) for TCP established connections. Defaults to 1200s if not set.

tcp_time_wait_timeout_sec: Option<i32>

Timeout (in seconds) for TCP connections that are in TIME_WAIT state. Defaults to 120s if not set.

tcp_transitory_idle_timeout_sec: Option<i32>

Timeout (in seconds) for TCP transitory connections. Defaults to 30s if not set.

udp_idle_timeout_sec: Option<i32>

Timeout (in seconds) for UDP connections. Defaults to 30s if not set.

Implementations

impl RouterNat

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn icmp_idle_timeout_sec(&self) -> i32

Returns the value of icmp_idle_timeout_sec, or the default value if icmp_idle_timeout_sec is unset.

pub fn udp_idle_timeout_sec(&self) -> i32

Returns the value of udp_idle_timeout_sec, or the default value if udp_idle_timeout_sec is unset.

pub fn min_ports_per_vm(&self) -> i32

Returns the value of min_ports_per_vm, or the default value if min_ports_per_vm is unset.

pub fn tcp_transitory_idle_timeout_sec(&self) -> i32

Returns the value of tcp_transitory_idle_timeout_sec, or the default value if tcp_transitory_idle_timeout_sec is unset.

pub fn tcp_established_idle_timeout_sec(&self) -> i32

Returns the value of tcp_established_idle_timeout_sec, or the default value if tcp_established_idle_timeout_sec is unset.

pub fn source_subnetwork_ip_ranges_to_nat(&self) -> &str

Returns the value of source_subnetwork_ip_ranges_to_nat, or the default value if source_subnetwork_ip_ranges_to_nat is unset.

pub fn enable_endpoint_independent_mapping(&self) -> bool

Returns the value of enable_endpoint_independent_mapping, or the default value if enable_endpoint_independent_mapping is unset.

pub fn nat_ip_allocate_option(&self) -> &str

Returns the value of nat_ip_allocate_option, or the default value if nat_ip_allocate_option is unset.

pub fn tcp_time_wait_timeout_sec(&self) -> i32

Returns the value of tcp_time_wait_timeout_sec, or the default value if tcp_time_wait_timeout_sec is unset.