Enum googapis::google::cloud::compute::v1::warnings::Code

#[repr(i32)]
pub enum Code {

    UndefinedCode,
    CleanupFailed,
    DeprecatedResourceUsed,
    DeprecatedTypeUsed,
    DiskSizeLargerThanImageSize,
    ExperimentalTypeUsed,
    ExternalApiWarning,
    FieldValueOverriden,
    InjectedKernelsDeprecated,
    LargeDeploymentWarning,
    MissingTypeDependency,
    NextHopAddressNotAssigned,
    NextHopCannotIpForward,
    NextHopInstanceNotFound,
    NextHopInstanceNotOnNetwork,
    NextHopNotRunning,
    NotCriticalError,
    NoResultsOnPage,
    PartialSuccess,
    RequiredTosAgreement,
    ResourceInUseByOtherResourceWarning,
    ResourceNotDeleted,
    SchemaValidationIgnored,
    SingleInstancePropertyTemplate,
    UndeclaredProperties,
    Unreachable,

}

[Output Only] A warning code, if applicable. For example, Compute Engine returns NO_RESULTS_ON_PAGE if there are no results in the response.

Variants

UndefinedCode

A value indicating that the enum field is not set.

CleanupFailed

Warning about failed cleanup of transient changes made by a failed operation.

DeprecatedResourceUsed

A link to a deprecated resource was created.

DeprecatedTypeUsed

When deploying and at least one of the resources has a type marked as deprecated

DiskSizeLargerThanImageSize

The user created a boot disk that is larger than image size.

ExperimentalTypeUsed

When deploying and at least one of the resources has a type marked as experimental

ExternalApiWarning

Warning that is present in an external api call

FieldValueOverriden

Warning that value of a field has been overridden. Deprecated unused field.

InjectedKernelsDeprecated

The operation involved use of an injected kernel, which is deprecated.

LargeDeploymentWarning

When deploying a deployment with a exceedingly large number of resources

MissingTypeDependency

A resource depends on a missing type

NextHopAddressNotAssigned

The route’s nextHopIp address is not assigned to an instance on the network.

NextHopCannotIpForward

The route’s next hop instance cannot ip forward.

NextHopInstanceNotFound

The route’s nextHopInstance URL refers to an instance that does not exist.

NextHopInstanceNotOnNetwork

The route’s nextHopInstance URL refers to an instance that is not on the same network as the route.

NextHopNotRunning

The route’s next hop instance does not have a status of RUNNING.

NotCriticalError

Error which is not critical. We decided to continue the process despite the mentioned error.

NoResultsOnPage

No results are present on a particular list page.

PartialSuccess

Success is reported, but some results may be missing due to errors

RequiredTosAgreement

The user attempted to use a resource that requires a TOS they have not accepted.

ResourceInUseByOtherResourceWarning

Warning that a resource is in use.

ResourceNotDeleted

One or more of the resources set to auto-delete could not be deleted because they were in use.

SchemaValidationIgnored

When a resource schema validation is ignored.

SingleInstancePropertyTemplate

Instance template used in instance group manager is valid as such, but its application does not make a lot of sense, because it allows only single instance in instance group.

UndeclaredProperties

When undeclared properties in the schema are present

Unreachable

A given scope cannot be reached.

Implementations

impl Code

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Code.

pub fn from_i32(value: i32) -> Option<Code>

Converts an i32 to a Code, or None if value is not a valid variant.