Module googapis::google::cloud::compute::v1small::address

Nested message and enum types in Address.

Enums

AddressType

The type of address to reserve, either INTERNAL or EXTERNAL. If unspecified, defaults to EXTERNAL.

IpVersion

The IP version that will be used by this address. Valid options are IPV4 or IPV6. This can only be specified for a global address.

NetworkTier

This signifies the networking tier used for configuring this address and can only take the following values: PREMIUM or STANDARD. Global forwarding rules can only be Premium Tier. Regional forwarding rules can be either Premium or Standard Tier. Standard Tier addresses applied to regional forwarding rules can be used with any external load balancer. Regional forwarding rules in Premium Tier can only be used with a network load balancer.

Purpose

The purpose of this resource, which can be one of the following values:

Status

[Output Only] The status of the address, which can be one of RESERVING, RESERVED, or IN_USE. An address that is RESERVING is currently in the process of being reserved. A RESERVED address is currently reserved and available to use. An IN_USE address is currently being used by another resource and is not available.