Struct googapis::google::cloud::contactcenterinsights::v1::contact_center_insights_client::ContactCenterInsightsClient

pub struct ContactCenterInsightsClient<T> { /* fields omitted */ }

An API that lets users analyze and explore their business conversation data.

Implementations

impl<T> ContactCenterInsightsClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> ContactCenterInsightsClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_conversation(
&mut self,
request: impl IntoRequest<CreateConversationRequest>
) -> Result<Response<Conversation>, Status>

Creates a conversation.

pub async fn update_conversation(
&mut self,
request: impl IntoRequest<UpdateConversationRequest>
) -> Result<Response<Conversation>, Status>

Updates a conversation.

pub async fn get_conversation(
&mut self,
request: impl IntoRequest<GetConversationRequest>
) -> Result<Response<Conversation>, Status>

Gets a conversation.

pub async fn list_conversations(
&mut self,
request: impl IntoRequest<ListConversationsRequest>
) -> Result<Response<ListConversationsResponse>, Status>

Lists conversations.

pub async fn delete_conversation(
&mut self,
request: impl IntoRequest<DeleteConversationRequest>
) -> Result<Response<()>, Status>

Deletes a conversation.

pub async fn create_analysis(
&mut self,
request: impl IntoRequest<CreateAnalysisRequest>
) -> Result<Response<Operation>, Status>

Creates an analysis. The long running operation is done when the analysis has completed.

pub async fn get_analysis(
&mut self,
request: impl IntoRequest<GetAnalysisRequest>
) -> Result<Response<Analysis>, Status>

Gets an analysis.

pub async fn list_analyses(
&mut self,
request: impl IntoRequest<ListAnalysesRequest>
) -> Result<Response<ListAnalysesResponse>, Status>

Lists analyses.

pub async fn delete_analysis(
&mut self,
request: impl IntoRequest<DeleteAnalysisRequest>
) -> Result<Response<()>, Status>

Deletes an analysis.

pub async fn export_insights_data(
&mut self,
request: impl IntoRequest<ExportInsightsDataRequest>
) -> Result<Response<Operation>, Status>

Export insights data to a destination defined in the request body.

pub async fn create_issue_model(
&mut self,
request: impl IntoRequest<CreateIssueModelRequest>
) -> Result<Response<Operation>, Status>

Creates an issue model.

pub async fn update_issue_model(
&mut self,
request: impl IntoRequest<UpdateIssueModelRequest>
) -> Result<Response<IssueModel>, Status>

Updates an issue model.

pub async fn get_issue_model(
&mut self,
request: impl IntoRequest<GetIssueModelRequest>
) -> Result<Response<IssueModel>, Status>

Gets an issue model.

pub async fn list_issue_models(
&mut self,
request: impl IntoRequest<ListIssueModelsRequest>
) -> Result<Response<ListIssueModelsResponse>, Status>

Lists issue models.

pub async fn delete_issue_model(
&mut self,
request: impl IntoRequest<DeleteIssueModelRequest>
) -> Result<Response<Operation>, Status>

Deletes an issue model.

pub async fn deploy_issue_model(
&mut self,
request: impl IntoRequest<DeployIssueModelRequest>
) -> Result<Response<Operation>, Status>

Deploys an issue model. Returns an error if a model is already deployed. An issue model can only be used in analysis after it has been deployed.

pub async fn undeploy_issue_model(
&mut self,
request: impl IntoRequest<UndeployIssueModelRequest>
) -> Result<Response<Operation>, Status>

Undeploys an issue model. An issue model can not be used in analysis after it has been undeployed.

pub async fn get_issue(
&mut self,
request: impl IntoRequest<GetIssueRequest>
) -> Result<Response<Issue>, Status>

Gets an issue.

pub async fn list_issues(
&mut self,
request: impl IntoRequest<ListIssuesRequest>
) -> Result<Response<ListIssuesResponse>, Status>

Lists issues.

pub async fn update_issue(
&mut self,
request: impl IntoRequest<UpdateIssueRequest>
) -> Result<Response<Issue>, Status>

Updates an issue.

pub async fn calculate_issue_model_stats(
&mut self,
request: impl IntoRequest<CalculateIssueModelStatsRequest>
) -> Result<Response<CalculateIssueModelStatsResponse>, Status>

Gets an issue model’s statistics.

pub async fn create_phrase_matcher(
&mut self,
request: impl IntoRequest<CreatePhraseMatcherRequest>
) -> Result<Response<PhraseMatcher>, Status>

Creates a phrase matcher.

pub async fn get_phrase_matcher(
&mut self,
request: impl IntoRequest<GetPhraseMatcherRequest>
) -> Result<Response<PhraseMatcher>, Status>

Gets a phrase matcher.

pub async fn list_phrase_matchers(
&mut self,
request: impl IntoRequest<ListPhraseMatchersRequest>
) -> Result<Response<ListPhraseMatchersResponse>, Status>

Lists phrase matchers.

pub async fn delete_phrase_matcher(
&mut self,
request: impl IntoRequest<DeletePhraseMatcherRequest>
) -> Result<Response<()>, Status>

Deletes a phrase matcher.

pub async fn update_phrase_matcher(
&mut self,
request: impl IntoRequest<UpdatePhraseMatcherRequest>
) -> Result<Response<PhraseMatcher>, Status>

Updates a phrase matcher.

pub async fn calculate_stats(
&mut self,
request: impl IntoRequest<CalculateStatsRequest>
) -> Result<Response<CalculateStatsResponse>, Status>

Gets conversation statistics.

pub async fn get_settings(
&mut self,
request: impl IntoRequest<GetSettingsRequest>
) -> Result<Response<Settings>, Status>

Gets project-level settings.

pub async fn update_settings(
&mut self,
request: impl IntoRequest<UpdateSettingsRequest>
) -> Result<Response<Settings>, Status>

Updates project-level settings.