Module googapis::google::cloud::contactcenterinsights::v1::conversation

Nested message and enum types in Conversation.

Modules

transcript

Nested message and enum types in Transcript.

Structs

CallMetadata

Call-specific metadata.

Transcript

A message representing the transcript of a conversation.

Enums

Expiration

A time to live expiration setting, can be either a specified timestamp or a duration from the time that the conversation creation request was received. Conversations with an expiration set will be removed up to 24 hours after the specified time.

Medium

Possible media for the conversation.

Metadata

Metadata that applies to the conversation.