Module googapis::google::cloud::datacatalog::v1

Modules

big_query_table_spec
Nested message and enum types in BigQueryTableSpec.

data_catalog_client
Generated client implementations.

data_source
Nested message and enum types in DataSource.

database_table_spec
Nested message and enum types in DatabaseTableSpec.

entry
Nested message and enum types in Entry.

export_taxonomies_request
Nested message and enum types in ExportTaxonomiesRequest.

field_type
Nested message and enum types in FieldType.

import_taxonomies_request
Nested message and enum types in ImportTaxonomiesRequest.

lookup_entry_request
Nested message and enum types in LookupEntryRequest.

policy_tag_manager_client
Generated client implementations.

policy_tag_manager_serialization_client
Generated client implementations.

search_catalog_request
Nested message and enum types in SearchCatalogRequest.

search_catalog_result
Nested message and enum types in SearchCatalogResult.

tag
Nested message and enum types in Tag.

tag_field
Nested message and enum types in TagField.

taxonomy
Nested message and enum types in Taxonomy.

Structs

BigQueryDateShardedSpec
Spec for a group of BigQuery tables with name pattern [prefix]YYYYMMDD. Context: https://cloud.google.com/bigquery/docs/partitioned-tables#partitioning_versus_sharding

BigQueryTableSpec
Describes a BigQuery table.

ColumnSchema
Representation of a column within a schema. Columns could be nested inside other columns.

CreateEntryGroupRequest
Request message for [CreateEntryGroup][google.cloud.datacatalog.v1.DataCatalog.CreateEntryGroup].

CreateEntryRequest
Request message for [CreateEntry][google.cloud.datacatalog.v1.DataCatalog.CreateEntry].

CreatePolicyTagRequest
Request message for [CreatePolicyTag][google.cloud.datacatalog.v1.PolicyTagManager.CreatePolicyTag].

CreateTagRequest
Request message for [CreateTag][google.cloud.datacatalog.v1.DataCatalog.CreateTag].

CreateTagTemplateFieldRequest
Request message for [CreateTagTemplateField][google.cloud.datacatalog.v1.DataCatalog.CreateTagTemplateField].

CreateTagTemplateRequest
Request message for [CreateTagTemplate][google.cloud.datacatalog.v1.DataCatalog.CreateTagTemplate].

CreateTaxonomyRequest
Request message for [CreateTaxonomy][google.cloud.datacatalog.v1.PolicyTagManager.CreateTaxonomy].

CrossRegionalSource
Cross-regional source used to import an existing taxonomy into a different region.

DataSource
Describes the physical location of an entry.

DatabaseTableSpec
Specification that applies to a table resource. Only valid for entries of TABLE type.

DeleteEntryGroupRequest
Request message for [DeleteEntryGroup][google.cloud.datacatalog.v1.DataCatalog.DeleteEntryGroup].

DeleteEntryRequest
Request message for [DeleteEntry][google.cloud.datacatalog.v1.DataCatalog.DeleteEntry].

DeletePolicyTagRequest
Request message for [DeletePolicyTag][google.cloud.datacatalog.v1.PolicyTagManager.DeletePolicyTag].

DeleteTagRequest
Request message for [DeleteTag][google.cloud.datacatalog.v1.DataCatalog.DeleteTag].

DeleteTagTemplateFieldRequest
Request message for [DeleteTagTemplateField][google.cloud.datacatalog.v1.DataCatalog.DeleteTagTemplateField].

DeleteTagTemplateRequest
Request message for [DeleteTagTemplate][google.cloud.datacatalog.v1.DataCatalog.DeleteTagTemplate].

DeleteTaxonomyRequest
Request message for [DeleteTaxonomy][google.cloud.datacatalog.v1.PolicyTagManager.DeleteTaxonomy].

Entry
Entry Metadata. A Data Catalog Entry resource represents another resource in Google Cloud Platform (such as a BigQuery dataset or a Pub/Sub topic) or outside of Google Cloud Platform. Clients can use the linked_resource field in the Entry resource to refer to the original resource ID of the source system.

EntryGroup
EntryGroup Metadata. An EntryGroup resource represents a logical grouping of zero or more Data Catalog [Entry][google.cloud.datacatalog.v1.Entry] resources.

ExportTaxonomiesRequest
Request message for [ExportTaxonomies][google.cloud.datacatalog.v1.PolicyTagManagerSerialization.ExportTaxonomies].

ExportTaxonomiesResponse
Response message for [ExportTaxonomies][google.cloud.datacatalog.v1.PolicyTagManagerSerialization.ExportTaxonomies].

FieldType

GcsFileSpec
Specifications of a single file in Cloud Storage.

GcsFilesetSpec
Describes a Cloud Storage fileset entry.

GetEntryGroupRequest
Request message for [GetEntryGroup][google.cloud.datacatalog.v1.DataCatalog.GetEntryGroup].

GetEntryRequest
Request message for [GetEntry][google.cloud.datacatalog.v1.DataCatalog.GetEntry].

GetPolicyTagRequest
Request message for [GetPolicyTag][google.cloud.datacatalog.v1.PolicyTagManager.GetPolicyTag].

GetTagTemplateRequest
Request message for [GetTagTemplate][google.cloud.datacatalog.v1.DataCatalog.GetTagTemplate].

GetTaxonomyRequest
Request message for [GetTaxonomy][google.cloud.datacatalog.v1.PolicyTagManager.GetTaxonomy].

ImportTaxonomiesRequest
Request message for [ImportTaxonomies][google.cloud.datacatalog.v1.PolicyTagManagerSerialization.ImportTaxonomies].

ImportTaxonomiesResponse
Response message for [ImportTaxonomies][google.cloud.datacatalog.v1.PolicyTagManagerSerialization.ImportTaxonomies].

InlineSource
Inline source containing taxonomies to import.

ListEntriesRequest
Request message for [ListEntries][google.cloud.datacatalog.v1.DataCatalog.ListEntries].

ListEntriesResponse
Response message for [ListEntries][google.cloud.datacatalog.v1.DataCatalog.ListEntries].

ListEntryGroupsRequest
Request message for [ListEntryGroups][google.cloud.datacatalog.v1.DataCatalog.ListEntryGroups].

ListEntryGroupsResponse
Response message for [ListEntryGroups][google.cloud.datacatalog.v1.DataCatalog.ListEntryGroups].

ListPolicyTagsRequest
Request message for [ListPolicyTags][google.cloud.datacatalog.v1.PolicyTagManager.ListPolicyTags].

ListPolicyTagsResponse
Response message for [ListPolicyTags][google.cloud.datacatalog.v1.PolicyTagManager.ListPolicyTags].

ListTagsRequest
Request message for [ListTags][google.cloud.datacatalog.v1.DataCatalog.ListTags].

ListTagsResponse
Response message for [ListTags][google.cloud.datacatalog.v1.DataCatalog.ListTags].

ListTaxonomiesRequest
Request message for [ListTaxonomies][google.cloud.datacatalog.v1.PolicyTagManager.ListTaxonomies].

ListTaxonomiesResponse
Response message for [ListTaxonomies][google.cloud.datacatalog.v1.PolicyTagManager.ListTaxonomies].

LookupEntryRequest
Request message for [LookupEntry][google.cloud.datacatalog.v1.DataCatalog.LookupEntry].

PolicyTag
Denotes one policy tag in a taxonomy (e.g. ssn). Policy tags can be defined in a hierarchy. For example, consider the following hierarchy:

RenameTagTemplateFieldEnumValueRequest
Request message for [RenameTagTemplateFieldEnumValue][google.cloud.datacatalog.v1.DataCatalog.RenameTagTemplateFieldEnumValue].

RenameTagTemplateFieldRequest
Request message for [RenameTagTemplateField][google.cloud.datacatalog.v1.DataCatalog.RenameTagTemplateField].

Schema
Represents a schema (e.g. BigQuery, GoogleSQL, Avro schema).

SearchCatalogRequest
Request message for [SearchCatalog][google.cloud.datacatalog.v1.DataCatalog.SearchCatalog].

SearchCatalogResponse
Response message for [SearchCatalog][google.cloud.datacatalog.v1.DataCatalog.SearchCatalog].

SearchCatalogResult
A result that appears in the response of a search request. Each result captures details of one entry that matches the search.

SerializedPolicyTag
Message representing one policy tag, including all its descendant policy tags, as a nested proto.

SerializedTaxonomy
Message representing a taxonomy, including its policy tags in hierarchy, as a nested proto. Used for taxonomy replacement, import, and export.

SystemTimestamps
Timestamps about this resource according to a particular system.

TableSpec
Normal BigQuery table spec.

Tag
Tags are used to attach custom metadata to Data Catalog resources. Tags conform to the specifications within their tag template.

TagField
Contains the value and supporting information for a field within a [Tag][google.cloud.datacatalog.v1.Tag].

TagTemplate
A tag template defines a tag, which can have one or more typed fields. The template is used to create and attach the tag to GCP resources. Tag template roles provide permissions to create, edit, and use the template. See, for example, the TagTemplate User role, which includes permission to use the tag template to tag resources.

TagTemplateField
The template for an individual field within a tag template.

Taxonomy
A taxonomy is a collection of hierarchical policy tags that classify data along a common axis. For instance a “data sensitivity” taxonomy could contain the following policy tags:

UpdateEntryGroupRequest
Request message for [UpdateEntryGroup][google.cloud.datacatalog.v1.DataCatalog.UpdateEntryGroup].

UpdateEntryRequest
Request message for [UpdateEntry][google.cloud.datacatalog.v1.DataCatalog.UpdateEntry].

UpdatePolicyTagRequest
Request message for [UpdatePolicyTag][google.cloud.datacatalog.v1.PolicyTagManager.UpdatePolicyTag].

UpdateTagRequest
Request message for [UpdateTag][google.cloud.datacatalog.v1.DataCatalog.UpdateTag].

UpdateTagTemplateFieldRequest
Request message for [UpdateTagTemplateField][google.cloud.datacatalog.v1.DataCatalog.UpdateTagTemplateField].

UpdateTagTemplateRequest
Request message for [UpdateTagTemplate][google.cloud.datacatalog.v1.DataCatalog.UpdateTagTemplate].

UpdateTaxonomyRequest
Request message for [UpdateTaxonomy][google.cloud.datacatalog.v1.PolicyTagManager.UpdateTaxonomy].

ViewSpec
Table view specification.

Enums

EntryType
Entry resources in Data Catalog can be of different types e.g. a BigQuery Table entry is of type TABLE. This enum describes all the possible types Data Catalog contains.

IntegratedSystem
This enum describes all the possible systems that Data Catalog integrates with.

SearchResultType
The different types of resources that can be returned in search.

TableSourceType
Table source type.