Struct googapis::google::cloud::datacatalog::v1::policy_tag_manager_client::PolicyTagManagerClient

pub struct PolicyTagManagerClient<T> { /* fields omitted */ }

Policy Tag Manager API service allows clients to manage their policy tags and taxonomies.

Policy tags are used to tag BigQuery columns and apply additional access control policies. A taxonomy is a hierarchical grouping of policy tags that classify data along a common axis.

Implementations

impl<T> PolicyTagManagerClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + Sync + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> PolicyTagManagerClient<InterceptedService<T, F>> where
F: FnMut(Request<()>) -> Result<Request<()>, Status>,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_taxonomy(
&mut self,
request: impl IntoRequest<CreateTaxonomyRequest>
) -> Result<Response<Taxonomy>, Status>

Creates a taxonomy in a specified project. The taxonomy is initially empty, i.e., does not contain policy tags.

pub async fn delete_taxonomy(
&mut self,
request: impl IntoRequest<DeleteTaxonomyRequest>
) -> Result<Response<()>, Status>

Deletes a taxonomy. This method will also delete all policy tags in this taxonomy, their associated policies, and the policy tags references from BigQuery columns.

pub async fn update_taxonomy(
&mut self,
request: impl IntoRequest<UpdateTaxonomyRequest>
) -> Result<Response<Taxonomy>, Status>

Updates a taxonomy. This method can update the taxonomy’s display name, description, and activated policy types.

pub async fn list_taxonomies(
&mut self,
request: impl IntoRequest<ListTaxonomiesRequest>
) -> Result<Response<ListTaxonomiesResponse>, Status>

Lists all taxonomies in a project in a particular location that the caller has permission to view.

pub async fn get_taxonomy(
&mut self,
request: impl IntoRequest<GetTaxonomyRequest>
) -> Result<Response<Taxonomy>, Status>

Gets a taxonomy.

pub async fn create_policy_tag(
&mut self,
request: impl IntoRequest<CreatePolicyTagRequest>
) -> Result<Response<PolicyTag>, Status>

Creates a policy tag in a taxonomy.

pub async fn delete_policy_tag(
&mut self,
request: impl IntoRequest<DeletePolicyTagRequest>
) -> Result<Response<()>, Status>

Deletes a policy tag. This method also deletes:

all of its descendant policy tags, if any

the policies associated with the policy tag and its descendants

references from BigQuery table schema of the policy tag and its descendants.

pub async fn update_policy_tag(
&mut self,
request: impl IntoRequest<UpdatePolicyTagRequest>
) -> Result<Response<PolicyTag>, Status>

Updates a policy tag. This method can update the policy tag’s display name, description, and parent policy tag.

pub async fn list_policy_tags(
&mut self,
request: impl IntoRequest<ListPolicyTagsRequest>
) -> Result<Response<ListPolicyTagsResponse>, Status>

Lists all policy tags in a taxonomy.

pub async fn get_policy_tag(
&mut self,
request: impl IntoRequest<GetPolicyTagRequest>
) -> Result<Response<PolicyTag>, Status>

Gets a policy tag.

pub async fn get_iam_policy(
&mut self,
request: impl IntoRequest<GetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Gets the IAM policy for a policy tag or a taxonomy.

pub async fn set_iam_policy(
&mut self,
request: impl IntoRequest<SetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Sets the IAM policy for a policy tag or a taxonomy.

pub async fn test_iam_permissions(
&mut self,
request: impl IntoRequest<TestIamPermissionsRequest>
) -> Result<Response<TestIamPermissionsResponse>, Status>

Returns the permissions that a caller has on a specified policy tag or taxonomy.