Module googapis::google::cloud::dataproc::v1

Modules

autoscaling_policy
Nested message and enum types in AutoscalingPolicy.

autoscaling_policy_service_client
Generated client implementations.

cluster_controller_client
Generated client implementations.

cluster_operation_status
Nested message and enum types in ClusterOperationStatus.

cluster_status
Nested message and enum types in ClusterStatus.

gce_cluster_config
Nested message and enum types in GceClusterConfig.

gke_cluster_config
Nested message and enum types in GkeClusterConfig.

hadoop_job
Nested message and enum types in HadoopJob.

hive_job
Nested message and enum types in HiveJob.

instance_group_config
Nested message and enum types in InstanceGroupConfig.

job
Nested message and enum types in Job.

job_controller_client
Generated client implementations.

job_status
Nested message and enum types in JobStatus.

lifecycle_config
Nested message and enum types in LifecycleConfig.

list_jobs_request
Nested message and enum types in ListJobsRequest.

logging_config
Nested message and enum types in LoggingConfig.

ordered_job
Nested message and enum types in OrderedJob.

parameter_validation
Nested message and enum types in ParameterValidation.

pig_job
Nested message and enum types in PigJob.

presto_job
Nested message and enum types in PrestoJob.

reservation_affinity
Nested message and enum types in ReservationAffinity.

spark_job
Nested message and enum types in SparkJob.

spark_sql_job
Nested message and enum types in SparkSqlJob.

workflow_metadata
Nested message and enum types in WorkflowMetadata.

workflow_node
Nested message and enum types in WorkflowNode.

workflow_template_placement
Nested message and enum types in WorkflowTemplatePlacement.

workflow_template_service_client
Generated client implementations.

yarn_application
Nested message and enum types in YarnApplication.

Structs

AcceleratorConfig
Specifies the type and number of accelerator cards attached to the instances of an instance. See GPUs on Compute Engine.

AutoscalingConfig
Autoscaling Policy config associated with the cluster.

AutoscalingPolicy
Describes an autoscaling policy for Dataproc cluster autoscaler.

BasicAutoscalingAlgorithm
Basic algorithm for autoscaling.

BasicYarnAutoscalingConfig
Basic autoscaling configurations for YARN.

CancelJobRequest
A request to cancel a job.

Cluster
Describes the identifying information, config, and status of a cluster of Compute Engine instances.

ClusterConfig
The cluster config.

ClusterMetrics
Contains cluster daemon metrics, such as HDFS and YARN stats.

ClusterOperation
The cluster operation triggered by a workflow.

ClusterOperationMetadata
Metadata describing the operation.

ClusterOperationStatus
The status of the operation.

ClusterSelector
A selector that chooses target cluster for jobs based on metadata.

ClusterStatus
The status of a cluster and its instances.

CreateAutoscalingPolicyRequest
A request to create an autoscaling policy.

CreateClusterRequest
A request to create a cluster.

CreateWorkflowTemplateRequest
A request to create a workflow template.

DeleteAutoscalingPolicyRequest
A request to delete an autoscaling policy.

DeleteClusterRequest
A request to delete a cluster.

DeleteJobRequest
A request to delete a job.

DeleteWorkflowTemplateRequest
A request to delete a workflow template.

DiagnoseClusterRequest
A request to collect cluster diagnostic information.

DiagnoseClusterResults
The location of diagnostic output.

DiskConfig
Specifies the config of disk options for a group of VM instances.

EncryptionConfig
Encryption settings for the cluster.

EndpointConfig
Endpoint config for this cluster

GceClusterConfig
Common config settings for resources of Compute Engine cluster instances, applicable to all instances in the cluster.

GetAutoscalingPolicyRequest
A request to fetch an autoscaling policy.

GetClusterRequest
Request to get the resource representation for a cluster in a project.

GetJobRequest
A request to get the resource representation for a job in a project.

GetWorkflowTemplateRequest
A request to fetch a workflow template.

GkeClusterConfig
The GKE config for this cluster.

HadoopJob
A Dataproc job for running Apache Hadoop MapReduce jobs on Apache Hadoop YARN.

HiveJob
A Dataproc job for running Apache Hive queries on YARN.

IdentityConfig
Identity related configuration, including service account based secure multi-tenancy user mappings.

InstanceGroupAutoscalingPolicyConfig
Configuration for the size bounds of an instance group, including its proportional size to other groups.

InstanceGroupConfig
The config settings for Compute Engine resources in an instance group, such as a master or worker group.

InstantiateInlineWorkflowTemplateRequest
A request to instantiate an inline workflow template.

InstantiateWorkflowTemplateRequest
A request to instantiate a workflow template.

Job
A Dataproc job resource.

JobMetadata
Job Operation metadata.

JobPlacement
Dataproc job config.

JobReference
Encapsulates the full scoping used to reference a job.

JobScheduling
Job scheduling options.

JobStatus
Dataproc job status.

KerberosConfig
Specifies Kerberos related configuration.

LifecycleConfig
Specifies the cluster auto-delete schedule configuration.

ListAutoscalingPoliciesRequest
A request to list autoscaling policies in a project.

ListAutoscalingPoliciesResponse
A response to a request to list autoscaling policies in a project.

ListClustersRequest
A request to list the clusters in a project.

ListClustersResponse
The list of all clusters in a project.

ListJobsRequest
A request to list jobs in a project.

ListJobsResponse
A list of jobs in a project.

ListWorkflowTemplatesRequest
A request to list workflow templates in a project.

ListWorkflowTemplatesResponse
A response to a request to list workflow templates in a project.

LoggingConfig
The runtime logging config of the job.

ManagedCluster
Cluster that is managed by the workflow.

ManagedGroupConfig
Specifies the resources used to actively manage an instance group.

MetastoreConfig
Specifies a Metastore configuration.

NodeGroupAffinity
Node Group Affinity for clusters using sole-tenant node groups.

NodeInitializationAction
Specifies an executable to run on a fully configured node and a timeout period for executable completion.

OrderedJob
A job executed by the workflow.

ParameterValidation
Configuration for parameter validation.

PigJob
A Dataproc job for running Apache Pig queries on YARN.

PrestoJob
A Dataproc job for running Presto queries. IMPORTANT: The Dataproc Presto Optional Component must be enabled when the cluster is created to submit a Presto job to the cluster.

PySparkJob
A Dataproc job for running Apache PySpark applications on YARN.

QueryList
A list of queries to run on a cluster.

RegexValidation
Validation based on regular expressions.

ReservationAffinity
Reservation Affinity for consuming Zonal reservation.

SecurityConfig
Security related configuration, including encryption, Kerberos, etc.

ShieldedInstanceConfig
Shielded Instance Config for clusters using Compute Engine Shielded VMs.

SoftwareConfig
Specifies the selection and config of software inside the cluster.

SparkJob
A Dataproc job for running Apache Spark applications on YARN.

SparkRJob
A Dataproc job for running Apache SparkR applications on YARN.

SparkSqlJob
A Dataproc job for running Apache Spark SQL queries.

StartClusterRequest
A request to start a cluster.

StopClusterRequest
A request to stop a cluster.

SubmitJobRequest
A request to submit a job.

TemplateParameter
A configurable parameter that replaces one or more fields in the template. Parameterizable fields:

UpdateAutoscalingPolicyRequest
A request to update an autoscaling policy.

UpdateClusterRequest
A request to update a cluster.

UpdateJobRequest
A request to update a job.

UpdateWorkflowTemplateRequest
A request to update a workflow template.

ValueValidation
Validation based on a list of allowed values.

WorkflowGraph
The workflow graph.

WorkflowMetadata
A Dataproc workflow template resource.

WorkflowNode
The workflow node.

WorkflowTemplate
A Dataproc workflow template resource.

WorkflowTemplatePlacement
Specifies workflow execution target.

YarnApplication
A YARN application created by a job. Application information is a subset of org.apache.hadoop.yarn.proto.YarnProtos.ApplicationReportProto.

Enums

Component
Cluster components that can be activated. Next ID: 16.