Module googapis::google::cloud::dataqna::v1alpha

Modules

annotated_string

Nested message and enum types in AnnotatedString.

auto_suggestion_service_client

Generated client implementations.

execution_info

Nested message and enum types in ExecutionInfo.

interpret_error

Nested message and enum types in InterpretError.

interpretation_structure

Nested message and enum types in InterpretationStructure.

question_service_client

Generated client implementations.

suggestion_info

Nested message and enum types in SuggestionInfo.

user_feedback

Nested message and enum types in UserFeedback.

Structs

AnnotatedString

Describes string annotation from both semantic and formatting perspectives. Example:

BigQueryJob

BigQuery job information. This can be used to query the BigQuery API and retrieve the current job’s status (using [jobs.get](https://cloud.google.com/bigquery/docs/reference/rest/v2/jobs/get)).

CreateQuestionRequest

Request to create a question resource.

DataQuery

Representation of the data query for the backend. This is provided for informational purposes only. Clients should not use it to send it to the backend directly, but rather use the execute RPC to trigger the execution. Using the execute RPC is needed in order to track the state of a question and report on it correctly to the data administrators.

DebugFlags

Configuriation of debug flags.

ExecuteQuestionRequest

Request to execute an interpretation.

ExecutionInfo

Information about the backend status (such as BigQuery) of the execution.

GetQuestionRequest

A request to get a previously created question.

GetUserFeedbackRequest

Request to get user feedback.

HumanReadable

Human readable interpretation.

InterpretError

Details on the failure to interpret the question.

Interpretation

An interpretation of a natural language query.

InterpretationStructure

Information about the interpretation structure that helps to understand and visualize the response.

Question

The question resource represents a natural language query, its settings, understanding generated by the system, and answer retrieval status. A question cannot be modified.

SuggestQueriesRequest

Request for query suggestions.

SuggestQueriesResponse

Response to SuggestQueries.

Suggestion

A suggestion for a query with a ranking score.

SuggestionInfo

Detailed information about the suggestion.

UpdateUserFeedbackRequest

Request to updates user feedback.

UserFeedback

Feedback provided by a user.

Enums

InterpretEntity

Query entities of an interpretation.

SuggestionType

The type of suggestion.