Module googapis::google::cloud::deploy::v1

Modules

cloud_deploy_client

Generated client implementations.

delivery_pipeline

Nested message and enum types in DeliveryPipeline.

execution_config

Nested message and enum types in ExecutionConfig.

release

Nested message and enum types in Release.

rollout

Nested message and enum types in Rollout.

target

Nested message and enum types in Target.

target_artifact

Nested message and enum types in TargetArtifact.

Structs

ApproveRolloutRequest

The request object used by ApproveRollout.

ApproveRolloutResponse

The response object from ApproveRollout.

BuildArtifact

Description of an a image to use during Skaffold rendering.

Config

Service-wide configuration.

CreateDeliveryPipelineRequest

The request object for CreateDeliveryPipeline.

CreateReleaseRequest

The request object for CreateRelease,

CreateRolloutRequest

CreateRolloutRequest is the request object used by CreateRollout.

CreateTargetRequest

The request object for CreateTarget.

DefaultPool

Execution using the default Cloud Build pool.

DeleteDeliveryPipelineRequest

The request object for DeleteDeliveryPipeline.

DeleteTargetRequest

The request object for DeleteTarget.

DeliveryPipeline

A DeliveryPipeline resource in the Google Cloud Deploy API.

ExecutionConfig

Configuration of the environment to use when calling Skaffold.

GetConfigRequest

Request to get a configuration.

GetDeliveryPipelineRequest

The request object for GetDeliveryPipeline

GetReleaseRequest

The request object for GetRelease.

GetRolloutRequest

GetRolloutRequest is the request object used by GetRollout.

GetTargetRequest

The request object for GetTarget.

GkeCluster

Information specifying a GKE Cluster.

ListDeliveryPipelinesRequest

The request object for ListDeliveryPipelines.

ListDeliveryPipelinesResponse

The response object from ListDeliveryPipelines.

ListReleasesRequest

The request object for ListReleases.

ListReleasesResponse

The response object from ListReleases.

ListRolloutsRequest

ListRolloutsRequest is the request object used by ListRollouts.

ListRolloutsResponse

ListRolloutsResponse is the response object reutrned by ListRollouts.

ListTargetsRequest

The request object for ListTargets.

ListTargetsResponse

The response object from ListTargets.

OperationMetadata

Represents the metadata of the long-running operation.

PipelineCondition

PipelineCondition contains all conditions relevant to a Delivery Pipeline.

PipelineReadyCondition

PipelineReadyCondition contains information around the status of the Pipeline.

PrivatePool

Execution using a private Cloud Build pool.

Release

A Release resource in the Google Cloud Deploy API.

Rollout

A Rollout resource in the Google Cloud Deploy API.

SerialPipeline

SerialPipeline defines a sequential set of stages for a DeliveryPipeline.

SkaffoldVersion

Details of a supported Skaffold version.

Stage

Stage specifies a location to which to deploy.

Target

A Target resource in the Google Cloud Deploy API.

TargetArtifact

The artifacts produced by a target render operation.

TargetsPresentCondition

TargetsPresentCondition contains information on any Targets defined in the Delivery Pipeline that do not actually exist.

UpdateDeliveryPipelineRequest

The request object for UpdateDeliveryPipeline.

UpdateTargetRequest

The request object for UpdateTarget.