Module googapis::google::cloud::dialogflow::v2

Modules

agent
Nested message and enum types in Agent.

agent_assistant_feedback
Nested message and enum types in AgentAssistantFeedback.

agent_assistant_record
Nested message and enum types in AgentAssistantRecord.

agents_client
Generated client implementations.

analyze_content_request
Nested message and enum types in AnalyzeContentRequest.

answer_feedback
Nested message and enum types in AnswerFeedback.

answer_record
Nested message and enum types in AnswerRecord.

answer_records_client
Generated client implementations.

automated_agent_reply
Nested message and enum types in AutomatedAgentReply.

batch_update_entity_types_request
Nested message and enum types in BatchUpdateEntityTypesRequest.

batch_update_intents_request
Nested message and enum types in BatchUpdateIntentsRequest.

contexts_client
Generated client implementations.

conversation
Nested message and enum types in Conversation.

conversation_event
Nested message and enum types in ConversationEvent.

conversation_profiles_client
Generated client implementations.

conversations_client
Generated client implementations.

document
Nested message and enum types in Document.

documents_client
Generated client implementations.

entity_type
Nested message and enum types in EntityType.

entity_types_client
Generated client implementations.

environment
Nested message and enum types in Environment.

environment_history
Nested message and enum types in EnvironmentHistory.

environments_client
Generated client implementations.

export_agent_response
Nested message and enum types in ExportAgentResponse.

fulfillment
Nested message and enum types in Fulfillment.

fulfillments_client
Generated client implementations.

human_agent_assistant_config
Nested message and enum types in HumanAgentAssistantConfig.

human_agent_handoff_config
Nested message and enum types in HumanAgentHandoffConfig.

import_agent_request
Nested message and enum types in ImportAgentRequest.

intent
Nested message and enum types in Intent.

intents_client
Generated client implementations.

knowledge_bases_client
Generated client implementations.

knowledge_operation_metadata
Nested message and enum types in KnowledgeOperationMetadata.

notification_config
Nested message and enum types in NotificationConfig.

participant
Nested message and enum types in Participant.

participants_client
Generated client implementations.

query_input
Nested message and enum types in QueryInput.

reload_document_request
Nested message and enum types in ReloadDocumentRequest.

restore_agent_request
Nested message and enum types in RestoreAgentRequest.

session_entity_type
Nested message and enum types in SessionEntityType.

session_entity_types_client
Generated client implementations.

sessions_client
Generated client implementations.

streaming_recognition_result
Nested message and enum types in StreamingRecognitionResult.

suggestion_feature
Nested message and enum types in SuggestionFeature.

suggestion_result
Nested message and enum types in SuggestionResult.

validation_error
Nested message and enum types in ValidationError.

version
Nested message and enum types in Version.

versions_client
Generated client implementations.

Structs

Agent
A Dialogflow agent is a virtual agent that handles conversations with your end-users. It is a natural language understanding module that understands the nuances of human language. Dialogflow translates end-user text or audio during a conversation to structured data that your apps and services can understand. You design and build a Dialogflow agent to handle the types of conversations required for your system.

AgentAssistantFeedback
Detail feedback of Agent Assist result.

AgentAssistantRecord
Represents a record of a human agent assist answer.

AnalyzeContentRequest
The request message for [Participants.AnalyzeContent][google.cloud.dialogflow.v2.Participants.AnalyzeContent].

AnalyzeContentResponse
The response message for [Participants.AnalyzeContent][google.cloud.dialogflow.v2.Participants.AnalyzeContent].

AnnotatedMessagePart
Represents a part of a message possibly annotated with an entity. The part can be an entity or purely a part of the message between two entities or message start/end.

AnswerFeedback
Represents feedback the customer has about the quality & correctness of a certain answer in a conversation.

AnswerRecord
Answer records are records to manage answer history and feedbacks for Dialogflow.

ArticleAnswer
Represents article answer.

AutomatedAgentConfig
Defines the Automated Agent to connect to a conversation.

AutomatedAgentReply
Represents a response from an automated agent.

BatchCreateEntitiesRequest
The request message for [EntityTypes.BatchCreateEntities][google.cloud.dialogflow.v2.EntityTypes.BatchCreateEntities].

BatchDeleteEntitiesRequest
The request message for [EntityTypes.BatchDeleteEntities][google.cloud.dialogflow.v2.EntityTypes.BatchDeleteEntities].

BatchDeleteEntityTypesRequest
The request message for [EntityTypes.BatchDeleteEntityTypes][google.cloud.dialogflow.v2.EntityTypes.BatchDeleteEntityTypes].

BatchDeleteIntentsRequest
The request message for [Intents.BatchDeleteIntents][google.cloud.dialogflow.v2.Intents.BatchDeleteIntents].

BatchUpdateEntitiesRequest
The request message for [EntityTypes.BatchUpdateEntities][google.cloud.dialogflow.v2.EntityTypes.BatchUpdateEntities].

BatchUpdateEntityTypesRequest
The request message for [EntityTypes.BatchUpdateEntityTypes][google.cloud.dialogflow.v2.EntityTypes.BatchUpdateEntityTypes].

BatchUpdateEntityTypesResponse
The response message for [EntityTypes.BatchUpdateEntityTypes][google.cloud.dialogflow.v2.EntityTypes.BatchUpdateEntityTypes].

BatchUpdateIntentsRequest

BatchUpdateIntentsResponse
The response message for [Intents.BatchUpdateIntents][google.cloud.dialogflow.v2.Intents.BatchUpdateIntents].

CompleteConversationRequest
The request message for [Conversations.CompleteConversation][google.cloud.dialogflow.v2.Conversations.CompleteConversation].

Context
Dialogflow contexts are similar to natural language context. If a person says to you “they are orange”, you need context in order to understand what “they” is referring to. Similarly, for Dialogflow to handle an end-user expression like that, it needs to be provided with context in order to correctly match an intent.

Conversation
Represents a conversation. A conversation is an interaction between an agent, including live agents and Dialogflow agents, and a support customer. Conversations can include phone calls and text-based chat sessions.

ConversationEvent
Represents a notification sent to Pub/Sub subscribers for conversation lifecycle events.

ConversationPhoneNumber
Represents a phone number for telephony integration. It allows for connecting a particular conversation over telephony.

ConversationProfile
Defines the services to connect to incoming Dialogflow conversations.

CreateContextRequest
The request message for [Contexts.CreateContext][google.cloud.dialogflow.v2.Contexts.CreateContext].

CreateConversationProfileRequest
The request message for [ConversationProfiles.CreateConversationProfile][google.cloud.dialogflow.v2.ConversationProfiles.CreateConversationProfile].

CreateConversationRequest
The request message for [Conversations.CreateConversation][google.cloud.dialogflow.v2.Conversations.CreateConversation].

CreateDocumentRequest
Request message for [Documents.CreateDocument][google.cloud.dialogflow.v2.Documents.CreateDocument].

CreateEntityTypeRequest
The request message for [EntityTypes.CreateEntityType][google.cloud.dialogflow.v2.EntityTypes.CreateEntityType].

CreateEnvironmentRequest
The request message for [Environments.CreateEnvironment][google.cloud.dialogflow.v2.Environments.CreateEnvironment].

CreateIntentRequest
The request message for [Intents.CreateIntent][google.cloud.dialogflow.v2.Intents.CreateIntent].

CreateKnowledgeBaseRequest
Request message for [KnowledgeBases.CreateKnowledgeBase][google.cloud.dialogflow.v2.KnowledgeBases.CreateKnowledgeBase].

CreateParticipantRequest
The request message for [Participants.CreateParticipant][google.cloud.dialogflow.v2.Participants.CreateParticipant].

CreateSessionEntityTypeRequest
The request message for [SessionEntityTypes.CreateSessionEntityType][google.cloud.dialogflow.v2.SessionEntityTypes.CreateSessionEntityType].

CreateVersionRequest
The request message for [Versions.CreateVersion][google.cloud.dialogflow.v2.Versions.CreateVersion].

DeleteAgentRequest
The request message for [Agents.DeleteAgent][google.cloud.dialogflow.v2.Agents.DeleteAgent].

DeleteAllContextsRequest
The request message for [Contexts.DeleteAllContexts][google.cloud.dialogflow.v2.Contexts.DeleteAllContexts].

DeleteContextRequest
The request message for [Contexts.DeleteContext][google.cloud.dialogflow.v2.Contexts.DeleteContext].

DeleteConversationProfileRequest
The request message for [ConversationProfiles.DeleteConversationProfile][google.cloud.dialogflow.v2.ConversationProfiles.DeleteConversationProfile].

DeleteDocumentRequest
Request message for [Documents.DeleteDocument][google.cloud.dialogflow.v2.Documents.DeleteDocument].

DeleteEntityTypeRequest
The request message for [EntityTypes.DeleteEntityType][google.cloud.dialogflow.v2.EntityTypes.DeleteEntityType].

DeleteEnvironmentRequest
The request message for [Environments.DeleteEnvironment][google.cloud.dialogflow.v2.Environments.DeleteEnvironment].

DeleteIntentRequest
The request message for [Intents.DeleteIntent][google.cloud.dialogflow.v2.Intents.DeleteIntent].

DeleteKnowledgeBaseRequest
Request message for [KnowledgeBases.DeleteKnowledgeBase][google.cloud.dialogflow.v2.KnowledgeBases.DeleteKnowledgeBase].

DeleteSessionEntityTypeRequest
The request message for [SessionEntityTypes.DeleteSessionEntityType][google.cloud.dialogflow.v2.SessionEntityTypes.DeleteSessionEntityType].

DeleteVersionRequest
The request message for [Versions.DeleteVersion][google.cloud.dialogflow.v2.Versions.DeleteVersion].

DetectIntentRequest
The request to detect user’s intent.

DetectIntentResponse
The message returned from the DetectIntent method.

Document
A knowledge document to be used by a [KnowledgeBase][google.cloud.dialogflow.v2.KnowledgeBase].

DtmfParameters
The message in the response that indicates the parameters of DTMF.

EntityType
Each intent parameter has a type, called the entity type, which dictates exactly how data from an end-user expression is extracted.

EntityTypeBatch
This message is a wrapper around a collection of entity types.

Environment
You can create multiple versions of your agent and publish them to separate environments.

EnvironmentHistory
The response message for [Environments.GetEnvironmentHistory][google.cloud.dialogflow.v2.Environments.GetEnvironmentHistory].

EventInput
Events allow for matching intents by event name instead of the natural language input. For instance, input <event: { name: "welcome_event", parameters: { name: "Sam" } }> can trigger a personalized welcome response. The parameter name may be used by the agent in the response: "Hello #welcome_event.name! What can I do for you today?".

ExportAgentRequest
The request message for [Agents.ExportAgent][google.cloud.dialogflow.v2.Agents.ExportAgent].

ExportAgentResponse
The response message for [Agents.ExportAgent][google.cloud.dialogflow.v2.Agents.ExportAgent].

FaqAnswer
Represents answer from “frequently asked questions”.

Fulfillment
By default, your agent responds to a matched intent with a static response. As an alternative, you can provide a more dynamic response by using fulfillment. When you enable fulfillment for an intent, Dialogflow responds to that intent by calling a service that you define. For example, if an end-user wants to schedule a haircut on Friday, your service can check your database and respond to the end-user with availability information for Friday.

GetAgentRequest
The request message for [Agents.GetAgent][google.cloud.dialogflow.v2.Agents.GetAgent].

GetContextRequest
The request message for [Contexts.GetContext][google.cloud.dialogflow.v2.Contexts.GetContext].

GetConversationProfileRequest
The request message for [ConversationProfiles.GetConversationProfile][google.cloud.dialogflow.v2.ConversationProfiles.GetConversationProfile].

GetConversationRequest
The request message for [Conversations.GetConversation][google.cloud.dialogflow.v2.Conversations.GetConversation].

GetDocumentRequest
Request message for [Documents.GetDocument][google.cloud.dialogflow.v2.Documents.GetDocument].

GetEntityTypeRequest
The request message for [EntityTypes.GetEntityType][google.cloud.dialogflow.v2.EntityTypes.GetEntityType].

GetEnvironmentHistoryRequest
The request message for [Environments.GetEnvironmentHistory][google.cloud.dialogflow.v2.Environments.GetEnvironmentHistory].

GetEnvironmentRequest
The request message for [Environments.GetEnvironment][google.cloud.dialogflow.v2.Environments.GetEnvironment].

GetFulfillmentRequest
The request message for [Fulfillments.GetFulfillment][google.cloud.dialogflow.v2.Fulfillments.GetFulfillment].

GetIntentRequest
The request message for [Intents.GetIntent][google.cloud.dialogflow.v2.Intents.GetIntent].

GetKnowledgeBaseRequest
Request message for [KnowledgeBases.GetKnowledgeBase][google.cloud.dialogflow.v2.KnowledgeBases.GetKnowledgeBase].

GetParticipantRequest
The request message for [Participants.GetParticipant][google.cloud.dialogflow.v2.Participants.GetParticipant].

GetSessionEntityTypeRequest
The request message for [SessionEntityTypes.GetSessionEntityType][google.cloud.dialogflow.v2.SessionEntityTypes.GetSessionEntityType].

GetValidationResultRequest
The request message for [Agents.GetValidationResult][google.cloud.dialogflow.v2.Agents.GetValidationResult].

GetVersionRequest
The request message for [Versions.GetVersion][google.cloud.dialogflow.v2.Versions.GetVersion].

HumanAgentAssistantConfig
Defines the Human Agent Assist to connect to a conversation.

HumanAgentAssistantEvent
Represents a notification sent to Cloud Pub/Sub subscribers for human agent assistant events in a specific conversation.

HumanAgentHandoffConfig
Defines the hand off to a live agent, typically on which external agent service provider to connect to a conversation.

ImportAgentRequest
The request message for [Agents.ImportAgent][google.cloud.dialogflow.v2.Agents.ImportAgent].

InputAudioConfig
Instructs the speech recognizer how to process the audio content.

Intent
An intent categorizes an end-user’s intention for one conversation turn. For each agent, you define many intents, where your combined intents can handle a complete conversation. When an end-user writes or says something, referred to as an end-user expression or end-user input, Dialogflow matches the end-user input to the best intent in your agent. Matching an intent is also known as intent classification.

IntentBatch
This message is a wrapper around a collection of intents.

KnowledgeBase
A knowledge base represents a collection of knowledge documents that you provide to Dialogflow. Your knowledge documents contain information that may be useful during conversations with end-users. Some Dialogflow features use knowledge bases when looking for a response to an end-user input.

KnowledgeOperationMetadata
Metadata in google::longrunning::Operation for Knowledge operations.

ListAnswerRecordsRequest
Request message for [AnswerRecords.ListAnswerRecords][google.cloud.dialogflow.v2.AnswerRecords.ListAnswerRecords].

ListAnswerRecordsResponse
Response message for [AnswerRecords.ListAnswerRecords][google.cloud.dialogflow.v2.AnswerRecords.ListAnswerRecords].

ListContextsRequest
The request message for [Contexts.ListContexts][google.cloud.dialogflow.v2.Contexts.ListContexts].

ListContextsResponse
The response message for [Contexts.ListContexts][google.cloud.dialogflow.v2.Contexts.ListContexts].

ListConversationProfilesRequest
The request message for [ConversationProfiles.ListConversationProfiles][google.cloud.dialogflow.v2.ConversationProfiles.ListConversationProfiles].

ListConversationProfilesResponse
The response message for [ConversationProfiles.ListConversationProfiles][google.cloud.dialogflow.v2.ConversationProfiles.ListConversationProfiles].

ListConversationsRequest
The request message for [Conversations.ListConversations][google.cloud.dialogflow.v2.Conversations.ListConversations].

ListConversationsResponse
The response message for [Conversations.ListConversations][google.cloud.dialogflow.v2.Conversations.ListConversations].

ListDocumentsRequest
Request message for [Documents.ListDocuments][google.cloud.dialogflow.v2.Documents.ListDocuments].

ListDocumentsResponse
Response message for [Documents.ListDocuments][google.cloud.dialogflow.v2.Documents.ListDocuments].

ListEntityTypesRequest
The request message for [EntityTypes.ListEntityTypes][google.cloud.dialogflow.v2.EntityTypes.ListEntityTypes].

ListEntityTypesResponse
The response message for [EntityTypes.ListEntityTypes][google.cloud.dialogflow.v2.EntityTypes.ListEntityTypes].

ListEnvironmentsRequest
The request message for [Environments.ListEnvironments][google.cloud.dialogflow.v2.Environments.ListEnvironments].

ListEnvironmentsResponse
The response message for [Environments.ListEnvironments][google.cloud.dialogflow.v2.Environments.ListEnvironments].

ListIntentsRequest
The request message for [Intents.ListIntents][google.cloud.dialogflow.v2.Intents.ListIntents].

ListIntentsResponse
The response message for [Intents.ListIntents][google.cloud.dialogflow.v2.Intents.ListIntents].

ListKnowledgeBasesRequest
Request message for [KnowledgeBases.ListKnowledgeBases][google.cloud.dialogflow.v2.KnowledgeBases.ListKnowledgeBases].

ListKnowledgeBasesResponse
Response message for [KnowledgeBases.ListKnowledgeBases][google.cloud.dialogflow.v2.KnowledgeBases.ListKnowledgeBases].

ListMessagesRequest
The request message for [Conversations.ListMessages][google.cloud.dialogflow.v2.Conversations.ListMessages].

ListMessagesResponse
The response message for [Conversations.ListMessages][google.cloud.dialogflow.v2.Conversations.ListMessages].

ListParticipantsRequest
The request message for [Participants.ListParticipants][google.cloud.dialogflow.v2.Participants.ListParticipants].

ListParticipantsResponse
The response message for [Participants.ListParticipants][google.cloud.dialogflow.v2.Participants.ListParticipants].

ListSessionEntityTypesRequest
The request message for [SessionEntityTypes.ListSessionEntityTypes][google.cloud.dialogflow.v2.SessionEntityTypes.ListSessionEntityTypes].

ListSessionEntityTypesResponse
The response message for [SessionEntityTypes.ListSessionEntityTypes][google.cloud.dialogflow.v2.SessionEntityTypes.ListSessionEntityTypes].

ListVersionsRequest
The request message for [Versions.ListVersions][google.cloud.dialogflow.v2.Versions.ListVersions].

ListVersionsResponse
The response message for [Versions.ListVersions][google.cloud.dialogflow.v2.Versions.ListVersions].

LoggingConfig
Defines logging behavior for conversation lifecycle events.

Message
Represents a message posted into a conversation.

MessageAnnotation
Represents the result of annotation for the message.

NotificationConfig
Defines notification behavior.

OriginalDetectIntentRequest
Represents the contents of the original request that was passed to the [Streaming]DetectIntent call.

OutputAudio
Represents the natural language speech audio to be played to the end user.

OutputAudioConfig
Instructs the speech synthesizer on how to generate the output audio content. If this audio config is supplied in a request, it overrides all existing text-to-speech settings applied to the agent.

Participant
Represents a conversation participant (human agent, virtual agent, end-user).

QueryInput
Represents the query input. It can contain either:

QueryParameters
Represents the parameters of the conversational query.

QueryResult
Represents the result of conversational query or event processing.

ReloadDocumentRequest
Request message for [Documents.ReloadDocument][google.cloud.dialogflow.v2.Documents.ReloadDocument].

RestoreAgentRequest
The request message for [Agents.RestoreAgent][google.cloud.dialogflow.v2.Agents.RestoreAgent].

SearchAgentsRequest
The request message for [Agents.SearchAgents][google.cloud.dialogflow.v2.Agents.SearchAgents].

SearchAgentsResponse
The response message for [Agents.SearchAgents][google.cloud.dialogflow.v2.Agents.SearchAgents].

Sentiment
The sentiment, such as positive/negative feeling or association, for a unit of analysis, such as the query text.

SentimentAnalysisRequestConfig
Configures the types of sentiment analysis to perform.

SentimentAnalysisResult
The result of sentiment analysis. Sentiment analysis inspects user input and identifies the prevailing subjective opinion, especially to determine a user’s attitude as positive, negative, or neutral. For [Participants.DetectIntent][], it needs to be configured in [DetectIntentRequest.query_params][google.cloud.dialogflow.v2.DetectIntentRequest.query_params]. For [Participants.StreamingDetectIntent][], it needs to be configured in [StreamingDetectIntentRequest.query_params][google.cloud.dialogflow.v2.StreamingDetectIntentRequest.query_params]. And for [Participants.AnalyzeContent][google.cloud.dialogflow.v2.Participants.AnalyzeContent] and [Participants.StreamingAnalyzeContent][google.cloud.dialogflow.v2.Participants.StreamingAnalyzeContent], it needs to be configured in [ConversationProfile.human_agent_assistant_config][google.cloud.dialogflow.v2.ConversationProfile.human_agent_assistant_config]

SessionEntityType
A session represents a conversation between a Dialogflow agent and an end-user. You can create special entities, called session entities, during a session. Session entities can extend or replace custom entity types and only exist during the session that they were created for. All session data, including session entities, is stored by Dialogflow for 20 minutes.

SetAgentRequest
The request message for [Agents.SetAgent][google.cloud.dialogflow.v2.Agents.SetAgent].

SpeechContext
Hints for the speech recognizer to help with recognition in a specific conversation state.

SpeechToTextConfig
Configures speech transcription for [ConversationProfile][google.cloud.dialogflow.v2.ConversationProfile].

SpeechWordInfo
Information for a word recognized by the speech recognizer.

StreamingDetectIntentRequest
The top-level message sent by the client to the [Sessions.StreamingDetectIntent][google.cloud.dialogflow.v2.Sessions.StreamingDetectIntent] method.

StreamingDetectIntentResponse
The top-level message returned from the StreamingDetectIntent method.

StreamingRecognitionResult
Contains a speech recognition result corresponding to a portion of the audio that is currently being processed or an indication that this is the end of the single requested utterance.

SuggestArticlesRequest
The request message for [Participants.SuggestArticles][google.cloud.dialogflow.v2.Participants.SuggestArticles].

SuggestArticlesResponse
The response message for [Participants.SuggestArticles][google.cloud.dialogflow.v2.Participants.SuggestArticles].

SuggestFaqAnswersRequest
The request message for [Participants.SuggestFaqAnswers][google.cloud.dialogflow.v2.Participants.SuggestFaqAnswers].

SuggestFaqAnswersResponse
The request message for [Participants.SuggestFaqAnswers][google.cloud.dialogflow.v2.Participants.SuggestFaqAnswers].

SuggestionFeature
The type of Human Agent Assistant API suggestion to perform, and the maximum number of results to return for that type. Multiple Feature objects can be specified in the features list.

SuggestionResult
One response of different type of suggestion response which is used in the response of [Participants.AnalyzeContent][google.cloud.dialogflow.v2.Participants.AnalyzeContent] and [Participants.AnalyzeContent][google.cloud.dialogflow.v2.Participants.AnalyzeContent], as well as [HumanAgentAssistantEvent][google.cloud.dialogflow.v2.HumanAgentAssistantEvent].

SynthesizeSpeechConfig
Configuration of how speech should be synthesized.

TextInput
Represents the natural language text to be processed.

TextToSpeechSettings
Instructs the speech synthesizer on how to generate the output audio content.

TrainAgentRequest
The request message for [Agents.TrainAgent][google.cloud.dialogflow.v2.Agents.TrainAgent].

UpdateAnswerRecordRequest
Request message for [AnswerRecords.UpdateAnswerRecord][google.cloud.dialogflow.v2.AnswerRecords.UpdateAnswerRecord].

UpdateContextRequest
The request message for [Contexts.UpdateContext][google.cloud.dialogflow.v2.Contexts.UpdateContext].

UpdateConversationProfileRequest
The request message for [ConversationProfiles.UpdateConversationProfile][google.cloud.dialogflow.v2.ConversationProfiles.UpdateConversationProfile].

UpdateDocumentRequest
Request message for [Documents.UpdateDocument][google.cloud.dialogflow.v2.Documents.UpdateDocument].

UpdateEntityTypeRequest
The request message for [EntityTypes.UpdateEntityType][google.cloud.dialogflow.v2.EntityTypes.UpdateEntityType].

UpdateEnvironmentRequest
The request message for [Environments.UpdateEnvironment][google.cloud.dialogflow.v2.Environments.UpdateEnvironment].

UpdateFulfillmentRequest
The request message for [Fulfillments.UpdateFulfillment][google.cloud.dialogflow.v2.Fulfillments.UpdateFulfillment].

UpdateIntentRequest
The request message for [Intents.UpdateIntent][google.cloud.dialogflow.v2.Intents.UpdateIntent].

UpdateKnowledgeBaseRequest
Request message for [KnowledgeBases.UpdateKnowledgeBase][google.cloud.dialogflow.v2.KnowledgeBases.UpdateKnowledgeBase].

UpdateParticipantRequest
The request message for [Participants.UpdateParticipant][google.cloud.dialogflow.v2.Participants.UpdateParticipant].

UpdateSessionEntityTypeRequest
The request message for [SessionEntityTypes.UpdateSessionEntityType][google.cloud.dialogflow.v2.SessionEntityTypes.UpdateSessionEntityType].

UpdateVersionRequest
The request message for [Versions.UpdateVersion][google.cloud.dialogflow.v2.Versions.UpdateVersion].

ValidationError
Represents a single validation error.

ValidationResult
Represents the output of agent validation.

Version
You can create multiple versions of your agent and publish them to separate environments.

VoiceSelectionParams
Description of which voice to use for speech synthesis.

WebhookRequest
The request message for a webhook call.

WebhookResponse
The response message for a webhook call.

Enums

AudioEncoding
Audio encoding of the audio content sent in the conversational query request. Refer to the Cloud Speech API documentation for more details.

IntentView
Represents the options for views of an intent. An intent can be a sizable object. Therefore, we provide a resource view that does not return training phrases in the response by default.

OutputAudioEncoding
Audio encoding of the output audio format in Text-To-Speech.

SpeechModelVariant
Variant of the specified [Speech model][google.cloud.dialogflow.v2.InputAudioConfig.model] to use.

SsmlVoiceGender
Gender of the voice as described in SSML voice element.