Struct googapis::google::cloud::dialogflow::v2::Intent

pub struct Intent {

    pub name: String,
    pub display_name: String,
    pub webhook_state: i32,
    pub priority: i32,
    pub is_fallback: bool,
    pub ml_disabled: bool,
    pub live_agent_handoff: bool,
    pub end_interaction: bool,
    pub input_context_names: Vec<String>,
    pub events: Vec<String>,
    pub training_phrases: Vec<TrainingPhrase>,
    pub action: String,
    pub output_contexts: Vec<Context>,
    pub reset_contexts: bool,
    pub parameters: Vec<Parameter>,
    pub messages: Vec<Message>,
    pub default_response_platforms: Vec<i32>,
    pub root_followup_intent_name: String,
    pub parent_followup_intent_name: String,
    pub followup_intent_info: Vec<FollowupIntentInfo>,

}

An intent categorizes an end-user’s intention for one conversation turn. For each agent, you define many intents, where your combined intents can handle a complete conversation. When an end-user writes or says something, referred to as an end-user expression or end-user input, Dialogflow matches the end-user input to the best intent in your agent. Matching an intent is also known as intent classification.

For more information, see the intent guide.

Fields
name: String

Optional. The unique identifier of this intent. Required for [Intents.UpdateIntent][google.cloud.dialogflow.v2.Intents.UpdateIntent] and [Intents.BatchUpdateIntents][google.cloud.dialogflow.v2.Intents.BatchUpdateIntents] methods. Format: projects/<Project ID>/agent/intents/<Intent ID>.

display_name: String

Required. The name of this intent.

webhook_state: i32

Optional. Indicates whether webhooks are enabled for the intent.

priority: i32

Optional. The priority of this intent. Higher numbers represent higher priorities.

If the supplied value is unspecified or 0, the service translates the value to 500,000, which corresponds to the Normal priority in the console.

If the supplied value is negative, the intent is ignored in runtime detect intent requests.

is_fallback: bool

Optional. Indicates whether this is a fallback intent.

ml_disabled: bool

Optional. Indicates whether Machine Learning is disabled for the intent. Note: If ml_disabled setting is set to true, then this intent is not taken into account during inference in ML ONLY match mode. Also, auto-markup in the UI is turned off.

live_agent_handoff: bool

Optional. Indicates that a live agent should be brought in to handle the interaction with the user. In most cases, when you set this flag to true, you would also want to set end_interaction to true as well. Default is false.

end_interaction: bool

Optional. Indicates that this intent ends an interaction. Some integrations (e.g., Actions on Google or Dialogflow phone gateway) use this information to close interaction with an end user. Default is false.

input_context_names: Vec<String>

Optional. The list of context names required for this intent to be triggered. Format: projects/<Project ID>/agent/sessions/-/contexts/<Context ID>.

events: Vec<String>

Optional. The collection of event names that trigger the intent. If the collection of input contexts is not empty, all of the contexts must be present in the active user session for an event to trigger this intent. Event names are limited to 150 characters.

training_phrases: Vec<TrainingPhrase>

Optional. The collection of examples that the agent is trained on.

action: String

Optional. The name of the action associated with the intent. Note: The action name must not contain whitespaces.

output_contexts: Vec<Context>

Optional. The collection of contexts that are activated when the intent is matched. Context messages in this collection should not set the parameters field. Setting the lifespan_count to 0 will reset the context when the intent is matched. Format: projects/<Project ID>/agent/sessions/-/contexts/<Context ID>.

reset_contexts: bool

Optional. Indicates whether to delete all contexts in the current session when this intent is matched.

parameters: Vec<Parameter>

Optional. The collection of parameters associated with the intent.

messages: Vec<Message>

Optional. The collection of rich messages corresponding to the Response field in the Dialogflow console.

default_response_platforms: Vec<i32>

Optional. The list of platforms for which the first responses will be copied from the messages in PLATFORM_UNSPECIFIED (i.e. default platform).

root_followup_intent_name: String

Output only. Read-only. The unique identifier of the root intent in the chain of followup intents. It identifies the correct followup intents chain for this intent. We populate this field only in the output.

Format: projects/<Project ID>/agent/intents/<Intent ID>.

parent_followup_intent_name: String

Read-only after creation. The unique identifier of the parent intent in the chain of followup intents. You can set this field when creating an intent, for example with [CreateIntent][google.cloud.dialogflow.v2.Intents.CreateIntent] or [BatchUpdateIntents][google.cloud.dialogflow.v2.Intents.BatchUpdateIntents], in order to make this intent a followup intent.

It identifies the parent followup intent. Format: projects/<Project ID>/agent/intents/<Intent ID>.

followup_intent_info: Vec<FollowupIntentInfo>

Output only. Read-only. Information about all followup intents that have this intent as a direct or indirect parent. We populate this field only in the output.

Implementations

impl Intent

pub fn webhook_state(&self) -> WebhookState

Returns the enum value of webhook_state, or the default if the field is set to an invalid enum value.

pub fn set_webhook_state(&mut self, value: WebhookState)

Sets webhook_state to the provided enum value.

pub fn default_response_platforms(
&self
) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<Platform>>

Returns an iterator which yields the valid enum values contained in default_response_platforms.

pub fn push_default_response_platforms(&mut self, value: Platform)

Appends the provided enum value to default_response_platforms.