Module googapis::google::cloud::documentai::v1beta3

Modules

batch_documents_input_config

Nested message and enum types in BatchDocumentsInputConfig.

batch_process_metadata

Nested message and enum types in BatchProcessMetadata.

batch_process_request

Nested message and enum types in BatchProcessRequest.

common_operation_metadata

Nested message and enum types in CommonOperationMetadata.

document

Nested message and enum types in Document.

document_output_config

Nested message and enum types in DocumentOutputConfig.

document_processor_service_client

Generated client implementations.

human_review_status

Nested message and enum types in HumanReviewStatus.

process_request

Nested message and enum types in ProcessRequest.

processor

Nested message and enum types in Processor.

processor_type

Nested message and enum types in ProcessorType.

review_document_operation_metadata

Nested message and enum types in ReviewDocumentOperationMetadata.

review_document_request

Nested message and enum types in ReviewDocumentRequest.

Structs

BatchDocumentsInputConfig

The common config to specify a set of documents used as input.

BatchProcessMetadata

The long running operation metadata for batch process method.

BatchProcessRequest

Request message for batch process document method.

BatchProcessResponse

Response message for batch process document method.

BoundingPoly

A bounding polygon for the detected image annotation.

CommonOperationMetadata

The common metadata for long running operations.

CreateProcessorRequest

Request message for create a processor. Notice this request is sent to a regionalized backend service, and if the processor type is not available on that region, the creation will fail.

DeleteProcessorMetadata

The long running operation metadata for delete processor method.

DeleteProcessorRequest

Request message for the delete processor method.

DisableProcessorMetadata

The long running operation metadata for disable processor method.

DisableProcessorRequest

Request message for the disable processor method.

DisableProcessorResponse

Response message for the disable processor method.

Document

Document represents the canonical document resource in Document Understanding AI. It is an interchange format that provides insights into documents and allows for collaboration between users and Document Understanding AI to iterate and optimize for quality.

DocumentOutputConfig

Config that controls the output of documents. All documents will be written as a JSON file.

EnableProcessorMetadata

The long running operation metadata for enable processor method.

EnableProcessorRequest

Request message for the enable processor method.

EnableProcessorResponse

Response message for the enable processor method.

FetchProcessorTypesRequest

Request message for fetch processor types.

FetchProcessorTypesResponse

Response message for fetch processor types.

GcsDocument

Specifies a document stored on Cloud Storage.

GcsDocuments

Specifies a set of documents on Cloud Storage.

GcsPrefix

Specifies all documents on Cloud Storage with a common prefix.

HumanReviewStatus

The status of human review on a processed document.

ListProcessorsRequest

Request message for list all processors belongs to a project.

ListProcessorsResponse

Response message for list processors.

NormalizedVertex

A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

ProcessRequest

Request message for the process document method.

ProcessResponse

Response message for the process document method.

Processor

The first-class citizen for DocumentAI. Each processor defines how to extract structural information from a document.

ProcessorType

A processor type is responsible for performing a certain document understanding task on a certain type of document.

RawDocument

Payload message of raw document content (bytes).

ReviewDocumentOperationMetadata

The long running operation metadata for review document method.

ReviewDocumentRequest

Request message for review document method. Next Id: 6.

ReviewDocumentResponse

Response message for review document method.

Vertex

A vertex represents a 2D point in the image. NOTE: the vertex coordinates are in the same scale as the original image.