Module googapis::google::cloud::gkebackup::logging::v1

Modules

logged_backup

Nested message and enum types in LoggedBackup.

logged_backup_plan

Nested message and enum types in LoggedBackupPlan.

logged_restore

Nested message and enum types in LoggedRestore.

restore_config

Nested message and enum types in RestoreConfig.

Structs

BackupChange

use case 2 A log entry when modification(creation, update, deletion) is made to a Backup

BackupPlanChange

use case 1 A log entry when modification(creation, update, deletion) is made to a BackupPlan

EncryptionKey

Encryption key. This only contains the key metadata, and no key material.

LoggedBackup

Backup as stored in Platform log. It’s used to log the details of a createBackup/updateBackup request, so only fields that can be taken from API calls are included here.

LoggedBackupPlan

BackupPlan as stored in Platform log. It’s used to log the details of a createBackupPlan/updateBackupPlan request, so only fields that can be taken from user input are included here.

LoggedRestore

Restore as stored in Platform log. It’s used to log the update details of a updateRestore request, so only mutable and non-output_only fields are included here..

LoggedRestorePlan

RestorePlan as stored in Platform log. It’s used to log the details of a createRestorePlan/updateRestorePlan request, so only fields that can be taken from user input are included here.

NamespacedName

NamespacedName

NamespacedNames

NamespacedNames

Namespaces

Namespaces, list of namespaces

RestoreChange

use case 4 A log entry when modification(creation, update, deletion) is made to a restore.

RestoreConfig

Configuration of a restore.

RestorePlanChange

use case 3 A log entry when modification(creation, update, deletion) is made to a restorePlan.

Enums

ChangeType

The type of changes this log is about.