Module googapis::google::cloud::gkehub::configmanagement::v1beta

Modules

sync_state
Nested message and enum types in SyncState.

Structs

BinauthzConfig
Configuration for Binauthz

BinauthzState
State for Binauthz

BinauthzVersion
The version of binauthz.

ConfigSync
Configuration for Config Sync

ConfigSyncDeploymentState
The state of ConfigSync’s deployment on a cluster

ConfigSyncState
State information for ConfigSync

ConfigSyncVersion
Specific versioning information pertaining to ConfigSync’s Pods

ErrorResource
Model for a config file in the git repo with an associated Sync error

GatekeeperDeploymentState
State of Policy Controller installation.

GitConfig
Git repo configuration for a single cluster.

GroupVersionKind
A Kubernetes object’s GVK

HierarchyControllerConfig
Configuration for Hierarchy Controller

HierarchyControllerDeploymentState
Deployment state for Hierarchy Controller

HierarchyControllerState
State for Hierarchy Controller

HierarchyControllerVersion
Version for Hierarchy Controller

InstallError
Errors pertaining to the installation of ACM

MembershipSpec
Anthos Config Management: Configuration for a single cluster. Intended to parallel the ConfigManagement CR.

MembershipState
Anthos Config Management: State for a single cluster.

OperatorState
State information for an ACM’s Operator

PolicyController
Configuration for Policy Controller

PolicyControllerState
State for PolicyControllerState.

PolicyControllerVersion
The build version of Gatekeeper Policy Controller is using.

SyncError
An ACM created error representing a problem syncing configurations

SyncState
State indicating an ACM’s progress syncing configurations to a cluster

Enums

DeploymentState
Enum representing the state of an ACM’s deployment on a cluster