Module googapis::google::cloud::gkehub::servicemesh::v1alpha

Modules

analysis_message_base

Nested message and enum types in AnalysisMessageBase.

Structs

AnalysisMessage

AnalysisMessage is a single message produced by an analyzer, and it used to communicate to the end user about the state of their Service Mesh configuration.

AnalysisMessageBase

AnalysisMessageBase describes some common information that is needed for all messages.

FeatureState

Service Mesh: State for the whole Hub, as analyzed by the Service Mesh Hub Controller.

MembershipState

Service Mesh: State for a single Membership, as analyzed by the Service Mesh Hub Controller.