Module googapis::google::cloud::gkehub::v1alpha

Modules

common_feature_spec
Nested message and enum types in CommonFeatureSpec.

common_feature_state
Nested message and enum types in CommonFeatureState.

feature_resource_state
Nested message and enum types in FeatureResourceState.

feature_state
Nested message and enum types in FeatureState.

gke_hub_client
Generated client implementations.

membership_feature_spec
Nested message and enum types in MembershipFeatureSpec.

membership_feature_state
Nested message and enum types in MembershipFeatureState.

Structs

CommonFeatureSpec
CommonFeatureSpec contains Hub-wide configuration information

CommonFeatureState
CommonFeatureState contains Hub-wide Feature status information.

CreateFeatureRequest
Request message for the GkeHub.CreateFeature method.

DeleteFeatureRequest
Request message for GkeHub.DeleteFeature method.

Feature
Feature represents the settings and status of any Hub Feature.

FeatureResourceState
FeatureResourceState describes the state of a Feature resource in the GkeHub API. See FeatureState for the “running state” of the Feature in the Hub and across Memberships.

FeatureState
FeatureState describes the high-level state of a Feature. It may be used to describe a Feature’s state at the environ-level, or per-membershop, depending on the context.

GetFeatureRequest
Request message for GkeHub.GetFeature method.

ListFeaturesRequest
Request message for GkeHub.ListFeatures method.

ListFeaturesResponse
Response message for the GkeHub.ListFeatures method.

MembershipFeatureSpec
MembershipFeatureSpec contains configuration information for a single Membership.

MembershipFeatureState
MembershipFeatureState contains Feature status information for a single Membership.

OperationMetadata
Represents the metadata of the long-running operation.

UpdateFeatureRequest
Request message for GkeHub.UpdateFeature method.