Enum googapis::google::cloud::language::v1::dependency_edge::Label

#[repr(i32)]
pub enum Label {

    Unknown,
    Abbrev,
    Acomp,
    Advcl,
    Advmod,
    Amod,
    Appos,
    Attr,
    Aux,
    Auxpass,
    Cc,
    Ccomp,
    Conj,
    Csubj,
    Csubjpass,
    Dep,
    Det,
    Discourse,
    Dobj,
    Expl,
    Goeswith,
    Iobj,
    Mark,
    Mwe,
    Mwv,
    Neg,
    Nn,
    Npadvmod,
    Nsubj,
    Nsubjpass,
    Num,
    Number,
    P,
    Parataxis,
    Partmod,
    Pcomp,
    Pobj,
    Poss,
    Postneg,
    Precomp,
    Preconj,
    Predet,
    Pref,
    Prep,
    Pronl,
    Prt,
    Ps,
    Quantmod,
    Rcmod,
    Rcmodrel,
    Rdrop,
    Ref,
    Remnant,
    Reparandum,
    Root,
    Snum,
    Suff,
    Tmod,
    Topic,
    Vmod,
    Vocative,
    Xcomp,
    Suffix,
    Title,
    Advphmod,
    Auxcaus,
    Auxvv,
    Dtmod,
    Foreign,
    Kw,
    List,
    Nomc,
    Nomcsubj,
    Nomcsubjpass,
    Numc,
    Cop,
    Dislocated,
    Asp,
    Gmod,
    Gobj,
    Infmod,
    Mes,
    Ncomp,

}

The parse label enum for the token.

Variants

Unknown

Unknown

Abbrev

Abbreviation modifier

Acomp

Adjectival complement

Advcl

Adverbial clause modifier

Advmod

Adverbial modifier

Amod

Adjectival modifier of an NP

Appos

Appositional modifier of an NP

Attr

Attribute dependent of a copular verb

Aux

Auxiliary (non-main) verb

Auxpass

Passive auxiliary

Cc

Coordinating conjunction

Ccomp

Clausal complement of a verb or adjective

Conj

Conjunct

Csubj

Clausal subject

Csubjpass

Clausal passive subject

Dep

Dependency (unable to determine)

Det

Determiner

Discourse

Discourse

Dobj

Direct object

Expl

Expletive

Goeswith

Goes with (part of a word in a text not well edited)

Iobj

Indirect object

Mark

Marker (word introducing a subordinate clause)

Mwe

Multi-word expression

Mwv

Multi-word verbal expression

Neg

Negation modifier

Nn

Noun compound modifier

Npadvmod

Noun phrase used as an adverbial modifier

Nsubj

Nominal subject

Nsubjpass

Passive nominal subject

Num

Numeric modifier of a noun

Number

Element of compound number

P

Punctuation mark

Parataxis

Parataxis relation

Partmod

Participial modifier

Pcomp

The complement of a preposition is a clause

Pobj

Object of a preposition

Poss

Possession modifier

Postneg

Postverbal negative particle

Precomp

Predicate complement

Preconj

Preconjunt

Predet

Predeterminer

Pref

Prefix

Prep

Prepositional modifier

Pronl

The relationship between a verb and verbal morpheme

Prt

Particle

Ps

Associative or possessive marker

Quantmod

Quantifier phrase modifier

Rcmod

Relative clause modifier

Rcmodrel

Complementizer in relative clause

Rdrop

Ellipsis without a preceding predicate

Ref

Referent

Remnant

Remnant

Reparandum

Reparandum

Root

Root

Snum

Suffix specifying a unit of number

Suff

Suffix

Tmod

Temporal modifier

Topic

Topic marker

Vmod

Clause headed by an infinite form of the verb that modifies a noun

Vocative

Vocative

Xcomp

Open clausal complement

Suffix

Name suffix

Title

Name title

Advphmod

Adverbial phrase modifier

Auxcaus

Causative auxiliary

Auxvv

Helper auxiliary

Dtmod

Rentaishi (Prenominal modifier)

Foreign

Foreign words

Kw

Keyword

List

List for chains of comparable items

Nomc

Nominalized clause

Nomcsubj

Nominalized clausal subject

Nomcsubjpass

Nominalized clausal passive

Numc

Compound of numeric modifier

Cop

Copula

Dislocated

Dislocated relation (for fronted/topicalized elements)

Asp

Aspect marker

Gmod

Genitive modifier

Gobj

Genitive object

Infmod

Infinitival modifier

Mes

Measure

Ncomp

Nominal complement of a noun

Implementations

impl Label

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Label.

pub fn from_i32(value: i32) -> Option<Label>

Converts an i32 to a Label, or None if value is not a valid variant.