Module googapis::google::cloud::language::v1beta2

Modules

annotate_text_request
Nested message and enum types in AnnotateTextRequest.

dependency_edge
Nested message and enum types in DependencyEdge.

document
Nested message and enum types in Document.

entity
Nested message and enum types in Entity.

entity_mention
Nested message and enum types in EntityMention.

language_service_client
Generated client implementations.

part_of_speech
Nested message and enum types in PartOfSpeech.

Structs

AnalyzeEntitiesRequest
The entity analysis request message.

AnalyzeEntitiesResponse
The entity analysis response message.

AnalyzeEntitySentimentRequest
The entity-level sentiment analysis request message.

AnalyzeEntitySentimentResponse
The entity-level sentiment analysis response message.

AnalyzeSentimentRequest
The sentiment analysis request message.

AnalyzeSentimentResponse
The sentiment analysis response message.

AnalyzeSyntaxRequest
The syntax analysis request message.

AnalyzeSyntaxResponse
The syntax analysis response message.

AnnotateTextRequest
The request message for the text annotation API, which can perform multiple analysis types (sentiment, entities, and syntax) in one call.

AnnotateTextResponse
The text annotations response message.

ClassificationCategory
Represents a category returned from the text classifier.

ClassifyTextRequest
The document classification request message.

ClassifyTextResponse
The document classification response message.

DependencyEdge
Represents dependency parse tree information for a token.

Document
Represents the input to API methods.

Entity
Represents a phrase in the text that is a known entity, such as a person, an organization, or location. The API associates information, such as salience and mentions, with entities.

EntityMention
Represents a mention for an entity in the text. Currently, proper noun mentions are supported.

PartOfSpeech
Represents part of speech information for a token.

Sentence
Represents a sentence in the input document.

Sentiment
Represents the feeling associated with the entire text or entities in the text. Next ID: 6

TextSpan
Represents an output piece of text.

Token
Represents the smallest syntactic building block of the text.

Enums

EncodingType
Represents the text encoding that the caller uses to process the output. Providing an EncodingType is recommended because the API provides the beginning offsets for various outputs, such as tokens and mentions, and languages that natively use different text encodings may access offsets differently.