Module googapis::google::cloud::language::v1beta2::part_of_speech

Nested message and enum types in PartOfSpeech.

Enums

Aspect

The characteristic of a verb that expresses time flow during an event.

Case

The grammatical function performed by a noun or pronoun in a phrase, clause, or sentence. In some languages, other parts of speech, such as adjective and determiner, take case inflection in agreement with the noun.

Form

Depending on the language, Form can be categorizing different forms of verbs, adjectives, adverbs, etc. For example, categorizing inflected endings of verbs and adjectives or distinguishing between short and long forms of adjectives and participles

Gender

Gender classes of nouns reflected in the behaviour of associated words.

Mood

The grammatical feature of verbs, used for showing modality and attitude.

Number

Count distinctions.

Person

The distinction between the speaker, second person, third person, etc.

Proper

This category shows if the token is part of a proper name.

Reciprocity

Reciprocal features of a pronoun.

Tag

The part of speech tags enum.

Tense

Time reference.

Voice

The relationship between the action that a verb expresses and the participants identified by its arguments.