Struct googapis::google::cloud::managedidentities::v1::managed_identities_service_client::ManagedIdentitiesServiceClient

pub struct ManagedIdentitiesServiceClient<T> { /* fields omitted */ }

API Overview

The managedidentites.googleapis.com service implements the Google Cloud Managed Identites API for identity services (e.g. Microsoft Active Directory).

The Managed Identities service provides methods to manage (create/read/update/delete) domains, reset managed identities admin password, add/remove domain controllers in GCP regions and add/remove VPC peering.

Data Model

The Managed Identities service exposes the following resources:

Locations as global, named as follows: projects/{project_id}/locations/global.

Domains, named as follows: /projects/{project_id}/locations/global/domain/{domain_name}.

The {domain_name} refers to fully qualified domain name in the customer project e.g. mydomain.myorganization.com, with the following restrictions:

Must contain only lowercase letters, numbers, periods and hyphens.

Must start with a letter.

Must contain between 2-64 characters.

Must end with a number or a letter.

Must not start with period.

First segement length (mydomain form example above) shouldn’t exceed 15 chars.

The last segment cannot be fully numeric.

Must be unique within the customer project.

Implementations

impl<T> ManagedIdentitiesServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + Sync + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> ManagedIdentitiesServiceClient<InterceptedService<T, F>> where
F: FnMut(Request<()>) -> Result<Request<()>, Status>,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_microsoft_ad_domain(
&mut self,
request: impl IntoRequest<CreateMicrosoftAdDomainRequest>
) -> Result<Response<Operation>, Status>

Creates a Microsoft AD domain.

pub async fn reset_admin_password(
&mut self,
request: impl IntoRequest<ResetAdminPasswordRequest>
) -> Result<Response<ResetAdminPasswordResponse>, Status>

Resets a domain’s administrator password.

pub async fn list_domains(
&mut self,
request: impl IntoRequest<ListDomainsRequest>
) -> Result<Response<ListDomainsResponse>, Status>

Lists domains in a project.

pub async fn get_domain(
&mut self,
request: impl IntoRequest<GetDomainRequest>
) -> Result<Response<Domain>, Status>

Gets information about a domain.

pub async fn update_domain(
&mut self,
request: impl IntoRequest<UpdateDomainRequest>
) -> Result<Response<Operation>, Status>

Updates the metadata and configuration of a domain.

pub async fn delete_domain(
&mut self,
request: impl IntoRequest<DeleteDomainRequest>
) -> Result<Response<Operation>, Status>

Deletes a domain.

pub async fn attach_trust(
&mut self,
request: impl IntoRequest<AttachTrustRequest>
) -> Result<Response<Operation>, Status>

Adds an AD trust to a domain.

pub async fn reconfigure_trust(
&mut self,
request: impl IntoRequest<ReconfigureTrustRequest>
) -> Result<Response<Operation>, Status>

Updates the DNS conditional forwarder.

pub async fn detach_trust(
&mut self,
request: impl IntoRequest<DetachTrustRequest>
) -> Result<Response<Operation>, Status>

Removes an AD trust.

pub async fn validate_trust(
&mut self,
request: impl IntoRequest<ValidateTrustRequest>
) -> Result<Response<Operation>, Status>

Validates a trust state, that the target domain is reachable, and that the target domain is able to accept incoming trust requests.