Module googapis::google::cloud::memcache::v1beta2

Modules

cloud_memcache_client
Generated client implementations.

instance
Nested message and enum types in Instance.

Structs

ApplyParametersRequest
Request for [ApplyParameters][google.cloud.memcache.v1beta2.CloudMemcache.ApplyParameters].

ApplySoftwareUpdateRequest
Request for [ApplySoftwareUpdate][google.cloud.memcache.v1beta2.CloudMemcache.ApplySoftwareUpdate].

CreateInstanceRequest
Request for [CreateInstance][google.cloud.memcache.v1beta2.CloudMemcache.CreateInstance].

DeleteInstanceRequest
Request for [DeleteInstance][google.cloud.memcache.v1beta2.CloudMemcache.DeleteInstance].

GetInstanceRequest
Request for [GetInstance][google.cloud.memcache.v1beta2.CloudMemcache.GetInstance].

Instance
A Memorystore for Memcached instance

ListInstancesRequest
Request for [ListInstances][google.cloud.memcache.v1beta2.CloudMemcache.ListInstances].

ListInstancesResponse
Response for [ListInstances][google.cloud.memcache.v1beta2.CloudMemcache.ListInstances].

LocationMetadata
Metadata for the given [google.cloud.location.Location][google.cloud.location.Location].

MemcacheParameters
The unique ID associated with this set of parameters. Users can use this id to determine if the parameters associated with the instance differ from the parameters associated with the nodes. A discrepancy between parameter ids can inform users that they may need to take action to apply parameters on nodes.

OperationMetadata
Represents the metadata of a long-running operation.

UpdateInstanceRequest
Request for [UpdateInstance][google.cloud.memcache.v1beta2.CloudMemcache.UpdateInstance].

UpdateParametersRequest
Request for [UpdateParameters][google.cloud.memcache.v1beta2.CloudMemcache.UpdateParameters].

ZoneMetadata

Enums

MemcacheVersion
Memcached versions supported by our service.