Module googapis::google::cloud::metastore::v1alpha

Modules

backup
Nested message and enum types in Backup.

database_dump_spec
Nested message and enum types in DatabaseDumpSpec.

dataproc_metastore_client
Generated client implementations.

export_metadata_request
Nested message and enum types in ExportMetadataRequest.

location_metadata
Nested message and enum types in LocationMetadata.

metadata_export
Nested message and enum types in MetadataExport.

metadata_import
Nested message and enum types in MetadataImport.

restore
Nested message and enum types in Restore.

secret
Nested message and enum types in Secret.

service
Nested message and enum types in Service.

Structs

Backup
The details of a backup resource.

CreateBackupRequest
Request message for [DataprocMetastore.CreateBackup][google.cloud.metastore.v1alpha.DataprocMetastore.CreateBackup].

CreateMetadataImportRequest
Request message for [DataprocMetastore.CreateMetadataImport][google.cloud.metastore.v1alpha.DataprocMetastore.CreateMetadataImport].

CreateServiceRequest
Request message for [DataprocMetastore.CreateService][google.cloud.metastore.v1alpha.DataprocMetastore.CreateService].

DataCatalogConfig
Specifies how metastore metadata should be integrated with the Data Catalog service.

DatabaseDumpSpec
The specification of database dump to import from or export to.

DeleteBackupRequest
Request message for [DataprocMetastore.DeleteBackup][google.cloud.metastore.v1alpha.DataprocMetastore.DeleteBackup].

DeleteServiceRequest
Request message for [DataprocMetastore.DeleteService][google.cloud.metastore.v1alpha.DataprocMetastore.DeleteService].

ExportMetadataRequest
Request message for [DataprocMetastore.ExportMetadata][google.cloud.metastore.v1alpha.DataprocMetastore.ExportMetadata].

GetBackupRequest
Request message for [DataprocMetastore.GetBackup][google.cloud.metastore.v1alpha.DataprocMetastore.GetBackup].

GetMetadataImportRequest
Request message for [DataprocMetastore.GetMetadataImport][google.cloud.metastore.v1alpha.DataprocMetastore.GetMetadataImport].

GetServiceRequest
Request message for [DataprocMetastore.GetService][google.cloud.metastore.v1alpha.DataprocMetastore.GetService].

HiveMetastoreConfig
Specifies configuration information specific to running Hive metastore software as the metastore service.

KerberosConfig
Configuration information for a Kerberos principal.

ListBackupsRequest
Request message for [DataprocMetastore.ListBackups][google.cloud.metastore.v1alpha.DataprocMetastore.ListBackups].

ListBackupsResponse
Response message for [DataprocMetastore.ListBackups][google.cloud.metastore.v1alpha.DataprocMetastore.ListBackups].

ListMetadataImportsRequest
Request message for [DataprocMetastore.ListMetadataImports][google.cloud.metastore.v1alpha.DataprocMetastore.ListMetadataImports].

ListMetadataImportsResponse
Response message for [DataprocMetastore.ListMetadataImports][google.cloud.metastore.v1alpha.DataprocMetastore.ListMetadataImports].

ListServicesRequest
Request message for [DataprocMetastore.ListServices][google.cloud.metastore.v1alpha.DataprocMetastore.ListServices].

ListServicesResponse
Response message for [DataprocMetastore.ListServices][google.cloud.metastore.v1alpha.DataprocMetastore.ListServices].

LocationMetadata
Metadata about the service in a location.

MaintenanceWindow
Maintenance window. This specifies when Dataproc Metastore may perform system maintenance operation to the service.

MetadataExport
The details of a metadata export operation.

MetadataImport
A metastore resource that imports metadata.

MetadataIntegration
Specifies how metastore metadata should be integrated with external services.

MetadataManagementActivity
The metadata management activities of the metastore service.

OperationMetadata
Represents the metadata of a long-running operation.

Restore
The details of a metadata restore operation.

RestoreServiceRequest
Request message for [DataprocMetastore.Restore][].

Secret
A securely stored value.

Service
A managed metastore service that serves metadata queries.

UpdateMetadataImportRequest
Request message for [DataprocMetastore.UpdateMetadataImport][google.cloud.metastore.v1alpha.DataprocMetastore.UpdateMetadataImport].

UpdateServiceRequest
Request message for [DataprocMetastore.UpdateService][google.cloud.metastore.v1alpha.DataprocMetastore.UpdateService].