Module googapis::google::cloud::networkconnectivity::v1

Modules

hub_service_client

Generated client implementations.

Structs

CreateHubRequest

Request for [HubService.CreateHub][google.cloud.networkconnectivity.v1.HubService.CreateHub] method.

CreateSpokeRequest

The request for [HubService.CreateSpoke][google.cloud.networkconnectivity.v1.HubService.CreateSpoke].

DeleteHubRequest

The request for [HubService.DeleteHub][google.cloud.networkconnectivity.v1.HubService.DeleteHub].

DeleteSpokeRequest

The request for [HubService.DeleteSpoke][google.cloud.networkconnectivity.v1.HubService.DeleteSpoke].

GetHubRequest

Request for [HubService.GetHub][google.cloud.networkconnectivity.v1.HubService.GetHub] method.

GetSpokeRequest

The request for [HubService.GetSpoke][google.cloud.networkconnectivity.v1.HubService.GetSpoke].

Hub

A hub is a collection of spokes. A single hub can contain spokes from multiple regions. However, all of a hub’s spokes must be associated with resources that reside in the same VPC network.

LinkedInterconnectAttachments

A collection of VLAN attachment resources. These resources should be redundant attachments that all advertise the same prefixes to Google Cloud. Alternatively, in active/passive configurations, all attachments should be capable of advertising the same prefixes.

LinkedRouterApplianceInstances

A collection of router appliance instances. If you have multiple router appliance instances connected to the same site, they should all be attached to the same spoke.

LinkedVpnTunnels

A collection of Cloud VPN tunnel resources. These resources should be redundant HA VPN tunnels that all advertise the same prefixes to Google Cloud. Alternatively, in a passive/active configuration, all tunnels should be capable of advertising the same prefixes.

ListHubsRequest

Request for [HubService.ListHubs][google.cloud.networkconnectivity.v1.HubService.ListHubs] method.

ListHubsResponse

Response for [HubService.ListHubs][google.cloud.networkconnectivity.v1.HubService.ListHubs] method.

ListSpokesRequest

The request for [HubService.ListSpokes][google.cloud.networkconnectivity.v1.HubService.ListSpokes].

ListSpokesResponse

The response for [HubService.ListSpokes][google.cloud.networkconnectivity.v1.HubService.ListSpokes].

OperationMetadata

Represents the metadata of the long-running operation.

RouterApplianceInstance

A router appliance instance is a Compute Engine virtual machine (VM) instance that acts as a BGP speaker. A router appliance instance is specified by the URI of the VM and the internal IP address of one of the VM’s network interfaces.

RoutingVpc

RoutingVPC contains information about the VPC network that is associated with a hub’s spokes.

Spoke

A spoke represents a connection between your Google Cloud network resources and a non-Google-Cloud network.

UpdateHubRequest

Request for [HubService.UpdateHub][google.cloud.networkconnectivity.v1.HubService.UpdateHub] method.

UpdateSpokeRequest

Request for [HubService.UpdateSpoke][google.cloud.networkconnectivity.v1.HubService.UpdateSpoke] method.

Enums

State

The State enum represents the lifecycle stage of a Network Connectivity Center resource.