Enum googapis::google::cloud::networkmanagement::v1beta1::step::StepInfo

pub enum StepInfo {

    Instance(InstanceInfo),
    Firewall(FirewallInfo),
    Route(RouteInfo),
    Endpoint(EndpointInfo),
    ForwardingRule(ForwardingRuleInfo),
    VpnGateway(VpnGatewayInfo),
    VpnTunnel(VpnTunnelInfo),
    Deliver(DeliverInfo),
    Forward(ForwardInfo),
    Abort(AbortInfo),
    Drop(DropInfo),
    LoadBalancer(LoadBalancerInfo),
    Network(NetworkInfo),
    GkeMaster(GkeMasterInfo),
    CloudSqlInstance(CloudSqlInstanceInfo),

}

Configuration or metadata associated with each step. The configuration is filtered based on viewer’s permission. If a viewer has no permission to view the configuration in this step, for non-final states a special state is populated (VIEWER_PERMISSION_MISSING), and for final state the configuration is cleared.

Variants

Instance(InstanceInfo)

Display information of a Compute Engine instance.

Tuple Fields of Instance

0: InstanceInfo

Firewall(FirewallInfo)

Display information of a Compute Engine firewall rule.

Tuple Fields of Firewall

0: FirewallInfo

Route(RouteInfo)

Display information of a Compute Engine route.

Tuple Fields of Route

0: RouteInfo

Endpoint(EndpointInfo)

Display information of the source and destination under analysis. The endpoint information in an intermediate state may differ with the initial input, as it might be modified by state like NAT, or Connection Proxy.

Tuple Fields of Endpoint

0: EndpointInfo

ForwardingRule(ForwardingRuleInfo)

Display information of a Compute Engine forwarding rule.

Tuple Fields of ForwardingRule

0: ForwardingRuleInfo

VpnGateway(VpnGatewayInfo)

Display information of a Compute Engine VPN gateway.

Tuple Fields of VpnGateway

0: VpnGatewayInfo

VpnTunnel(VpnTunnelInfo)

Display information of a Compute Engine VPN tunnel.

Tuple Fields of VpnTunnel

0: VpnTunnelInfo

Deliver(DeliverInfo)

Display information of the final state “deliver” and reason.

Tuple Fields of Deliver

0: DeliverInfo

Forward(ForwardInfo)

Display information of the final state “forward” and reason.

Tuple Fields of Forward

0: ForwardInfo

Abort(AbortInfo)

Display information of the final state “abort” and reason.

Tuple Fields of Abort

0: AbortInfo

Drop(DropInfo)

Display information of the final state “drop” and reason.

Tuple Fields of Drop

0: DropInfo

LoadBalancer(LoadBalancerInfo)

Display information of the load balancers.

Tuple Fields of LoadBalancer

0: LoadBalancerInfo

Network(NetworkInfo)

Display information of a Google Cloud network.

Tuple Fields of Network

0: NetworkInfo

GkeMaster(GkeMasterInfo)

Display information of a Google Kubernetes Engine cluster master.

Tuple Fields of GkeMaster

0: GkeMasterInfo

CloudSqlInstance(CloudSqlInstanceInfo)

Display information of a Cloud SQL instance.

Tuple Fields of CloudSqlInstance

0: CloudSqlInstanceInfo

Implementations

impl StepInfo

pub fn encode<B>(&self, buf: &mut B) where
B: BufMut,

pub fn merge<B>(
field: &mut Option<StepInfo>,
tag: u32,
wire_type: WireType,
buf: &mut B,
ctx: DecodeContext
) -> Result<(), DecodeError> where
B: Buf,

pub fn encoded_len(&self) -> usize