Module googapis::google::cloud::networkservices::v1beta1

Modules

endpoint_matcher
Nested message and enum types in EndpointMatcher.

endpoint_policy
Nested message and enum types in EndpointPolicy.

network_services_client
Generated client implementations.

Structs

CreateEndpointPolicyRequest
Request used with the CreateEndpointPolicy method.

DeleteEndpointPolicyRequest
Request used with the DeleteEndpointPolicy method.

EndpointMatcher
A definition of a matcher that selects endpoints to which the policies should be applied.

EndpointPolicy
EndpointPolicy is a resource that helps apply desired configuration on the endpoints that match specific criteria. For example, this resource can be used to apply “authentication config” an all endpoints that serve on port 8080.

GetEndpointPolicyRequest
Request used with the GetEndpointPolicy method.

ListEndpointPoliciesRequest
Request used with the ListEndpointPolicies method.

ListEndpointPoliciesResponse
Response returned by the ListEndpointPolicies method.

OperationMetadata
Represents the metadata of the long-running operation.

TrafficPortSelector
Specification of a port-based selector.

UpdateEndpointPolicyRequest
Request used with the UpdateEndpointPolicy method.