Module googapis::google::cloud::notebooks::v1

Modules

environment

Nested message and enum types in Environment.

event

Nested message and enum types in Event.

execution

Nested message and enum types in Execution.

execution_template

Nested message and enum types in ExecutionTemplate.

get_instance_health_response

Nested message and enum types in GetInstanceHealthResponse.

instance

Nested message and enum types in Instance.

local_disk

Nested message and enum types in LocalDisk.

local_disk_initialize_params

Nested message and enum types in LocalDiskInitializeParams.

managed_notebook_service_client

Generated client implementations.

notebook_service_client

Generated client implementations.

reservation_affinity

Nested message and enum types in ReservationAffinity.

runtime

Nested message and enum types in Runtime.

runtime_accelerator_config

Nested message and enum types in RuntimeAcceleratorConfig.

runtime_access_config

Nested message and enum types in RuntimeAccessConfig.

schedule

Nested message and enum types in Schedule.

virtual_machine_config

Nested message and enum types in VirtualMachineConfig.

vm_image

Nested message and enum types in VmImage.

Structs

ContainerImage

Definition of a container image for starting a notebook instance with the environment installed in a container.

CreateEnvironmentRequest

Request for creating a notebook environment.

CreateExecutionRequest

Request to create notebook execution

CreateInstanceRequest

Request for creating a notebook instance.

CreateRuntimeRequest

Request for creating a Managed Notebook Runtime.

CreateScheduleRequest

Request for created scheduled notebooks

DeleteEnvironmentRequest

Request for deleting a notebook environment.

DeleteExecutionRequest

Request for deleting a scheduled notebook execution

DeleteInstanceRequest

Request for deleting a notebook instance.

DeleteRuntimeRequest

Request for deleting a Managed Notebook Runtime.

DeleteScheduleRequest

Request for deleting an Schedule

EncryptionConfig

Represents a custom encryption key configuration that can be applied to a resource. This will encrypt all disks in Virtual Machine.

Environment

Definition of a software environment that is used to start a notebook instance.

Event

The definition of an Event for a managed / semi-managed notebook instance.

Execution

The definition of a single executed notebook.

ExecutionTemplate

The description a notebook execution workload.

GetEnvironmentRequest

Request for getting a notebook environment.

GetExecutionRequest

Request for getting scheduled notebook execution

GetInstanceHealthRequest

Request for checking if a notebook instance is healthy.

GetInstanceHealthResponse

Response for checking if a notebook instance is healthy.

GetInstanceRequest

Request for getting a notebook instance.

GetRuntimeRequest

Request for getting a Managed Notebook Runtime.

GetScheduleRequest

Request for getting scheduled notebook.

Instance

The definition of a notebook instance.

InstanceConfig

Notebook instance configurations that can be updated.

IsInstanceUpgradeableRequest

Request for checking if a notebook instance is upgradeable.

IsInstanceUpgradeableResponse

Response for checking if a notebook instance is upgradeable.

ListEnvironmentsRequest

Request for listing environments.

ListEnvironmentsResponse

Response for listing environments.

ListExecutionsRequest

Request for listing scheduled notebook executions.

ListExecutionsResponse

Response for listing scheduled notebook executions

ListInstancesRequest

Request for listing notebook instances.

ListInstancesResponse

Response for listing notebook instances.

ListRuntimesRequest

Request for listing Managed Notebook Runtimes.

ListRuntimesResponse

Response for listing Managed Notebook Runtimes.

ListSchedulesRequest

Request for listing scheduled notebook job.

ListSchedulesResponse

Response for listing scheduled notebook job.

LocalDisk

A Local attached disk resource.

LocalDiskInitializeParams

Input only. Specifies the parameters for a new disk that will be created alongside the new instance. Use initialization parameters to create boot disks or local SSDs attached to the new runtime. This property is mutually exclusive with the source property; you can only define one or the other, but not both.

OperationMetadata

Represents the metadata of the long-running operation.

RegisterInstanceRequest

Request for registering a notebook instance.

ReportInstanceInfoRequest

Request for notebook instances to report information to Notebooks API.

ReportRuntimeEventRequest

Request for reporting a Managed Notebook Event.

ReservationAffinity

Reservation Affinity for consuming Zonal reservation.

ResetInstanceRequest

Request for reseting a notebook instance

ResetRuntimeRequest

Request for reseting a Managed Notebook Runtime.

RollbackInstanceRequest

Request for rollbacking a notebook instance

Runtime

The definition of a Runtime for a managed notebook instance.

RuntimeAcceleratorConfig

Definition of the types of hardware accelerators that can be used. Definition of the types of hardware accelerators that can be used. See Compute Engine AcceleratorTypes. Examples:

RuntimeAccessConfig

Specifies the login configuration for Runtime

RuntimeMetrics

Contains runtime daemon metrics, such as OS and kernels and sessions stats.

RuntimeShieldedInstanceConfig

A set of Shielded Instance options. Check Images using supported Shielded VM features. Not all combinations are valid.

RuntimeSoftwareConfig

Specifies the selection and configuration of software inside the runtime. The properties to set on runtime. Properties keys are specified in key:value format, for example:

Schedule

The definition of a schedule.

SetInstanceAcceleratorRequest

Request for setting instance accelerator.

SetInstanceLabelsRequest

Request for setting instance labels.

SetInstanceMachineTypeRequest

Request for setting instance machine type.

StartInstanceRequest

Request for starting a notebook instance

StartRuntimeRequest

Request for starting a Managed Notebook Runtime.

StopInstanceRequest

Request for stopping a notebook instance

StopRuntimeRequest

Request for stopping a Managed Notebook Runtime.

SwitchRuntimeRequest

Request for switching a Managed Notebook Runtime.

TriggerScheduleRequest

Request for created scheduled notebooks

UpdateInstanceConfigRequest

Request for updating instance configurations.

UpdateShieldedInstanceConfigRequest

Request for updating the Shielded Instance config for a notebook instance. You can only use this method on a stopped instance

UpgradeInstanceInternalRequest

Request for upgrading a notebook instance from within the VM

UpgradeInstanceRequest

Request for upgrading a notebook instance

VirtualMachine

Runtime using Virtual Machine for computing.

VirtualMachineConfig

The config settings for virtual machine.

VmImage

Definition of a custom Compute Engine virtual machine image for starting a notebook instance with the environment installed directly on the VM.