Module googapis::google::cloud::notebooks::v1::execution_template

Nested message and enum types in ExecutionTemplate.

Structs

DataprocParameters

Parameters used in Dataproc JobType executions.

SchedulerAcceleratorConfig

Definition of a hardware accelerator. Note that not all combinations of type and core_count are valid. Check GPUs on Compute Engine to find a valid combination. TPUs are not supported.

Enums

JobParameters

Parameters for an execution type. NOTE: There are currently no extra parameters for VertexAI jobs.

JobType

The backend used for this execution.

ScaleTier

Required. Specifies the machine types, the number of replicas for workers and parameter servers.

SchedulerAcceleratorType

Hardware accelerator types for AI Platform Training jobs.