Module googapis::google::cloud::orchestration::airflow::service::v1

Modules

check_upgrade_response

Nested message and enum types in CheckUpgradeResponse.

environment

Nested message and enum types in Environment.

environments_client

Generated client implementations.

image_versions_client

Generated client implementations.

ip_allocation_policy

Nested message and enum types in IPAllocationPolicy.

operation_metadata

Nested message and enum types in OperationMetadata.

web_server_network_access_control

Nested message and enum types in WebServerNetworkAccessControl.

Structs

CheckUpgradeResponse

Message containing information about the result of an upgrade check operation.

CreateEnvironmentRequest

Create a new environment.

DatabaseConfig

The configuration of Cloud SQL instance that is used by the Apache Airflow software.

DeleteEnvironmentRequest

Delete an environment.

EncryptionConfig

The encryption options for the Cloud Composer environment and its dependencies.

Environment

An environment for running orchestration tasks.

EnvironmentConfig

Configuration information for an environment.

GetEnvironmentRequest

Get an environment.

ImageVersion

ImageVersion information

IpAllocationPolicy

Configuration for controlling how IPs are allocated in the GKE cluster running the Apache Airflow software.

ListEnvironmentsRequest

List environments in a project and location.

ListEnvironmentsResponse

The environments in a project and location.

ListImageVersionsRequest

List ImageVersions in a project and location.

ListImageVersionsResponse

The ImageVersions in a project and location.

NodeConfig

The configuration information for the Kubernetes Engine nodes running the Apache Airflow software.

OperationMetadata

Metadata describing an operation.

PrivateClusterConfig

Configuration options for the private GKE cluster in a Cloud Composer environment.

PrivateEnvironmentConfig

The configuration information for configuring a Private IP Cloud Composer environment.

SoftwareConfig

Specifies the selection and configuration of software inside the environment.

UpdateEnvironmentRequest

Update an environment.

WebServerConfig

The configuration settings for the Airflow web server App Engine instance.

WebServerNetworkAccessControl

Network-level access control policy for the Airflow web server.