Module googapis::google::cloud::orchestration::airflow::service::v1beta1

Modules

check_upgrade_response
Nested message and enum types in CheckUpgradeResponse.

environment
Nested message and enum types in Environment.

environment_config
Nested message and enum types in EnvironmentConfig.

environments_client
Generated client implementations.

image_versions_client
Generated client implementations.

operation_metadata
Nested message and enum types in OperationMetadata.

web_server_network_access_control
Nested message and enum types in WebServerNetworkAccessControl.

workloads_config
Nested message and enum types in WorkloadsConfig.

Structs

CheckUpgradeRequest
Request to check whether image upgrade will succeed.

CheckUpgradeResponse
Message containing information about the result of an upgrade check operation.

CreateEnvironmentRequest
Create a new environment.

DatabaseConfig
The configuration of Cloud SQL instance that is used by the Apache Airflow software.

DeleteEnvironmentRequest
Delete an environment.

EncryptionConfig
The encryption options for the Cloud Composer environment and its dependencies.

Environment
An environment for running orchestration tasks.

EnvironmentConfig
Configuration information for an environment.

GetEnvironmentRequest
Get an environment.

ImageVersion
Image Version information

IpAllocationPolicy
Configuration for controlling how IPs are allocated in the GKE cluster.

ListEnvironmentsRequest
List environments in a project and location.

ListEnvironmentsResponse
The environments in a project and location.

ListImageVersionsRequest
List ImageVersions in a project and location.

ListImageVersionsResponse
The ImageVersions in a project and location.

MaintenanceWindow
The configuration settings for Cloud Composer maintenance window.

NodeConfig
The configuration information for the Kubernetes Engine nodes running the Apache Airflow software.

OperationMetadata
Metadata describing an operation.

PrivateClusterConfig
Configuration options for the private GKE cluster in a Cloud Composer environment.

PrivateEnvironmentConfig
The configuration information for configuring a Private IP Cloud Composer environment.

RestartWebServerRequest
Restart Airflow web server.

SoftwareConfig
Specifies the selection and configuration of software inside the environment.

UpdateEnvironmentRequest
Update an environment.

WebServerConfig
The configuration settings for the Airflow web server App Engine instance.

WebServerNetworkAccessControl
Network-level access control policy for the Airflow web server.

WorkloadsConfig
The Kubernetes workloads configuration for GKE cluster associated with the Cloud Composer environment. Supported for Cloud Composer environments in versions composer-2..-airflow-..* and newer.