Module googapis::google::cloud::osconfig::agentendpoint::v1::inventory

Nested message and enum types in Inventory.

Modules

software_package

Nested message and enum types in SoftwarePackage.

windows_update_package

Nested message and enum types in WindowsUpdatePackage.

Structs

OsInfo

Operating system information for the VM.

SoftwarePackage

Software package information of the operating system.

VersionedPackage

Information related to the a standard versioned package. This includes package info for APT, Yum, Zypper, and Googet package managers.

WindowsApplication

Details about Windows Application - based on Windows Registry. All fields in this message are taken from: https://docs.microsoft.com/en-us/windows/win32/msi/uninstall-registry-key

WindowsQuickFixEngineeringPackage

Information related to a Quick Fix Engineering package. Fields are taken from Windows QuickFixEngineering Interface and match the source names: https://docs.microsoft.com/en-us/windows/win32/cimwin32prov/win32-quickfixengineering

WindowsUpdatePackage

Details related to a Windows Update package. Field data and names are taken from Windows Update API IUpdate Interface: https://docs.microsoft.com/en-us/windows/win32/api/_wua/ Descriptive fields like title, and description are localized based on the locale of the VM being updated.

ZypperPatch

Details related to a Zypper Patch.