Module googapis::google::cloud::osconfig::agentendpoint::v1beta

Modules

agent_endpoint_service_client

Generated client implementations.

apply_patches_task_output

Nested message and enum types in ApplyPatchesTaskOutput.

apply_patches_task_progress

Nested message and enum types in ApplyPatchesTaskProgress.

apt_repository

Nested message and enum types in AptRepository.

apt_settings

Nested message and enum types in AptSettings.

effective_guest_policy

Nested message and enum types in EffectiveGuestPolicy.

exec_step_config

Nested message and enum types in ExecStepConfig.

exec_step_task_output

Nested message and enum types in ExecStepTaskOutput.

exec_step_task_progress

Nested message and enum types in ExecStepTaskProgress.

package

Nested message and enum types in Package.

package_repository

Nested message and enum types in PackageRepository.

patch_config

Nested message and enum types in PatchConfig.

report_task_complete_request

Nested message and enum types in ReportTaskCompleteRequest.

report_task_progress_request

Nested message and enum types in ReportTaskProgressRequest.

software_recipe

Nested message and enum types in SoftwareRecipe.

task

Nested message and enum types in Task.

windows_update_settings

Nested message and enum types in WindowsUpdateSettings.

Structs

ApplyPatchesTask

Message which instructs agent to apply patches.

ApplyPatchesTaskOutput

Information reported from the agent about applying patches execution.

ApplyPatchesTaskProgress

Information reported from the agent about applying patches execution.

AptRepository

Represents a single Apt package repository. This repository is added to a repo file that is stored at /etc/apt/sources.list.d/google_osconfig.list.

AptSettings

Apt patching will be performed by executing apt-get update && apt-get upgrade. Additional options can be set to control how this is executed.

EffectiveGuestPolicy

The effective guest policy assigned to the instance.

ExecStep

A step that runs an executable for a PatchJob.

ExecStepConfig

Common configurations for an ExecStep.

ExecStepTask

Message which instructs agent to execute the following command.

ExecStepTaskOutput

Information reported from the agent about the exec step execution.

ExecStepTaskProgress

Information reported from the agent about the exec step execution.

GcsObject

GCS object representation.

GooRepository

Represents a Goo package repository. These is added to a repo file that is stored at C:/ProgramData/GooGet/repos/google_osconfig.repo.

GooSettings

Googet patching is performed by running googet update.

LookupEffectiveGuestPolicyRequest

A request message for getting effective policy assigned to the instance.

Package

Package is a reference to the software package to be installed or removed. The agent on the VM instance uses the system package manager to apply the config.

PackageRepository

A package repository.

PatchConfig

Patch configuration specifications. Contains details on how to apply patches to a VM instance.

ReceiveTaskNotificationRequest

A request message to receive task notifications.

ReceiveTaskNotificationResponse

The streaming rpc message that notifies the agent when it has a task that it needs to perform on the VM instance.

RegisterAgentRequest

The request message for registering the agent.

RegisterAgentResponse

The response message after the agent registered.

ReportTaskCompleteRequest

A request message for signaling the completion of a task execution.

ReportTaskCompleteResponse

The response message after the agent signaled the current task complete.

ReportTaskProgressRequest

A request message for reporting the progress of current task.

ReportTaskProgressResponse

The response message after the agent reported the current task progress.

RetryStrategy

The strategy for retrying failed patches during the patch window.

SoftwareRecipe

A software recipe is a set of instructions for installing and configuring a piece of software. It consists of a set of artifacts that are downloaded, and a set of steps that install, configure, and/or update the software.

StartNextTaskRequest

A request message for signaling the start of a task execution.

StartNextTaskResponse

A response message that contains the details of the task to work on.

Task

A unit of work to be performed by the agent.

WindowsUpdateSettings

Windows patching is performed using the Windows Update Agent.

YumRepository

Represents a single Yum package repository. This repository is added to a repo file that is stored at /etc/yum.repos.d/google_osconfig.repo.

YumSettings

Yum patching will be performed by executing yum update. Additional options can be set to control how this is executed.

ZypperRepository

Represents a single Zypper package repository. This repository is added to a repo file that is stored at /etc/zypp/repos.d/google_osconfig.repo.

ZypperSettings

Zypper patching is performed by running zypper patch. See also https://en.opensuse.org/SDB:Zypper_manual.

Enums

DesiredState

The desired state that the OS Config agent will maintain on the VM.

TaskDirective

Specifies the current agent behavior.

TaskType

Specifies the type of task to perform.