Module googapis::google::cloud::osconfig::v1

Modules

apt_settings
Nested message and enum types in AptSettings.

exec_step_config
Nested message and enum types in ExecStepConfig.

fixed_or_percent
Nested message and enum types in FixedOrPercent.

instance
Nested message and enum types in Instance.

inventory
Nested message and enum types in Inventory.

monthly_schedule
Nested message and enum types in MonthlySchedule.

os_config_service_client
Generated client implementations.

patch_config
Nested message and enum types in PatchConfig.

patch_deployment
Nested message and enum types in PatchDeployment.

patch_instance_filter
Nested message and enum types in PatchInstanceFilter.

patch_job
Nested message and enum types in PatchJob.

patch_rollout
Nested message and enum types in PatchRollout.

recurring_schedule
Nested message and enum types in RecurringSchedule.

windows_update_settings
Nested message and enum types in WindowsUpdateSettings.

Structs

AptSettings
Apt patching is completed by executing apt-get update && apt-get upgrade. Additional options can be set to control how this is executed.

CancelPatchJobRequest
Message for canceling a patch job.

CreatePatchDeploymentRequest
A request message for creating a patch deployment.

DeletePatchDeploymentRequest
A request message for deleting a patch deployment.

ExecStep
A step that runs an executable for a PatchJob.

ExecStepConfig
Common configurations for an ExecStep.

ExecutePatchJobRequest
A request message to initiate patching across Compute Engine instances.

FixedOrPercent
Message encapsulating a value that can be either absolute (“fixed”) or relative (“percent”) to a value.

GcsObject
Cloud Storage object representation.

GetPatchDeploymentRequest
A request message for retrieving a patch deployment.

GetPatchJobRequest
Request to get an active or completed patch job.

GooSettings
Googet patching is performed by running googet update.

Instance
Namespace for instance state enums.

Inventory
The inventory details of a VM.

ListPatchDeploymentsRequest
A request message for listing patch deployments.

ListPatchDeploymentsResponse
A response message for listing patch deployments.

ListPatchJobInstanceDetailsRequest
Request to list details for all instances that are part of a patch job.

ListPatchJobInstanceDetailsResponse
A response message for listing the instances details for a patch job.

ListPatchJobsRequest
A request message for listing patch jobs.

ListPatchJobsResponse
A response message for listing patch jobs.

MonthlySchedule
Represents a monthly schedule. An example of a valid monthly schedule is “on the third Tuesday of the month” or “on the 15th of the month”.

OneTimeSchedule
Sets the time for a one time patch deployment. Timestamp is in RFC3339 text format.

PatchConfig
Patch configuration specifications. Contains details on how to apply the patch(es) to a VM instance.

PatchDeployment
Patch deployments are configurations that individual patch jobs use to complete a patch. These configurations include instance filter, package repository settings, and a schedule. For more information about creating and managing patch deployments, see Scheduling patch jobs.

PatchInstanceFilter
A filter to target VM instances for patching. The targeted VMs must meet all criteria specified. So if both labels and zones are specified, the patch job targets only VMs with those labels and in those zones.

PatchJob
A high level representation of a patch job that is either in progress or has completed.

PatchJobInstanceDetails
Patch details for a VM instance. For more information about reviewing VM instance details, see Listing all VM instance details for a specific patch job.

PatchRollout
Patch rollout configuration specifications. Contains details on the concurrency control when applying patch(es) to all targeted VMs.

RecurringSchedule
Sets the time for recurring patch deployments.

WeekDayOfMonth
Represents one week day in a month. An example is “the 4th Sunday”.

WeeklySchedule
Represents a weekly schedule.

WindowsUpdateSettings
Windows patching is performed using the Windows Update Agent.

YumSettings
Yum patching is performed by executing yum update. Additional options can be set to control how this is executed.

ZypperSettings
Zypper patching is performed by running zypper patch. See also https://en.opensuse.org/SDB:Zypper_manual.