Struct googapis::google::cloud::osconfig::v1alpha::os_config_zonal_service_client::OsConfigZonalServiceClient

pub struct OsConfigZonalServiceClient<T> { /* fields omitted */ }

Zonal OS Config API

The OS Config service is the server-side component that allows users to manage package installations and patch jobs for Compute Engine VM instances.

Implementations

impl<T> OsConfigZonalServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> OsConfigZonalServiceClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_os_policy_assignment(
&mut self,
request: impl IntoRequest<CreateOsPolicyAssignmentRequest>
) -> Result<Response<Operation>, Status>

Create an OS policy assignment.

This method also creates the first revision of the OS policy assignment.

This method returns a long running operation (LRO) that contains the rollout details. The rollout can be cancelled by cancelling the LRO.

For more information, see Method: projects.locations.osPolicyAssignments.operations.cancel.

pub async fn update_os_policy_assignment(
&mut self,
request: impl IntoRequest<UpdateOsPolicyAssignmentRequest>
) -> Result<Response<Operation>, Status>

Update an existing OS policy assignment.

This method creates a new revision of the OS policy assignment.

This method returns a long running operation (LRO) that contains the rollout details. The rollout can be cancelled by cancelling the LRO.

For more information, see Method: projects.locations.osPolicyAssignments.operations.cancel.

pub async fn get_os_policy_assignment(
&mut self,
request: impl IntoRequest<GetOsPolicyAssignmentRequest>
) -> Result<Response<OsPolicyAssignment>, Status>

Retrieve an existing OS policy assignment.

This method always returns the latest revision. In order to retrieve a previous revision of the assignment, also provide the revision ID in the name parameter.

pub async fn list_os_policy_assignments(
&mut self,
request: impl IntoRequest<ListOsPolicyAssignmentsRequest>
) -> Result<Response<ListOsPolicyAssignmentsResponse>, Status>

List the OS policy assignments under the parent resource.

For each OS policy assignment, the latest revision is returned.

pub async fn list_os_policy_assignment_revisions(
&mut self,
request: impl IntoRequest<ListOsPolicyAssignmentRevisionsRequest>
) -> Result<Response<ListOsPolicyAssignmentRevisionsResponse>, Status>

List the OS policy assignment revisions for a given OS policy assignment.

pub async fn delete_os_policy_assignment(
&mut self,
request: impl IntoRequest<DeleteOsPolicyAssignmentRequest>
) -> Result<Response<Operation>, Status>

Delete the OS policy assignment.

This method creates a new revision of the OS policy assignment.

This method returns a long running operation (LRO) that contains the rollout details. The rollout can be cancelled by cancelling the LRO.

If the LRO completes and is not cancelled, all revisions associated with the OS policy assignment are deleted.

For more information, see Method: projects.locations.osPolicyAssignments.operations.cancel.

pub async fn get_instance_os_policies_compliance(
&mut self,
request: impl IntoRequest<GetInstanceOsPoliciesComplianceRequest>
) -> Result<Response<InstanceOsPoliciesCompliance>, Status>

Get OS policies compliance data for the specified Compute Engine VM instance.

pub async fn list_instance_os_policies_compliances(
&mut self,
request: impl IntoRequest<ListInstanceOsPoliciesCompliancesRequest>
) -> Result<Response<ListInstanceOsPoliciesCompliancesResponse>, Status>

List OS policies compliance data for all Compute Engine VM instances in the specified zone.

pub async fn get_inventory(
&mut self,
request: impl IntoRequest<GetInventoryRequest>
) -> Result<Response<Inventory>, Status>

Get inventory data for the specified VM instance. If the VM has no associated inventory, the message NOT_FOUND is returned.

pub async fn list_inventories(
&mut self,
request: impl IntoRequest<ListInventoriesRequest>
) -> Result<Response<ListInventoriesResponse>, Status>

List inventory data for all VM instances in the specified zone.

pub async fn get_vulnerability_report(
&mut self,
request: impl IntoRequest<GetVulnerabilityReportRequest>
) -> Result<Response<VulnerabilityReport>, Status>

Gets the vulnerability report for the specified VM instance. Only VMs with inventory data have vulnerability reports associated with them.

pub async fn list_vulnerability_reports(
&mut self,
request: impl IntoRequest<ListVulnerabilityReportsRequest>
) -> Result<Response<ListVulnerabilityReportsResponse>, Status>

List vulnerability reports for all VM instances in the specified zone.