Module googapis::google::cloud::osconfig::v1beta

Modules

apt_repository

Nested message and enum types in AptRepository.

apt_settings

Nested message and enum types in AptSettings.

assignment

Nested message and enum types in Assignment.

effective_guest_policy

Nested message and enum types in EffectiveGuestPolicy.

exec_step_config

Nested message and enum types in ExecStepConfig.

instance

Nested message and enum types in Instance.

monthly_schedule

Nested message and enum types in MonthlySchedule.

os_config_service_client

Generated client implementations.

package

Nested message and enum types in Package.

package_repository

Nested message and enum types in PackageRepository.

patch_config

Nested message and enum types in PatchConfig.

patch_deployment

Nested message and enum types in PatchDeployment.

patch_instance_filter

Nested message and enum types in PatchInstanceFilter.

patch_job

Nested message and enum types in PatchJob.

recurring_schedule

Nested message and enum types in RecurringSchedule.

software_recipe

Nested message and enum types in SoftwareRecipe.

windows_update_settings

Nested message and enum types in WindowsUpdateSettings.

Structs

AptRepository

Represents a single Apt package repository. This repository is added to a repo file that is stored at /etc/apt/sources.list.d/google_osconfig.list.

AptSettings

Apt patching is completed by executing apt-get update && apt-get upgrade. Additional options can be set to control how this is executed.

Assignment

An assignment represents the group or groups of VM instances that the policy applies to.

CancelPatchJobRequest

Message for canceling a patch job.

CreateGuestPolicyRequest

A request message for creating a guest policy.

CreatePatchDeploymentRequest

A request message for creating a patch deployment.

DeleteGuestPolicyRequest

A request message for deleting a guest policy.

DeletePatchDeploymentRequest

A request message for deleting a patch deployment.

EffectiveGuestPolicy

The effective guest policy that applies to a VM instance.

ExecStep

A step that runs an executable for a PatchJob.

ExecStepConfig

Common configurations for an ExecStep.

ExecutePatchJobRequest

A request message to initiate patching across Google Compute Engine instances.

GcsObject

Google Cloud Storage object representation.

GetGuestPolicyRequest

A request message for retrieving a guest policy.

GetPatchDeploymentRequest

A request message for retrieving a patch deployment.

GetPatchJobRequest

Request to get an active or completed patch job.

GooRepository

Represents a Goo package repository. These is added to a repo file that is stored at C:/ProgramData/GooGet/repos/google_osconfig.repo.

GooSettings

Googet patching is performed by running googet update.

GuestPolicy

An OS Config resource representing a guest configuration policy. These policies represent the desired state for VM instance guest environments including packages to install or remove, package repository configurations, and software to install.

Instance

Namespace for instance state enums.

ListGuestPoliciesRequest

A request message for listing guest policies.

ListGuestPoliciesResponse

A response message for listing guest policies.

ListPatchDeploymentsRequest

A request message for listing patch deployments.

ListPatchDeploymentsResponse

A response message for listing patch deployments.

ListPatchJobInstanceDetailsRequest

Request to list details for all instances that are part of a patch job.

ListPatchJobInstanceDetailsResponse

A response message for listing the instances details for a patch job.

ListPatchJobsRequest

A request message for listing patch jobs.

ListPatchJobsResponse

A response message for listing patch jobs.

LookupEffectiveGuestPolicyRequest

A request message for getting the effective guest policy assigned to the instance.

MonthlySchedule

Represents a monthly schedule. An example of a valid monthly schedule is “on the third Tuesday of the month” or “on the 15th of the month”.

OneTimeSchedule

Sets the time for a one time patch deployment. Timestamp is in target=“_blank”>RFC3339 text format.

Package

Package is a reference to the software package to be installed or removed. The agent on the VM instance uses the system package manager to apply the config.

PackageRepository

A package repository.

PatchConfig

Patch configuration specifications. Contains details on how to apply the patch(es) to a VM instance.

PatchDeployment

Patch deployments are configurations that individual patch jobs use to complete a patch. These configurations include instance filter, package repository settings, and a schedule. For more information about creating and managing patch deployments, see Scheduling patch jobs.

PatchInstanceFilter

A filter to target VM instances for patching. The targeted VMs must meet all criteria specified. So if both labels and zones are specified, the patch job targets only VMs with those labels and in those zones.

PatchJob

A high level representation of a patch job that is either in progress or has completed.

PatchJobInstanceDetails

Patch details for a VM instance. For more information about reviewing VM instance details, see Listing all VM instance details for a specific patch job.

RecurringSchedule

Sets the time for recurring patch deployments.

SoftwareRecipe

A software recipe is a set of instructions for installing and configuring a piece of software. It consists of a set of artifacts that are downloaded, and a set of steps that install, configure, and/or update the software.

UpdateGuestPolicyRequest

A request message for updating a guest policy.

WeekDayOfMonth

Represents one week day in a month. An example is “the 4th Sunday”.

WeeklySchedule

Represents a weekly schedule.

WindowsUpdateSettings

Windows patching is performed using the Windows Update Agent.

YumRepository

Represents a single Yum package repository. This repository is added to a repo file that is stored at /etc/yum.repos.d/google_osconfig.repo.

YumSettings

Yum patching is performed by executing yum update. Additional options can be set to control how this is executed.

ZypperRepository

Represents a single Zypper package repository. This repository is added to a repo file that is stored at /etc/zypp/repos.d/google_osconfig.repo.

ZypperSettings

Zypper patching is performed by running zypper patch. See also https://en.opensuse.org/SDB:Zypper_manual.

Enums

DesiredState

The desired state that the OS Config agent maintains on the VM instance.