Module googapis::google::cloud::recaptchaenterprise::v1

Modules

annotate_assessment_request
Nested message and enum types in AnnotateAssessmentRequest.

key
Nested message and enum types in Key.

recaptcha_enterprise_service_client
Generated client implementations.

risk_analysis
Nested message and enum types in RiskAnalysis.

token_properties
Nested message and enum types in TokenProperties.

web_key_settings
Nested message and enum types in WebKeySettings.

Structs

AndroidKeySettings
Settings specific to keys that can be used by Android apps.

AnnotateAssessmentRequest
The request message to annotate an Assessment.

AnnotateAssessmentResponse
Empty response for AnnotateAssessment.

Assessment
A recaptcha assessment resource.

CreateAssessmentRequest
The create assessment request message.

CreateKeyRequest
The create key request message.

DeleteKeyRequest
The delete key request message.

Event

GetKeyRequest
The get key request message.

IosKeySettings
Settings specific to keys that can be used by iOS apps.

Key
A key used to identify and configure applications (web and/or mobile) that use reCAPTCHA Enterprise.

ListKeysRequest
The list keys request message.

ListKeysResponse
Response to request to list keys in a project.

RiskAnalysis
Risk analysis result for an event.

TokenProperties

UpdateKeyRequest
The update key request message.

WebKeySettings
Settings specific to keys that can be used by websites.