Module googapis::google::cloud::redis::v1

Modules

cloud_redis_client
Generated client implementations.

failover_instance_request
Nested message and enum types in FailoverInstanceRequest.

input_config
Nested message and enum types in InputConfig.

instance
Nested message and enum types in Instance.

output_config
Nested message and enum types in OutputConfig.

Structs

CreateInstanceRequest
Request for [CreateInstance][google.cloud.redis.v1.CloudRedis.CreateInstance].

DeleteInstanceRequest
Request for [DeleteInstance][google.cloud.redis.v1.CloudRedis.DeleteInstance].

ExportInstanceRequest
Request for [Export][google.cloud.redis.v1.CloudRedis.ExportInstance].

FailoverInstanceRequest
Request for [Failover][google.cloud.redis.v1.CloudRedis.FailoverInstance].

GcsDestination
The Cloud Storage location for the output content

GcsSource
The Cloud Storage location for the input content

GetInstanceRequest
Request for [GetInstance][google.cloud.redis.v1.CloudRedis.GetInstance].

ImportInstanceRequest
Request for [Import][google.cloud.redis.v1.CloudRedis.ImportInstance].

InputConfig
The input content

Instance
A Google Cloud Redis instance.

ListInstancesRequest
Request for [ListInstances][google.cloud.redis.v1.CloudRedis.ListInstances].

ListInstancesResponse
Response for [ListInstances][google.cloud.redis.v1.CloudRedis.ListInstances].

LocationMetadata
This location metadata represents additional configuration options for a given location where a Redis instance may be created. All fields are output only. It is returned as content of the google.cloud.location.Location.metadata field.

OperationMetadata
Represents the v1 metadata of the long-running operation.

OutputConfig
The output content

UpdateInstanceRequest
Request for [UpdateInstance][google.cloud.redis.v1.CloudRedis.UpdateInstance].

UpgradeInstanceRequest
Request for [UpgradeInstance][google.cloud.redis.v1.CloudRedis.UpgradeInstance].

ZoneMetadata
Defines specific information for a particular zone. Currently empty and reserved for future use only.