Struct googapis::google::cloud::retail::v2::SearchRequest

pub struct SearchRequest {

    pub placement: String,
    pub branch: String,
    pub query: String,
    pub visitor_id: String,
    pub user_info: Option<UserInfo>,
    pub page_size: i32,
    pub page_token: String,
    pub offset: i32,
    pub filter: String,
    pub canonical_filter: String,
    pub order_by: String,
    pub facet_specs: Vec<FacetSpec>,
    pub dynamic_facet_spec: Option<DynamicFacetSpec>,
    pub boost_spec: Option<BoostSpec>,
    pub query_expansion_spec: Option<QueryExpansionSpec>,
    pub variant_rollup_keys: Vec<String>,
    pub page_categories: Vec<String>,
    pub search_mode: i32,

}

Request message for [SearchService.Search][google.cloud.retail.v2.SearchService.Search] method.

Fields
placement: String

Required. The resource name of the search engine placement, such as projects/*/locations/global/catalogs/default_catalog/placements/default_search. This field is used to identify the serving configuration name and the set of models that will be used to make the search.

branch: String

The branch resource name, such as projects/*/locations/global/catalogs/default_catalog/branches/0.

Use “default_branch” as the branch ID or leave this field empty, to search products under the default branch.

query: String

Raw search query.

visitor_id: String

Required. A unique identifier for tracking visitors. For example, this could be implemented with an HTTP cookie, which should be able to uniquely identify a visitor on a single device. This unique identifier should not change if the visitor logs in or out of the website.

The field must be a UTF-8 encoded string with a length limit of 128 characters. Otherwise, an INVALID_ARGUMENT error is returned.

user_info: Option<UserInfo>

User information.

page_size: i32

Maximum number of [Product][google.cloud.retail.v2.Product]s to return. If unspecified, defaults to a reasonable value. The maximum allowed value is 120. Values above 120 will be coerced to 120.

If this field is negative, an INVALID_ARGUMENT is returned.

page_token: String

A page token [SearchResponse.next_page_token][google.cloud.retail.v2.SearchResponse.next_page_token], received from a previous [SearchService.Search][google.cloud.retail.v2.SearchService.Search] call. Provide this to retrieve the subsequent page.

When paginating, all other parameters provided to [SearchService.Search][google.cloud.retail.v2.SearchService.Search] must match the call that provided the page token. Otherwise, an INVALID_ARGUMENT error is returned.

offset: i32

A 0-indexed integer that specifies the current offset (that is, starting result location, amongst the [Product][google.cloud.retail.v2.Product]s deemed by the API as relevant) in search results. This field is only considered if [page_token][google.cloud.retail.v2.SearchRequest.page_token] is unset.

If this field is negative, an INVALID_ARGUMENT is returned.

filter: String

The filter syntax consists of an expression language for constructing a predicate from one or more fields of the products being filtered. Filter expression is case-sensitive. See more details at this user guide.

If this field is unrecognizable, an INVALID_ARGUMENT is returned.

canonical_filter: String

The filter applied to every search request when quality improvement such as query expansion is needed. For example, if a query does not have enough results, an expanded query with [SearchRequest.canonical_filter][google.cloud.retail.v2.SearchRequest.canonical_filter] will be returned as a supplement of the original query. This field is strongly recommended to achieve high search quality.

See [SearchRequest.filter][google.cloud.retail.v2.SearchRequest.filter] for more details about filter syntax.

order_by: String

The order in which products are returned. Products can be ordered by a field in an [Product][google.cloud.retail.v2.Product] object. Leave it unset if ordered by relevance. OrderBy expression is case-sensitive. See more details at this user guide.

If this field is unrecognizable, an INVALID_ARGUMENT is returned.

facet_specs: Vec<FacetSpec>

Facet specifications for faceted search. If empty, no facets are returned.

A maximum of 100 values are allowed. Otherwise, an INVALID_ARGUMENT error is returned.

dynamic_facet_spec: Option<DynamicFacetSpec>

The specification for dynamically generated facets. Notice that only textual facets can be dynamically generated.

This feature requires additional allowlisting. Contact Retail Search support team if you are interested in using dynamic facet feature.

boost_spec: Option<BoostSpec>

Boost specification to boost certain products. See more details at this user guide.

Notice that if both [ServingConfig.boost_control_ids][] and [SearchRequest.boost_spec] are set, the boost conditions from both places are evaluated. If a search request matches multiple boost conditions, the final boost score is equal to the sum of the boost scores from all matched boost conditions.

query_expansion_spec: Option<QueryExpansionSpec>

The query expansion specification that specifies the conditions under which query expansion will occur. See more details at this user guide.

variant_rollup_keys: Vec<String>

The keys to fetch and rollup the matching [variant][google.cloud.retail.v2.Product.Type.VARIANT] [Product][google.cloud.retail.v2.Product]s attributes. The attributes from all the matching [variant][google.cloud.retail.v2.Product.Type.VARIANT] [Product][google.cloud.retail.v2.Product]s are merged and de-duplicated. Notice that rollup [variant][google.cloud.retail.v2.Product.Type.VARIANT] [Product][google.cloud.retail.v2.Product]s attributes will lead to extra query latency. Maximum number of keys is 10.

For [FulfillmentInfo][google.cloud.retail.v2.FulfillmentInfo], a fulfillment type and a fulfillment ID must be provided in the format of “fulfillmentType.fulfillmentId”. E.g., in “pickupInStore.store123”, “pickupInStore” is fulfillment type and “store123” is the store ID.

Supported keys are:

colorFamilies

price

originalPrice

discount

variantId

inventory(place_id,price)

inventory(place_id,attributes.key), where key is any key in the [Product.inventories.attributes][] map.

attributes.key, where key is any key in the [Product.attributes][google.cloud.retail.v2.Product.attributes] map.

pickupInStore.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2.FulfillmentInfo.type] “pickup-in-store”.

shipToStore.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2.FulfillmentInfo.type] “ship-to-store”.

sameDayDelivery.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2.FulfillmentInfo.type] “same-day-delivery”.

nextDayDelivery.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2.FulfillmentInfo.type] “next-day-delivery”.

customFulfillment1.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2.FulfillmentInfo.type] “custom-type-1”.

customFulfillment2.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2.FulfillmentInfo.type] “custom-type-2”.

customFulfillment3.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2.FulfillmentInfo.type] “custom-type-3”.

customFulfillment4.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2.FulfillmentInfo.type] “custom-type-4”.

customFulfillment5.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2.FulfillmentInfo.type] “custom-type-5”.

If this field is set to an invalid value other than these, an INVALID_ARGUMENT error is returned.

page_categories: Vec<String>

The categories associated with a category page. Required for category navigation queries to achieve good search quality. The format should be the same as [UserEvent.page_categories][google.cloud.retail.v2.UserEvent.page_categories];

To represent full path of category, use ‘>’ sign to separate different hierarchies. If ‘>’ is part of the category name, please replace it with other character(s).

Category pages include special pages such as sales or promotions. For instance, a special sale page may have the category hierarchy: “pageCategories” : [“Sales > 2017 Black Friday Deals”].

search_mode: i32

The search mode of the search request. If not specified, a single search request triggers both product search and faceted search.

Implementations

impl SearchRequest

pub fn search_mode(&self) -> SearchMode

Returns the enum value of search_mode, or the default if the field is set to an invalid enum value.

pub fn set_search_mode(&mut self, value: SearchMode)

Sets search_mode to the provided enum value.