Module googapis::google::cloud::retail::v2alpha

Modules

big_query_source
Nested message and enum types in BigQuerySource.

catalog_service_client
Generated client implementations.

complete_query_response
Nested message and enum types in CompleteQueryResponse.

completion_data_input_config
Nested message and enum types in CompletionDataInputConfig.

completion_service_client
Generated client implementations.

export_errors_config
Nested message and enum types in ExportErrorsConfig.

import_errors_config
Nested message and enum types in ImportErrorsConfig.

import_products_request
Nested message and enum types in ImportProductsRequest.

interval
Nested message and enum types in Interval.

predict_response
Nested message and enum types in PredictResponse.

prediction_service_client
Generated client implementations.

price_info
Nested message and enum types in PriceInfo.

product
Nested message and enum types in Product.

product_input_config
Nested message and enum types in ProductInputConfig.

product_service_client
Generated client implementations.

rejoin_user_events_request
Nested message and enum types in RejoinUserEventsRequest.

search_request
Nested message and enum types in SearchRequest.

search_response
Nested message and enum types in SearchResponse.

search_service_client
Generated client implementations.

user_event_input_config
Nested message and enum types in UserEventInputConfig.

user_event_service_client
Generated client implementations.

Structs

AddFulfillmentPlacesMetadata
Metadata related to the progress of the AddFulfillmentPlaces operation. Currently empty because there is no meaningful metadata populated from the [AddFulfillmentPlaces][] method.

AddFulfillmentPlacesRequest
Request message for [AddFulfillmentPlaces][] method.

AddFulfillmentPlacesResponse
Response of the RemoveFulfillmentPlacesRequest. Currently empty because there is no meaningful response populated from the [AddFulfillmentPlaces][] method.

Audience
An intended audience of the [Product][google.cloud.retail.v2alpha.Product] for whom it’s sold.

BigQuerySource
BigQuery source import data from.

Catalog
The catalog configuration.

CollectUserEventRequest
Request message for CollectUserEvent method.

ColorInfo
The color information of a [Product][google.cloud.retail.v2alpha.Product].

CompleteQueryRequest
Auto-complete parameters.

CompleteQueryResponse
Response of the auto-complete query.

CompletionDataInputConfig
The input config source for completion data.

CompletionDetail
Detailed completion information including completion attribution token and clicked completion info.

CreateProductRequest
Request message for [CreateProduct][] method.

CustomAttribute
A custom attribute that is not explicitly modeled in [Product][google.cloud.retail.v2alpha.Product].

DeleteProductRequest
Request message for [DeleteProduct][] method.

ExportErrorsConfig
Configuration of destination for Export related errors.

ExportMetadata
Metadata related to the progress of the Export operation. This will be returned by the google.longrunning.Operation.metadata field.

ExportProductsResponse
Response of the ExportProductsRequest. If the long running operation is done, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

ExportUserEventsResponse
Response of the ExportUserEventsRequest. If the long running operation was successful, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

FulfillmentInfo
Fulfillment information, such as the store IDs for in-store pickup or region IDs for different shipping methods.

GcsSource
Google Cloud Storage location for input content. format.

GetDefaultBranchRequest
Request message to show which branch is currently the default branch.

GetDefaultBranchResponse
Response message of [CatalogService.GetDefaultBranch][google.cloud.retail.v2alpha.CatalogService.GetDefaultBranch].

GetProductRequest
Request message for [GetProduct][] method.

Image
[Product][google.cloud.retail.v2alpha.Product] thumbnail/detail image.

ImportCompletionDataRequest
Request message for ImportCompletionData methods.

ImportCompletionDataResponse
Response of the [ImportCompletionDataRequest][google.cloud.retail.v2alpha.ImportCompletionDataRequest]. If the long running operation is done, this message is returned by the google.longrunning.Operations.response field if the operation is successful.

ImportErrorsConfig
Configuration of destination for Import related errors.

ImportMetadata
Metadata related to the progress of the Import operation. This will be returned by the google.longrunning.Operation.metadata field.

ImportProductsRequest
Request message for Import methods.

ImportProductsResponse
Response of the [ImportProductsRequest][google.cloud.retail.v2alpha.ImportProductsRequest]. If the long running operation is done, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

ImportUserEventsRequest
Request message for the ImportUserEvents request.

ImportUserEventsResponse
Response of the ImportUserEventsRequest. If the long running operation was successful, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

Interval
A floating point interval.

ListCatalogsRequest
Request for [CatalogService.ListCatalogs][google.cloud.retail.v2alpha.CatalogService.ListCatalogs] method.

ListCatalogsResponse
Response for [CatalogService.ListCatalogs][google.cloud.retail.v2alpha.CatalogService.ListCatalogs] method.

ListProductsRequest
Request message for [ProductService.ListProducts][google.cloud.retail.v2alpha.ProductService.ListProducts] method.

ListProductsResponse
Response message for [ProductService.ListProducts][google.cloud.retail.v2alpha.ProductService.ListProducts] method.

MerchantCenterLink
Represents a link between a Merchant Center account and a branch. Once a link is established, products from the linked merchant center account will be streamed to the linked branch.

MerchantCenterLinkingConfig
Configures Merchant Center linking. Links contained in the config will be used to sync data from a Merchant Center account to a Cloud Retail branch.

PredictRequest
Request message for Predict method.

PredictResponse
Response message for predict method.

PriceInfo
The price information of a [Product][google.cloud.retail.v2alpha.Product].

Product
Product captures all metadata information of items to be recommended or searched.

ProductDetail
Detailed product information associated with a user event.

ProductInlineSource
The inline source for the input config for ImportProducts method.

ProductInputConfig
The input config source for products.

ProductLevelConfig
Configures what level the product should be uploaded with regards to how users will be send events and how predictions will be made.

Promotion
Promotion information.

PurchaseTransaction
A transaction represents the entire purchase transaction.

PurgeMetadata
Metadata related to the progress of the Purge operation. This will be returned by the google.longrunning.Operation.metadata field.

PurgeUserEventsRequest
Request message for PurgeUserEvents method.

PurgeUserEventsResponse
Response of the PurgeUserEventsRequest. If the long running operation is successfully done, then this message is returned by the google.longrunning.Operations.response field.

Rating
The rating of a [Product][google.cloud.retail.v2alpha.Product].

RejoinUserEventsMetadata
Metadata for RejoinUserEvents method.

RejoinUserEventsRequest
Request message for RejoinUserEvents method.

RejoinUserEventsResponse
Response message for RejoinUserEvents method.

RemoveFulfillmentPlacesMetadata
Metadata related to the progress of the RemoveFulfillmentPlaces operation. Currently empty because there is no meaningful metadata populated from the [RemoveFulfillmentPlaces][] method.

RemoveFulfillmentPlacesRequest
Request message for [RemoveFulfillmentPlaces][] method.

RemoveFulfillmentPlacesResponse
Response of the RemoveFulfillmentPlacesRequest. Currently empty because there is no meaningful response populated from the [RemoveFulfillmentPlaces][] method.

SearchRequest
Request message for [SearchService.Search][google.cloud.retail.v2alpha.SearchService.Search] method.

SearchResponse
Response message for [SearchService.Search][google.cloud.retail.v2alpha.SearchService.Search] method.

SetDefaultBranchRequest
Request message to set a specified branch as new default_branch.

SetInventoryMetadata
Metadata related to the progress of the SetInventory operation. Currently empty because there is no meaningful metadata populated from the [SetInventory][] method.

SetInventoryRequest
Request message for [SetInventory][] method.

SetInventoryResponse
Response of the SetInventoryRequest. Currently empty because there is no meaningful response populated from the [SetInventory][] method.

UpdateCatalogRequest
Request for [CatalogService.UpdateCatalog][google.cloud.retail.v2alpha.CatalogService.UpdateCatalog] method.

UpdateProductRequest
Request message for [UpdateProduct][] method.

UserEvent
UserEvent captures all metadata information Retail API needs to know about how end users interact with customers’ website.

UserEventImportSummary
A summary of import result. The UserEventImportSummary summarizes the import status for user events.

UserEventInlineSource
The inline source for the input config for ImportUserEvents method.

UserEventInputConfig
The input config source for user events.

UserInfo
Information of an end user.

WriteUserEventRequest
Request message for WriteUserEvent method.