Struct googapis::google::cloud::retail::v2alpha::search_request::facet_spec::FacetKey

pub struct FacetKey {
    pub key: String,
    pub intervals: Vec<Interval>,
    pub restricted_values: Vec<String>,
    pub prefixes: Vec<String>,
    pub contains: Vec<String>,
    pub order_by: String,
    pub query: String,
}

Specifies how a facet is computed.

Fields
key: String

Required. Supported textual and numerical facet keys in [Product][google.cloud.retail.v2alpha.Product] object, over which the facet values are computed. Facet key is case-sensitive.

Allowed facet keys when [FacetKey.query][google.cloud.retail.v2alpha.SearchRequest.FacetSpec.FacetKey.query] is not specified:

textual_field =
# The [Product.brands][google.cloud.retail.v2alpha.Product.brands].
“brands”;
# The [Product.categories][google.cloud.retail.v2alpha.Product.categories].
“categories”;
# The [Audience.genders][google.cloud.retail.v2alpha.Audience.genders].
| “genders”;
# The [Audience.age_groups][google.cloud.retail.v2alpha.Audience.age_groups].
| “ageGroups”;
# The [Product.availability][google.cloud.retail.v2alpha.Product.availability]. Value is one of
# “IN_STOCK”, “OUT_OF_STOCK”, PREORDER“, “BACKORDER”.
| “availability”;
# The [ColorInfo.color_families][google.cloud.retail.v2alpha.ColorInfo.color_families].
| “colorFamilies”;
# The [ColorInfo.colors][google.cloud.retail.v2alpha.ColorInfo.colors].
| “colors”;
# The [Product.sizes][google.cloud.retail.v2alpha.Product.sizes].
| “sizes”;
# The [Product.materials][google.cloud.retail.v2alpha.Product.materials].
| “materials”;
# The [Product.patterns][google.cloud.retail.v2alpha.Product.patterns].
| “patterns”;
# The [Product.conditions][google.cloud.retail.v2alpha.Product.conditions].
| “conditions”;
# The textual custom attribute in [Product][google.cloud.retail.v2alpha.Product] object. Key can
# be any key in the [Product.attributes][google.cloud.retail.v2alpha.Product.attributes] map
# if the attribute values are textual.
# map.
| “attributes.key”;
*# The [FulfillmentInfo.ids][] for type # [FulfillmentInfo.Type.PICKUP_IN_STORE][].
| “pickupInStore”;
*# The [FulfillmentInfo.ids][] for type # [FulfillmentInfo.Type.SHIP_TO_STORE][].
| “shipToStore”;
*# The [FulfillmentInfo.ids][] for type # [FulfillmentInfo.Type.SAME_DAY_DELIVERY][].
| “sameDayDelivery”;
*# The [FulfillmentInfo.ids][] for type # [FulfillmentInfo.Type.NEXT_DAY_DELIVERY][].
| “nextDayDelivery”;
*# The [FulfillmentInfo.ids][] for type # [FulfillmentInfo.Type.CUSTOM_TYPE_1][].
| “customFulfillment1”;
*# The [FulfillmentInfo.ids][] for type # [FulfillmentInfo.Type.CUSTOM_TYPE_2][].
| “customFulfillment2”;
*# The [FulfillmentInfo.ids][] for type # [FulfillmentInfo.Type.CUSTOM_TYPE_3][].
| “customFulfillment3”;
*# The [FulfillmentInfo.ids][] for type # [FulfillmentInfo.Type.CUSTOM_TYPE_4][].
| “customFulfillment4”;
*# The [FulfillmentInfo.ids][] for type # [FulfillmentInfo.Type.CUSTOM_TYPE_5][].
| “customFulfillment5”;

numerical_field =
# The [PriceInfo.price][google.cloud.retail.v2alpha.PriceInfo.price].
“price”;
# The discount. Computed by (original_price-price)/price
“discount”;
# The [Rating.average_rating][google.cloud.retail.v2alpha.Rating.average_rating].
“rating”;
# The [Rating.rating_count][google.cloud.retail.v2alpha.Rating.rating_count].
“ratingCount”;
# The numerical custom attribute in [Product][google.cloud.retail.v2alpha.Product] object. Key can
# be any key in the [Product.attributes][google.cloud.retail.v2alpha.Product.attributes] map
# if the attribute values are numerical.
| “attributes.key”;

intervals: Vec<Interval>

Set only if values should be bucketized into intervals. Must be set for facets with numerical values. Must not be set for facet with text values. Maximum number of intervals is 30.

restricted_values: Vec<String>

Only get facet for the given restricted values. For example, when using “pickupInStore” as key and set restricted values to [“store123”, “store456”], only facets for “store123” and “store456” are returned. Only supported on textual fields and fulfillments. Maximum is 20.

Must be set for the fulfillment facet keys:

pickupInStore

shipToStore

sameDayDelivery

nextDayDelivery

customFulfillment1

customFulfillment2

customFulfillment3

customFulfillment4

customFulfillment5

prefixes: Vec<String>

Only get facet values that start with the given string prefix. For example, suppose “categories” has three values “Women > Shoe”, “Women > Dress” and “Men > Shoe”. If set “prefixes” to “Women”, the “categories” facet will give only “Women > Shoe” and “Women > Dress”. Only supported on textual fields. Maximum is 10.

contains: Vec<String>

Only get facet values that contains the given strings. For example, suppose “categories” has three values “Women > Shoe”, “Women > Dress” and “Men > Shoe”. If set “contains” to “Shoe”, the “categories” facet will give only “Women > Shoe” and “Men > Shoe”. Only supported on textual fields. Maximum is 10.

order_by: String

The order in which [Facet.values][] are returned.

Allowed values are:

“count desc”, which means order by [Facet.FacetValue.count][] descending.

“value desc”, which means order by [Facet.FacetValue.value][] descending. Only applies to textual facets.

If not set, textual values are sorted in natural order; numerical intervals are sorted in the order given by [FacetSpec.FacetKey.intervals][google.cloud.retail.v2alpha.SearchRequest.FacetSpec.FacetKey.intervals]; [FulfillmentInfo.ids][] are sorted in the order given by [FacetSpec.FacetKey.restricted_values][google.cloud.retail.v2alpha.SearchRequest.FacetSpec.FacetKey.restricted_values].

query: String

The query that is used to compute facet for the given facet key. When provided, it will override the default behavior of facet computation. The query syntax is the same as a filter expression. See [SearchRequest.filter][google.cloud.retail.v2alpha.SearchRequest.filter] for detail syntax and limitations. Notice that there is no limitation on [FacetKey.key][google.cloud.retail.v2alpha.SearchRequest.FacetSpec.FacetKey.key] when query is specified.

In the response, [FacetValue.value][] will be always “1” and [FacetValue.count][] will be the number of results that matches the query.

For example, you can set a customized facet for “shipToStore”, where [FacetKey.key][google.cloud.retail.v2alpha.SearchRequest.FacetSpec.FacetKey.key] is “customizedShipToStore”, and [FacetKey.query][google.cloud.retail.v2alpha.SearchRequest.FacetSpec.FacetKey.query] is “availability: ANY("IN_STOCK") AND shipToStore: ANY("123")”. Then the facet will count the products that are both in stock and ship to store “123”.