Module googapis::google::cloud::security::privateca::v1::ca_pool::issuance_policy::allowed_key_type

Nested message and enum types in AllowedKeyType.

Modules

ec_key_type
Nested message and enum types in EcKeyType.

Structs

EcKeyType
Describes an Elliptic Curve key that may be used in a [Certificate][google.cloud.security.privateca.v1.Certificate] issued from a [CaPool][google.cloud.security.privateca.v1.CaPool].

RsaKeyType
Describes an RSA key that may be used in a [Certificate][google.cloud.security.privateca.v1.Certificate] issued from a [CaPool][google.cloud.security.privateca.v1.CaPool].

Enums

KeyType