Module googapis::google::cloud::servicedirectory::v1beta1

Modules

lookup_service_client
Generated client implementations.

registration_service_client
Generated client implementations.

Structs

CreateEndpointRequest
The request message for [RegistrationService.CreateEndpoint][google.cloud.servicedirectory.v1beta1.RegistrationService.CreateEndpoint].

CreateNamespaceRequest
The request message for [RegistrationService.CreateNamespace][google.cloud.servicedirectory.v1beta1.RegistrationService.CreateNamespace].

CreateServiceRequest
The request message for [RegistrationService.CreateService][google.cloud.servicedirectory.v1beta1.RegistrationService.CreateService].

DeleteEndpointRequest
The request message for [RegistrationService.DeleteEndpoint][google.cloud.servicedirectory.v1beta1.RegistrationService.DeleteEndpoint].

DeleteNamespaceRequest
The request message for [RegistrationService.DeleteNamespace][google.cloud.servicedirectory.v1beta1.RegistrationService.DeleteNamespace].

DeleteServiceRequest
The request message for [RegistrationService.DeleteService][google.cloud.servicedirectory.v1beta1.RegistrationService.DeleteService].

Endpoint
An individual endpoint that provides a [service][google.cloud.servicedirectory.v1beta1.Service]. The service must already exist to create an endpoint.

GetEndpointRequest
The request message for [RegistrationService.GetEndpoint][google.cloud.servicedirectory.v1beta1.RegistrationService.GetEndpoint]. This should not be used to lookup endpoints at runtime. Instead, use the resolve method.

GetNamespaceRequest
The request message for [RegistrationService.GetNamespace][google.cloud.servicedirectory.v1beta1.RegistrationService.GetNamespace].

GetServiceRequest
The request message for [RegistrationService.GetService][google.cloud.servicedirectory.v1beta1.RegistrationService.GetService]. This should not be used for looking up a service. Insead, use the resolve method as it contains all endpoints and associated metadata.

ListEndpointsRequest
The request message for [RegistrationService.ListEndpoints][google.cloud.servicedirectory.v1beta1.RegistrationService.ListEndpoints].

ListEndpointsResponse
The response message for [RegistrationService.ListEndpoints][google.cloud.servicedirectory.v1beta1.RegistrationService.ListEndpoints].

ListNamespacesRequest
The request message for [RegistrationService.ListNamespaces][google.cloud.servicedirectory.v1beta1.RegistrationService.ListNamespaces].

ListNamespacesResponse
The response message for [RegistrationService.ListNamespaces][google.cloud.servicedirectory.v1beta1.RegistrationService.ListNamespaces].

ListServicesRequest
The request message for [RegistrationService.ListServices][google.cloud.servicedirectory.v1beta1.RegistrationService.ListServices].

ListServicesResponse
The response message for [RegistrationService.ListServices][google.cloud.servicedirectory.v1beta1.RegistrationService.ListServices].

Namespace
A container for [services][google.cloud.servicedirectory.v1beta1.Service]. Namespaces allow administrators to group services together and define permissions for a collection of services.

ResolveServiceRequest
The request message for [LookupService.ResolveService][google.cloud.servicedirectory.v1beta1.LookupService.ResolveService]. Looks up a service by its name, returns the service and its endpoints.

ResolveServiceResponse
The response message for [LookupService.ResolveService][google.cloud.servicedirectory.v1beta1.LookupService.ResolveService].

Service
An individual service. A service contains a name and optional metadata. A service must exist before [endpoints][google.cloud.servicedirectory.v1beta1.Endpoint] can be added to it.

UpdateEndpointRequest
The request message for [RegistrationService.UpdateEndpoint][google.cloud.servicedirectory.v1beta1.RegistrationService.UpdateEndpoint].

UpdateNamespaceRequest
The request message for [RegistrationService.UpdateNamespace][google.cloud.servicedirectory.v1beta1.RegistrationService.UpdateNamespace].

UpdateServiceRequest
The request message for [RegistrationService.UpdateService][google.cloud.servicedirectory.v1beta1.RegistrationService.UpdateService].