Module googapis::google::cloud::shell::v1

Modules

cloud_shell_error_details
Nested message and enum types in CloudShellErrorDetails.

cloud_shell_service_client
Generated client implementations.

environment
Nested message and enum types in Environment.

start_environment_metadata
Nested message and enum types in StartEnvironmentMetadata.

Structs

AddPublicKeyMetadata
Message included in the metadata field of operations returned from [AddPublicKey][google.cloud.shell.v1.CloudShellService.AddPublicKey].

AddPublicKeyRequest
Request message for [AddPublicKey][google.cloud.shell.v1.CloudShellService.AddPublicKey].

AddPublicKeyResponse
Response message for [AddPublicKey][google.cloud.shell.v1.CloudShellService.AddPublicKey].

AuthorizeEnvironmentMetadata
Message included in the metadata field of operations returned from [AuthorizeEnvironment][google.cloud.shell.v1.CloudShellService.AuthorizeEnvironment].

AuthorizeEnvironmentRequest
Request message for [AuthorizeEnvironment][google.cloud.shell.v1.CloudShellService.AuthorizeEnvironment].

AuthorizeEnvironmentResponse
Response message for [AuthorizeEnvironment][google.cloud.shell.v1.CloudShellService.AuthorizeEnvironment].

CloudShellErrorDetails
Cloud-shell specific information that will be included as details in failure responses.

CreateEnvironmentMetadata
Message included in the metadata field of operations returned from [CreateEnvironment][google.cloud.shell.v1.CloudShellService.CreateEnvironment].

DeleteEnvironmentMetadata
Message included in the metadata field of operations returned from [DeleteEnvironment][google.cloud.shell.v1.CloudShellService.DeleteEnvironment].

Environment
A Cloud Shell environment, which is defined as the combination of a Docker image specifying what is installed on the environment and a home directory containing the user’s data that will remain across sessions. Each user has at least an environment with the ID “default”.

GetEnvironmentRequest
Request message for [GetEnvironment][google.cloud.shell.v1.CloudShellService.GetEnvironment].

RemovePublicKeyMetadata
Message included in the metadata field of operations returned from [RemovePublicKey][google.cloud.shell.v1.CloudShellService.RemovePublicKey].

RemovePublicKeyRequest
Request message for [RemovePublicKey][google.cloud.shell.v1.CloudShellService.RemovePublicKey].

RemovePublicKeyResponse
Response message for [RemovePublicKey][google.cloud.shell.v1.CloudShellService.RemovePublicKey].

StartEnvironmentMetadata
Message included in the metadata field of operations returned from [StartEnvironment][google.cloud.shell.v1.CloudShellService.StartEnvironment].

StartEnvironmentRequest
Request message for [StartEnvironment][google.cloud.shell.v1.CloudShellService.StartEnvironment].

StartEnvironmentResponse
Message included in the response field of operations returned from [StartEnvironment][google.cloud.shell.v1.CloudShellService.StartEnvironment] once the operation is complete.