Struct googapis::google::cloud::sql::v1beta4::DatabaseInstance

pub struct DatabaseInstance {

    pub kind: String,
    pub state: i32,
    pub database_version: i32,
    pub settings: Option<Settings>,
    pub etag: String,
    pub failover_replica: Option<SqlFailoverReplica>,
    pub master_instance_name: String,
    pub replica_names: Vec<String>,
    pub max_disk_size: Option<i64>,
    pub current_disk_size: Option<i64>,
    pub ip_addresses: Vec<IpMapping>,
    pub server_ca_cert: Option<SslCert>,
    pub instance_type: i32,
    pub project: String,
    pub ipv6_address: String,
    pub service_account_email_address: String,
    pub on_premises_configuration: Option<OnPremisesConfiguration>,
    pub replica_configuration: Option<ReplicaConfiguration>,
    pub backend_type: i32,
    pub self_link: String,
    pub suspension_reason: Vec<i32>,
    pub connection_name: String,
    pub name: String,
    pub region: String,
    pub gce_zone: String,
    pub secondary_gce_zone: String,
    pub disk_encryption_configuration: Option<DiskEncryptionConfiguration>,
    pub disk_encryption_status: Option<DiskEncryptionStatus>,
    pub root_password: String,
    pub scheduled_maintenance: Option<SqlScheduledMaintenance>,
    pub satisfies_pzs: Option<bool>,
    pub out_of_disk_report: Option<SqlOutOfDiskReport>,

}

A Cloud SQL instance resource.

Fields
kind: String

This is always sql#instance.

state: i32

The current serving state of the Cloud SQL instance. This can be one of the following.
SQL_INSTANCE_STATE_UNSPECIFIED: The state of the instance is unknown.
RUNNABLE: The instance is running, or has been stopped by owner.
SUSPENDED: The instance is not available, for example due to problems with billing.
PENDING_DELETE: The instance is being deleted.
PENDING_CREATE: The instance is being created.
MAINTENANCE: The instance is down for maintenance.
FAILED: The instance creation failed.

database_version: i32

The database engine type and version. The databaseVersion field cannot be changed after instance creation.
MySQL instances: MYSQL_8_0, MYSQL_5_7 (default), or MYSQL_5_6.
PostgreSQL instances: POSTGRES_9_6, POSTGRES_10, POSTGRES_11, POSTGRES_12, or POSTGRES_13 (default).
SQL Server instances: SQLSERVER_2017_STANDARD (default), SQLSERVER_2017_ENTERPRISE, SQLSERVER_2017_EXPRESS, or SQLSERVER_2017_WEB.

settings: Option<Settings>

The user settings.

etag: String

This field is deprecated and will be removed from a future version of the API. Use the settings.settingsVersion field instead.

failover_replica: Option<SqlFailoverReplica>

The name and status of the failover replica. This property is applicable only to Second Generation instances.

master_instance_name: String

The name of the instance which will act as primary in the replication setup.

replica_names: Vec<String>

The replicas of the instance.

max_disk_size: Option<i64>

👎 Deprecated

The maximum disk size of the instance in bytes.

current_disk_size: Option<i64>

👎 Deprecated

The current disk usage of the instance in bytes. This property has been deprecated. Use the “cloudsql.googleapis.com/database/disk/bytes_used” metric in Cloud Monitoring API instead. Please see this announcement for details.

ip_addresses: Vec<IpMapping>

The assigned IP addresses for the instance.

server_ca_cert: Option<SslCert>

SSL configuration.

instance_type: i32

The instance type. This can be one of the following.
CLOUD_SQL_INSTANCE: A Cloud SQL instance that is not replicating from a primary instance.
ON_PREMISES_INSTANCE: An instance running on the customer’s premises.
READ_REPLICA_INSTANCE: A Cloud SQL instance configured as a read-replica.

project: String

The project ID of the project containing the Cloud SQL instance. The Google apps domain is prefixed if applicable.

ipv6_address: String

👎 Deprecated

The IPv6 address assigned to the instance. (Deprecated) This property was applicable only to First Generation instances.

service_account_email_address: String

The service account email address assigned to the instance.
This property is read-only.

on_premises_configuration: Option<OnPremisesConfiguration>

Configuration specific to on-premises instances.

replica_configuration: Option<ReplicaConfiguration>

Configuration specific to failover replicas and read replicas.

backend_type: i32

SECOND_GEN: Cloud SQL database instance.
EXTERNAL: A database server that is not managed by Google.
This property is read-only; use the tier property in the settings object to determine the database type.

self_link: String

The URI of this resource.

suspension_reason: Vec<i32>

If the instance state is SUSPENDED, the reason for the suspension.

connection_name: String

Connection name of the Cloud SQL instance used in connection strings.

name: String

Name of the Cloud SQL instance. This does not include the project ID.

region: String

The geographical region. Can be
us-central (FIRST_GEN instances only)
us-central1 (SECOND_GEN instances only)
asia-east1 or europe-west1.
Defaults to us-central or us-central1 depending on the instance type. The region cannot be changed after instance creation.

gce_zone: String

The Compute Engine zone that the instance is currently serving from. This value could be different from the zone that was specified when the instance was created if the instance has failed over to its secondary zone.

secondary_gce_zone: String

The Compute Engine zone that the failover instance is currently serving from for a regional instance. This value could be different from the zone that was specified when the instance was created if the instance has failed over to its secondary/failover zone. Reserved for future use.

disk_encryption_configuration: Option<DiskEncryptionConfiguration>

Disk encryption configuration specific to an instance. Applies only to Second Generation instances.

disk_encryption_status: Option<DiskEncryptionStatus>

Disk encryption status specific to an instance. Applies only to Second Generation instances.

root_password: String

Initial root password. Use only on creation.

scheduled_maintenance: Option<SqlScheduledMaintenance>

The start time of any upcoming scheduled maintenance for this instance.

satisfies_pzs: Option<bool>

The status indicating if instance satisfiesPzs. Reserved for future use.

out_of_disk_report: Option<SqlOutOfDiskReport>

This field represents the report generated by the proactive database wellness job for OutOfDisk issues. Writers: – the proactive database wellness job for OOD. Readers: – the proactive database wellness job

Implementations

impl DatabaseInstance

pub fn state(&self) -> SqlInstanceState

Returns the enum value of state, or the default if the field is set to an invalid enum value.

pub fn set_state(&mut self, value: SqlInstanceState)

Sets state to the provided enum value.

pub fn database_version(&self) -> SqlDatabaseVersion

Returns the enum value of database_version, or the default if the field is set to an invalid enum value.

pub fn set_database_version(&mut self, value: SqlDatabaseVersion)

Sets database_version to the provided enum value.

pub fn instance_type(&self) -> SqlInstanceType

Returns the enum value of instance_type, or the default if the field is set to an invalid enum value.

pub fn set_instance_type(&mut self, value: SqlInstanceType)

Sets instance_type to the provided enum value.

pub fn backend_type(&self) -> SqlBackendType

Returns the enum value of backend_type, or the default if the field is set to an invalid enum value.

pub fn set_backend_type(&mut self, value: SqlBackendType)

Sets backend_type to the provided enum value.

pub fn suspension_reason(
&self
) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<SqlSuspensionReason>>

Returns an iterator which yields the valid enum values contained in suspension_reason.

pub fn push_suspension_reason(&mut self, value: SqlSuspensionReason)

Appends the provided enum value to suspension_reason.