Module googapis::google::cloud::talent::v4

Modules

batch_operation_metadata
Nested message and enum types in BatchOperationMetadata.

client_event
Nested message and enum types in ClientEvent.

commute_filter
Nested message and enum types in CommuteFilter.

company
Nested message and enum types in Company.

company_service_client
Generated client implementations.

compensation_filter
Nested message and enum types in CompensationFilter.

compensation_info
Nested message and enum types in CompensationInfo.

complete_query_request
Nested message and enum types in CompleteQueryRequest.

complete_query_response
Nested message and enum types in CompleteQueryResponse.

completion_client
Generated client implementations.

device_info
Nested message and enum types in DeviceInfo.

event_service_client
Generated client implementations.

job
Nested message and enum types in Job.

job_event
Nested message and enum types in JobEvent.

job_service_client
Generated client implementations.

location
Nested message and enum types in Location.

location_filter
Nested message and enum types in LocationFilter.

search_jobs_request
Nested message and enum types in SearchJobsRequest.

search_jobs_response
Nested message and enum types in SearchJobsResponse.

tenant_service_client
Generated client implementations.

Structs

BatchCreateJobsRequest
Request to create a batch of jobs.

BatchCreateJobsResponse
The result of [JobService.BatchCreateJobs][google.cloud.talent.v4.JobService.BatchCreateJobs]. It’s used to replace [google.longrunning.Operation.response][google.longrunning.Operation.response] in case of success.

BatchDeleteJobsRequest
Request to delete a batch of jobs.

BatchDeleteJobsResponse
The result of [JobService.BatchDeleteJobs][google.cloud.talent.v4.JobService.BatchDeleteJobs]. It’s used to replace [google.longrunning.Operation.response][google.longrunning.Operation.response] in case of success.

BatchOperationMetadata
Metadata used for long running operations returned by CTS batch APIs. It’s used to replace [google.longrunning.Operation.metadata][google.longrunning.Operation.metadata].

BatchUpdateJobsRequest
Request to update a batch of jobs.

BatchUpdateJobsResponse
The result of [JobService.BatchUpdateJobs][google.cloud.talent.v4.JobService.BatchUpdateJobs]. It’s used to replace [google.longrunning.Operation.response][google.longrunning.Operation.response] in case of success.

ClientEvent
An event issued when an end user interacts with the application that implements Cloud Talent Solution. Providing this information improves the quality of results for the API clients, enabling the service to perform optimally. The number of events sent must be consistent with other calls, such as job searches, issued to the service by the client.

CommuteFilter
Parameters needed for commute search.

Company
A Company resource represents a company in the service. A company is the entity that owns job postings, that is, the hiring entity responsible for employing applicants for the job position.

CompensationFilter
Filter on job compensation type and amount.

CompensationInfo
Job compensation details.

CompleteQueryRequest
Auto-complete parameters.

CompleteQueryResponse
Response of auto-complete query.

CreateClientEventRequest
The report event request.

CreateCompanyRequest
The Request of the CreateCompany method.

CreateJobRequest
Create job request.

CreateTenantRequest
The Request of the CreateTenant method.

CustomAttribute
Custom attribute values that are either filterable or non-filterable.

DeleteCompanyRequest
Request to delete a company.

DeleteJobRequest
Delete job request.

DeleteTenantRequest
Request to delete a tenant.

DeviceInfo
Device information collected from the job seeker, candidate, or other entity conducting the job search. Providing this information improves the quality of the search results across devices.

GetCompanyRequest
Request for getting a company by name.

GetJobRequest
Get job request.

GetTenantRequest
Request for getting a tenant by name.

HistogramQuery
The histogram request.

HistogramQueryResult
Histogram result that matches [HistogramQuery][google.cloud.talent.v4.HistogramQuery] specified in searches.

Job
A Job resource represents a job posting (also referred to as a “job listing” or “job requisition”). A job belongs to a [Company][google.cloud.talent.v4.Company], which is the hiring entity responsible for the job.

JobEvent
An event issued when a job seeker interacts with the application that implements Cloud Talent Solution.

JobQuery
The query required to perform a search query.

JobResult
Mutation result of a job from a batch operation.

ListCompaniesRequest
List companies for which the client has ACL visibility.

ListCompaniesResponse
The List companies response object.

ListJobsRequest
List jobs request.

ListJobsResponse
List jobs response.

ListTenantsRequest
List tenants for which the client has ACL visibility.

ListTenantsResponse
The List tenants response object.

Location
A resource that represents a location with full geographic information.

LocationFilter
Geographic region of the search.

RequestMetadata
Meta information related to the job searcher or entity conducting the job search. This information is used to improve the performance of the service.

ResponseMetadata
Additional information returned to client, such as debugging information.

SearchJobsRequest
The Request body of the SearchJobs call.

SearchJobsResponse
Response for SearchJob method.

SpellingCorrection
Spell check result.

Tenant
A Tenant resource represents a tenant in the service. A tenant is a group or entity that shares common access with specific privileges for resources like jobs. Customer may create multiple tenants to provide data isolation for different groups.

TimestampRange
Message representing a period of time between two timestamps.

UpdateCompanyRequest
Request for updating a specified company.

UpdateJobRequest
Update job request.

UpdateTenantRequest
Request for updating a specified tenant.

Enums

CommuteMethod
Method for commute.

CompanySize
An enum that represents the size of the company.

DegreeType
Educational degree level defined in International Standard Classification of Education (ISCED).

EmploymentType
An enum that represents the employment type of a job.

HtmlSanitization
Option for HTML content sanitization on user input fields, for example, job description. By setting this option, user can determine whether and how sanitization is performed on these fields.

JobBenefit
An enum that represents employee benefits included with the job.

JobCategory
An enum that represents the categorization or primary focus of specific role. This value is different than the “industry” associated with a role, which is related to the categorization of the company listing the job.

JobLevel
An enum that represents the required experience level required for the job.

JobView
An enum that specifies the job attributes that are returned in the [MatchingJob.job][google.cloud.talent.v4.SearchJobsResponse.MatchingJob.job] or [ListJobsResponse.jobs][google.cloud.talent.v4.ListJobsResponse.jobs] fields.

PostingRegion
An enum that represents the job posting region. In most cases, job postings don’t need to specify a region. If a region is given, jobs are eligible for searches in the specified region.

Visibility
Deprecated. All resources are only visible to the owner.