Module googapis::google::cloud::talent::v4beta1

Modules

address

Nested message and enum types in Address.

application

Nested message and enum types in Application.

application_service_client

Generated client implementations.

batch_operation_metadata

Nested message and enum types in BatchOperationMetadata.

client_event

Nested message and enum types in ClientEvent.

commute_filter

Nested message and enum types in CommuteFilter.

company

Nested message and enum types in Company.

company_service_client

Generated client implementations.

compensation_filter

Nested message and enum types in CompensationFilter.

compensation_info

Nested message and enum types in CompensationInfo.

complete_query_request

Nested message and enum types in CompleteQueryRequest.

complete_query_response

Nested message and enum types in CompleteQueryResponse.

completion_client

Generated client implementations.

device_info

Nested message and enum types in DeviceInfo.

education_record

Nested message and enum types in EducationRecord.

employer_filter

Nested message and enum types in EmployerFilter.

event_service_client

Generated client implementations.

job

Nested message and enum types in Job.

job_event

Nested message and enum types in JobEvent.

job_operation_result

Nested message and enum types in JobOperationResult.

job_service_client

Generated client implementations.

location

Nested message and enum types in Location.

location_filter

Nested message and enum types in LocationFilter.

person_name

Nested message and enum types in PersonName.

phone

Nested message and enum types in Phone.

profile_event

Nested message and enum types in ProfileEvent.

profile_service_client

Generated client implementations.

resume

Nested message and enum types in Resume.

search_jobs_request

Nested message and enum types in SearchJobsRequest.

search_jobs_response

Nested message and enum types in SearchJobsResponse.

tenant

Nested message and enum types in Tenant.

tenant_service_client

Generated client implementations.

time_filter

Nested message and enum types in TimeFilter.

Structs

Activity

Resource that represents an individual or collaborative activity participated in by a candidate, for example, an open-source project, a class assignment, and so on.

AdditionalContactInfo

Resource that represents contact information other than phone, email, URI and addresses.

Address

Resource that represents a address.

Application

Resource that represents a job application record of a candidate.

ApplicationDateFilter

Application Date Range Filter.

ApplicationJobFilter

Filter on the job information of Application.

ApplicationOutcomeNotesFilter

Outcome Notes Filter.

AvailabilityFilter

Filter on availability signals.

AvailabilitySignal

Candidate availability signal.

BatchCreateJobsRequest

Request to create a batch of jobs.

BatchDeleteJobsRequest

Batch delete jobs request.

BatchOperationMetadata

Metadata used for long running operations returned by CTS batch APIs. It’s used to replace [google.longrunning.Operation.metadata][google.longrunning.Operation.metadata].

BatchUpdateJobsRequest

Request to update a batch of jobs.

CandidateAvailabilityFilter

Deprecated. Use AvailabilityFilter instead.

Certification

Resource that represents a license or certification.

ClientEvent

An event issued when an end user interacts with the application that implements Cloud Talent Solution. Providing this information improves the quality of results for the API clients, enabling the service to perform optimally. The number of events sent must be consistent with other calls, such as job searches, issued to the service by the client.

CommuteFilter

Parameters needed for commute search.

Company

A Company resource represents a company in the service. A company is the entity that owns job postings, that is, the hiring entity responsible for employing applicants for the job position.

CompensationFilter

Filter on job compensation type and amount.

CompensationInfo

Job compensation details.

CompleteQueryRequest

Auto-complete parameters.

CompleteQueryResponse

Response of auto-complete query.

CreateApplicationRequest

The Request of the CreateApplication method.

CreateClientEventRequest

The report event request.

CreateCompanyRequest

The Request of the CreateCompany method.

CreateJobRequest

Create job request.

CreateProfileRequest

Create profile request.

CreateTenantRequest

The Request of the CreateTenant method.

CustomAttribute

Custom attribute values that are either filterable or non-filterable.

Degree

Resource that represents a degree pursuing or acquired by a candidate.

DeleteApplicationRequest

Request to delete a application.

DeleteCompanyRequest

Request to delete a company.

DeleteJobRequest

Delete job request.

DeleteProfileRequest

Delete profile request.

DeleteTenantRequest

Request to delete a tenant.

DeviceInfo

Device information collected from the job seeker, candidate, or other entity conducting the job search. Providing this information improves the quality of the search results across devices.

EducationFilter

Education filter of the search.

EducationRecord

Resource that represents an education record of a candidate.

Email

Resource that represents a person’s email address.

EmployerFilter

Employer filter of the search.

EmploymentRecord

Resource that represents an employment record of a candidate.

GetApplicationRequest

Request for getting a application by name.

GetCompanyRequest

Request for getting a company by name.

GetJobRequest

Get job request.

GetProfileRequest

Get profile request.

GetTenantRequest

Request for getting a tenant by name.

HistogramQuery

The histogram request.

HistogramQueryResult

Histogram result that matches [HistogramQuery][google.cloud.talent.v4beta1.HistogramQuery] specified in searches.

Interview

Details of an interview.

Job

A Job resource represents a job posting (also referred to as a “job listing” or “job requisition”). A job belongs to a [Company][google.cloud.talent.v4beta1.Company], which is the hiring entity responsible for the job.

JobEvent

An event issued when a job seeker interacts with the application that implements Cloud Talent Solution.

JobOperationResult

The result of [JobService.BatchCreateJobs][google.cloud.talent.v4beta1.JobService.BatchCreateJobs] or [JobService.BatchUpdateJobs][google.cloud.talent.v4beta1.JobService.BatchUpdateJobs] APIs. It’s used to replace [google.longrunning.Operation.response][google.longrunning.Operation.response] in case of success.

JobQuery

The query required to perform a search query.

JobTitleFilter

Job title of the search.

ListApplicationsRequest

List applications for which the client has ACL visibility.

ListApplicationsResponse

The List applications response object.

ListCompaniesRequest

List companies for which the client has ACL visibility.

ListCompaniesResponse

The List companies response object.

ListJobsRequest

List jobs request.

ListJobsResponse

List jobs response.

ListProfilesRequest

List profiles request.

ListProfilesResponse

The List profiles response object.

ListTenantsRequest

List tenants for which the client has ACL visibility.

ListTenantsResponse

The List tenants response object.

Location

A resource that represents a location with full geographic information.

LocationFilter

Geographic region of the search.

Patent

Resource that represents the patent acquired by a candidate.

PersonName

Resource that represents the name of a person.

PersonNameFilter

Filter on person name.

PersonalUri

Resource that represents a valid URI for a personal use.

Phone

Resource that represents a person’s telephone number.

Profile

A resource that represents the profile for a job candidate (also referred to as a “single-source profile”).

ProfileEvent

An event issued when a profile searcher interacts with the application that implements Cloud Talent Solution.

ProfileQuery

Filters to apply when performing the search query.

Publication

Resource that represents a publication resource of a candidate.

Rating

The details of the score received for an assessment or interview.

RequestMetadata

Meta information related to the job searcher or entity conducting the job search. This information is used to improve the performance of the service.

ResponseMetadata

Additional information returned to client, such as debugging information.

Resume

Resource that represents a resume.

SearchJobsRequest

The Request body of the SearchJobs call.

SearchJobsResponse

Response for SearchJob method.

SearchProfilesRequest

The request body of the SearchProfiles call.

SearchProfilesResponse

Response of SearchProfiles method.

Skill

Resource that represents a skill of a candidate.

SkillFilter

Skill filter of the search.

SpellingCorrection

Spell check result.

SummarizedProfile

Profile entry with metadata inside [SearchProfilesResponse][google.cloud.talent.v4beta1.SearchProfilesResponse].

Tenant

A Tenant resource represents a tenant in the service. A tenant is a group or entity that shares common access with specific privileges for resources like profiles. Customer may create multiple tenants to provide data isolation for different groups.

TimeFilter

Filter on create timestamp or update timestamp of profiles.

TimestampRange

Message representing a period of time between two timestamps.

UpdateApplicationRequest

Request for updating a specified application.

UpdateCompanyRequest

Request for updating a specified company.

UpdateJobRequest

Update job request.

UpdateProfileRequest

Update profile request

UpdateTenantRequest

Request for updating a specified tenant.

WorkExperienceFilter

Work experience filter.

Enums

AvailabilitySignalType

The type of candidate availability signal.

CommuteMethod

Method for commute.

CompanySize

An enum that represents the size of the company.

ContactInfoUsage

Enum that represents the usage of the contact information.

DegreeType

Educational degree level defined in International Standard Classification of Education (ISCED).

EmploymentType

An enum that represents the employment type of a job.

HtmlSanitization

Option for HTML content sanitization on user input fields, for example, job description. By setting this option, user can determine whether and how sanitization is performed on these fields.

JobBenefit

An enum that represents employee benefits included with the job.

JobCategory

An enum that represents the categorization or primary focus of specific role. This value is different than the “industry” associated with a role, which is related to the categorization of the company listing the job.

JobLevel

An enum that represents the required experience level required for the job.

JobView

An enum that specifies the job attributes that are returned in the [MatchingJob.job][google.cloud.talent.v4beta1.SearchJobsResponse.MatchingJob.job] or [ListJobsResponse.jobs][google.cloud.talent.v4beta1.ListJobsResponse.jobs] fields.

Outcome

The overall outcome /decision / result indicator.

PostingRegion

An enum that represents the job posting region. In most cases, job postings don’t need to specify a region. If a region is given, jobs are eligible for searches in the specified region.

SkillProficiencyLevel

Enum that represents the skill proficiency level.

Visibility

Deprecated. All resources are only visible to the owner.