Module googapis::google::cloud::talent::v4beta1::search_jobs_request

Nested message and enum types in SearchJobsRequest.

Modules

custom_ranking_info
Nested message and enum types in CustomRankingInfo.

Structs

CustomRankingInfo
Custom ranking information for [SearchJobsRequest][google.cloud.talent.v4beta1.SearchJobsRequest].

Enums

DiversificationLevel
Controls whether highly similar jobs are returned next to each other in the search results. Jobs are identified as highly similar based on their titles, job categories, and locations. Highly similar results are clustered so that only one representative job of the cluster is displayed to the job seeker higher up in the results, with the other jobs being displayed lower down in the results.

SearchMode
A string-represented enumeration of the job search mode. The service operate differently for different modes of service.