Struct googapis::google::cloud::tasks::v2beta3::cloud_tasks_client::CloudTasksClient

pub struct CloudTasksClient<T> { /* fields omitted */ }

Cloud Tasks allows developers to manage the execution of background work in their applications.

Implementations

impl<T> CloudTasksClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + Sync + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> CloudTasksClient<InterceptedService<T, F>> where
F: FnMut(Request<()>) -> Result<Request<()>, Status>,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn list_queues(
&mut self,
request: impl IntoRequest<ListQueuesRequest>
) -> Result<Response<ListQueuesResponse>, Status>

Lists queues.

Queues are returned in lexicographical order.

pub async fn get_queue(
&mut self,
request: impl IntoRequest<GetQueueRequest>
) -> Result<Response<Queue>, Status>

Gets a queue.

pub async fn create_queue(
&mut self,
request: impl IntoRequest<CreateQueueRequest>
) -> Result<Response<Queue>, Status>

Creates a queue.

Queues created with this method allow tasks to live for a maximum of 31 days. After a task is 31 days old, the task will be deleted regardless of whether it was dispatched or not.

WARNING: Using this method may have unintended side effects if you are using an App Engine queue.yaml or queue.xml file to manage your queues. Read Overview of Queue Management and queue.yaml before using this method.

pub async fn update_queue(
&mut self,
request: impl IntoRequest<UpdateQueueRequest>
) -> Result<Response<Queue>, Status>

Updates a queue.

This method creates the queue if it does not exist and updates the queue if it does exist.

Queues created with this method allow tasks to live for a maximum of 31 days. After a task is 31 days old, the task will be deleted regardless of whether it was dispatched or not.

WARNING: Using this method may have unintended side effects if you are using an App Engine queue.yaml or queue.xml file to manage your queues. Read Overview of Queue Management and queue.yaml before using this method.

pub async fn delete_queue(
&mut self,
request: impl IntoRequest<DeleteQueueRequest>
) -> Result<Response<()>, Status>

Deletes a queue.

This command will delete the queue even if it has tasks in it.

Note: If you delete a queue, a queue with the same name can’t be created for 7 days.

WARNING: Using this method may have unintended side effects if you are using an App Engine queue.yaml or queue.xml file to manage your queues. Read Overview of Queue Management and queue.yaml before using this method.

pub async fn purge_queue(
&mut self,
request: impl IntoRequest<PurgeQueueRequest>
) -> Result<Response<Queue>, Status>

Purges a queue by deleting all of its tasks.

All tasks created before this method is called are permanently deleted.

Purge operations can take up to one minute to take effect. Tasks might be dispatched before the purge takes effect. A purge is irreversible.

pub async fn pause_queue(
&mut self,
request: impl IntoRequest<PauseQueueRequest>
) -> Result<Response<Queue>, Status>

Pauses the queue.

If a queue is paused then the system will stop dispatching tasks until the queue is resumed via [ResumeQueue][google.cloud.tasks.v2beta3.CloudTasks.ResumeQueue]. Tasks can still be added when the queue is paused. A queue is paused if its [state][google.cloud.tasks.v2beta3.Queue.state] is [PAUSED][google.cloud.tasks.v2beta3.Queue.State.PAUSED].

pub async fn resume_queue(
&mut self,
request: impl IntoRequest<ResumeQueueRequest>
) -> Result<Response<Queue>, Status>

Resume a queue.

This method resumes a queue after it has been [PAUSED][google.cloud.tasks.v2beta3.Queue.State.PAUSED] or [DISABLED][google.cloud.tasks.v2beta3.Queue.State.DISABLED]. The state of a queue is stored in the queue’s [state][google.cloud.tasks.v2beta3.Queue.state]; after calling this method it will be set to [RUNNING][google.cloud.tasks.v2beta3.Queue.State.RUNNING].

WARNING: Resuming many high-QPS queues at the same time can lead to target overloading. If you are resuming high-QPS queues, follow the 500/50/5 pattern described in Managing Cloud Tasks Scaling Risks.

pub async fn get_iam_policy(
&mut self,
request: impl IntoRequest<GetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Gets the access control policy for a [Queue][google.cloud.tasks.v2beta3.Queue]. Returns an empty policy if the resource exists and does not have a policy set.

Authorization requires the following Google IAM permission on the specified resource parent:

cloudtasks.queues.getIamPolicy

pub async fn set_iam_policy(
&mut self,
request: impl IntoRequest<SetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Sets the access control policy for a [Queue][google.cloud.tasks.v2beta3.Queue]. Replaces any existing policy.

Note: The Cloud Console does not check queue-level IAM permissions yet. Project-level permissions are required to use the Cloud Console.

Authorization requires the following Google IAM permission on the specified resource parent:

cloudtasks.queues.setIamPolicy

pub async fn test_iam_permissions(
&mut self,
request: impl IntoRequest<TestIamPermissionsRequest>
) -> Result<Response<TestIamPermissionsResponse>, Status>

Returns permissions that a caller has on a [Queue][google.cloud.tasks.v2beta3.Queue]. If the resource does not exist, this will return an empty set of permissions, not a [NOT_FOUND][google.rpc.Code.NOT_FOUND] error.

Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may “fail open” without warning.

pub async fn list_tasks(
&mut self,
request: impl IntoRequest<ListTasksRequest>
) -> Result<Response<ListTasksResponse>, Status>

Lists the tasks in a queue.

By default, only the [BASIC][google.cloud.tasks.v2beta3.Task.View.BASIC] view is retrieved due to performance considerations; [response_view][google.cloud.tasks.v2beta3.ListTasksRequest.response_view] controls the subset of information which is returned.

The tasks may be returned in any order. The ordering may change at any time.

pub async fn get_task(
&mut self,
request: impl IntoRequest<GetTaskRequest>
) -> Result<Response<Task>, Status>

Gets a task.

pub async fn create_task(
&mut self,
request: impl IntoRequest<CreateTaskRequest>
) -> Result<Response<Task>, Status>

Creates a task and adds it to a queue.

Tasks cannot be updated after creation; there is no UpdateTask command.

The maximum task size is 100KB.

pub async fn delete_task(
&mut self,
request: impl IntoRequest<DeleteTaskRequest>
) -> Result<Response<()>, Status>

Deletes a task.

A task can be deleted if it is scheduled or dispatched. A task cannot be deleted if it has executed successfully or permanently failed.

pub async fn run_task(
&mut self,
request: impl IntoRequest<RunTaskRequest>
) -> Result<Response<Task>, Status>

Forces a task to run now.

When this method is called, Cloud Tasks will dispatch the task, even if the task is already running, the queue has reached its [RateLimits][google.cloud.tasks.v2beta3.RateLimits] or is [PAUSED][google.cloud.tasks.v2beta3.Queue.State.PAUSED].

This command is meant to be used for manual debugging. For example, [RunTask][google.cloud.tasks.v2beta3.CloudTasks.RunTask] can be used to retry a failed task after a fix has been made or to manually force a task to be dispatched now.

The dispatched task is returned. That is, the task that is returned contains the [status][Task.status] after the task is dispatched but before the task is received by its target.

If Cloud Tasks receives a successful response from the task’s target, then the task will be deleted; otherwise the task’s [schedule_time][google.cloud.tasks.v2beta3.Task.schedule_time] will be reset to the time that [RunTask][google.cloud.tasks.v2beta3.CloudTasks.RunTask] was called plus the retry delay specified in the queue’s [RetryConfig][google.cloud.tasks.v2beta3.RetryConfig].

[RunTask][google.cloud.tasks.v2beta3.CloudTasks.RunTask] returns [NOT_FOUND][google.rpc.Code.NOT_FOUND] when it is called on a task that has already succeeded or permanently failed.