Struct googapis::google::cloud::tpu::v2alpha1::Node

pub struct Node {

    pub name: String,
    pub description: String,
    pub accelerator_type: String,
    pub state: i32,
    pub health_description: String,
    pub runtime_version: String,
    pub network_config: Option<NetworkConfig>,
    pub cidr_block: String,
    pub service_account: Option<ServiceAccount>,
    pub create_time: Option<Timestamp>,
    pub scheduling_config: Option<SchedulingConfig>,
    pub network_endpoints: Vec<NetworkEndpoint>,
    pub health: i32,
    pub labels: HashMap<String, String>,
    pub metadata: HashMap<String, String>,
    pub tags: Vec<String>,
    pub id: i64,
    pub data_disks: Vec<AttachedDisk>,
    pub api_version: i32,
    pub symptoms: Vec<Symptom>,

}

A TPU instance.

Fields
name: String

Output only. Immutable. The name of the TPU.

description: String

The user-supplied description of the TPU. Maximum of 512 characters.

accelerator_type: String

Required. The type of hardware accelerators associated with this node.

state: i32

Output only. The current state for the TPU Node.

health_description: String

Output only. If this field is populated, it contains a description of why the TPU Node is unhealthy.

runtime_version: String

Required. The runtime version running in the Node.

network_config: Option<NetworkConfig>

Network configurations for the TPU node.

cidr_block: String

The CIDR block that the TPU node will use when selecting an IP address. This CIDR block must be a /29 block; the Compute Engine networks API forbids a smaller block, and using a larger block would be wasteful (a node can only consume one IP address). Errors will occur if the CIDR block has already been used for a currently existing TPU node, the CIDR block conflicts with any subnetworks in the user’s provided network, or the provided network is peered with another network that is using that CIDR block.

service_account: Option<ServiceAccount>

The Google Cloud Platform Service Account to be used by the TPU node VMs. If None is specified, the default compute service account will be used.

create_time: Option<Timestamp>

Output only. The time when the node was created.

scheduling_config: Option<SchedulingConfig>

The scheduling options for this node.

network_endpoints: Vec<NetworkEndpoint>

Output only. The network endpoints where TPU workers can be accessed and sent work. It is recommended that runtime clients of the node reach out to the 0th entry in this map first.

health: i32

The health status of the TPU node.

labels: HashMap<String, String>

Resource labels to represent user-provided metadata.

metadata: HashMap<String, String>

Custom metadata to apply to the TPU Node. Can set startup-script and shutdown-script

tags: Vec<String>

Tags to apply to the TPU Node. Tags are used to identify valid sources or targets for network firewalls.

id: i64

Output only. The unique identifier for the TPU Node.

data_disks: Vec<AttachedDisk>

The additional data disks for the Node.

api_version: i32

Output only. The API version that created this Node.

symptoms: Vec<Symptom>

Output only. The Symptoms that have occurred to the TPU Node.

Implementations

impl Node

pub fn state(&self) -> State

Returns the enum value of state, or the default if the field is set to an invalid enum value.

pub fn set_state(&mut self, value: State)

Sets state to the provided enum value.

pub fn health(&self) -> Health

Returns the enum value of health, or the default if the field is set to an invalid enum value.

pub fn set_health(&mut self, value: Health)

Sets health to the provided enum value.

pub fn api_version(&self) -> ApiVersion

Returns the enum value of api_version, or the default if the field is set to an invalid enum value.

pub fn set_api_version(&mut self, value: ApiVersion)

Sets api_version to the provided enum value.