Module googapis::google::cloud::translation::v3beta1

Modules

batch_document_input_config
Nested message and enum types in BatchDocumentInputConfig.

batch_document_output_config
Nested message and enum types in BatchDocumentOutputConfig.

batch_translate_document_metadata
Nested message and enum types in BatchTranslateDocumentMetadata.

batch_translate_metadata
Nested message and enum types in BatchTranslateMetadata.

create_glossary_metadata
Nested message and enum types in CreateGlossaryMetadata.

delete_glossary_metadata
Nested message and enum types in DeleteGlossaryMetadata.

detect_language_request
Nested message and enum types in DetectLanguageRequest.

document_input_config
Nested message and enum types in DocumentInputConfig.

document_output_config
Nested message and enum types in DocumentOutputConfig.

glossary
Nested message and enum types in Glossary.

glossary_input_config
Nested message and enum types in GlossaryInputConfig.

input_config
Nested message and enum types in InputConfig.

output_config
Nested message and enum types in OutputConfig.

translation_service_client
Generated client implementations.

Structs

BatchDocumentInputConfig
Input configuration for BatchTranslateDocument request.

BatchDocumentOutputConfig
Output configuration for BatchTranslateDocument request.

BatchTranslateDocumentMetadata
State metadata for the batch translation operation.

BatchTranslateDocumentRequest
The BatchTranslateDocument request.

BatchTranslateDocumentResponse
Stored in the [google.longrunning.Operation.response][google.longrunning.Operation.response] field returned by BatchTranslateDocument if at least one document is translated successfully.

BatchTranslateMetadata
State metadata for the batch translation operation.

BatchTranslateResponse
Stored in the [google.longrunning.Operation.response][google.longrunning.Operation.response] field returned by BatchTranslateText if at least one sentence is translated successfully.

BatchTranslateTextRequest
The batch translation request.

CreateGlossaryMetadata
Stored in the [google.longrunning.Operation.metadata][google.longrunning.Operation.metadata] field returned by CreateGlossary.

CreateGlossaryRequest
Request message for CreateGlossary.

DeleteGlossaryMetadata
Stored in the [google.longrunning.Operation.metadata][google.longrunning.Operation.metadata] field returned by DeleteGlossary.

DeleteGlossaryRequest
Request message for DeleteGlossary.

DeleteGlossaryResponse
Stored in the [google.longrunning.Operation.response][google.longrunning.Operation.response] field returned by DeleteGlossary.

DetectLanguageRequest
The request message for language detection.

DetectLanguageResponse
The response message for language detection.

DetectedLanguage
The response message for language detection.

DocumentInputConfig
A document translation request input config.

DocumentOutputConfig
A document translation request output config.

DocumentTranslation
A translated document message.

GcsDestination
The Google Cloud Storage location for the output content.

GcsSource
The Google Cloud Storage location for the input content.

GetGlossaryRequest
Request message for GetGlossary.

GetSupportedLanguagesRequest
The request message for discovering supported languages.

Glossary
Represents a glossary built from user provided data.

GlossaryInputConfig
Input configuration for glossaries.

InputConfig
Input configuration for BatchTranslateText request.

ListGlossariesRequest
Request message for ListGlossaries.

ListGlossariesResponse
Response message for ListGlossaries.

OutputConfig
Output configuration for BatchTranslateText request.

SupportedLanguage
A single supported language response corresponds to information related to one supported language.

SupportedLanguages
The response message for discovering supported languages.

TranslateDocumentRequest
A document translation request.

TranslateDocumentResponse
A translated document response message.

TranslateTextGlossaryConfig
Configures which glossary should be used for a specific target language, and defines options for applying that glossary.

TranslateTextRequest
The request message for synchronous translation.

TranslateTextResponse

Translation
A single translation response.