Module googapis::google::cloud::video::transcoder::v1beta1

Modules

audio_stream

Nested message and enum types in AudioStream.

elementary_stream

Nested message and enum types in ElementaryStream.

encryption

Nested message and enum types in Encryption.

job

Nested message and enum types in Job.

manifest

Nested message and enum types in Manifest.

overlay

Nested message and enum types in Overlay.

preprocessing_config

Nested message and enum types in PreprocessingConfig.

sprite_sheet

Nested message and enum types in SpriteSheet.

text_stream

Nested message and enum types in TextStream.

transcoder_service_client

Generated client implementations.

video_stream

Nested message and enum types in VideoStream.

Structs

AdBreak

Ad break.

AudioStream

Audio stream resource.

CreateJobRequest

Request message for TranscoderService.CreateJob.

CreateJobTemplateRequest

Request message for TranscoderService.CreateJobTemplate.

DeleteJobRequest

Request message for TranscoderService.DeleteJob.

DeleteJobTemplateRequest

Request message for TranscoderService.DeleteJobTemplate.

EditAtom

Edit atom.

ElementaryStream

Encoding of an input file such as an audio, video, or text track. Elementary streams must be packaged before mapping and sharing between different output formats.

Encryption

Encryption settings.

FailureDetail

Additional information about the reasons for the failure.

GetJobRequest

Request message for TranscoderService.GetJob.

GetJobTemplateRequest

Request message for TranscoderService.GetJobTemplate.

Input

Input asset.

Job

Transcoding job resource.

JobConfig

Job configuration

JobTemplate

Transcoding job template resource.

ListJobTemplatesRequest

Request message for TranscoderService.ListJobTemplates.

ListJobTemplatesResponse

Response message for TranscoderService.ListJobTemplates.

ListJobsRequest

Request message for TranscoderService.ListJobs. The parent location from which to retrieve the collection of jobs.

ListJobsResponse

Response message for TranscoderService.ListJobs.

Manifest

Manifest configuration.

MuxStream

Multiplexing settings for output stream.

Output

Location of output file(s) in a Cloud Storage bucket.

Overlay

Overlay configuration.

PreprocessingConfig

Preprocessing configurations.

Progress

Estimated fractional progress for each step, from 0 to 1.

PubsubDestination

A Pub/Sub destination.

SegmentSettings

Segment settings for "ts", "fmp4" and "vtt".

SpriteSheet

Sprite sheet configuration.

TextStream

Encoding of a text stream. For example, closed captions or subtitles.

VideoStream

Video stream resource.