Enum googapis::google::cloud::vision::v1p2beta1::face_annotation::landmark::Type

#[repr(i32)]
pub enum Type {

    UnknownLandmark,
    LeftEye,
    RightEye,
    LeftOfLeftEyebrow,
    RightOfLeftEyebrow,
    LeftOfRightEyebrow,
    RightOfRightEyebrow,
    MidpointBetweenEyes,
    NoseTip,
    UpperLip,
    LowerLip,
    MouthLeft,
    MouthRight,
    MouthCenter,
    NoseBottomRight,
    NoseBottomLeft,
    NoseBottomCenter,
    LeftEyeTopBoundary,
    LeftEyeRightCorner,
    LeftEyeBottomBoundary,
    LeftEyeLeftCorner,
    RightEyeTopBoundary,
    RightEyeRightCorner,
    RightEyeBottomBoundary,
    RightEyeLeftCorner,
    LeftEyebrowUpperMidpoint,
    RightEyebrowUpperMidpoint,
    LeftEarTragion,
    RightEarTragion,
    LeftEyePupil,
    RightEyePupil,
    ForeheadGlabella,
    ChinGnathion,
    ChinLeftGonion,
    ChinRightGonion,

}

Face landmark (feature) type. Left and right are defined from the vantage of the viewer of the image without considering mirror projections typical of photos. So, LEFT_EYE, typically, is the person’s right eye.

Variants

UnknownLandmark

Unknown face landmark detected. Should not be filled.

LeftEye

Left eye.

RightEye

Right eye.

LeftOfLeftEyebrow

Left of left eyebrow.

RightOfLeftEyebrow

Right of left eyebrow.

LeftOfRightEyebrow

Left of right eyebrow.

RightOfRightEyebrow

Right of right eyebrow.

MidpointBetweenEyes

Midpoint between eyes.

NoseTip

Nose tip.

UpperLip

Upper lip.

LowerLip

Lower lip.

MouthLeft

Mouth left.

MouthRight

Mouth right.

MouthCenter

Mouth center.

NoseBottomRight

Nose, bottom right.

NoseBottomLeft

Nose, bottom left.

NoseBottomCenter

Nose, bottom center.

LeftEyeTopBoundary

Left eye, top boundary.

LeftEyeRightCorner

Left eye, right corner.

LeftEyeBottomBoundary

Left eye, bottom boundary.

LeftEyeLeftCorner

Left eye, left corner.

RightEyeTopBoundary

Right eye, top boundary.

RightEyeRightCorner

Right eye, right corner.

RightEyeBottomBoundary

Right eye, bottom boundary.

RightEyeLeftCorner

Right eye, left corner.

LeftEyebrowUpperMidpoint

Left eyebrow, upper midpoint.

RightEyebrowUpperMidpoint

Right eyebrow, upper midpoint.

LeftEarTragion

Left ear tragion.

RightEarTragion

Right ear tragion.

LeftEyePupil

Left eye pupil.

RightEyePupil

Right eye pupil.

ForeheadGlabella

Forehead glabella.

ChinGnathion

Chin gnathion.

ChinLeftGonion

Chin left gonion.

ChinRightGonion

Chin right gonion.

Implementations

impl Type

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Type.

pub fn from_i32(value: i32) -> Option<Type>

Converts an i32 to a Type, or None if value is not a valid variant.