Struct googapis::google::container::v1::Cluster

pub struct Cluster {

    pub name: String,
    pub description: String,
    pub initial_node_count: i32,
    pub node_config: Option<NodeConfig>,
    pub master_auth: Option<MasterAuth>,
    pub logging_service: String,
    pub monitoring_service: String,
    pub network: String,
    pub cluster_ipv4_cidr: String,
    pub addons_config: Option<AddonsConfig>,
    pub subnetwork: String,
    pub node_pools: Vec<NodePool>,
    pub locations: Vec<String>,
    pub enable_kubernetes_alpha: bool,
    pub resource_labels: HashMap<String, String>,
    pub label_fingerprint: String,
    pub legacy_abac: Option<LegacyAbac>,
    pub network_policy: Option<NetworkPolicy>,
    pub ip_allocation_policy: Option<IpAllocationPolicy>,
    pub master_authorized_networks_config: Option<MasterAuthorizedNetworksConfig>,
    pub maintenance_policy: Option<MaintenancePolicy>,
    pub binary_authorization: Option<BinaryAuthorization>,
    pub autoscaling: Option<ClusterAutoscaling>,
    pub network_config: Option<NetworkConfig>,
    pub default_max_pods_constraint: Option<MaxPodsConstraint>,
    pub resource_usage_export_config: Option<ResourceUsageExportConfig>,
    pub authenticator_groups_config: Option<AuthenticatorGroupsConfig>,
    pub private_cluster_config: Option<PrivateClusterConfig>,
    pub database_encryption: Option<DatabaseEncryption>,
    pub vertical_pod_autoscaling: Option<VerticalPodAutoscaling>,
    pub shielded_nodes: Option<ShieldedNodes>,
    pub release_channel: Option<ReleaseChannel>,
    pub workload_identity_config: Option<WorkloadIdentityConfig>,
    pub self_link: String,
    pub zone: String,
    pub endpoint: String,
    pub initial_cluster_version: String,
    pub current_master_version: String,
    pub current_node_version: String,
    pub create_time: String,
    pub status: i32,
    pub status_message: String,
    pub node_ipv4_cidr_size: i32,
    pub services_ipv4_cidr: String,
    pub instance_group_urls: Vec<String>,
    pub current_node_count: i32,
    pub expire_time: String,
    pub location: String,
    pub enable_tpu: bool,
    pub tpu_ipv4_cidr_block: String,
    pub conditions: Vec<StatusCondition>,

}

A Google Kubernetes Engine cluster.

Fields
name: String

The name of this cluster. The name must be unique within this project and location (e.g. zone or region), and can be up to 40 characters with the following restrictions:

Lowercase letters, numbers, and hyphens only.

Must start with a letter.

Must end with a number or a letter.

description: String

An optional description of this cluster.

initial_node_count: i32

👎 Deprecated

The number of nodes to create in this cluster. You must ensure that your Compute Engine resource quota is sufficient for this number of instances. You must also have available firewall and routes quota. For requests, this field should only be used in lieu of a “node_pool” object, since this configuration (along with the “node_config”) will be used to create a “NodePool” object with an auto-generated name. Do not use this and a node_pool at the same time.

This field is deprecated, use node_pool.initial_node_count instead.

node_config: Option<NodeConfig>

👎 Deprecated

Parameters used in creating the cluster’s nodes. For requests, this field should only be used in lieu of a “node_pool” object, since this configuration (along with the “initial_node_count”) will be used to create a “NodePool” object with an auto-generated name. Do not use this and a node_pool at the same time. For responses, this field will be populated with the node configuration of the first node pool. (For configuration of each node pool, see node_pool.config)

If unspecified, the defaults are used. This field is deprecated, use node_pool.config instead.

master_auth: Option<MasterAuth>

The authentication information for accessing the master endpoint. If unspecified, the defaults are used: For clusters before v1.12, if master_auth is unspecified, username will be set to “admin”, a random password will be generated, and a client certificate will be issued.

logging_service: String

The logging service the cluster should use to write logs. Currently available options:

logging.googleapis.com/kubernetes - The Cloud Logging service with a Kubernetes-native resource model

logging.googleapis.com - The legacy Cloud Logging service (no longer available as of GKE 1.15).

none - no logs will be exported from the cluster.

If left as an empty string,logging.googleapis.com/kubernetes will be used for GKE 1.14+ or logging.googleapis.com for earlier versions.

monitoring_service: String

The monitoring service the cluster should use to write metrics. Currently available options:

“monitoring.googleapis.com/kubernetes” - The Cloud Monitoring service with a Kubernetes-native resource model

monitoring.googleapis.com - The legacy Cloud Monitoring service (no longer available as of GKE 1.15).

none - No metrics will be exported from the cluster.

If left as an empty string,monitoring.googleapis.com/kubernetes will be used for GKE 1.14+ or monitoring.googleapis.com for earlier versions.

network: String

The name of the Google Compute Engine [network](https://cloud.google.com/compute/docs/networks-and-firewalls#networks) to which the cluster is connected. If left unspecified, the default network will be used.

cluster_ipv4_cidr: String

The IP address range of the container pods in this cluster, in [CIDR](http://en.wikipedia.org/wiki/Classless_Inter-Domain_Routing) notation (e.g. 10.96.0.0/14). Leave blank to have one automatically chosen or specify a /14 block in 10.0.0.0/8.

addons_config: Option<AddonsConfig>

Configurations for the various addons available to run in the cluster.

subnetwork: String

The name of the Google Compute Engine [subnetwork](https://cloud.google.com/compute/docs/subnetworks) to which the cluster is connected.

node_pools: Vec<NodePool>

The node pools associated with this cluster. This field should not be set if “node_config” or “initial_node_count” are specified.

locations: Vec<String>

The list of Google Compute Engine [zones](https://cloud.google.com/compute/docs/zones#available) in which the cluster’s nodes should be located.

This field provides a default value if [NodePool.Locations](https://cloud.google.com/kubernetes-engine/docs/reference/rest/v1/projects.locations.clusters.nodePools#NodePool.FIELDS.locations) are not specified during node pool creation.

Warning: changing cluster locations will update the [NodePool.Locations](https://cloud.google.com/kubernetes-engine/docs/reference/rest/v1/projects.locations.clusters.nodePools#NodePool.FIELDS.locations) of all node pools and will result in nodes being added and/or removed.

enable_kubernetes_alpha: bool

Kubernetes alpha features are enabled on this cluster. This includes alpha API groups (e.g. v1alpha1) and features that may not be production ready in the kubernetes version of the master and nodes. The cluster has no SLA for uptime and master/node upgrades are disabled. Alpha enabled clusters are automatically deleted thirty days after creation.

resource_labels: HashMap<String, String>

The resource labels for the cluster to use to annotate any related Google Compute Engine resources.

label_fingerprint: String

The fingerprint of the set of labels for this cluster.

legacy_abac: Option<LegacyAbac>

Configuration for the legacy ABAC authorization mode.

network_policy: Option<NetworkPolicy>

Configuration options for the NetworkPolicy feature.

ip_allocation_policy: Option<IpAllocationPolicy>

Configuration for cluster IP allocation.

master_authorized_networks_config: Option<MasterAuthorizedNetworksConfig>

The configuration options for master authorized networks feature.

maintenance_policy: Option<MaintenancePolicy>

Configure the maintenance policy for this cluster.

binary_authorization: Option<BinaryAuthorization>

Configuration for Binary Authorization.

autoscaling: Option<ClusterAutoscaling>

Cluster-level autoscaling configuration.

network_config: Option<NetworkConfig>

Configuration for cluster networking.

default_max_pods_constraint: Option<MaxPodsConstraint>

The default constraint on the maximum number of pods that can be run simultaneously on a node in the node pool of this cluster. Only honored if cluster created with IP Alias support.

resource_usage_export_config: Option<ResourceUsageExportConfig>

Configuration for exporting resource usages. Resource usage export is disabled when this config is unspecified.

authenticator_groups_config: Option<AuthenticatorGroupsConfig>

Configuration controlling RBAC group membership information.

private_cluster_config: Option<PrivateClusterConfig>

Configuration for private cluster.

database_encryption: Option<DatabaseEncryption>

Configuration of etcd encryption.

vertical_pod_autoscaling: Option<VerticalPodAutoscaling>

Cluster-level Vertical Pod Autoscaling configuration.

shielded_nodes: Option<ShieldedNodes>

Shielded Nodes configuration.

release_channel: Option<ReleaseChannel>

Release channel configuration.

workload_identity_config: Option<WorkloadIdentityConfig>

Configuration for the use of Kubernetes Service Accounts in GCP IAM policies.

self_link: String

[Output only] Server-defined URL for the resource.

zone: String

👎 Deprecated

[Output only] The name of the Google Compute Engine [zone](https://cloud.google.com/compute/docs/zones#available) in which the cluster resides. This field is deprecated, use location instead.

endpoint: String

[Output only] The IP address of this cluster’s master endpoint. The endpoint can be accessed from the internet at <https://username:password@endpoint/.>

See the masterAuth property of this resource for username and password information.

initial_cluster_version: String

The initial Kubernetes version for this cluster. Valid versions are those found in validMasterVersions returned by getServerConfig. The version can be upgraded over time; such upgrades are reflected in currentMasterVersion and currentNodeVersion.

Users may specify either explicit versions offered by Kubernetes Engine or version aliases, which have the following behavior:

“latest”: picks the highest valid Kubernetes version

“1.X”: picks the highest valid patch+gke.N patch in the 1.X version

“1.X.Y”: picks the highest valid gke.N patch in the 1.X.Y version

“1.X.Y-gke.N”: picks an explicit Kubernetes version

“”,“-”: picks the default Kubernetes version

current_master_version: String

[Output only] The current software version of the master endpoint.

current_node_version: String

👎 Deprecated

[Output only] Deprecated, use [NodePools.version](https://cloud.google.com/kubernetes-engine/docs/reference/rest/v1/projects.locations.clusters.nodePools) instead. The current version of the node software components. If they are currently at multiple versions because they’re in the process of being upgraded, this reflects the minimum version of all nodes.

create_time: String

[Output only] The time the cluster was created, in [RFC3339](https://www.ietf.org/rfc/rfc3339.txt) text format.

status: i32

[Output only] The current status of this cluster.

status_message: String

👎 Deprecated

[Output only] Deprecated. Use conditions instead. Additional information about the current status of this cluster, if available.

node_ipv4_cidr_size: i32

[Output only] The size of the address space on each node for hosting containers. This is provisioned from within the container_ipv4_cidr range. This field will only be set when cluster is in route-based network mode.

services_ipv4_cidr: String

[Output only] The IP address range of the Kubernetes services in this cluster, in [CIDR](http://en.wikipedia.org/wiki/Classless_Inter-Domain_Routing) notation (e.g. 1.2.3.4/29). Service addresses are typically put in the last /16 from the container CIDR.

instance_group_urls: Vec<String>

👎 Deprecated

Deprecated. Use node_pools.instance_group_urls.

current_node_count: i32

👎 Deprecated

[Output only] The number of nodes currently in the cluster. Deprecated. Call Kubernetes API directly to retrieve node information.

expire_time: String

[Output only] The time the cluster will be automatically deleted in [RFC3339](https://www.ietf.org/rfc/rfc3339.txt) text format.

location: String

[Output only] The name of the Google Compute Engine [zone](https://cloud.google.com/compute/docs/regions-zones/regions-zones#available) or [region](https://cloud.google.com/compute/docs/regions-zones/regions-zones#available) in which the cluster resides.

enable_tpu: bool

Enable the ability to use Cloud TPUs in this cluster.

tpu_ipv4_cidr_block: String

[Output only] The IP address range of the Cloud TPUs in this cluster, in [CIDR](http://en.wikipedia.org/wiki/Classless_Inter-Domain_Routing) notation (e.g. 1.2.3.4/29).

conditions: Vec<StatusCondition>

Which conditions caused the current cluster state.

Implementations

impl Cluster

pub fn status(&self) -> Status

Returns the enum value of status, or the default if the field is set to an invalid enum value.

pub fn set_status(&mut self, value: Status)

Sets status to the provided enum value.