Module googapis::google::container::v1alpha1

Modules

cluster
Nested message and enum types in Cluster.

cluster_manager_client
Generated client implementations.

maintenance_window
Nested message and enum types in MaintenanceWindow.

master_authorized_networks_config
Nested message and enum types in MasterAuthorizedNetworksConfig.

network_policy
Nested message and enum types in NetworkPolicy.

node_pool
Nested message and enum types in NodePool.

node_taint
Nested message and enum types in NodeTaint.

operation
Nested message and enum types in Operation.

set_master_auth_request
Nested message and enum types in SetMasterAuthRequest.

Structs

AcceleratorConfig
AcceleratorConfig represents a Hardware Accelerator request.

AddonsConfig
Configuration for the addons that can be automatically spun up in the cluster, enabling additional functionality.

AutoUpgradeOptions
AutoUpgradeOptions defines the set of options for the user to control how the Auto Upgrades will proceed.

CancelOperationRequest
CancelOperationRequest cancels a single operation.

ClientCertificateConfig
Configuration for client certificates on the cluster.

Cluster
A Google Container Engine cluster.

ClusterUpdate
ClusterUpdate describes an update to the cluster. Exactly one update can be applied to a cluster with each request, so at most one field can be provided.

CompleteIpRotationRequest
CompleteIPRotationRequest moves the cluster master back into single-IP mode.

CreateClusterRequest
CreateClusterRequest creates a cluster.

CreateNodePoolRequest
CreateNodePoolRequest creates a node pool for a cluster.

DailyMaintenanceWindow
Time window specified for daily maintenance operations.

DeleteClusterRequest
DeleteClusterRequest deletes a cluster.

DeleteNodePoolRequest
DeleteNodePoolRequest deletes a node pool for a cluster.

GetClusterRequest
GetClusterRequest gets the settings of a cluster.

GetNodePoolRequest
GetNodePoolRequest retrieves a node pool for a cluster.

GetOperationRequest
GetOperationRequest gets a single operation.

GetServerConfigRequest
Gets the current Container Engine service configuration.

HorizontalPodAutoscaling
Configuration options for the horizontal pod autoscaling feature, which increases or decreases the number of replica pods a replication controller has based on the resource usage of the existing pods.

HttpLoadBalancing
Configuration options for the HTTP (L7) load balancing controller addon, which makes it easy to set up HTTP load balancers for services in a cluster.

IpAllocationPolicy
Configuration for controlling how IPs are allocated in the cluster.

KubernetesDashboard
Configuration for the Kubernetes Dashboard.

ListClustersRequest
ListClustersRequest lists clusters.

ListClustersResponse
ListClustersResponse is the result of ListClustersRequest.

ListNodePoolsRequest
ListNodePoolsRequest lists the node pool(s) for a cluster.

ListNodePoolsResponse
ListNodePoolsResponse is the result of ListNodePoolsRequest.

ListOperationsRequest
ListOperationsRequest lists operations.

ListOperationsResponse
ListOperationsResponse is the result of ListOperationsRequest.

MaintenancePolicy
MaintenancePolicy defines the maintenance policy to be used for the cluster.

MaintenanceWindow
MaintenanceWindow defines the maintenance window to be used for the cluster.

MasterAuth
The authentication information for accessing the master endpoint. Authentication can be done using HTTP basic auth or using client certificates.

MasterAuthorizedNetworksConfig
Configuration options for the master authorized networks feature. Enabled master authorized networks will disallow all external traffic to access Kubernetes master through HTTPS except traffic from the given CIDR blocks, Google Compute Engine Public IPs and Google Prod IPs.

NetworkPolicy
Configuration options for the NetworkPolicy feature. https://kubernetes.io/docs/concepts/services-networking/networkpolicies/

NetworkPolicyConfig
Configuration for NetworkPolicy. This only tracks whether the addon is enabled or not on the Master, it does not track whether network policy is enabled for the nodes.

NodeConfig
Parameters that describe the nodes in a cluster.

NodeManagement
NodeManagement defines the set of node management services turned on for the node pool.

NodePool
NodePool contains the name and configuration for a cluster’s node pool. Node pools are a set of nodes (i.e. VM’s), with a common configuration and specification, under the control of the cluster master. They may have a set of Kubernetes labels applied to them, which may be used to reference them during pod scheduling. They may also be resized up or down, to accommodate the workload.

NodePoolAutoscaling
NodePoolAutoscaling contains information required by cluster autoscaler to adjust the size of the node pool to the current cluster usage.

NodeTaint
Kubernetes taint is comprised of three fields: key, value, and effect. Effect can only be one of three types: NoSchedule, PreferNoSchedule or NoExecute.

Operation
This operation resource represents operations that may have happened or are happening on the cluster. All fields are output only.

PodSecurityPolicyConfig
Configuration for the PodSecurityPolicy feature.

RollbackNodePoolUpgradeRequest
RollbackNodePoolUpgradeRequest rollbacks the previously Aborted or Failed NodePool upgrade. This will be an no-op if the last upgrade successfully completed.

ServerConfig
Container Engine service configuration.

SetAddonsConfigRequest
SetAddonsRequest sets the addons associated with the cluster.

SetLabelsRequest
SetLabelsRequest sets the Google Cloud Platform labels on a Google Container Engine cluster, which will in turn set them for Google Compute Engine resources used by that cluster

SetLegacyAbacRequest
SetLegacyAbacRequest enables or disables the ABAC authorization mechanism for a cluster.

SetLocationsRequest
SetLocationsRequest sets the locations of the cluster.

SetLoggingServiceRequest
SetLoggingServiceRequest sets the logging service of a cluster.

SetMaintenancePolicyRequest
SetMaintenancePolicyRequest sets the maintenance policy for a cluster.

SetMasterAuthRequest
SetMasterAuthRequest updates the admin password of a cluster.

SetMonitoringServiceRequest
SetMonitoringServiceRequest sets the monitoring service of a cluster.

SetNetworkPolicyRequest
SetNetworkPolicyRequest enables/disables network policy for a cluster.

SetNodePoolAutoscalingRequest
SetNodePoolAutoscalingRequest sets the autoscaler settings of a node pool.

SetNodePoolManagementRequest
SetNodePoolManagementRequest sets the node management properties of a node pool.

SetNodePoolSizeRequest
SetNodePoolSizeRequest sets the size a node pool.

StartIpRotationRequest
StartIPRotationRequest creates a new IP for the cluster and then performs a node upgrade on each node pool to point to the new IP.

UpdateClusterRequest
UpdateClusterRequest updates the settings of a cluster.

UpdateMasterRequest
UpdateMasterRequest updates the master of the cluster.

UpdateNodePoolRequest
SetNodePoolVersionRequest updates the version of a node pool.