Struct googapis::google::dataflow::v1beta3::Job

pub struct Job {

    pub id: String,
    pub project_id: String,
    pub name: String,
    pub type: i32,
    pub environment: Option<Environment>,
    pub steps: Vec<Step>,
    pub steps_location: String,
    pub current_state: i32,
    pub current_state_time: Option<Timestamp>,
    pub requested_state: i32,
    pub execution_info: Option<JobExecutionInfo>,
    pub create_time: Option<Timestamp>,
    pub replace_job_id: String,
    pub transform_name_mapping: HashMap<String, String>,
    pub client_request_id: String,
    pub replaced_by_job_id: String,
    pub temp_files: Vec<String>,
    pub labels: HashMap<String, String>,
    pub location: String,
    pub pipeline_description: Option<PipelineDescription>,
    pub stage_states: Vec<ExecutionStageState>,
    pub job_metadata: Option<JobMetadata>,
    pub start_time: Option<Timestamp>,
    pub created_from_snapshot_id: String,
    pub satisfies_pzs: bool,

}

Defines a job to be run by the Cloud Dataflow service. nextID: 26

Fields
id: String

The unique ID of this job.

This field is set by the Cloud Dataflow service when the Job is created, and is immutable for the life of the job.

project_id: String

The ID of the Cloud Platform project that the job belongs to.

name: String

The user-specified Cloud Dataflow job name.

Only one Job with a given name may exist in a project at any given time. If a caller attempts to create a Job with the same name as an already-existing Job, the attempt returns the existing Job.

The name must match the regular expression [a-z]([-a-z0-9]{0,38}[a-z0-9])?

type: i32

The type of Cloud Dataflow job.

environment: Option<Environment>

The environment for the job.

steps: Vec<Step>

Exactly one of step or steps_location should be specified.

The top-level steps that constitute the entire job. Only retrieved with JOB_VIEW_ALL.

steps_location: String

The Cloud Storage location where the steps are stored.

current_state: i32

The current state of the job.

Jobs are created in the JOB_STATE_STOPPED state unless otherwise specified.

A job in the JOB_STATE_RUNNING state may asynchronously enter a terminal state. After a job has reached a terminal state, no further state updates may be made.

This field may be mutated by the Cloud Dataflow service; callers cannot mutate it.

current_state_time: Option<Timestamp>

The timestamp associated with the current state.

requested_state: i32

The job’s requested state.

UpdateJob may be used to switch between the JOB_STATE_STOPPED and JOB_STATE_RUNNING states, by setting requested_state. UpdateJob may also be used to directly set a job’s requested state to JOB_STATE_CANCELLED or JOB_STATE_DONE, irrevocably terminating the job if it has not already reached a terminal state.

execution_info: Option<JobExecutionInfo>

Deprecated.

create_time: Option<Timestamp>

The timestamp when the job was initially created. Immutable and set by the Cloud Dataflow service.

replace_job_id: String

If this job is an update of an existing job, this field is the job ID of the job it replaced.

When sending a CreateJobRequest, you can update a job by specifying it here. The job named here is stopped, and its intermediate state is transferred to this job.

transform_name_mapping: HashMap<String, String>

The map of transform name prefixes of the job to be replaced to the corresponding name prefixes of the new job.

client_request_id: String

The client’s unique identifier of the job, re-used across retried attempts. If this field is set, the service will ensure its uniqueness. The request to create a job will fail if the service has knowledge of a previously submitted job with the same client’s ID and job name. The caller may use this field to ensure idempotence of job creation across retried attempts to create a job. By default, the field is empty and, in that case, the service ignores it.

replaced_by_job_id: String

If another job is an update of this job (and thus, this job is in JOB_STATE_UPDATED), this field contains the ID of that job.

temp_files: Vec<String>

A set of files the system should be aware of that are used for temporary storage. These temporary files will be removed on job completion. No duplicates are allowed. No file patterns are supported.

The supported files are:

Google Cloud Storage:

storage.googleapis.com/{bucket}/{object} bucket.storage.googleapis.com/{object}

labels: HashMap<String, String>

User-defined labels for this job.

The labels map can contain no more than 64 entries. Entries of the labels map are UTF8 strings that comply with the following restrictions:

Keys must conform to regexp: [\p{Ll}\p{Lo}][\p{Ll}\p{Lo}\p{N}_-]{0,62}

Values must conform to regexp: [\p{Ll}\p{Lo}\p{N}_-]{0,63}

Both keys and values are additionally constrained to be <= 128 bytes in size.

location: String

The [regional endpoint] (https://cloud.google.com/dataflow/docs/concepts/regional-endpoints) that contains this job.

pipeline_description: Option<PipelineDescription>

Preliminary field: The format of this data may change at any time. A description of the user pipeline and stages through which it is executed. Created by Cloud Dataflow service. Only retrieved with JOB_VIEW_DESCRIPTION or JOB_VIEW_ALL.

stage_states: Vec<ExecutionStageState>

This field may be mutated by the Cloud Dataflow service; callers cannot mutate it.

job_metadata: Option<JobMetadata>

This field is populated by the Dataflow service to support filtering jobs by the metadata values provided here. Populated for ListJobs and all GetJob views SUMMARY and higher.

start_time: Option<Timestamp>

The timestamp when the job was started (transitioned to JOB_STATE_PENDING). Flexible resource scheduling jobs are started with some delay after job creation, so start_time is unset before start and is updated when the job is started by the Cloud Dataflow service. For other jobs, start_time always equals to create_time and is immutable and set by the Cloud Dataflow service.

created_from_snapshot_id: String

If this is specified, the job’s initial state is populated from the given snapshot.

satisfies_pzs: bool

Reserved for future use. This field is set only in responses from the server; it is ignored if it is set in any requests.

Implementations

impl Job

pub fn type(&self) -> JobType

Returns the enum value of type, or the default if the field is set to an invalid enum value.

pub fn set_type(&mut self, value: JobType)

Sets type to the provided enum value.

pub fn current_state(&self) -> JobState

Returns the enum value of current_state, or the default if the field is set to an invalid enum value.

pub fn set_current_state(&mut self, value: JobState)

Sets current_state to the provided enum value.

pub fn requested_state(&self) -> JobState

Returns the enum value of requested_state, or the default if the field is set to an invalid enum value.

pub fn set_requested_state(&mut self, value: JobState)

Sets requested_state to the provided enum value.