Module googapis::google::datastore::admin::v1beta1

Modules

common_metadata
Nested message and enum types in CommonMetadata.

datastore_admin_client
Generated client implementations.

Structs

CommonMetadata
Metadata common to all Datastore Admin operations.

EntityFilter
Identifies a subset of entities in a project. This is specified as combinations of kinds and namespaces (either or both of which may be all, as described in the following examples). Example usage:

ExportEntitiesMetadata
Metadata for ExportEntities operations.

ExportEntitiesRequest
The request for [google.datastore.admin.v1beta1.DatastoreAdmin.ExportEntities][google.datastore.admin.v1beta1.DatastoreAdmin.ExportEntities].

ExportEntitiesResponse
The response for [google.datastore.admin.v1beta1.DatastoreAdmin.ExportEntities][google.datastore.admin.v1beta1.DatastoreAdmin.ExportEntities].

ImportEntitiesMetadata
Metadata for ImportEntities operations.

ImportEntitiesRequest
The request for [google.datastore.admin.v1beta1.DatastoreAdmin.ImportEntities][google.datastore.admin.v1beta1.DatastoreAdmin.ImportEntities].

Progress
Measures the progress of a particular metric.

Enums

OperationType
Operation types.