Module googapis::google::devtools::cloudbuild::v1::build_options

Nested message and enum types in BuildOptions.

Structs

PoolOption

Details about how a build should be executed on a WorkerPool.

Enums

LogStreamingOption

Specifies the behavior when writing build logs to Google Cloud Storage.

LoggingMode

Specifies the logging mode.

MachineType

Supported Compute Engine machine types. For more information, see Machine types.

SubstitutionOption

Specifies the behavior when there is an error in the substitution checks.

VerifyOption

Specifies the manner in which the build should be verified, if at all.