Struct googapis::google::devtools::cloudbuild::v1::cloud_build_client::CloudBuildClient

pub struct CloudBuildClient<T> { /* fields omitted */ }

Creates and manages builds on Google Cloud Platform.

The main concept used by this API is a Build, which describes the location of the source to build, how to build the source, and where to store the built artifacts, if any.

A user can list previously-requested builds or get builds by their ID to determine the status of the build.

Implementations

impl<T> CloudBuildClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + Sync + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> CloudBuildClient<InterceptedService<T, F>> where
F: FnMut(Request<()>) -> Result<Request<()>, Status>,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_build(
&mut self,
request: impl IntoRequest<CreateBuildRequest>
) -> Result<Response<Operation>, Status>

Starts a build with the specified configuration.

This method returns a long-running Operation, which includes the build ID. Pass the build ID to GetBuild to determine the build status (such as SUCCESS or FAILURE).

pub async fn get_build(
&mut self,
request: impl IntoRequest<GetBuildRequest>
) -> Result<Response<Build>, Status>

Returns information about a previously requested build.

The Build that is returned includes its status (such as SUCCESS, FAILURE, or WORKING), and timing information.

pub async fn list_builds(
&mut self,
request: impl IntoRequest<ListBuildsRequest>
) -> Result<Response<ListBuildsResponse>, Status>

Lists previously requested builds.

Previously requested builds may still be in-progress, or may have finished successfully or unsuccessfully.

pub async fn cancel_build(
&mut self,
request: impl IntoRequest<CancelBuildRequest>
) -> Result<Response<Build>, Status>

Cancels a build in progress.

pub async fn retry_build(
&mut self,
request: impl IntoRequest<RetryBuildRequest>
) -> Result<Response<Operation>, Status>

Creates a new build based on the specified build.

This method creates a new build using the original build request, which may or may not result in an identical build.

For triggered builds:

Triggered builds resolve to a precise revision; therefore a retry of a triggered build will result in a build that uses the same revision.

For non-triggered builds that specify RepoSource:

If the original build built from the tip of a branch, the retried build will build from the tip of that branch, which may not be the same revision as the original build.

If the original build specified a commit sha or revision ID, the retried build will use the identical source.

For builds that specify StorageSource:

If the original build pulled source from Google Cloud Storage without specifying the generation of the object, the new build will use the current object, which may be different from the original build source.

If the original build pulled source from Cloud Storage and specified the generation of the object, the new build will attempt to use the same object, which may or may not be available depending on the bucket’s lifecycle management settings.

pub async fn create_build_trigger(
&mut self,
request: impl IntoRequest<CreateBuildTriggerRequest>
) -> Result<Response<BuildTrigger>, Status>

Creates a new BuildTrigger.

This API is experimental.

pub async fn get_build_trigger(
&mut self,
request: impl IntoRequest<GetBuildTriggerRequest>
) -> Result<Response<BuildTrigger>, Status>

Returns information about a BuildTrigger.

This API is experimental.

pub async fn list_build_triggers(
&mut self,
request: impl IntoRequest<ListBuildTriggersRequest>
) -> Result<Response<ListBuildTriggersResponse>, Status>

Lists existing BuildTriggers.

This API is experimental.

pub async fn delete_build_trigger(
&mut self,
request: impl IntoRequest<DeleteBuildTriggerRequest>
) -> Result<Response<()>, Status>

Deletes a BuildTrigger by its project ID and trigger ID.

This API is experimental.

pub async fn update_build_trigger(
&mut self,
request: impl IntoRequest<UpdateBuildTriggerRequest>
) -> Result<Response<BuildTrigger>, Status>

Updates a BuildTrigger by its project ID and trigger ID.

This API is experimental.

pub async fn run_build_trigger(
&mut self,
request: impl IntoRequest<RunBuildTriggerRequest>
) -> Result<Response<Operation>, Status>

Runs a BuildTrigger at a particular source revision.

pub async fn receive_trigger_webhook(
&mut self,
request: impl IntoRequest<ReceiveTriggerWebhookRequest>
) -> Result<Response<ReceiveTriggerWebhookResponse>, Status>

ReceiveTriggerWebhook [Experimental] is called when the API receives a webhook request targeted at a specific trigger.

pub async fn create_worker_pool(
&mut self,
request: impl IntoRequest<CreateWorkerPoolRequest>
) -> Result<Response<Operation>, Status>

Creates a WorkerPool.

pub async fn get_worker_pool(
&mut self,
request: impl IntoRequest<GetWorkerPoolRequest>
) -> Result<Response<WorkerPool>, Status>

Returns details of a WorkerPool.

pub async fn delete_worker_pool(
&mut self,
request: impl IntoRequest<DeleteWorkerPoolRequest>
) -> Result<Response<Operation>, Status>

Deletes a WorkerPool.

pub async fn update_worker_pool(
&mut self,
request: impl IntoRequest<UpdateWorkerPoolRequest>
) -> Result<Response<Operation>, Status>

Updates a WorkerPool.

pub async fn list_worker_pools(
&mut self,
request: impl IntoRequest<ListWorkerPoolsRequest>
) -> Result<Response<ListWorkerPoolsResponse>, Status>

Lists WorkerPools.