Struct googapis::google::devtools::testing::v1::TestMatrix

pub struct TestMatrix {

    pub test_matrix_id: String,
    pub project_id: String,
    pub client_info: Option<ClientInfo>,
    pub test_specification: Option<TestSpecification>,
    pub environment_matrix: Option<EnvironmentMatrix>,
    pub test_executions: Vec<TestExecution>,
    pub result_storage: Option<ResultStorage>,
    pub state: i32,
    pub timestamp: Option<Timestamp>,
    pub invalid_matrix_details: i32,
    pub flaky_test_attempts: i32,
    pub outcome_summary: i32,
    pub fail_fast: bool,

}

TestMatrix captures all details about a test. It contains the environment configuration, test specification, test executions and overall state and outcome.

Fields
test_matrix_id: String

Output only. Unique id set by the service.

project_id: String

The cloud project that owns the test matrix.

client_info: Option<ClientInfo>

Information about the client which invoked the test.

test_specification: Option<TestSpecification>

Required. How to run the test.

environment_matrix: Option<EnvironmentMatrix>

Required. The devices the tests are being executed on.

test_executions: Vec<TestExecution>

Output only. The list of test executions that the service creates for this matrix.

result_storage: Option<ResultStorage>

Required. Where the results for the matrix are written.

state: i32

Output only. Indicates the current progress of the test matrix.

timestamp: Option<Timestamp>

Output only. The time this test matrix was initially created.

invalid_matrix_details: i32

Output only. Describes why the matrix is considered invalid. Only useful for matrices in the INVALID state.

flaky_test_attempts: i32

The number of times a TestExecution should be re-attempted if one or more of its test cases fail for any reason. The maximum number of reruns allowed is 10.

Default is 0, which implies no reruns.

outcome_summary: i32

Output Only. The overall outcome of the test. Only set when the test matrix state is FINISHED.

fail_fast: bool

If true, only a single attempt at most will be made to run each execution/shard in the matrix. Flaky test attempts are not affected.

Normally, 2 or more attempts are made if a potential infrastructure issue is detected.

This feature is for latency sensitive workloads. The incidence of execution failures may be significantly greater for fail-fast matrices and support is more limited because of that expectation.

Implementations

impl TestMatrix

pub fn state(&self) -> TestState

Returns the enum value of state, or the default if the field is set to an invalid enum value.

pub fn set_state(&mut self, value: TestState)

Sets state to the provided enum value.

pub fn invalid_matrix_details(&self) -> InvalidMatrixDetails

Returns the enum value of invalid_matrix_details, or the default if the field is set to an invalid enum value.

pub fn set_invalid_matrix_details(&mut self, value: InvalidMatrixDetails)

Sets invalid_matrix_details to the provided enum value.

pub fn outcome_summary(&self) -> OutcomeSummary

Returns the enum value of outcome_summary, or the default if the field is set to an invalid enum value.

pub fn set_outcome_summary(&mut self, value: OutcomeSummary)

Sets outcome_summary to the provided enum value.