Module googapis::google::example::showcase::v1beta2

Modules

block_request

Nested message and enum types in BlockRequest.

compliance_client

Generated client implementations.

echo_client

Generated client implementations.

echo_request

Nested message and enum types in EchoRequest.

issue

Nested message and enum types in Issue.

report_session_response

Nested message and enum types in ReportSessionResponse.

sequence

Nested message and enum types in Sequence.

sequence_report

Nested message and enum types in SequenceReport.

sequence_service_client

Generated client implementations.

session

Nested message and enum types in Session.

test

Nested message and enum types in Test.

testing_client

Generated client implementations.

wait_request

Nested message and enum types in WaitRequest.

Structs

AttemptSequenceRequest

BlockRequest

The request for Block method.

BlockResponse

The response for Block method.

ComplianceData

ComplianceData is a message used for testing REST transcoding of different data types. scalar types

ComplianceDataChild

ComplianceDataGrandchild

CreateSequenceRequest

CreateSessionRequest

The request for the CreateSession method.

DeleteSessionRequest

Request for the DeleteSession method.

DeleteTestRequest

Request message for deleting a test.

EchoRequest

The request message used for the Echo, Collect and Chat methods. If content or opt are set in this message then the request will succeed. If status is set in this message then the status will be returned as an error.

EchoResponse

The response message for the Echo methods.

ExpandRequest

The request message for the Expand method.

GetSequenceReportRequest

GetSessionRequest

The request for the GetSession method.

Issue

An issue found in the test.

ListSessionsRequest

The request for the ListSessions method.

ListSessionsResponse

Response for the ListSessions method.

ListTestsRequest

The request for the ListTests method.

ListTestsResponse

The response for the ListTests method.

PagedExpandRequest

The request for the PagedExpand method.

PagedExpandResponse

The response for the PagedExpand method.

RepeatRequest

RepeatResponse

ReportSessionRequest

Request message for reporting on a session.

ReportSessionResponse

Response message for reporting on a session.

Sequence

SequenceReport

Session

A session is a suite of tests, generally being made in the context of testing code generation.

Test

TestRun

A TestRun is the result of running a Test.

VerifyTestRequest

VerifyTestResponse

WaitRequest

The request for Wait method.

WaitResponse

The result of the Wait operation.

Enums

Severity

A severity enum used to test enum capabilities in GAPIC surfaces.