Struct googapis::google::genomics::v1::ReadGroup

pub struct ReadGroup {
    pub id: String,
    pub dataset_id: String,
    pub name: String,
    pub description: String,
    pub sample_id: String,
    pub experiment: Option<Experiment>,
    pub predicted_insert_size: i32,
    pub programs: Vec<Program>,
    pub reference_set_id: String,
    pub info: HashMap<String, ListValue>,
}

A read group is all the data that’s processed the same way by the sequencer.

Fields
id: String

The server-generated read group ID, unique for all read groups. Note: This is different than the @RG ID field in the SAM spec. For that value, see [name][google.genomics.v1.ReadGroup.name].

dataset_id: String

The dataset to which this read group belongs.

name: String

The read group name. This corresponds to the @RG ID field in the SAM spec.

description: String

A free-form text description of this read group.

sample_id: String

A client-supplied sample identifier for the reads in this read group.

experiment: Option<Experiment>

The experiment used to generate this read group.

predicted_insert_size: i32

The predicted insert size of this read group. The insert size is the length the sequenced DNA fragment from end-to-end, not including the adapters.

programs: Vec<Program>

The programs used to generate this read group. Programs are always identical for all read groups within a read group set. For this reason, only the first read group in a returned set will have this field populated.

reference_set_id: String

The reference set the reads in this read group are aligned to.

info: HashMap<String, ListValue>

A map of additional read group information. This must be of the form map<string, string[]> (string key mapping to a list of string values).