Module googapis::google::home::enterprise::sdm::v1

Modules

smart_device_management_service_client

Generated client implementations.

Structs

Device

Device resource represents an instance of enterprise managed device in the property.

ExecuteDeviceCommandRequest

Request message for SmartDeviceManagementService.ExecuteDeviceCommand

ExecuteDeviceCommandResponse

Response message for SmartDeviceManagementService.ExecuteDeviceCommand

GetDeviceRequest

Request message for SmartDeviceManagementService.GetDevice

GetRoomRequest

Request message for SmartDeviceManagementService.GetRoom

GetStructureRequest

Request message for SmartDeviceManagementService.GetStructure

ListDevicesRequest

Request message for SmartDeviceManagementService.ListDevices

ListDevicesResponse

Response message for SmartDeviceManagementService.ListDevices

ListRoomsRequest

Request message for SmartDeviceManagementService.ListRooms

ListRoomsResponse

Response message for SmartDeviceManagementService.ListRooms

ListStructuresRequest

Request message for SmartDeviceManagementService.ListStructures

ListStructuresResponse

Response message for SmartDeviceManagementService.ListStructures

ParentRelation

Represents device relationships, for instance, structure/room to which the device is assigned to. For now this is only filled in the enterprise flow.

Room

Room resource represents an instance of sub-space within a structure such as rooms in a hotel suite or rental apartment.

Structure

Structure resource represents an instance of enterprise managed home or hotel room.