Module googapis::google::iam::admin::v1

Modules

audit_data
Nested message and enum types in AuditData.

iam_client
Generated client implementations.

lint_policy_request
Nested message and enum types in LintPolicyRequest.

lint_result
Nested message and enum types in LintResult.

list_service_account_keys_request
Nested message and enum types in ListServiceAccountKeysRequest.

permission
Nested message and enum types in Permission.

query_auditable_services_response
Nested message and enum types in QueryAuditableServicesResponse.

role
Nested message and enum types in Role.

Structs

AuditData
Audit log information specific to Cloud IAM admin APIs. This message is serialized as an Any type in the ServiceData message of an AuditLog message.

CreateRoleRequest
The request to create a new role.

CreateServiceAccountKeyRequest
The service account key create request.

CreateServiceAccountRequest
The service account create request.

DeleteRoleRequest
The request to delete an existing role.

DeleteServiceAccountKeyRequest
The service account key delete request.

DeleteServiceAccountRequest
The service account delete request.

DisableServiceAccountRequest
The service account disable request.

EnableServiceAccountRequest
The service account enable request.

GetRoleRequest
The request to get the definition of an existing role.

GetServiceAccountKeyRequest
The service account key get by id request.

GetServiceAccountRequest
The service account get request.

LintPolicyRequest
The request to lint a Cloud IAM policy object.

LintPolicyResponse
The response of a lint operation. An empty response indicates the operation was able to fully execute and no lint issue was found.

LintResult
Structured response of a single validation unit.

ListRolesRequest
The request to get all roles defined under a resource.

ListRolesResponse
The response containing the roles defined under a resource.

ListServiceAccountKeysRequest
The service account keys list request.

ListServiceAccountKeysResponse
The service account keys list response.

ListServiceAccountsRequest
The service account list request.

ListServiceAccountsResponse
The service account list response.

PatchServiceAccountRequest
The request for [PatchServiceAccount][google.iam.admin.v1.PatchServiceAccount].

Permission
A permission which can be included by a role.

QueryAuditableServicesRequest
A request to get the list of auditable services for a resource.

QueryAuditableServicesResponse
A response containing a list of auditable services for a resource.

QueryGrantableRolesRequest
The grantable role query request.

QueryGrantableRolesResponse
The grantable role query response.

QueryTestablePermissionsRequest
A request to get permissions which can be tested on a resource.

QueryTestablePermissionsResponse
The response containing permissions which can be tested on a resource.

Role
A role in the Identity and Access Management API.

ServiceAccount
An IAM service account.

ServiceAccountKey
Represents a service account key.

SignBlobRequest
Deprecated. Migrate to Service Account Credentials API.

SignBlobResponse
Deprecated. Migrate to Service Account Credentials API.

SignJwtRequest
Deprecated. Migrate to Service Account Credentials API.

SignJwtResponse
Deprecated. Migrate to Service Account Credentials API.

UndeleteRoleRequest
The request to undelete an existing role.

UndeleteServiceAccountRequest
The service account undelete request.

UndeleteServiceAccountResponse

UpdateRoleRequest
The request to update a role.

UploadServiceAccountKeyRequest
The service account key upload request.

Enums

RoleView
A view for Role objects.

ServiceAccountKeyAlgorithm
Supported key algorithms.

ServiceAccountKeyOrigin
Service Account Key Origin.

ServiceAccountPrivateKeyType
Supported private key output formats.

ServiceAccountPublicKeyType
Supported public key output formats.