Struct googapis::google::logging::v2::config_service_v2_client::ConfigServiceV2Client

pub struct ConfigServiceV2Client<T> { /* fields omitted */ }

Service for configuring sinks used to route log entries.

Implementations

impl<T> ConfigServiceV2Client<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + Sync + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> ConfigServiceV2Client<InterceptedService<T, F>> where
F: FnMut(Request<()>) -> Result<Request<()>, Status>,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn list_buckets(
&mut self,
request: impl IntoRequest<ListBucketsRequest>
) -> Result<Response<ListBucketsResponse>, Status>

Lists buckets.

pub async fn get_bucket(
&mut self,
request: impl IntoRequest<GetBucketRequest>
) -> Result<Response<LogBucket>, Status>

Gets a bucket.

pub async fn create_bucket(
&mut self,
request: impl IntoRequest<CreateBucketRequest>
) -> Result<Response<LogBucket>, Status>

Creates a bucket that can be used to store log entries. Once a bucket has been created, the region cannot be changed.

pub async fn update_bucket(
&mut self,
request: impl IntoRequest<UpdateBucketRequest>
) -> Result<Response<LogBucket>, Status>

Updates a bucket. This method replaces the following fields in the existing bucket with values from the new bucket: retention_period

If the retention period is decreased and the bucket is locked, FAILED_PRECONDITION will be returned.

If the bucket has a LifecycleState of DELETE_REQUESTED, FAILED_PRECONDITION will be returned.

A buckets region may not be modified after it is created.

pub async fn delete_bucket(
&mut self,
request: impl IntoRequest<DeleteBucketRequest>
) -> Result<Response<()>, Status>

Deletes a bucket. Moves the bucket to the DELETE_REQUESTED state. After 7 days, the bucket will be purged and all logs in the bucket will be permanently deleted.

pub async fn undelete_bucket(
&mut self,
request: impl IntoRequest<UndeleteBucketRequest>
) -> Result<Response<()>, Status>

Undeletes a bucket. A bucket that has been deleted may be undeleted within the grace period of 7 days.

pub async fn list_views(
&mut self,
request: impl IntoRequest<ListViewsRequest>
) -> Result<Response<ListViewsResponse>, Status>

Lists views on a bucket.

pub async fn get_view(
&mut self,
request: impl IntoRequest<GetViewRequest>
) -> Result<Response<LogView>, Status>

Gets a view.

pub async fn create_view(
&mut self,
request: impl IntoRequest<CreateViewRequest>
) -> Result<Response<LogView>, Status>

Creates a view over logs in a bucket. A bucket may contain a maximum of 50 views.

pub async fn update_view(
&mut self,
request: impl IntoRequest<UpdateViewRequest>
) -> Result<Response<LogView>, Status>

Updates a view. This method replaces the following fields in the existing view with values from the new view: filter.

pub async fn delete_view(
&mut self,
request: impl IntoRequest<DeleteViewRequest>
) -> Result<Response<()>, Status>

Deletes a view from a bucket.

pub async fn list_sinks(
&mut self,
request: impl IntoRequest<ListSinksRequest>
) -> Result<Response<ListSinksResponse>, Status>

Lists sinks.

pub async fn get_sink(
&mut self,
request: impl IntoRequest<GetSinkRequest>
) -> Result<Response<LogSink>, Status>

Gets a sink.

pub async fn create_sink(
&mut self,
request: impl IntoRequest<CreateSinkRequest>
) -> Result<Response<LogSink>, Status>

Creates a sink that exports specified log entries to a destination. The export of newly-ingested log entries begins immediately, unless the sink’s writer_identity is not permitted to write to the destination. A sink can export log entries only from the resource owning the sink.

pub async fn update_sink(
&mut self,
request: impl IntoRequest<UpdateSinkRequest>
) -> Result<Response<LogSink>, Status>

Updates a sink. This method replaces the following fields in the existing sink with values from the new sink: destination, and filter.

The updated sink might also have a new writer_identity; see the unique_writer_identity field.

pub async fn delete_sink(
&mut self,
request: impl IntoRequest<DeleteSinkRequest>
) -> Result<Response<()>, Status>

Deletes a sink. If the sink has a unique writer_identity, then that service account is also deleted.

pub async fn list_exclusions(
&mut self,
request: impl IntoRequest<ListExclusionsRequest>
) -> Result<Response<ListExclusionsResponse>, Status>

Lists all the exclusions in a parent resource.

pub async fn get_exclusion(
&mut self,
request: impl IntoRequest<GetExclusionRequest>
) -> Result<Response<LogExclusion>, Status>

Gets the description of an exclusion.

pub async fn create_exclusion(
&mut self,
request: impl IntoRequest<CreateExclusionRequest>
) -> Result<Response<LogExclusion>, Status>

Creates a new exclusion in a specified parent resource. Only log entries belonging to that resource can be excluded. You can have up to 10 exclusions in a resource.

pub async fn update_exclusion(
&mut self,
request: impl IntoRequest<UpdateExclusionRequest>
) -> Result<Response<LogExclusion>, Status>

Changes one or more properties of an existing exclusion.

pub async fn delete_exclusion(
&mut self,
request: impl IntoRequest<DeleteExclusionRequest>
) -> Result<Response<()>, Status>

Deletes an exclusion.

pub async fn get_cmek_settings(
&mut self,
request: impl IntoRequest<GetCmekSettingsRequest>
) -> Result<Response<CmekSettings>, Status>

Gets the Logs Router CMEK settings for the given resource.

Note: CMEK for the Logs Router can currently only be configured for GCP organizations. Once configured, it applies to all projects and folders in the GCP organization.

See Enabling CMEK for Logs Router for more information.

pub async fn update_cmek_settings(
&mut self,
request: impl IntoRequest<UpdateCmekSettingsRequest>
) -> Result<Response<CmekSettings>, Status>

Updates the Logs Router CMEK settings for the given resource.

Note: CMEK for the Logs Router can currently only be configured for GCP organizations. Once configured, it applies to all projects and folders in the GCP organization.

[UpdateCmekSettings][google.logging.v2.ConfigServiceV2.UpdateCmekSettings] will fail if 1) kms_key_name is invalid, or 2) the associated service account does not have the required roles/cloudkms.cryptoKeyEncrypterDecrypter role assigned for the key, or 3) access to the key is disabled.

See Enabling CMEK for Logs Router for more information.