Struct googapis::google::logging::v2::LogEntry

pub struct LogEntry {

    pub log_name: String,
    pub resource: Option<MonitoredResource>,
    pub timestamp: Option<Timestamp>,
    pub receive_timestamp: Option<Timestamp>,
    pub severity: i32,
    pub insert_id: String,
    pub http_request: Option<HttpRequest>,
    pub labels: HashMap<String, String>,
    pub operation: Option<LogEntryOperation>,
    pub trace: String,
    pub span_id: String,
    pub trace_sampled: bool,
    pub source_location: Option<LogEntrySourceLocation>,
    pub payload: Option<Payload>,

}

An individual entry in a log.

Fields
log_name: String

Required. The resource name of the log to which this log entry belongs:

"projects/\[PROJECT_ID]/logs/[LOG_ID\]"
"organizations/\[ORGANIZATION_ID]/logs/[LOG_ID\]"
"billingAccounts/\[BILLING_ACCOUNT_ID]/logs/[LOG_ID\]"
"folders/\[FOLDER_ID]/logs/[LOG_ID\]"

A project number may be used in place of PROJECT_ID. The project number is translated to its corresponding PROJECT_ID internally and the log_name field will contain PROJECT_ID in queries and exports.

\[LOG_ID\] must be URL-encoded within log_name. Example: "organizations/1234567890/logs/cloudresourcemanager.googleapis.com%2Factivity". \[LOG_ID\] must be less than 512 characters long and can only include the following characters: upper and lower case alphanumeric characters, forward-slash, underscore, hyphen, and period.

For backward compatibility, if log_name begins with a forward-slash, such as /projects/..., then the log entry is ingested as usual but the forward-slash is removed. Listing the log entry will not show the leading slash and filtering for a log name with a leading slash will never return any results.

resource: Option<MonitoredResource>

Required. The monitored resource that produced this log entry.

Example: a log entry that reports a database error would be associated with the monitored resource designating the particular database that reported the error.

timestamp: Option<Timestamp>

Optional. The time the event described by the log entry occurred. This time is used to compute the log entry’s age and to enforce the logs retention period. If this field is omitted in a new log entry, then Logging assigns it the current time. Timestamps have nanosecond accuracy, but trailing zeros in the fractional seconds might be omitted when the timestamp is displayed.

Incoming log entries must have timestamps that don’t exceed the logs retention period in the past, and that don’t exceed 24 hours in the future. Log entries outside those time boundaries aren’t ingested by Logging.

receive_timestamp: Option<Timestamp>

Output only. The time the log entry was received by Logging.

severity: i32

Optional. The severity of the log entry. The default value is LogSeverity.DEFAULT.

insert_id: String

Optional. A unique identifier for the log entry. If you provide a value, then Logging considers other log entries in the same project, with the same timestamp, and with the same insert_id to be duplicates which are removed in a single query result. However, there are no guarantees of de-duplication in the export of logs.

If the insert_id is omitted when writing a log entry, the Logging API assigns its own unique identifier in this field.

In queries, the insert_id is also used to order log entries that have the same log_name and timestamp values.

http_request: Option<HttpRequest>

Optional. Information about the HTTP request associated with this log entry, if applicable.

labels: HashMap<String, String>

Optional. A set of user-defined (key, value) data that provides additional information about the log entry.

operation: Option<LogEntryOperation>

Optional. Information about an operation associated with the log entry, if applicable.

trace: String

Optional. Resource name of the trace associated with the log entry, if any. If it contains a relative resource name, the name is assumed to be relative to //tracing.googleapis.com. Example: projects/my-projectid/traces/06796866738c859f2f19b7cfb3214824

span_id: String

Optional. The span ID within the trace associated with the log entry.

For Trace spans, this is the same format that the Trace API v2 uses: a 16-character hexadecimal encoding of an 8-byte array, such as 000000000000004a.

trace_sampled: bool

Optional. The sampling decision of the trace associated with the log entry.

True means that the trace resource name in the trace field was sampled for storage in a trace backend. False means that the trace was not sampled for storage when this log entry was written, or the sampling decision was unknown at the time. A non-sampled trace value is still useful as a request correlation identifier. The default is False.

source_location: Option<LogEntrySourceLocation>

Optional. Source code location information associated with the log entry, if any.

payload: Option<Payload>

The log entry payload, which can be one of multiple types.

Implementations

impl LogEntry

pub fn severity(&self) -> LogSeverity

Returns the enum value of severity, or the default if the field is set to an invalid enum value.

pub fn set_severity(&mut self, value: LogSeverity)

Sets severity to the provided enum value.