Module googapis::google::maps::playablelocations::v3

Modules

impression

Nested message and enum types in Impression.

playable_locations_client

Generated client implementations.

player_report

Nested message and enum types in PlayerReport.

sample

Structs

Impression

Encapsulates impression event details.

LogImpressionsRequest

A request for logging impressions.

LogImpressionsResponse

A response for the [LogImpressions][google.maps.playablelocations.v3.PlayableLocations.LogImpressions] method. This method returns no data upon success.

LogPlayerReportsRequest

A request for logging your player’s bad location reports.

LogPlayerReportsResponse

A response for the [LogPlayerReports][google.maps.playablelocations.v3.PlayableLocations.LogPlayerReports] method.

PlayerReport

A report submitted by a player about a playable location that is considered inappropriate for use in the game.

SamplePlayableLocationsRequest

Life of a query:

SamplePlayableLocationsResponse

Response for the [SamplePlayableLocations][google.maps.playablelocations.v3.PlayableLocations.SamplePlayableLocations] method.