Struct googapis::google::maps::routes::v1::RouteLeg

pub struct RouteLeg {
    pub distance_meters: i32,
    pub duration: Option<Duration>,
    pub static_duration: Option<Duration>,
    pub polyline: Option<Polyline>,
    pub start_location: Option<Location>,
    pub end_location: Option<Location>,
    pub steps: Vec<RouteLegStep>,
    pub travel_advisory: Option<RouteLegTravelAdvisory>,
}

Encapsulates a segment between non-via waypoints.

Fields
distance_meters: i32

The travel distance of the route leg, in meters.

duration: Option<Duration>

The length of time needed to navigate the leg. If the route_preference is set to TRAFFIC_UNAWARE, then this value is the same as static_duration. If the route_preference is either TRAFFIC_AWARE or TRAFFIC_AWARE_OPTIMAL, then this value is calculated taking traffic conditions into account.

static_duration: Option<Duration>

The duration of traveling through the leg, calculated without taking traffic conditions into consideration.

polyline: Option<Polyline>

The overall polyline for this leg. This includes that each step’s polyline.

start_location: Option<Location>

The start location of this leg. This might be different from the provided origin. For example, when the provided origin is not near a road, this is a point on the road.

end_location: Option<Location>

The end location of this leg. This might be different from the provided destination. For example, when the provided destination is not near a road, this is a point on the road.

steps: Vec<RouteLegStep>

An array of steps denoting segments within this leg. Each step represents one navigation instruction.

travel_advisory: Option<RouteLegTravelAdvisory>

Encapsulates the additional information that the user should be informed about, such as possible traffic zone restriction etc. on a route leg.