Module googapis::google::monitoring::dashboard::v1

Modules

aggregation

Nested message and enum types in Aggregation.

chart_options

Nested message and enum types in ChartOptions.

column_layout

Nested message and enum types in ColumnLayout.

dashboard

Nested message and enum types in Dashboard.

dashboards_service_client

Generated client implementations.

mosaic_layout

Nested message and enum types in MosaicLayout.

pick_time_series_filter

Nested message and enum types in PickTimeSeriesFilter.

row_layout

Nested message and enum types in RowLayout.

scorecard

Nested message and enum types in Scorecard.

statistical_time_series_filter

Nested message and enum types in StatisticalTimeSeriesFilter.

text

Nested message and enum types in Text.

threshold

Nested message and enum types in Threshold.

time_series_filter

Nested message and enum types in TimeSeriesFilter.

time_series_filter_ratio

Nested message and enum types in TimeSeriesFilterRatio.

time_series_query

Nested message and enum types in TimeSeriesQuery.

widget

Nested message and enum types in Widget.

xy_chart

Nested message and enum types in XyChart.

Structs

Aggregation

Describes how to combine multiple time series to provide a different view of the data. Aggregation of time series is done in two steps. First, each time series in the set is aligned to the same time interval boundaries, then the set of time series is optionally reduced in number.

AlertChart

A chart that displays alert policy data.

ChartOptions

Options to control visual rendering of a chart.

ColumnLayout

A simplified layout that divides the available space into vertical columns and arranges a set of widgets vertically in each column.

CreateDashboardRequest

The CreateDashboard request.

Dashboard

A Google Stackdriver dashboard. Dashboards define the content and layout of pages in the Stackdriver web application.

DeleteDashboardRequest

The DeleteDashboard request.

GetDashboardRequest

The GetDashboard request.

GridLayout

A basic layout divides the available space into vertical columns of equal width and arranges a list of widgets using a row-first strategy.

ListDashboardsRequest

The ListDashboards request.

ListDashboardsResponse

The ListDashboards request.

MosaicLayout

A mosaic layout divides the available space into a grid of blocks, and overlays the grid with tiles. Unlike GridLayout, tiles may span multiple grid blocks and can be placed at arbitrary locations in the grid.

PickTimeSeriesFilter

Describes a ranking-based time series filter. Each input time series is ranked with an aligner. The filter will allow up to num_time_series time series to pass through it, selecting them based on the relative ranking.

RowLayout

A simplified layout that divides the available space into rows and arranges a set of widgets horizontally in each row.

Scorecard

A widget showing the latest value of a metric, and how this value relates to one or more thresholds.

StatisticalTimeSeriesFilter

A filter that ranks streams based on their statistical relation to other streams in a request. Note: This field is deprecated and completely ignored by the API.

Text

A widget that displays textual content.

Threshold

Defines a threshold for categorizing time series values.

TimeSeriesFilter

A filter that defines a subset of time series data that is displayed in a widget. Time series data is fetched using the [ListTimeSeries](https://cloud.google.com/monitoring/api/ref_v3/rest/v3/projects.timeSeries/list) method.

TimeSeriesFilterRatio

A pair of time series filters that define a ratio computation. The output time series is the pair-wise division of each aligned element from the numerator and denominator time series.

TimeSeriesQuery

TimeSeriesQuery collects the set of supported methods for querying time series data from the Stackdriver metrics API.

UpdateDashboardRequest

The UpdateDashboard request.

Widget

Widget contains a single dashboard component and configuration of how to present the component in the dashboard.

XyChart

A chart that displays data on a 2D (X and Y axes) plane.

Enums

SparkChartType

Defines the possible types of spark chart supported by the Scorecard.