Module googapis::google::monitoring::metricsscope::v1

Modules

metrics_scopes_client

Generated client implementations.

operation_metadata

Nested message and enum types in OperationMetadata.

Structs

CreateMonitoredProjectRequest

Request for the CreateMonitoredProject method.

DeleteMonitoredProjectRequest

Request for the DeleteMonitoredProject method.

GetMetricsScopeRequest

Request for the GetMetricsScope method.

ListMetricsScopesByMonitoredProjectRequest

Request for the ListMetricsScopesByMonitoredProject method.

ListMetricsScopesByMonitoredProjectResponse

Response for the ListMetricsScopesByMonitoredProject method.

MetricsScope

Represents a Metrics Scope in Cloud Monitoring, which specifies one or more Google projects and zero or more AWS accounts to monitor together.

MonitoredProject

A project being monitored by a Metrics Scope.

OperationMetadata

Contains metadata for longrunning operation for the edit Metrics Scope endpoints.