Struct googapis::google::monitoring::v3::service_monitoring_service_client::ServiceMonitoringServiceClient

pub struct ServiceMonitoringServiceClient<T> { /* fields omitted */ }

The Cloud Monitoring Service-Oriented Monitoring API has endpoints for managing and querying aspects of a workspace’s services. These include the Service’s monitored resources, its Service-Level Objectives, and a taxonomy of categorized Health Metrics.

Implementations

impl<T> ServiceMonitoringServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + Sync + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> ServiceMonitoringServiceClient<InterceptedService<T, F>> where
F: FnMut(Request<()>) -> Result<Request<()>, Status>,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_service(
&mut self,
request: impl IntoRequest<CreateServiceRequest>
) -> Result<Response<Service>, Status>

Create a Service.

pub async fn get_service(
&mut self,
request: impl IntoRequest<GetServiceRequest>
) -> Result<Response<Service>, Status>

Get the named Service.

pub async fn list_services(
&mut self,
request: impl IntoRequest<ListServicesRequest>
) -> Result<Response<ListServicesResponse>, Status>

List Services for this workspace.

pub async fn update_service(
&mut self,
request: impl IntoRequest<UpdateServiceRequest>
) -> Result<Response<Service>, Status>

Update this Service.

pub async fn delete_service(
&mut self,
request: impl IntoRequest<DeleteServiceRequest>
) -> Result<Response<()>, Status>

Soft delete this Service.

pub async fn create_service_level_objective(
&mut self,
request: impl IntoRequest<CreateServiceLevelObjectiveRequest>
) -> Result<Response<ServiceLevelObjective>, Status>

Create a ServiceLevelObjective for the given Service.

pub async fn get_service_level_objective(
&mut self,
request: impl IntoRequest<GetServiceLevelObjectiveRequest>
) -> Result<Response<ServiceLevelObjective>, Status>

Get a ServiceLevelObjective by name.

pub async fn list_service_level_objectives(
&mut self,
request: impl IntoRequest<ListServiceLevelObjectivesRequest>
) -> Result<Response<ListServiceLevelObjectivesResponse>, Status>

List the ServiceLevelObjectives for the given Service.

pub async fn update_service_level_objective(
&mut self,
request: impl IntoRequest<UpdateServiceLevelObjectiveRequest>
) -> Result<Response<ServiceLevelObjective>, Status>

Update the given ServiceLevelObjective.

pub async fn delete_service_level_objective(
&mut self,
request: impl IntoRequest<DeleteServiceLevelObjectiveRequest>
) -> Result<Response<()>, Status>

Delete the given ServiceLevelObjective.