Module googapis::google::monitoring::v3::uptime_check_config

Nested message and enum types in UptimeCheckConfig.

Modules

content_matcher

Nested message and enum types in ContentMatcher.

http_check

Nested message and enum types in HttpCheck.

Structs

ContentMatcher

Optional. Used to perform content matching. This allows matching based on substrings and regular expressions, together with their negations. Only the first 4 MB of an HTTP or HTTPS check’s response (and the first 1 MB of a TCP check’s response) are examined for purposes of content matching.

HttpCheck

Information involved in an HTTP/HTTPS Uptime check request.

ResourceGroup

The resource submessage for group checks. It can be used instead of a monitored resource, when multiple resources are being monitored.

TcpCheck

Information required for a TCP Uptime check request.

Enums

CheckRequestType

The type of Uptime check request.

Resource

The resource the check is checking. Required.