Module googapis::google::privacy::dlp::v2::privacy_metric::k_map_estimation_config

Nested message and enum types in KMapEstimationConfig.

Modules

auxiliary_table
Nested message and enum types in AuxiliaryTable.

tagged_field
Nested message and enum types in TaggedField.

Structs

AuxiliaryTable
An auxiliary table contains statistical information on the relative frequency of different quasi-identifiers values. It has one or several quasi-identifiers columns, and one column that indicates the relative frequency of each quasi-identifier tuple. If a tuple is present in the data but not in the auxiliary table, the corresponding relative frequency is assumed to be zero (and thus, the tuple is highly reidentifiable).

TaggedField
A column with a semantic tag attached.