Module googapis::google::pubsub::v1

Modules

publisher_client
Generated client implementations.

push_config
Nested message and enum types in PushConfig.

schema
Nested message and enum types in Schema.

schema_service_client
Generated client implementations.

seek_request
Nested message and enum types in SeekRequest.

streaming_pull_response
Nested message and enum types in StreamingPullResponse.

subscriber_client
Generated client implementations.

validate_message_request
Nested message and enum types in ValidateMessageRequest.

Structs

AcknowledgeRequest
Request for the Acknowledge method.

CreateSchemaRequest
Request for the CreateSchema method.

CreateSnapshotRequest
Request for the CreateSnapshot method.

DeadLetterPolicy
Dead lettering is done on a best effort basis. The same message might be dead lettered multiple times.

DeleteSchemaRequest
Request for the DeleteSchema method.

DeleteSnapshotRequest
Request for the DeleteSnapshot method.

DeleteSubscriptionRequest
Request for the DeleteSubscription method.

DeleteTopicRequest
Request for the DeleteTopic method.

DetachSubscriptionRequest
Request for the DetachSubscription method.

DetachSubscriptionResponse
Response for the DetachSubscription method. Reserved for future use.

ExpirationPolicy
A policy that specifies the conditions for resource expiration (i.e., automatic resource deletion).

GetSchemaRequest
Request for the GetSchema method.

GetSnapshotRequest
Request for the GetSnapshot method.

GetSubscriptionRequest
Request for the GetSubscription method.

GetTopicRequest
Request for the GetTopic method.

ListSchemasRequest
Request for the ListSchemas method.

ListSchemasResponse
Response for the ListSchemas method.

ListSnapshotsRequest
Request for the ListSnapshots method.

ListSnapshotsResponse
Response for the ListSnapshots method.

ListSubscriptionsRequest
Request for the ListSubscriptions method.

ListSubscriptionsResponse
Response for the ListSubscriptions method.

ListTopicSnapshotsRequest
Request for the ListTopicSnapshots method.

ListTopicSnapshotsResponse
Response for the ListTopicSnapshots method.

ListTopicSubscriptionsRequest
Request for the ListTopicSubscriptions method.

ListTopicSubscriptionsResponse
Response for the ListTopicSubscriptions method.

ListTopicsRequest
Request for the ListTopics method.

ListTopicsResponse
Response for the ListTopics method.

MessageStoragePolicy
A policy constraining the storage of messages published to the topic.

ModifyAckDeadlineRequest
Request for the ModifyAckDeadline method.

ModifyPushConfigRequest
Request for the ModifyPushConfig method.

PublishRequest
Request for the Publish method.

PublishResponse
Response for the Publish method.

PubsubMessage
A message that is published by publishers and consumed by subscribers. The message must contain either a non-empty data field or at least one attribute. Note that client libraries represent this object differently depending on the language. See the corresponding client library documentation for more information. See [quotas and limits] (https://cloud.google.com/pubsub/quotas) for more information about message limits.

PullRequest
Request for the Pull method.

PullResponse
Response for the Pull method.

PushConfig
Configuration for a push delivery endpoint.

ReceivedMessage
A message and its corresponding acknowledgment ID.

RetryPolicy
A policy that specifies how Cloud Pub/Sub retries message delivery.

Schema
A schema resource.

SchemaSettings
Settings for validating messages published against a schema.

SeekRequest
Request for the Seek method.

SeekResponse
Response for the Seek method (this response is empty).

Snapshot
A snapshot resource. Snapshots are used in Seek operations, which allow you to manage message acknowledgments in bulk. That is, you can set the acknowledgment state of messages in an existing subscription to the state captured by a snapshot.

StreamingPullRequest
Request for the StreamingPull streaming RPC method. This request is used to establish the initial stream as well as to stream acknowledgements and ack deadline modifications from the client to the server.

StreamingPullResponse
Response for the StreamingPull method. This response is used to stream messages from the server to the client.

Subscription
A subscription resource.

Topic
A topic resource.

UpdateSnapshotRequest
Request for the UpdateSnapshot method.

UpdateSubscriptionRequest
Request for the UpdateSubscription method.

UpdateTopicRequest
Request for the UpdateTopic method.

ValidateMessageRequest
Request for the ValidateMessage method.

ValidateMessageResponse
Response for the ValidateMessage method.

ValidateSchemaRequest
Request for the ValidateSchema method.

ValidateSchemaResponse
Response for the ValidateSchema method.

Enums

Encoding
Possible encoding types for messages.

SchemaView
View of Schema object fields to be returned by GetSchema and ListSchemas.