Module googapis::google::spanner::v1

Modules

commit_request
Nested message and enum types in CommitRequest.

commit_response
Nested message and enum types in CommitResponse.

execute_batch_dml_request
Nested message and enum types in ExecuteBatchDmlRequest.

execute_sql_request
Nested message and enum types in ExecuteSqlRequest.

key_range
Nested message and enum types in KeyRange.

mutation
Nested message and enum types in Mutation.

plan_node
Nested message and enum types in PlanNode.

request_options
Nested message and enum types in RequestOptions.

result_set_stats
Nested message and enum types in ResultSetStats.

spanner_client
Generated client implementations.

struct_type
Nested message and enum types in StructType.

transaction_options
Nested message and enum types in TransactionOptions.

transaction_selector
Nested message and enum types in TransactionSelector.

Structs

BatchCreateSessionsRequest
The request for [BatchCreateSessions][google.spanner.v1.Spanner.BatchCreateSessions].

BatchCreateSessionsResponse
The response for [BatchCreateSessions][google.spanner.v1.Spanner.BatchCreateSessions].

BeginTransactionRequest
The request for [BeginTransaction][google.spanner.v1.Spanner.BeginTransaction].

CommitRequest
The request for [Commit][google.spanner.v1.Spanner.Commit].

CommitResponse
The response for [Commit][google.spanner.v1.Spanner.Commit].

CreateSessionRequest
The request for [CreateSession][google.spanner.v1.Spanner.CreateSession].

DeleteSessionRequest
The request for [DeleteSession][google.spanner.v1.Spanner.DeleteSession].

ExecuteBatchDmlRequest
The request for [ExecuteBatchDml][google.spanner.v1.Spanner.ExecuteBatchDml].

ExecuteBatchDmlResponse
The response for [ExecuteBatchDml][google.spanner.v1.Spanner.ExecuteBatchDml]. Contains a list of [ResultSet][google.spanner.v1.ResultSet] messages, one for each DML statement that has successfully executed, in the same order as the statements in the request. If a statement fails, the status in the response body identifies the cause of the failure.

ExecuteSqlRequest
The request for [ExecuteSql][google.spanner.v1.Spanner.ExecuteSql] and [ExecuteStreamingSql][google.spanner.v1.Spanner.ExecuteStreamingSql].

GetSessionRequest
The request for [GetSession][google.spanner.v1.Spanner.GetSession].

KeyRange
KeyRange represents a range of rows in a table or index.

KeySet
KeySet defines a collection of Cloud Spanner keys and/or key ranges. All the keys are expected to be in the same table or index. The keys need not be sorted in any particular way.

ListSessionsRequest
The request for [ListSessions][google.spanner.v1.Spanner.ListSessions].

ListSessionsResponse
The response for [ListSessions][google.spanner.v1.Spanner.ListSessions].

Mutation
A modification to one or more Cloud Spanner rows. Mutations can be applied to a Cloud Spanner database by sending them in a [Commit][google.spanner.v1.Spanner.Commit] call.

PartialResultSet
Partial results from a streaming read or SQL query. Streaming reads and SQL queries better tolerate large result sets, large rows, and large values, but are a little trickier to consume.

Partition
Information returned for each partition returned in a PartitionResponse.

PartitionOptions
Options for a PartitionQueryRequest and PartitionReadRequest.

PartitionQueryRequest
The request for [PartitionQuery][google.spanner.v1.Spanner.PartitionQuery]

PartitionReadRequest
The request for [PartitionRead][google.spanner.v1.Spanner.PartitionRead]

PartitionResponse
The response for [PartitionQuery][google.spanner.v1.Spanner.PartitionQuery] or [PartitionRead][google.spanner.v1.Spanner.PartitionRead]

PlanNode
Node information for nodes appearing in a [QueryPlan.plan_nodes][google.spanner.v1.QueryPlan.plan_nodes].

QueryPlan
Contains an ordered list of nodes appearing in the query plan.

ReadRequest
The request for [Read][google.spanner.v1.Spanner.Read] and [StreamingRead][google.spanner.v1.Spanner.StreamingRead].

RequestOptions
Common request options for various APIs.

ResultSet
Results from [Read][google.spanner.v1.Spanner.Read] or [ExecuteSql][google.spanner.v1.Spanner.ExecuteSql].

ResultSetMetadata
Metadata about a [ResultSet][google.spanner.v1.ResultSet] or [PartialResultSet][google.spanner.v1.PartialResultSet].

ResultSetStats
Additional statistics about a [ResultSet][google.spanner.v1.ResultSet] or [PartialResultSet][google.spanner.v1.PartialResultSet].

RollbackRequest
The request for [Rollback][google.spanner.v1.Spanner.Rollback].

Session
A session in the Cloud Spanner API.

StructType
StructType defines the fields of a [STRUCT][google.spanner.v1.TypeCode.STRUCT] type.

Transaction
A transaction.

TransactionOptions
Transactions

TransactionSelector
This message is used to select the transaction in which a [Read][google.spanner.v1.Spanner.Read] or [ExecuteSql][google.spanner.v1.Spanner.ExecuteSql] call runs.

Type
Type indicates the type of a Cloud Spanner value, as might be stored in a table cell or returned from an SQL query.

Enums

TypeCode
TypeCode is used as part of [Type][google.spanner.v1.Type] to indicate the type of a Cloud Spanner value.