Module googapis::google::storage::v1

Modules

bucket
Nested message and enum types in Bucket.

common_enums
Nested message and enum types in CommonEnums.

compose_object_request
Nested message and enum types in ComposeObjectRequest.

insert_object_request
Nested message and enum types in InsertObjectRequest.

list_channels_response
Nested message and enum types in ListChannelsResponse.

object
Nested message and enum types in Object.

service_constants
Nested message and enum types in ServiceConstants.

storage_client
Generated client implementations.

Structs

Bucket
A bucket.

BucketAccessControl
An access-control entry.

Channel
An notification channel used to watch for resource changes.

ChecksummedData
Message used to convey content being read or written, along with its checksum.

CommonEnums
A collection of enums used in multiple places throughout the API.

CommonObjectRequestParams
Parameters that can be passed to any object request.

CommonRequestParams
Parameters that can be passed to any request.

ComposeObjectRequest
Request message for ComposeObject.

ContentRange
Specifies a requested range of bytes to download.

CopyObjectRequest
Request message for CopyObject.

CreateHmacKeyRequest

CreateHmacKeyResponse
Create hmac response. The only time the secret for an HMAC will be returned.

DeleteBucketAccessControlRequest
Request message for DeleteBucketAccessControl.

DeleteBucketRequest
Request message for DeleteBucket.

DeleteDefaultObjectAccessControlRequest
Request message for DeleteDefaultObjectAccessControl.

DeleteHmacKeyRequest
Request object to delete a given HMAC key.

DeleteNotificationRequest
Request message for DeleteNotification.

DeleteObjectAccessControlRequest
Request message for DeleteObjectAccessControl.

DeleteObjectRequest
Message for deleting an object. Either bucket and object or upload_id must be set (but not both).

GetBucketAccessControlRequest
Request message for GetBucketAccessControl.

GetBucketRequest
Request message for GetBucket.

GetDefaultObjectAccessControlRequest
Request message for GetDefaultObjectAccessControl.

GetHmacKeyRequest
Request object to get metadata on a given HMAC key.

GetIamPolicyRequest
A wrapper around the IAM get policy request to support our common_request_params.

GetNotificationRequest
Request message for GetNotification.

GetObjectAccessControlRequest
Request message for GetObjectAccessControl.

GetObjectMediaRequest
Request message for GetObjectMedia.

GetObjectMediaResponse
Response message for GetObject.

GetObjectRequest
Request message for GetObject.

GetProjectServiceAccountRequest
Request message for GetProjectServiceAccount.

HmacKeyMetadata
Hmac Key Metadata, which includes all information other than the secret.

InsertBucketAccessControlRequest
Request message for InsertBucketAccessControl.

InsertBucketRequest
Request message for InsertBucket.

InsertDefaultObjectAccessControlRequest
Request message for InsertDefaultObjectAccessControl.

InsertNotificationRequest
Request message for InsertNotification.

InsertObjectAccessControlRequest
Request message for InsertObjectAccessControl.

InsertObjectRequest
Message for writing an object.

InsertObjectSpec
Describes an attempt to insert an object, possibly over multiple requests.

ListBucketAccessControlsRequest
Request message for ListBucketAccessControl.

ListBucketAccessControlsResponse
The response to a call to BucketAccessControls.ListBucketAccessControls.

ListBucketsRequest
Request message for ListBuckets.

ListBucketsResponse
The result of a call to Buckets.ListBuckets

ListChannelsRequest
Request message for ListChannels.

ListChannelsResponse
The result of a call to Channels.ListChannels

ListDefaultObjectAccessControlsRequest
Request message for ListDefaultObjectAccessControls.

ListHmacKeysRequest
Request to fetch a list of HMAC keys under a given project.

ListHmacKeysResponse
Hmac key list response with next page information.

ListNotificationsRequest
Request message for ListNotifications.

ListNotificationsResponse
The result of a call to Notifications.ListNotifications

ListObjectAccessControlsRequest
Request message for ListObjectAccessControls.

ListObjectAccessControlsResponse
The result of a call to ObjectAccessControls.ListObjectAccessControls.

ListObjectsRequest
Request message for ListObjects.

ListObjectsResponse
The result of a call to Objects.ListObjects

LockRetentionPolicyRequest
Request message for LockRetentionPolicy.

Notification
A subscription to receive Google PubSub notifications.

Object
An object.

ObjectAccessControl
An access-control entry.

ObjectChecksums
Message used for storing full (not subrange) object checksums.

Owner
The owner of a specific resource.

PatchBucketAccessControlRequest
Request for PatchBucketAccessControl.

PatchBucketRequest
Request for PatchBucket method.

PatchDefaultObjectAccessControlRequest
Request message for PatchDefaultObjectAccessControl.

PatchObjectAccessControlRequest
Request message for PatchObjectAccessControl.

PatchObjectRequest
Request message for PatchObject.

ProjectTeam
Represents the Viewers, Editors, or Owners of a given project.

QueryWriteStatusRequest
Request object for ByteStream.QueryWriteStatus.

QueryWriteStatusResponse
Response object for ByteStream.QueryWriteStatus.

RewriteObjectRequest
Request message for RewriteObject.

RewriteResponse
A rewrite response.

ServiceAccount
A subscription to receive Google PubSub notifications.

ServiceConstants
Shared constants.

SetIamPolicyRequest
A wrapper around the IAM set policy request to support our common_request_params.

StartResumableWriteRequest
Request message StartResumableWrite.

StartResumableWriteResponse
Response object for ByteStream.StartResumableWrite.

StopChannelRequest
Request message for StopChannel.

TestIamPermissionsRequest
A wrapper around the IAM test iam permissions request to support our common_request_params.

UpdateBucketAccessControlRequest
Request for UpdateBucketAccessControl.

UpdateBucketRequest
Request for UpdateBucket method.

UpdateDefaultObjectAccessControlRequest
Request message for UpdateDefaultObjectAccessControl.

UpdateHmacKeyRequest
Request object to update an HMAC key state.

UpdateObjectAccessControlRequest
Request message for UpdateObjectAccessControl.

UpdateObjectRequest
Request message for UpdateObject.

WatchAllObjectsRequest
Request message for WatchAllObjects.