Module googapis::google::storage::v2::bucket

Nested message and enum types in Bucket.

Modules

iam_config

Nested message and enum types in IamConfig.

lifecycle

Nested message and enum types in Lifecycle.

Structs

Billing

Billing properties of a bucket.

Cors

Cross-Origin Response sharing (CORS) properties for a bucket. For more on Cloud Storage and CORS, see https://cloud.google.com/storage/docs/cross-origin. For more on CORS in general, see https://tools.ietf.org/html/rfc6454.

Encryption

Encryption properties of a bucket.

IamConfig

Bucket restriction options.

Lifecycle

Lifecycle properties of a bucket. For more information, see https://cloud.google.com/storage/docs/lifecycle.

Logging

Logging-related properties of a bucket.

RetentionPolicy

Retention policy properties of a bucket.

Versioning

Properties of a bucket related to versioning. For more on Cloud Storage versioning, see https://cloud.google.com/storage/docs/object-versioning.

Website

Properties of a bucket related to accessing the contents as a static website. For more on hosting a static website via Cloud Storage, see https://cloud.google.com/storage/docs/hosting-static-website.