Module googapis::google::storagetransfer::v1

Modules

notification_config

Nested message and enum types in NotificationConfig.

storage_transfer_service_client

Generated client implementations.

transfer_job

Nested message and enum types in TransferJob.

transfer_operation

Nested message and enum types in TransferOperation.

transfer_spec

Nested message and enum types in TransferSpec.

Structs

AwsAccessKey

AWS access key (see AWS Security Credentials).

AwsS3Data

An AwsS3Data resource can be a data source, but not a data sink. In an AwsS3Data resource, an object’s name is the S3 object’s key name.

AzureBlobStorageData

An AzureBlobStorageData resource can be a data source, but not a data sink. An AzureBlobStorageData resource represents one Azure container. The storage account determines the Azure endpoint. In an AzureBlobStorageData resource, a blobs’s name is the Azure Blob Storage blob’s key name.

AzureCredentials

Azure credentials

CreateTransferJobRequest

Request passed to CreateTransferJob.

ErrorLogEntry

An entry describing an error that has occurred.

ErrorSummary

A summary of errors by error code, plus a count and sample error log entries.

GcsData

In a GcsData resource, an object’s name is the Cloud Storage object’s name and its “last modification time” refers to the object’s updated property of Cloud Storage objects, which changes when the content or the metadata of the object is updated.

GetGoogleServiceAccountRequest

Request passed to GetGoogleServiceAccount.

GetTransferJobRequest

Request passed to GetTransferJob.

GoogleServiceAccount

Google service account

HttpData

An HttpData resource specifies a list of objects on the web to be transferred over HTTP. The information of the objects to be transferred is contained in a file referenced by a URL. The first line in the file must be "TsvHttpData-1.0", which specifies the format of the file. Subsequent lines specify the information of the list of objects, one object per list entry. Each entry has the following tab-delimited fields:

ListTransferJobsRequest

projectId, jobNames, and jobStatuses are query parameters that can be specified when listing transfer jobs.

ListTransferJobsResponse

Response from ListTransferJobs.

NotificationConfig

Specification to configure notifications published to Cloud Pub/Sub. Notifications will be published to the customer-provided topic using the following PubsubMessage.attributes:

ObjectConditions

Conditions that determine which objects will be transferred. Applies only to Cloud Data Sources such as S3, Azure, and Cloud Storage.

PauseTransferOperationRequest

Request passed to PauseTransferOperation.

ResumeTransferOperationRequest

Request passed to ResumeTransferOperation.

RunTransferJobRequest

Request passed to RunTransferJob.

Schedule

Transfers can be scheduled to recur or to run just once.

TransferCounters

A collection of counters that report the progress of a transfer operation.

TransferJob

This resource represents the configuration of a transfer job that runs periodically.

TransferOperation

A description of the execution of a transfer.

TransferOptions

TransferOptions define the actions to be performed on objects in a transfer.

TransferSpec

Configuration for running a transfer.

UpdateTransferJobRequest

Request passed to UpdateTransferJob.