Module googapis::google::streetview::publish::v1

Modules

photo

Nested message and enum types in Photo.

street_view_publish_service_client

Generated client implementations.

upload_ref

Nested message and enum types in UploadRef.

Structs

BatchDeletePhotosRequest

Request to delete multiple [Photos][google.streetview.publish.v1.Photo].

BatchDeletePhotosResponse

Response to batch delete of one or more [Photos][google.streetview.publish.v1.Photo].

BatchGetPhotosRequest

Request to get one or more [Photos][google.streetview.publish.v1.Photo]. By default

BatchGetPhotosResponse

Response to batch get of [Photos][google.streetview.publish.v1.Photo].

BatchUpdatePhotosRequest

Request to update the metadata of photos. Updating the pixels of photos is not supported.

BatchUpdatePhotosResponse

Response to batch update of metadata of one or more [Photos][google.streetview.publish.v1.Photo].

Connection

A connection is the link from a source photo to a destination photo.

CreatePhotoRequest

Request to create a [Photo][google.streetview.publish.v1.Photo].

DeletePhotoRequest

Request to delete a [Photo][google.streetview.publish.v1.Photo].

GetPhotoRequest

Request to get a [Photo][google.streetview.publish.v1.Photo].

Level

Level information containing level number and its corresponding name.

ListPhotosRequest

Request to list all photos that belong to the user sending the request.

ListPhotosResponse

Response to list all photos that belong to a user.

Photo

Photo is used to store 360 photos along with photo metadata.

PhotoId

Identifier for a [Photo][google.streetview.publish.v1.Photo].

PhotoResponse

Response payload for a single [Photo][google.streetview.publish.v1.Photo] in batch operations including [BatchGetPhotos][google.streetview.publish.v1.StreetViewPublishService.BatchGetPhotos] and [BatchUpdatePhotos][google.streetview.publish.v1.StreetViewPublishService.BatchUpdatePhotos].

Place

Place metadata for an entity.

Pose

Raw pose measurement for an entity.

UpdatePhotoRequest

Request to update the metadata of a [Photo][google.streetview.publish.v1.Photo]. Updating the pixels of a photo is not supported.

UploadRef

Upload reference for media files.

Enums

PhotoView

Specifies which view of the [Photo][google.streetview.publish.v1.Photo] to include in the response.