Module googapis::grafeas::v1beta1::source

Modules

alias_context

Nested message and enum types in AliasContext.

cloud_repo_source_context

Nested message and enum types in CloudRepoSourceContext.

gerrit_source_context

Nested message and enum types in GerritSourceContext.

repo_id

Nested message and enum types in RepoId.

source_context

Nested message and enum types in SourceContext.

Structs

AliasContext

An alias to a repo revision.

CloudRepoSourceContext

A CloudRepoSourceContext denotes a particular revision in a Google Cloud Source Repo.

GerritSourceContext

A SourceContext referring to a Gerrit project.

GitSourceContext

A GitSourceContext denotes a particular revision in a third party Git repository (e.g., GitHub).

ProjectRepoId

Selects a repo using a Google Cloud Platform project ID (e.g., winged-cargo-31) and a repo name within that project.

RepoId

A unique identifier for a Cloud Repo.

SourceContext

A SourceContext is a reference to a tree of files. A SourceContext together with a path point to a unique revision of a single file or directory.