Struct googapis::google::analytics::admin::v1alpha::analytics_admin_service_client::AnalyticsAdminServiceClient

pub struct AnalyticsAdminServiceClient<T> { /* fields omitted */ }

Service Interface for the Analytics Admin API (GA4).

Implementations

impl<T> AnalyticsAdminServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + Sync + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> AnalyticsAdminServiceClient<InterceptedService<T, F>> where
F: FnMut(Request<()>) -> Result<Request<()>, Status>,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn get_account(
&mut self,
request: impl IntoRequest<GetAccountRequest>
) -> Result<Response<Account>, Status>

Lookup for a single Account.

pub async fn list_accounts(
&mut self,
request: impl IntoRequest<ListAccountsRequest>
) -> Result<Response<ListAccountsResponse>, Status>

Returns all accounts accessible by the caller.

Note that these accounts might not currently have GA4 properties. Soft-deleted (ie: “trashed”) accounts are excluded by default. Returns an empty list if no relevant accounts are found.

pub async fn delete_account(
&mut self,
request: impl IntoRequest<DeleteAccountRequest>
) -> Result<Response<()>, Status>

Marks target Account as soft-deleted (ie: “trashed”) and returns it.

This API does not have a method to restore soft-deleted accounts. However, they can be restored using the Trash Can UI.

If the accounts are not restored before the expiration time, the account and all child resources (eg: Properties, GoogleAdsLinks, Streams, UserLinks) will be permanently purged. https://support.google.com/analytics/answer/6154772

Returns an error if the target is not found.

pub async fn update_account(
&mut self,
request: impl IntoRequest<UpdateAccountRequest>
) -> Result<Response<Account>, Status>

Updates an account.

pub async fn provision_account_ticket(
&mut self,
request: impl IntoRequest<ProvisionAccountTicketRequest>
) -> Result<Response<ProvisionAccountTicketResponse>, Status>

Requests a ticket for creating an account.

pub async fn list_account_summaries(
&mut self,
request: impl IntoRequest<ListAccountSummariesRequest>
) -> Result<Response<ListAccountSummariesResponse>, Status>

Returns summaries of all accounts accessible by the caller.

pub async fn get_property(
&mut self,
request: impl IntoRequest<GetPropertyRequest>
) -> Result<Response<Property>, Status>

Lookup for a single “GA4” Property.

pub async fn list_properties(
&mut self,
request: impl IntoRequest<ListPropertiesRequest>
) -> Result<Response<ListPropertiesResponse>, Status>

Returns child Properties under the specified parent Account.

Only “GA4” properties will be returned. Properties will be excluded if the caller does not have access. Soft-deleted (ie: “trashed”) properties are excluded by default. Returns an empty list if no relevant properties are found.

pub async fn create_property(
&mut self,
request: impl IntoRequest<CreatePropertyRequest>
) -> Result<Response<Property>, Status>

Creates an “GA4” property with the specified location and attributes.

pub async fn delete_property(
&mut self,
request: impl IntoRequest<DeletePropertyRequest>
) -> Result<Response<Property>, Status>

Marks target Property as soft-deleted (ie: “trashed”) and returns it.

This API does not have a method to restore soft-deleted properties. However, they can be restored using the Trash Can UI.

If the properties are not restored before the expiration time, the Property and all child resources (eg: GoogleAdsLinks, Streams, UserLinks) will be permanently purged. https://support.google.com/analytics/answer/6154772

Returns an error if the target is not found, or is not an GA4 Property.

pub async fn update_property(
&mut self,
request: impl IntoRequest<UpdatePropertyRequest>
) -> Result<Response<Property>, Status>

Updates a property.

pub async fn get_user_link(
&mut self,
request: impl IntoRequest<GetUserLinkRequest>
) -> Result<Response<UserLink>, Status>

Gets information about a user’s link to an account or property.

pub async fn batch_get_user_links(
&mut self,
request: impl IntoRequest<BatchGetUserLinksRequest>
) -> Result<Response<BatchGetUserLinksResponse>, Status>

Gets information about multiple users’ links to an account or property.

pub async fn list_user_links(
&mut self,
request: impl IntoRequest<ListUserLinksRequest>
) -> Result<Response<ListUserLinksResponse>, Status>

Lists all user links on an account or property.

pub async fn audit_user_links(
&mut self,
request: impl IntoRequest<AuditUserLinksRequest>
) -> Result<Response<AuditUserLinksResponse>, Status>

Lists all user links on an account or property, including implicit ones that come from effective permissions granted by groups or organization admin roles.

If a returned user link does not have direct permissions, they cannot be removed from the account or property directly with the DeleteUserLink command. They have to be removed from the group/etc that gives them permissions, which is currently only usable/discoverable in the GA or GMP UIs.

pub async fn create_user_link(
&mut self,
request: impl IntoRequest<CreateUserLinkRequest>
) -> Result<Response<UserLink>, Status>

Creates a user link on an account or property.

If the user with the specified email already has permissions on the account or property, then the user’s existing permissions will be unioned with the permissions specified in the new UserLink.

pub async fn batch_create_user_links(
&mut self,
request: impl IntoRequest<BatchCreateUserLinksRequest>
) -> Result<Response<BatchCreateUserLinksResponse>, Status>

Creates information about multiple users’ links to an account or property.

This method is transactional. If any UserLink cannot be created, none of the UserLinks will be created.

pub async fn update_user_link(
&mut self,
request: impl IntoRequest<UpdateUserLinkRequest>
) -> Result<Response<UserLink>, Status>

Updates a user link on an account or property.

pub async fn batch_update_user_links(
&mut self,
request: impl IntoRequest<BatchUpdateUserLinksRequest>
) -> Result<Response<BatchUpdateUserLinksResponse>, Status>

Updates information about multiple users’ links to an account or property.

pub async fn delete_user_link(
&mut self,
request: impl IntoRequest<DeleteUserLinkRequest>
) -> Result<Response<()>, Status>

Deletes a user link on an account or property.

pub async fn batch_delete_user_links(
&mut self,
request: impl IntoRequest<BatchDeleteUserLinksRequest>
) -> Result<Response<()>, Status>

Deletes information about multiple users’ links to an account or property.

pub async fn get_web_data_stream(
&mut self,
request: impl IntoRequest<GetWebDataStreamRequest>
) -> Result<Response<WebDataStream>, Status>

Lookup for a single WebDataStream

pub async fn delete_web_data_stream(
&mut self,
request: impl IntoRequest<DeleteWebDataStreamRequest>
) -> Result<Response<()>, Status>

Deletes a web stream on a property.

pub async fn update_web_data_stream(
&mut self,
request: impl IntoRequest<UpdateWebDataStreamRequest>
) -> Result<Response<WebDataStream>, Status>

Updates a web stream on a property.

pub async fn create_web_data_stream(
&mut self,
request: impl IntoRequest<CreateWebDataStreamRequest>
) -> Result<Response<WebDataStream>, Status>

Creates a web stream with the specified location and attributes.

pub async fn list_web_data_streams(
&mut self,
request: impl IntoRequest<ListWebDataStreamsRequest>
) -> Result<Response<ListWebDataStreamsResponse>, Status>

Returns child web data streams under the specified parent property.

Web data streams will be excluded if the caller does not have access. Returns an empty list if no relevant web data streams are found.

pub async fn get_ios_app_data_stream(
&mut self,
request: impl IntoRequest<GetIosAppDataStreamRequest>
) -> Result<Response<IosAppDataStream>, Status>

Lookup for a single IosAppDataStream

pub async fn delete_ios_app_data_stream(
&mut self,
request: impl IntoRequest<DeleteIosAppDataStreamRequest>
) -> Result<Response<()>, Status>

Deletes an iOS app stream on a property.

pub async fn update_ios_app_data_stream(
&mut self,
request: impl IntoRequest<UpdateIosAppDataStreamRequest>
) -> Result<Response<IosAppDataStream>, Status>

Updates an iOS app stream on a property.

pub async fn list_ios_app_data_streams(
&mut self,
request: impl IntoRequest<ListIosAppDataStreamsRequest>
) -> Result<Response<ListIosAppDataStreamsResponse>, Status>

Returns child iOS app data streams under the specified parent property.

iOS app data streams will be excluded if the caller does not have access. Returns an empty list if no relevant iOS app data streams are found.

pub async fn get_android_app_data_stream(
&mut self,
request: impl IntoRequest<GetAndroidAppDataStreamRequest>
) -> Result<Response<AndroidAppDataStream>, Status>

Lookup for a single AndroidAppDataStream

pub async fn delete_android_app_data_stream(
&mut self,
request: impl IntoRequest<DeleteAndroidAppDataStreamRequest>
) -> Result<Response<()>, Status>

Deletes an android app stream on a property.

pub async fn update_android_app_data_stream(
&mut self,
request: impl IntoRequest<UpdateAndroidAppDataStreamRequest>
) -> Result<Response<AndroidAppDataStream>, Status>

Updates an android app stream on a property.

pub async fn list_android_app_data_streams(
&mut self,
request: impl IntoRequest<ListAndroidAppDataStreamsRequest>
) -> Result<Response<ListAndroidAppDataStreamsResponse>, Status>

Returns child android app streams under the specified parent property.

Android app streams will be excluded if the caller does not have access. Returns an empty list if no relevant android app streams are found.

pub async fn get_enhanced_measurement_settings(
&mut self,
request: impl IntoRequest<GetEnhancedMeasurementSettingsRequest>
) -> Result<Response<EnhancedMeasurementSettings>, Status>

Returns the singleton enhanced measurement settings for this web stream. Note that the stream must enable enhanced measurement for these settings to take effect.

pub async fn update_enhanced_measurement_settings(
&mut self,
request: impl IntoRequest<UpdateEnhancedMeasurementSettingsRequest>
) -> Result<Response<EnhancedMeasurementSettings>, Status>

Updates the singleton enhanced measurement settings for this web stream. Note that the stream must enable enhanced measurement for these settings to take effect.

pub async fn create_firebase_link(
&mut self,
request: impl IntoRequest<CreateFirebaseLinkRequest>
) -> Result<Response<FirebaseLink>, Status>

Creates a FirebaseLink.

Properties can have at most one FirebaseLink.

pub async fn update_firebase_link(
&mut self,
request: impl IntoRequest<UpdateFirebaseLinkRequest>
) -> Result<Response<FirebaseLink>, Status>

Updates a FirebaseLink on a property

pub async fn delete_firebase_link(
&mut self,
request: impl IntoRequest<DeleteFirebaseLinkRequest>
) -> Result<Response<()>, Status>

Deletes a FirebaseLink on a property

pub async fn list_firebase_links(
&mut self,
request: impl IntoRequest<ListFirebaseLinksRequest>
) -> Result<Response<ListFirebaseLinksResponse>, Status>

Lists FirebaseLinks on a property. Properties can have at most one FirebaseLink.

pub async fn get_global_site_tag(
&mut self,
request: impl IntoRequest<GetGlobalSiteTagRequest>
) -> Result<Response<GlobalSiteTag>, Status>

Returns the Site Tag for the specified web stream. Site Tags are immutable singletons.

pub async fn create_google_ads_link(
&mut self,
request: impl IntoRequest<CreateGoogleAdsLinkRequest>
) -> Result<Response<GoogleAdsLink>, Status>

Creates a GoogleAdsLink.

pub async fn update_google_ads_link(
&mut self,
request: impl IntoRequest<UpdateGoogleAdsLinkRequest>
) -> Result<Response<GoogleAdsLink>, Status>

Updates a GoogleAdsLink on a property

pub async fn delete_google_ads_link(
&mut self,
request: impl IntoRequest<DeleteGoogleAdsLinkRequest>
) -> Result<Response<()>, Status>

Deletes a GoogleAdsLink on a property

pub async fn list_google_ads_links(
&mut self,
request: impl IntoRequest<ListGoogleAdsLinksRequest>
) -> Result<Response<ListGoogleAdsLinksResponse>, Status>

Lists GoogleAdsLinks on a property.

pub async fn get_data_sharing_settings(
&mut self,
request: impl IntoRequest<GetDataSharingSettingsRequest>
) -> Result<Response<DataSharingSettings>, Status>

Get data sharing settings on an account. Data sharing settings are singletons.

pub async fn get_measurement_protocol_secret(
&mut self,
request: impl IntoRequest<GetMeasurementProtocolSecretRequest>
) -> Result<Response<MeasurementProtocolSecret>, Status>

Lookup for a single “GA4” MeasurementProtocolSecret.

pub async fn list_measurement_protocol_secrets(
&mut self,
request: impl IntoRequest<ListMeasurementProtocolSecretsRequest>
) -> Result<Response<ListMeasurementProtocolSecretsResponse>, Status>

Returns child MeasurementProtocolSecrets under the specified parent Property.

pub async fn create_measurement_protocol_secret(
&mut self,
request: impl IntoRequest<CreateMeasurementProtocolSecretRequest>
) -> Result<Response<MeasurementProtocolSecret>, Status>

Creates a measurement protocol secret.

pub async fn delete_measurement_protocol_secret(
&mut self,
request: impl IntoRequest<DeleteMeasurementProtocolSecretRequest>
) -> Result<Response<()>, Status>

Deletes target MeasurementProtocolSecret.

pub async fn update_measurement_protocol_secret(
&mut self,
request: impl IntoRequest<UpdateMeasurementProtocolSecretRequest>
) -> Result<Response<MeasurementProtocolSecret>, Status>

Updates a measurement protocol secret.

pub async fn search_change_history_events(
&mut self,
request: impl IntoRequest<SearchChangeHistoryEventsRequest>
) -> Result<Response<SearchChangeHistoryEventsResponse>, Status>

Searches through all changes to an account or its children given the specified set of filters.

pub async fn get_google_signals_settings(
&mut self,
request: impl IntoRequest<GetGoogleSignalsSettingsRequest>
) -> Result<Response<GoogleSignalsSettings>, Status>

Lookup for Google Signals settings for a property.

pub async fn update_google_signals_settings(
&mut self,
request: impl IntoRequest<UpdateGoogleSignalsSettingsRequest>
) -> Result<Response<GoogleSignalsSettings>, Status>

Updates Google Signals settings for a property.

pub async fn create_conversion_event(
&mut self,
request: impl IntoRequest<CreateConversionEventRequest>
) -> Result<Response<ConversionEvent>, Status>

Creates a conversion event with the specified attributes.

pub async fn get_conversion_event(
&mut self,
request: impl IntoRequest<GetConversionEventRequest>
) -> Result<Response<ConversionEvent>, Status>

Retrieve a single conversion event.

pub async fn delete_conversion_event(
&mut self,
request: impl IntoRequest<DeleteConversionEventRequest>
) -> Result<Response<()>, Status>

Deletes a conversion event in a property.

pub async fn list_conversion_events(
&mut self,
request: impl IntoRequest<ListConversionEventsRequest>
) -> Result<Response<ListConversionEventsResponse>, Status>

Returns a list of conversion events in the specified parent property.

Returns an empty list if no conversion events are found.

pub async fn create_custom_dimension(
&mut self,
request: impl IntoRequest<CreateCustomDimensionRequest>
) -> Result<Response<CustomDimension>, Status>

Creates a CustomDimension.

pub async fn update_custom_dimension(
&mut self,
request: impl IntoRequest<UpdateCustomDimensionRequest>
) -> Result<Response<CustomDimension>, Status>

Updates a CustomDimension on a property.

pub async fn list_custom_dimensions(
&mut self,
request: impl IntoRequest<ListCustomDimensionsRequest>
) -> Result<Response<ListCustomDimensionsResponse>, Status>

Lists CustomDimensions on a property.

pub async fn archive_custom_dimension(
&mut self,
request: impl IntoRequest<ArchiveCustomDimensionRequest>
) -> Result<Response<()>, Status>

Archives a CustomDimension on a property.

pub async fn get_custom_dimension(
&mut self,
request: impl IntoRequest<GetCustomDimensionRequest>
) -> Result<Response<CustomDimension>, Status>

Lookup for a single CustomDimension.

pub async fn create_custom_metric(
&mut self,
request: impl IntoRequest<CreateCustomMetricRequest>
) -> Result<Response<CustomMetric>, Status>

Creates a CustomMetric.

pub async fn update_custom_metric(
&mut self,
request: impl IntoRequest<UpdateCustomMetricRequest>
) -> Result<Response<CustomMetric>, Status>

Updates a CustomMetric on a property.

pub async fn list_custom_metrics(
&mut self,
request: impl IntoRequest<ListCustomMetricsRequest>
) -> Result<Response<ListCustomMetricsResponse>, Status>

Lists CustomMetrics on a property.

pub async fn archive_custom_metric(
&mut self,
request: impl IntoRequest<ArchiveCustomMetricRequest>
) -> Result<Response<()>, Status>

Archives a CustomMetric on a property.

pub async fn get_custom_metric(
&mut self,
request: impl IntoRequest<GetCustomMetricRequest>
) -> Result<Response<CustomMetric>, Status>

Lookup for a single CustomMetric.