Module googapis::google::api

Modules

backend_rule
Nested message and enum types in BackendRule.

billing
Nested message and enum types in Billing.

distribution
Nested message and enum types in Distribution.

expr

http_rule
Nested message and enum types in HttpRule.

jwt_location
Nested message and enum types in JwtLocation.

label_descriptor
Nested message and enum types in LabelDescriptor.

logging
Nested message and enum types in Logging.

metric_descriptor
Nested message and enum types in MetricDescriptor.

monitoring
Nested message and enum types in Monitoring.

property
Nested message and enum types in Property.

resource_descriptor
Nested message and enum types in ResourceDescriptor.

servicecontrol

servicemanagement

serviceusage

Structs

Advice
Generated advice about this change, used for providing more information about how a change will affect the existing service.

AuthProvider
Configuration for an authentication provider, including support for JSON Web Token (JWT).

AuthRequirement
User-defined authentication requirements, including support for JSON Web Token (JWT).

Authentication
Authentication defines the authentication configuration for API methods provided by an API service.

AuthenticationRule
Authentication rules for the service.

Backend
Backend defines the backend configuration for a service.

BackendRule
A backend rule provides configuration for an individual API element.

Billing
Billing related configuration of the service.

ConfigChange
Output generated from semantically comparing two versions of a service configuration.

Context
Context defines which contexts an API requests.

ContextRule
A context rule provides information about the context for an individual API element.

Control
Selects and configures the service controller used by the service. The service controller handles features like abuse, quota, billing, logging, monitoring, etc.

CustomHttpPattern
A custom pattern is used for defining custom HTTP verb.

Distribution
Distribution contains summary statistics for a population of values. It optionally contains a histogram representing the distribution of those values across a set of buckets.

Documentation
Documentation provides the information for describing a service.

DocumentationRule
A documentation rule provides information about individual API elements.

Endpoint
Endpoint describes a network endpoint of a service that serves a set of APIs. It is commonly known as a service endpoint. A service may expose any number of service endpoints, and all service endpoints share the same service definition, such as quota limits and monitoring metrics.

Http
Defines the HTTP configuration for an API service. It contains a list of [HttpRule][google.api.HttpRule], each specifying the mapping of an RPC method to one or more HTTP REST API methods.

HttpBody
Message that represents an arbitrary HTTP body. It should only be used for payload formats that can’t be represented as JSON, such as raw binary or an HTML page.

HttpRule
gRPC Transcoding

JwtLocation
Specifies a location to extract JWT from an API request.

LabelDescriptor
A description of a label.

LogDescriptor
A description of a log type. Example in YAML format:

Logging
Logging configuration of the service.

Metric
A specific metric, identified by specifying values for all of the labels of a [MetricDescriptor][google.api.MetricDescriptor].

MetricDescriptor
Defines a metric type and its schema. Once a metric descriptor is created, deleting or altering it stops data collection and makes the metric type’s existing data unusable.

MetricRule
Bind API methods to metrics. Binding a method to a metric causes that metric’s configured quota behaviors to apply to the method call.

MonitoredResource
An object representing a resource that can be used for monitoring, logging, billing, or other purposes. Examples include virtual machine instances, databases, and storage devices such as disks. The type field identifies a [MonitoredResourceDescriptor][google.api.MonitoredResourceDescriptor] object that describes the resource’s schema. Information in the labels field identifies the actual resource and its attributes according to the schema. For example, a particular Compute Engine VM instance could be represented by the following object, because the [MonitoredResourceDescriptor][google.api.MonitoredResourceDescriptor] for "gce_instance" has labels "instance_id" and "zone":

MonitoredResourceDescriptor
An object that describes the schema of a [MonitoredResource][google.api.MonitoredResource] object using a type name and a set of labels. For example, the monitored resource descriptor for Google Compute Engine VM instances has a type of "gce_instance" and specifies the use of the labels "instance_id" and "zone" to identify particular VM instances.

MonitoredResourceMetadata
Auxiliary metadata for a [MonitoredResource][google.api.MonitoredResource] object. [MonitoredResource][google.api.MonitoredResource] objects contain the minimum set of information to uniquely identify a monitored resource instance. There is some other useful auxiliary metadata. Monitoring and Logging use an ingestion pipeline to extract metadata for cloud resources of all types, and store the metadata in this message.

Monitoring
Monitoring configuration of the service.

OAuthRequirements
OAuth scopes are a way to define data and permissions on data. For example, there are scopes defined for “Read-only access to Google Calendar” and “Access to Cloud Platform”. Users can consent to a scope for an application, giving it permission to access that data on their behalf.

Page
Represents a documentation page. A page can contain subpages to represent nested documentation set structure.

ProjectProperties
A descriptor for defining project properties for a service. One service may have many consumer projects, and the service may want to behave differently depending on some properties on the project. For example, a project may be associated with a school, or a business, or a government agency, a business type property on the project may affect how a service responds to the client. This descriptor defines which properties are allowed to be set on a project.

Property
Defines project properties.

Quota
Quota configuration helps to achieve fairness and budgeting in service usage.

QuotaLimit
QuotaLimit defines a specific limit that applies over a specified duration for a limit type. There can be at most one limit for a duration and limit type combination defined within a QuotaGroup.

ResourceDescriptor
A simple descriptor of a resource type.

ResourceReference
Defines a proto annotation that describes a string field that refers to an API resource.

Service
Service is the root object of Google service configuration schema. It describes basic information about a service, such as the name and the title, and delegates other aspects to sub-sections. Each sub-section is either a proto message or a repeated proto message that configures a specific aspect, such as auth. See each proto message definition for details.

SourceInfo
Source information used to create a Service Config

SystemParameter
Define a parameter’s name and location. The parameter may be passed as either an HTTP header or a URL query parameter, and if both are passed the behavior is implementation-dependent.

SystemParameterRule
Define a system parameter rule mapping system parameter definitions to methods.

SystemParameters
System parameter configuration

Usage
Configuration controlling usage of a service.

UsageRule
Usage configuration rules for the service.

Visibility
Visibility defines restrictions for the visibility of service elements. Restrictions are specified using visibility labels (e.g., PREVIEW) that are elsewhere linked to users and projects.

VisibilityRule
A visibility rule provides visibility configuration for an individual API element.

Enums

ChangeType
Classifies set of possible modifications to an object in the service configuration.

ErrorReason
Defines the supported values for google.rpc.ErrorInfo.reason for the googleapis.com error domain. This error domain is reserved for Service Infrastructure. For each error info of this domain, the metadata key “service” refers to the logical identifier of an API service, such as “pubsub.googleapis.com”. The “consumer” refers to the entity that consumes an API Service. It typically is a Google project that owns the client application or the server resource, such as “projects/123”. Other metadata keys are specific to each error reason. For more information, see the definition of the specific error reason.

FieldBehavior
An indicator of the behavior of a given field (for example, that a field is required in requests, or given as output but ignored as input). This does not change the behavior in protocol buffers itself; it only denotes the behavior and may affect how API tooling handles the field.

LaunchStage
The launch stage as defined by Google Cloud Platform Launch Stages.