Module googapis::google::api::expr::v1alpha1

Modules

conformance_service_client
Generated client implementations.

constant
Nested message and enum types in Constant.

decl
Nested message and enum types in Decl.

eval_request
Nested message and enum types in EvalRequest.

eval_state
Nested message and enum types in EvalState.

explain
Nested message and enum types in Explain.

expr
Nested message and enum types in Expr.

expr_value
Nested message and enum types in ExprValue.

issue_details
Nested message and enum types in IssueDetails.

map_value
Nested message and enum types in MapValue.

type
Nested message and enum types in Type.

value
Nested message and enum types in Value.

Structs

CheckRequest
Request message for the Check method.

CheckResponse
Response message for the Check method.

CheckedExpr
A CEL expression which has been successfully type checked.

Constant
Represents a primitive literal.

Decl
Represents a declaration of a named value or function.

EnumValue
An enum value.

ErrorSet
A set of errors.

EvalRequest
Request message for the Eval method.

EvalResponse
Response message for the Eval method.

EvalState
The state of an evaluation.

Explain
Values of intermediate expressions produced when evaluating expression. Deprecated, use EvalState instead.

Expr
An abstract representation of a common expression.

ExprValue
The value of an evaluated expression.

IssueDetails
Warnings or errors in service execution are represented by [google.rpc.Status][google.rpc.Status] messages, with the following message in the details field.

ListValue
A list.

MapValue
A map.

ParseRequest
Request message for the Parse method.

ParseResponse
Response message for the Parse method.

ParsedExpr
An expression together with source information as returned by the parser.

Reference
Describes a resolved reference to a declaration.

SourceInfo
Source information collected at parse time.

SourcePosition
A specific position in source.

Type
Represents a CEL type.

UnknownSet
A set of expressions for which the value is unknown.

Value
Represents a CEL value.