Module googapis::google::api::serviceusage::v1

Modules

batch_enable_services_response
Nested message and enum types in BatchEnableServicesResponse.

disable_service_request
Nested message and enum types in DisableServiceRequest.

service_usage_client
Generated client implementations.

Structs

BatchEnableServicesRequest
Request message for the BatchEnableServices method.

BatchEnableServicesResponse
Response message for the BatchEnableServices method. This response message is assigned to the response field of the returned Operation when that operation is done.

BatchGetServicesRequest
Request message for the BatchGetServices method.

BatchGetServicesResponse
Response message for the BatchGetServices method.

DisableServiceRequest
Request message for the DisableService method.

DisableServiceResponse
Response message for the DisableService method. This response message is assigned to the response field of the returned Operation when that operation is done.

EnableServiceRequest
Request message for the EnableService method.

EnableServiceResponse
Response message for the EnableService method. This response message is assigned to the response field of the returned Operation when that operation is done.

GetServiceRequest
Request message for the GetService method.

ListServicesRequest
Request message for the ListServices method.

ListServicesResponse
Response message for the ListServices method.

OperationMetadata
The operation metadata returned for the batchend services operation.

Service
A service that is available for use by the consumer.

ServiceConfig
The configuration of the service.

Enums

State
Whether or not a service has been enabled for use by a consumer.