Enum googapis::google::bigtable::v2::row_filter::Filter

pub enum Filter {

    Chain(Chain),
    Interleave(Interleave),
    Condition(Box<Condition>),
    Sink(bool),
    PassAllFilter(bool),
    BlockAllFilter(bool),
    RowKeyRegexFilter(Vec<u8>),
    RowSampleFilter(f64),
    FamilyNameRegexFilter(String),
    ColumnQualifierRegexFilter(Vec<u8>),
    ColumnRangeFilter(ColumnRange),
    TimestampRangeFilter(TimestampRange),
    ValueRegexFilter(Vec<u8>),
    ValueRangeFilter(ValueRange),
    CellsPerRowOffsetFilter(i32),
    CellsPerRowLimitFilter(i32),
    CellsPerColumnLimitFilter(i32),
    StripValueTransformer(bool),
    ApplyLabelTransformer(String),

}

Which of the possible RowFilter types to apply. If none are set, this RowFilter returns all cells in the input row.

Variants

Chain(Chain)

Applies several RowFilters to the data in sequence, progressively narrowing the results.

Interleave(Interleave)

Applies several RowFilters to the data in parallel and combines the results.

Condition(Box<Condition>)

Applies one of two possible RowFilters to the data based on the output of a predicate RowFilter.

Sink(bool)

ADVANCED USE ONLY. Hook for introspection into the RowFilter. Outputs all cells directly to the output of the read rather than to any parent filter. Consider the following example:

Chain(
  FamilyRegex("A"),
  Interleave(
    All(),
    Chain(Label("foo"), Sink())
  ),
  QualifierRegex("B")
)

                    A,A,1,w
                    A,B,2,x
                    B,B,4,z
                       |
                FamilyRegex("A")
                       |
                    A,A,1,w
                    A,B,2,x
                       |
          +------------+-------------+
          |                          |
        All()                    Label(foo)
          |                          |
       A,A,1,w              A,A,1,w,labels:[foo]
       A,B,2,x              A,B,2,x,labels:[foo]
          |                          |
          |                        Sink() --------------+
          |                          |                  |
          +------------+      x------+          A,A,1,w,labels:[foo]
                       |                        A,B,2,x,labels:[foo]
                    A,A,1,w                             |
                    A,B,2,x                             |
                       |                                |
               QualifierRegex("B")                      |
                       |                                |
                    A,B,2,x                             |
                       |                                |
                       +--------------------------------+
                       |
                    A,A,1,w,labels:[foo]
                    A,B,2,x,labels:[foo]  // could be switched
                    A,B,2,x               // could be switched

Despite being excluded by the qualifier filter, a copy of every cell that reaches the sink is present in the final result.

As with an [Interleave][google.bigtable.v2.RowFilter.Interleave], duplicate cells are possible, and appear in an unspecified mutual order. In this case we have a duplicate with column “A:B” and timestamp 2, because one copy passed through the all filter while the other was passed through the label and sink. Note that one copy has label “foo”, while the other does not.

Cannot be used within the predicate_filter, true_filter, or false_filter of a [Condition][google.bigtable.v2.RowFilter.Condition].

PassAllFilter(bool)

Matches all cells, regardless of input. Functionally equivalent to leaving filter unset, but included for completeness.

BlockAllFilter(bool)

Does not match any cells, regardless of input. Useful for temporarily disabling just part of a filter.

RowKeyRegexFilter(Vec<u8>)

Matches only cells from rows whose keys satisfy the given RE2 regex. In other words, passes through the entire row when the key matches, and otherwise produces an empty row. Note that, since row keys can contain arbitrary bytes, the \C escape sequence must be used if a true wildcard is desired. The . character will not match the new line character \n, which may be present in a binary key.

RowSampleFilter(f64)

Matches all cells from a row with probability p, and matches no cells from the row with probability 1-p.

FamilyNameRegexFilter(String)

Matches only cells from columns whose families satisfy the given RE2 regex. For technical reasons, the regex must not contain the : character, even if it is not being used as a literal. Note that, since column families cannot contain the new line character \n, it is sufficient to use . as a full wildcard when matching column family names.

ColumnQualifierRegexFilter(Vec<u8>)

Matches only cells from columns whose qualifiers satisfy the given RE2 regex. Note that, since column qualifiers can contain arbitrary bytes, the \C escape sequence must be used if a true wildcard is desired. The . character will not match the new line character \n, which may be present in a binary qualifier.

ColumnRangeFilter(ColumnRange)

Matches only cells from columns within the given range.

TimestampRangeFilter(TimestampRange)

Matches only cells with timestamps within the given range.

ValueRegexFilter(Vec<u8>)

Matches only cells with values that satisfy the given regular expression. Note that, since cell values can contain arbitrary bytes, the \C escape sequence must be used if a true wildcard is desired. The . character will not match the new line character \n, which may be present in a binary value.

ValueRangeFilter(ValueRange)

Matches only cells with values that fall within the given range.

CellsPerRowOffsetFilter(i32)

Skips the first N cells of each row, matching all subsequent cells. If duplicate cells are present, as is possible when using an Interleave, each copy of the cell is counted separately.

CellsPerRowLimitFilter(i32)

Matches only the first N cells of each row. If duplicate cells are present, as is possible when using an Interleave, each copy of the cell is counted separately.

CellsPerColumnLimitFilter(i32)

Matches only the most recent N cells within each column. For example, if N=2, this filter would match column foo:bar at timestamps 10 and 9, skip all earlier cells in foo:bar, and then begin matching again in column foo:bar2. If duplicate cells are present, as is possible when using an Interleave, each copy of the cell is counted separately.

StripValueTransformer(bool)

Replaces each cell’s value with the empty string.

ApplyLabelTransformer(String)

Applies the given label to all cells in the output row. This allows the client to determine which results were produced from which part of the filter.

Values must be at most 15 characters in length, and match the RE2 pattern [a-z0-9\\-]+

Due to a technical limitation, it is not currently possible to apply multiple labels to a cell. As a result, a Chain may have no more than one sub-filter which contains a apply_label_transformer. It is okay for an Interleave to contain multiple apply_label_transformers, as they will be applied to separate copies of the input. This may be relaxed in the future.

Implementations

impl Filter

pub fn encode<B>(&self, buf: &mut B) where
B: BufMut,

pub fn merge<B>(
field: &mut Option<Filter>,
tag: u32,
wire_type: WireType,
buf: &mut B,
ctx: DecodeContext
) -> Result<(), DecodeError> where
B: Buf,

pub fn encoded_len(&self) -> usize