Module googapis::google::cloud::bigquery::storage::v1

Modules

append_rows_request

Nested message and enum types in AppendRowsRequest.

append_rows_response

Nested message and enum types in AppendRowsResponse.

arrow_serialization_options

Nested message and enum types in ArrowSerializationOptions.

big_query_read_client

Generated client implementations.

big_query_write_client

Generated client implementations.

read_rows_response

Nested message and enum types in ReadRowsResponse.

read_session

Nested message and enum types in ReadSession.

storage_error

Nested message and enum types in StorageError.

stream_stats

Nested message and enum types in StreamStats.

table_field_schema

Nested message and enum types in TableFieldSchema.

write_stream

Nested message and enum types in WriteStream.

Structs

AppendRowsRequest

Request message for AppendRows.

AppendRowsResponse

Response message for AppendRows.

ArrowRecordBatch

Arrow RecordBatch.

ArrowSchema

Arrow schema as specified in https://arrow.apache.org/docs/python/api/datatypes.html and serialized to bytes using IPC: https://arrow.apache.org/docs/format/Columnar.html#serialization-and-interprocess-communication-ipc

ArrowSerializationOptions

Contains options specific to Arrow Serialization.

AvroRows

Avro rows.

AvroSchema

Avro schema.

BatchCommitWriteStreamsRequest

Request message for BatchCommitWriteStreams.

BatchCommitWriteStreamsResponse

Response message for BatchCommitWriteStreams.

CreateReadSessionRequest

Request message for CreateReadSession.

CreateWriteStreamRequest

Request message for CreateWriteStream.

FinalizeWriteStreamRequest

Request message for invoking FinalizeWriteStream.

FinalizeWriteStreamResponse

Response message for FinalizeWriteStream.

FlushRowsRequest

Request message for FlushRows.

FlushRowsResponse

Respond message for FlushRows.

GetWriteStreamRequest

Request message for GetWriteStreamRequest.

ProtoRows

ProtoSchema

ProtoSchema describes the schema of the serialized protocol buffer data rows.

ReadRowsRequest

Request message for ReadRows.

ReadRowsResponse

Response from calling ReadRows may include row data, progress and throttling information.

ReadSession

Information about the ReadSession.

ReadStream

Information about a single stream that gets data out of the storage system. Most of the information about ReadStream instances is aggregated, making ReadStream lightweight.

SplitReadStreamRequest

Request message for SplitReadStream.

SplitReadStreamResponse

Response message for SplitReadStream.

StorageError

Structured custom BigQuery Storage error message. The error can be attached as error details in the returned rpc Status. In particular, the use of error codes allows more structured error handling, and reduces the need to evaluate unstructured error text strings.

StreamStats

Estimated stream statistics for a given read Stream.

TableFieldSchema

TableFieldSchema defines a single field/column within a table schema.

TableSchema

Schema of a table.

ThrottleState

Information on if the current connection is being throttled.

WriteStream

Information about a single stream that gets data inside the storage system.

Enums

DataFormat

Data format for input or output data.