Module googapis::google::cloud::datastream::v1alpha1

Modules

connection_profile
Nested message and enum types in ConnectionProfile.

datastream_client
Generated client implementations.

destination_config
Nested message and enum types in DestinationConfig.

discover_connection_profile_request
Nested message and enum types in DiscoverConnectionProfileRequest.

discover_connection_profile_response
Nested message and enum types in DiscoverConnectionProfileResponse.

forward_ssh_tunnel_connectivity
Nested message and enum types in ForwardSshTunnelConnectivity.

gcs_destination_config
Nested message and enum types in GcsDestinationConfig.

json_file_format
Nested message and enum types in JsonFileFormat.

private_connection
Nested message and enum types in PrivateConnection.

source_config
Nested message and enum types in SourceConfig.

stream
Nested message and enum types in Stream.

validation
Nested message and enum types in Validation.

validation_message
Nested message and enum types in ValidationMessage.

Structs

AvroFileFormat
AVRO file format configuration.

ConnectionProfile

CreateConnectionProfileRequest

CreatePrivateConnectionRequest

CreateRouteRequest
route creation request

CreateStreamRequest

DeleteConnectionProfileRequest

DeletePrivateConnectionRequest

DeleteRouteRequest
route deletion request

DeleteStreamRequest

DestinationConfig
The configuration of the stream destination.

DiscoverConnectionProfileRequest
Request message for ‘discover’ ConnectionProfile request.

DiscoverConnectionProfileResponse

Error
Represent a user-facing Error.

FetchErrorsRequest
Request message for ‘FetchErrors’ request.

FetchErrorsResponse
Response message for a ‘FetchErrors’ response.

FetchStaticIpsRequest
Request message for ‘FetchStaticIps’ request.

FetchStaticIpsResponse
Response message for a ‘FetchStaticIps’ response.

ForwardSshTunnelConnectivity
Forward SSH Tunnel connectivity.

GcsDestinationConfig
Google Cloud Storage destination configuration

GcsProfile
Cloud Storage bucket profile.

GetConnectionProfileRequest

GetPrivateConnectionRequest

GetRouteRequest
route get request

GetStreamRequest

JsonFileFormat
JSON file format configuration.

ListConnectionProfilesRequest

ListConnectionProfilesResponse

ListPrivateConnectionsRequest

ListPrivateConnectionsResponse

ListRoutesRequest
route list request

ListRoutesResponse
route list response

ListStreamsRequest

ListStreamsResponse

MysqlColumn
MySQL Column.

MysqlDatabase
MySQL database.

MysqlProfile
MySQL database profile.

MysqlRdbms
MySQL database structure

MysqlSourceConfig
MySQL source configuration

MysqlSslConfig
MySQL SSL configuration information.

MysqlTable
MySQL table.

NoConnectivitySettings
No connectivity settings.

OperationMetadata
Represents the metadata of the long-running operation.

OracleColumn
Oracle Column.

OracleProfile
Oracle database profile.

OracleRdbms
Oracle database structure.

OracleSchema
Oracle schema.

OracleSourceConfig
Oracle data source configuration

OracleTable
Oracle table.

PrivateConnection
The PrivateConnection resource is used to establish private connectivity between Datastream and a customer’s network.

PrivateConnectivity
Private Connectivity

Route
The Route resource is the child of the PrivateConnection resource. It used to define a route for a PrivateConnection setup.

SourceConfig
The configuration of the stream source.

StaticServiceIpConnectivity
Static IP address connectivity.

Stream

UpdateConnectionProfileRequest

UpdateStreamRequest

Validation

ValidationMessage
Represent user-facing validation result message.

ValidationResult
Contains the current validation results.

VpcPeeringConfig
The VPC Peering configuration is used to create VPC peering between Datastream and the consumer’s VPC.

Enums

GcsFileFormat
File format in Cloud Storage.

SchemaFileFormat
Schema file format.