Module googapis::google::cloud::documentai::v1beta2

Modules

document

Nested message and enum types in Document.

document_understanding_service_client

Generated client implementations.

input_config

Nested message and enum types in InputConfig.

operation_metadata

Nested message and enum types in OperationMetadata.

output_config

Nested message and enum types in OutputConfig.

Structs

AutoMlParams

Parameters to control AutoML model prediction behavior.

BatchProcessDocumentsRequest

Request to batch process documents as an asynchronous operation. The output is written to Cloud Storage as JSON in the [Document] format.

BatchProcessDocumentsResponse

Response to an batch document processing request. This is returned in the LRO Operation after the operation is complete.

BoundingPoly

A bounding polygon for the detected image annotation.

Document

Document represents the canonical document resource in Document Understanding AI. It is an interchange format that provides insights into documents and allows for collaboration between users and Document Understanding AI to iterate and optimize for quality.

EntityExtractionParams

Parameters to control entity extraction behavior.

FormExtractionParams

Parameters to control form extraction behavior.

GcsDestination

The Google Cloud Storage location where the output file will be written to.

GcsSource

The Google Cloud Storage location where the input file will be read from.

InputConfig

The desired input location and metadata.

KeyValuePairHint

User-provided hint for key value pair.

NormalizedVertex

A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

OcrParams

Parameters to control Optical Character Recognition (OCR) behavior.

OperationMetadata

Contains metadata for the BatchProcessDocuments operation.

OutputConfig

The desired output location and metadata.

ProcessDocumentRequest

Request to process one document.

ProcessDocumentResponse

Response to a single document processing request.

TableBoundHint

A hint for a table bounding box on the page for table parsing.

TableExtractionParams

Parameters to control table extraction behavior.

Vertex

A vertex represents a 2D point in the image. NOTE: the vertex coordinates are in the same scale as the original image.