Module googapis::google::cloud::functions::v1

Modules

cloud_function

Nested message and enum types in CloudFunction.

cloud_functions_service_client

Generated client implementations.

failure_policy

Nested message and enum types in FailurePolicy.

https_trigger

Nested message and enum types in HttpsTrigger.

secret_volume

Nested message and enum types in SecretVolume.

Structs

CallFunctionRequest

Request for the CallFunction method.

CallFunctionResponse

Response of CallFunction method.

CloudFunction

Describes a Cloud Function that contains user computation executed in response to an event. It encapsulate function and triggers configurations. Next tag: 36

CreateFunctionRequest

Request for the CreateFunction method.

DeleteFunctionRequest

Request for the DeleteFunction method.

EventTrigger

Describes EventTrigger, used to request events be sent from another service.

FailurePolicy

Describes the policy in case of function’s execution failure. If empty, then defaults to ignoring failures (i.e. not retrying them).

GenerateDownloadUrlRequest

Request of GenerateDownloadUrl method.

GenerateDownloadUrlResponse

Response of GenerateDownloadUrl method.

GenerateUploadUrlRequest

Request of GenerateSourceUploadUrl method.

GenerateUploadUrlResponse

Response of GenerateSourceUploadUrl method.

GetFunctionRequest

Request for the GetFunction method.

HttpsTrigger

Describes HttpsTrigger, could be used to connect web hooks to function.

ListFunctionsRequest

Request for the ListFunctions method.

ListFunctionsResponse

Response for the ListFunctions method.

OperationMetadataV1

Metadata describing an [Operation][google.longrunning.Operation]

SecretEnvVar

Configuration for a secret environment variable. It has the information necessary to fetch the secret value from secret manager and expose it as an environment variable. Secret value is not a part of the configuration. Secret values are only fetched when a new clone starts.

SecretVolume

Configuration for a secret volume. It has the information necessary to fetch the secret value from secret manager and make it available as files mounted at the requested paths within the application container. Secret value is not a part of the configuration. Every filesystem read operation performs a lookup in secret manager to retrieve the secret value.

SourceRepository

Describes SourceRepository, used to represent parameters related to source repository where a function is hosted.

UpdateFunctionRequest

Request for the UpdateFunction method.

Enums

CloudFunctionStatus

Describes the current stage of a deployment.

OperationType

A type of an operation.