Module googapis::google::cloud::iap::v1

Modules

identity_aware_proxy_admin_service_client
Generated client implementations.

identity_aware_proxy_o_auth_service_client
Generated client implementations.

Structs

AccessDeniedPageSettings
Custom content configuration for access denied page. IAP allows customers to define a custom URI to use as the error page when access is denied to users. If IAP prevents access to this page, the default IAP error page will be displayed instead.

AccessSettings
Access related settings for IAP protected apps.

ApplicationSettings
Wrapper over application specific settings for IAP.

Brand
OAuth brand data. NOTE: Only contains a portion of the data that describes a brand.

CorsSettings
Allows customers to configure HTTP request paths that’ll allow HTTP OPTIONS call to bypass authentication and authorization.

CreateBrandRequest
The request sent to CreateBrand.

CreateIdentityAwareProxyClientRequest
The request sent to CreateIdentityAwareProxyClient.

CsmSettings
Configuration for RCTokens generated for CSM workloads protected by IAP. RCTokens are IAP generated JWTs that can be verified at the application. The RCToken is primarily used for ISTIO deployments, and can be scoped to a single mesh by configuring the audience field accordingly

DeleteIdentityAwareProxyClientRequest
The request sent to DeleteIdentityAwareProxyClient.

GcipSettings
Allows customers to configure tenant_id for GCIP instance per-app.

GetBrandRequest
The request sent to GetBrand.

GetIapSettingsRequest
The request sent to GetIapSettings.

GetIdentityAwareProxyClientRequest
The request sent to GetIdentityAwareProxyClient.

IapSettings
The IAP configurable settings.

IdentityAwareProxyClient
Contains the data that describes an Identity Aware Proxy owned client.

ListBrandsRequest
The request sent to ListBrands.

ListBrandsResponse
Response message for ListBrands.

ListIdentityAwareProxyClientsRequest
The request sent to ListIdentityAwareProxyClients.

ListIdentityAwareProxyClientsResponse
Response message for ListIdentityAwareProxyClients.

OAuthSettings
Configuration for OAuth login&consent flow behavior as well as for OAuth Credentials.

ResetIdentityAwareProxyClientSecretRequest
The request sent to ResetIdentityAwareProxyClientSecret.

UpdateIapSettingsRequest
The request sent to UpdateIapSettings.