Struct googapis::google::cloud::kms::v1::key_management_service_client::KeyManagementServiceClient

pub struct KeyManagementServiceClient<T> { /* fields omitted */ }

Google Cloud Key Management Service

Manages cryptographic keys and operations using those keys. Implements a REST model with the following objects:

[KeyRing][google.cloud.kms.v1.KeyRing]

[CryptoKey][google.cloud.kms.v1.CryptoKey]

[CryptoKeyVersion][google.cloud.kms.v1.CryptoKeyVersion]

[ImportJob][google.cloud.kms.v1.ImportJob]

If you are using manual gRPC libraries, see Using gRPC with Cloud KMS.

Implementations

impl<T> KeyManagementServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + Sync + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> KeyManagementServiceClient<InterceptedService<T, F>> where
F: FnMut(Request<()>) -> Result<Request<()>, Status>,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn list_key_rings(
&mut self,
request: impl IntoRequest<ListKeyRingsRequest>
) -> Result<Response<ListKeyRingsResponse>, Status>

Lists [KeyRings][google.cloud.kms.v1.KeyRing].

pub async fn list_crypto_keys(
&mut self,
request: impl IntoRequest<ListCryptoKeysRequest>
) -> Result<Response<ListCryptoKeysResponse>, Status>

Lists [CryptoKeys][google.cloud.kms.v1.CryptoKey].

pub async fn list_crypto_key_versions(
&mut self,
request: impl IntoRequest<ListCryptoKeyVersionsRequest>
) -> Result<Response<ListCryptoKeyVersionsResponse>, Status>

Lists [CryptoKeyVersions][google.cloud.kms.v1.CryptoKeyVersion].

pub async fn list_import_jobs(
&mut self,
request: impl IntoRequest<ListImportJobsRequest>
) -> Result<Response<ListImportJobsResponse>, Status>

Lists [ImportJobs][google.cloud.kms.v1.ImportJob].

pub async fn get_key_ring(
&mut self,
request: impl IntoRequest<GetKeyRingRequest>
) -> Result<Response<KeyRing>, Status>

Returns metadata for a given [KeyRing][google.cloud.kms.v1.KeyRing].

pub async fn get_crypto_key(
&mut self,
request: impl IntoRequest<GetCryptoKeyRequest>
) -> Result<Response<CryptoKey>, Status>

Returns metadata for a given [CryptoKey][google.cloud.kms.v1.CryptoKey], as well as its [primary][google.cloud.kms.v1.CryptoKey.primary] [CryptoKeyVersion][google.cloud.kms.v1.CryptoKeyVersion].

pub async fn get_crypto_key_version(
&mut self,
request: impl IntoRequest<GetCryptoKeyVersionRequest>
) -> Result<Response<CryptoKeyVersion>, Status>

Returns metadata for a given [CryptoKeyVersion][google.cloud.kms.v1.CryptoKeyVersion].

pub async fn get_public_key(
&mut self,
request: impl IntoRequest<GetPublicKeyRequest>
) -> Result<Response<PublicKey>, Status>

Returns the public key for the given [CryptoKeyVersion][google.cloud.kms.v1.CryptoKeyVersion]. The [CryptoKey.purpose][google.cloud.kms.v1.CryptoKey.purpose] must be [ASYMMETRIC_SIGN][google.cloud.kms.v1.CryptoKey.CryptoKeyPurpose.ASYMMETRIC_SIGN] or [ASYMMETRIC_DECRYPT][google.cloud.kms.v1.CryptoKey.CryptoKeyPurpose.ASYMMETRIC_DECRYPT].

pub async fn get_import_job(
&mut self,
request: impl IntoRequest<GetImportJobRequest>
) -> Result<Response<ImportJob>, Status>

Returns metadata for a given [ImportJob][google.cloud.kms.v1.ImportJob].

pub async fn create_key_ring(
&mut self,
request: impl IntoRequest<CreateKeyRingRequest>
) -> Result<Response<KeyRing>, Status>

Create a new [KeyRing][google.cloud.kms.v1.KeyRing] in a given Project and Location.

pub async fn create_crypto_key(
&mut self,
request: impl IntoRequest<CreateCryptoKeyRequest>
) -> Result<Response<CryptoKey>, Status>

Create a new [CryptoKey][google.cloud.kms.v1.CryptoKey] within a [KeyRing][google.cloud.kms.v1.KeyRing].

[CryptoKey.purpose][google.cloud.kms.v1.CryptoKey.purpose] and [CryptoKey.version_template.algorithm][google.cloud.kms.v1.CryptoKeyVersionTemplate.algorithm] are required.

pub async fn create_crypto_key_version(
&mut self,
request: impl IntoRequest<CreateCryptoKeyVersionRequest>
) -> Result<Response<CryptoKeyVersion>, Status>

Create a new [CryptoKeyVersion][google.cloud.kms.v1.CryptoKeyVersion] in a [CryptoKey][google.cloud.kms.v1.CryptoKey].

The server will assign the next sequential id. If unset, [state][google.cloud.kms.v1.CryptoKeyVersion.state] will be set to [ENABLED][google.cloud.kms.v1.CryptoKeyVersion.CryptoKeyVersionState.ENABLED].

pub async fn import_crypto_key_version(
&mut self,
request: impl IntoRequest<ImportCryptoKeyVersionRequest>
) -> Result<Response<CryptoKeyVersion>, Status>

Imports a new [CryptoKeyVersion][google.cloud.kms.v1.CryptoKeyVersion] into an existing [CryptoKey][google.cloud.kms.v1.CryptoKey] using the wrapped key material provided in the request.

The version ID will be assigned the next sequential id within the [CryptoKey][google.cloud.kms.v1.CryptoKey].

pub async fn create_import_job(
&mut self,
request: impl IntoRequest<CreateImportJobRequest>
) -> Result<Response<ImportJob>, Status>

Create a new [ImportJob][google.cloud.kms.v1.ImportJob] within a [KeyRing][google.cloud.kms.v1.KeyRing].

[ImportJob.import_method][google.cloud.kms.v1.ImportJob.import_method] is required.

pub async fn update_crypto_key(
&mut self,
request: impl IntoRequest<UpdateCryptoKeyRequest>
) -> Result<Response<CryptoKey>, Status>

Update a [CryptoKey][google.cloud.kms.v1.CryptoKey].

pub async fn update_crypto_key_version(
&mut self,
request: impl IntoRequest<UpdateCryptoKeyVersionRequest>
) -> Result<Response<CryptoKeyVersion>, Status>

Update a [CryptoKeyVersion][google.cloud.kms.v1.CryptoKeyVersion]’s metadata.

[state][google.cloud.kms.v1.CryptoKeyVersion.state] may be changed between [ENABLED][google.cloud.kms.v1.CryptoKeyVersion.CryptoKeyVersionState.ENABLED] and [DISABLED][google.cloud.kms.v1.CryptoKeyVersion.CryptoKeyVersionState.DISABLED] using this method. See [DestroyCryptoKeyVersion][google.cloud.kms.v1.KeyManagementService.DestroyCryptoKeyVersion] and [RestoreCryptoKeyVersion][google.cloud.kms.v1.KeyManagementService.RestoreCryptoKeyVersion] to move between other states.

pub async fn encrypt(
&mut self,
request: impl IntoRequest<EncryptRequest>
) -> Result<Response<EncryptResponse>, Status>

Encrypts data, so that it can only be recovered by a call to [Decrypt][google.cloud.kms.v1.KeyManagementService.Decrypt]. The [CryptoKey.purpose][google.cloud.kms.v1.CryptoKey.purpose] must be [ENCRYPT_DECRYPT][google.cloud.kms.v1.CryptoKey.CryptoKeyPurpose.ENCRYPT_DECRYPT].

pub async fn decrypt(
&mut self,
request: impl IntoRequest<DecryptRequest>
) -> Result<Response<DecryptResponse>, Status>

Decrypts data that was protected by [Encrypt][google.cloud.kms.v1.KeyManagementService.Encrypt]. The [CryptoKey.purpose][google.cloud.kms.v1.CryptoKey.purpose] must be [ENCRYPT_DECRYPT][google.cloud.kms.v1.CryptoKey.CryptoKeyPurpose.ENCRYPT_DECRYPT].

pub async fn asymmetric_sign(
&mut self,
request: impl IntoRequest<AsymmetricSignRequest>
) -> Result<Response<AsymmetricSignResponse>, Status>

Signs data using a [CryptoKeyVersion][google.cloud.kms.v1.CryptoKeyVersion] with [CryptoKey.purpose][google.cloud.kms.v1.CryptoKey.purpose] ASYMMETRIC_SIGN, producing a signature that can be verified with the public key retrieved from [GetPublicKey][google.cloud.kms.v1.KeyManagementService.GetPublicKey].

pub async fn asymmetric_decrypt(
&mut self,
request: impl IntoRequest<AsymmetricDecryptRequest>
) -> Result<Response<AsymmetricDecryptResponse>, Status>

Decrypts data that was encrypted with a public key retrieved from [GetPublicKey][google.cloud.kms.v1.KeyManagementService.GetPublicKey] corresponding to a [CryptoKeyVersion][google.cloud.kms.v1.CryptoKeyVersion] with [CryptoKey.purpose][google.cloud.kms.v1.CryptoKey.purpose] ASYMMETRIC_DECRYPT.

pub async fn update_crypto_key_primary_version(
&mut self,
request: impl IntoRequest<UpdateCryptoKeyPrimaryVersionRequest>
) -> Result<Response<CryptoKey>, Status>

Update the version of a [CryptoKey][google.cloud.kms.v1.CryptoKey] that will be used in [Encrypt][google.cloud.kms.v1.KeyManagementService.Encrypt].

Returns an error if called on a key whose purpose is not [ENCRYPT_DECRYPT][google.cloud.kms.v1.CryptoKey.CryptoKeyPurpose.ENCRYPT_DECRYPT].

pub async fn destroy_crypto_key_version(
&mut self,
request: impl IntoRequest<DestroyCryptoKeyVersionRequest>
) -> Result<Response<CryptoKeyVersion>, Status>

Schedule a [CryptoKeyVersion][google.cloud.kms.v1.CryptoKeyVersion] for destruction.

Upon calling this method, [CryptoKeyVersion.state][google.cloud.kms.v1.CryptoKeyVersion.state] will be set to [DESTROY_SCHEDULED][google.cloud.kms.v1.CryptoKeyVersion.CryptoKeyVersionState.DESTROY_SCHEDULED] and [destroy_time][google.cloud.kms.v1.CryptoKeyVersion.destroy_time] will be set to a time 24 hours in the future, at which point the [state][google.cloud.kms.v1.CryptoKeyVersion.state] will be changed to [DESTROYED][google.cloud.kms.v1.CryptoKeyVersion.CryptoKeyVersionState.DESTROYED], and the key material will be irrevocably destroyed.

Before the [destroy_time][google.cloud.kms.v1.CryptoKeyVersion.destroy_time] is reached, [RestoreCryptoKeyVersion][google.cloud.kms.v1.KeyManagementService.RestoreCryptoKeyVersion] may be called to reverse the process.

pub async fn restore_crypto_key_version(
&mut self,
request: impl IntoRequest<RestoreCryptoKeyVersionRequest>
) -> Result<Response<CryptoKeyVersion>, Status>

Restore a [CryptoKeyVersion][google.cloud.kms.v1.CryptoKeyVersion] in the [DESTROY_SCHEDULED][google.cloud.kms.v1.CryptoKeyVersion.CryptoKeyVersionState.DESTROY_SCHEDULED] state.

Upon restoration of the CryptoKeyVersion, [state][google.cloud.kms.v1.CryptoKeyVersion.state] will be set to [DISABLED][google.cloud.kms.v1.CryptoKeyVersion.CryptoKeyVersionState.DISABLED], and [destroy_time][google.cloud.kms.v1.CryptoKeyVersion.destroy_time] will be cleared.