Enum googapis::google::cloud::networkmanagement::v1beta1::step::StepInfo

pub enum StepInfo {

    Instance(InstanceInfo),
    Firewall(FirewallInfo),
    Route(RouteInfo),
    Endpoint(EndpointInfo),
    ForwardingRule(ForwardingRuleInfo),
    VpnGateway(VpnGatewayInfo),
    VpnTunnel(VpnTunnelInfo),
    Deliver(DeliverInfo),
    Forward(ForwardInfo),
    Abort(AbortInfo),
    Drop(DropInfo),
    LoadBalancer(LoadBalancerInfo),
    Network(NetworkInfo),
    GkeMaster(GkeMasterInfo),
    CloudSqlInstance(CloudSqlInstanceInfo),

}

Configuration or metadata associated with each step. The configuration is filtered based on viewer’s permission. If a viewer has no permission to view the configuration in this step, for non-final states a special state is populated (VIEWER_PERMISSION_MISSING), and for final state the configuration is cleared.

Variants

Instance(InstanceInfo)

Display information of a Compute Engine instance.

Firewall(FirewallInfo)

Display information of a Compute Engine firewall rule.

Route(RouteInfo)

Display information of a Compute Engine route.

Endpoint(EndpointInfo)

Display information of the source and destination under analysis. The endpoint information in an intermediate state may differ with the initial input, as it might be modified by state like NAT, or Connection Proxy.

ForwardingRule(ForwardingRuleInfo)

Display information of a Compute Engine forwarding rule.

VpnGateway(VpnGatewayInfo)

Display information of a Compute Engine VPN gateway.

VpnTunnel(VpnTunnelInfo)

Display information of a Compute Engine VPN tunnel.

Deliver(DeliverInfo)

Display information of the final state “deliver” and reason.

Forward(ForwardInfo)

Display information of the final state “forward” and reason.

Abort(AbortInfo)

Display information of the final state “abort” and reason.

Drop(DropInfo)

Display information of the final state “drop” and reason.

LoadBalancer(LoadBalancerInfo)

Display information of the load balancers.

Network(NetworkInfo)

Display information of a Google Cloud network.

GkeMaster(GkeMasterInfo)

Display information of a Google Kubernetes Engine cluster master.

CloudSqlInstance(CloudSqlInstanceInfo)

Display information of a Cloud SQL instance.

Implementations

impl StepInfo

pub fn encode<B>(&self, buf: &mut B) where
B: BufMut,

pub fn merge<B>(
field: &mut Option<StepInfo>,
tag: u32,
wire_type: WireType,
buf: &mut B,
ctx: DecodeContext
) -> Result<(), DecodeError> where
B: Buf,

pub fn encoded_len(&self) -> usize