Module googapis::google::cloud::osconfig::v1

Modules

apt_settings

Nested message and enum types in AptSettings.

cvs_sv3

Nested message and enum types in CVSSv3.

exec_step_config

Nested message and enum types in ExecStepConfig.

fixed_or_percent

Nested message and enum types in FixedOrPercent.

instance

Nested message and enum types in Instance.

inventory

Nested message and enum types in Inventory.

monthly_schedule

Nested message and enum types in MonthlySchedule.

os_config_service_client

Generated client implementations.

os_config_zonal_service_client

Generated client implementations.

os_policy

Nested message and enum types in OSPolicy.

os_policy_assignment

Nested message and enum types in OSPolicyAssignment.

os_policy_assignment_operation_metadata

Nested message and enum types in OSPolicyAssignmentOperationMetadata.

os_policy_assignment_report

Nested message and enum types in OSPolicyAssignmentReport.

patch_config

Nested message and enum types in PatchConfig.

patch_deployment

Nested message and enum types in PatchDeployment.

patch_instance_filter

Nested message and enum types in PatchInstanceFilter.

patch_job

Nested message and enum types in PatchJob.

patch_rollout

Nested message and enum types in PatchRollout.

recurring_schedule

Nested message and enum types in RecurringSchedule.

vulnerability_report

Nested message and enum types in VulnerabilityReport.

windows_update_settings

Nested message and enum types in WindowsUpdateSettings.

Structs

AptSettings

Apt patching is completed by executing apt-get update && apt-get upgrade. Additional options can be set to control how this is executed.

CancelPatchJobRequest

Message for canceling a patch job.

CreateOsPolicyAssignmentRequest

A request message to create an OS policy assignment

CreatePatchDeploymentRequest

A request message for creating a patch deployment.

CvsSv3

Common Vulnerability Scoring System version 3. For details, see https://www.first.org/cvss/specification-document

DeleteOsPolicyAssignmentRequest

A request message for deleting a OS policy assignment.

DeletePatchDeploymentRequest

A request message for deleting a patch deployment.

ExecStep

A step that runs an executable for a PatchJob.

ExecStepConfig

Common configurations for an ExecStep.

ExecutePatchJobRequest

A request message to initiate patching across Compute Engine instances.

FixedOrPercent

Message encapsulating a value that can be either absolute (“fixed”) or relative (“percent”) to a value.

GcsObject

Cloud Storage object representation.

GetInventoryRequest

A request message for getting inventory data for the specified VM.

GetOsPolicyAssignmentReportRequest

Get a report of the OS policy assignment for a VM instance.

GetOsPolicyAssignmentRequest

A request message to get an OS policy assignment

GetPatchDeploymentRequest

A request message for retrieving a patch deployment.

GetPatchJobRequest

Request to get an active or completed patch job.

GetVulnerabilityReportRequest

A request message for getting the vulnerability report for the specified VM.

GooSettings

Googet patching is performed by running googet update.

Instance

Namespace for instance state enums.

Inventory

This API resource represents the available inventory data for a Compute Engine virtual machine (VM) instance at a given point in time.

ListInventoriesRequest

A request message for listing inventory data for all VMs in the specified location.

ListInventoriesResponse

A response message for listing inventory data for all VMs in a specified location.

ListOsPolicyAssignmentReportsRequest

List the OS policy assignment reports for VM instances.

ListOsPolicyAssignmentReportsResponse

A response message for listing OS Policy assignment reports including the page of results and page token.

ListOsPolicyAssignmentRevisionsRequest

A request message to list revisions for a OS policy assignment

ListOsPolicyAssignmentRevisionsResponse

A response message for listing all revisions for a OS policy assignment.

ListOsPolicyAssignmentsRequest

A request message to list OS policy assignments for a parent resource

ListOsPolicyAssignmentsResponse

A response message for listing all assignments under given parent.

ListPatchDeploymentsRequest

A request message for listing patch deployments.

ListPatchDeploymentsResponse

A response message for listing patch deployments.

ListPatchJobInstanceDetailsRequest

Request to list details for all instances that are part of a patch job.

ListPatchJobInstanceDetailsResponse

A response message for listing the instances details for a patch job.

ListPatchJobsRequest

A request message for listing patch jobs.

ListPatchJobsResponse

A response message for listing patch jobs.

ListVulnerabilityReportsRequest

A request message for listing vulnerability reports for all VM instances in the specified location.

ListVulnerabilityReportsResponse

A response message for listing vulnerability reports for all VM instances in the specified location.

MonthlySchedule

Represents a monthly schedule. An example of a valid monthly schedule is “on the third Tuesday of the month” or “on the 15th of the month”.

OneTimeSchedule

Sets the time for a one time patch deployment. Timestamp is in [RFC3339](https://www.ietf.org/rfc/rfc3339.txt) text format.

OsPolicy

An OS policy defines the desired state configuration for a VM.

OsPolicyAssignment

OS policy assignment is an API resource that is used to apply a set of OS policies to a dynamically targeted group of Compute Engine VM instances.

OsPolicyAssignmentOperationMetadata

OS policy assignment operation metadata provided by OS policy assignment API methods that return long running operations.

OsPolicyAssignmentReport

A report of the OS policy assignment status for a given instance.

PatchConfig

Patch configuration specifications. Contains details on how to apply the patch(es) to a VM instance.

PatchDeployment

Patch deployments are configurations that individual patch jobs use to complete a patch. These configurations include instance filter, package repository settings, and a schedule. For more information about creating and managing patch deployments, see Scheduling patch jobs.

PatchInstanceFilter

A filter to target VM instances for patching. The targeted VMs must meet all criteria specified. So if both labels and zones are specified, the patch job targets only VMs with those labels and in those zones.

PatchJob

A high level representation of a patch job that is either in progress or has completed.

PatchJobInstanceDetails

Patch details for a VM instance. For more information about reviewing VM instance details, see Listing all VM instance details for a specific patch job.

PatchRollout

Patch rollout configuration specifications. Contains details on the concurrency control when applying patch(es) to all targeted VMs.

RecurringSchedule

Sets the time for recurring patch deployments.

UpdateOsPolicyAssignmentRequest

A request message to update an OS policy assignment

VulnerabilityReport

This API resource represents the vulnerability report for a specified Compute Engine virtual machine (VM) instance at a given point in time.

WeekDayOfMonth

Represents one week day in a month. An example is “the 4th Sunday”.

WeeklySchedule

Represents a weekly schedule.

WindowsUpdateSettings

Windows patching is performed using the Windows Update Agent.

YumSettings

Yum patching is performed by executing yum update. Additional options can be set to control how this is executed.

ZypperSettings

Zypper patching is performed by running zypper patch. See also https://en.opensuse.org/SDB:Zypper_manual.

Enums

InventoryView

The view for inventory objects.