Module googapis::google::cloud::tpu::v1

Modules

node

Nested message and enum types in Node.

symptom

Nested message and enum types in Symptom.

tpu_client

Generated client implementations.

Structs

AcceleratorType

A accelerator type that a Node can be configured with.

CreateNodeRequest

Request for [CreateNode][google.cloud.tpu.v1.Tpu.CreateNode].

DeleteNodeRequest

Request for [DeleteNode][google.cloud.tpu.v1.Tpu.DeleteNode].

GetAcceleratorTypeRequest

Request for [GetAcceleratorType][google.cloud.tpu.v1.Tpu.GetAcceleratorType].

GetNodeRequest

Request for [GetNode][google.cloud.tpu.v1.Tpu.GetNode].

GetTensorFlowVersionRequest

Request for [GetTensorFlowVersion][google.cloud.tpu.v1.Tpu.GetTensorFlowVersion].

ListAcceleratorTypesRequest

Request for [ListAcceleratorTypes][google.cloud.tpu.v1.Tpu.ListAcceleratorTypes].

ListAcceleratorTypesResponse

Response for [ListAcceleratorTypes][google.cloud.tpu.v1.Tpu.ListAcceleratorTypes].

ListNodesRequest

Request for [ListNodes][google.cloud.tpu.v1.Tpu.ListNodes].

ListNodesResponse

Response for [ListNodes][google.cloud.tpu.v1.Tpu.ListNodes].

ListTensorFlowVersionsRequest

Request for [ListTensorFlowVersions][google.cloud.tpu.v1.Tpu.ListTensorFlowVersions].

ListTensorFlowVersionsResponse

Response for [ListTensorFlowVersions][google.cloud.tpu.v1.Tpu.ListTensorFlowVersions].

NetworkEndpoint

A network endpoint over which a TPU worker can be reached.

Node

A TPU instance.

OperationMetadata

Metadata describing an [Operation][google.longrunning.Operation]

ReimageNodeRequest

Request for [ReimageNode][google.cloud.tpu.v1.Tpu.ReimageNode].

SchedulingConfig

Sets the scheduling options for this node.

StartNodeRequest

Request for [StartNode][google.cloud.tpu.v1.Tpu.StartNode].

StopNodeRequest

Request for [StopNode][google.cloud.tpu.v1.Tpu.StopNode].

Symptom

A Symptom instance.

TensorFlowVersion

A tensorflow version that a Node can be configured with.