Module googapis::google::devtools::artifactregistry::v1beta2

Modules

artifact_registry_client
Generated client implementations.

hash
Nested message and enum types in Hash.

repository
Nested message and enum types in Repository.

Structs

CreateRepositoryRequest
The request to create a new repository.

CreateTagRequest
The request to create a new tag.

DeletePackageRequest
The request to delete a package.

DeleteRepositoryRequest
The request to delete a repository.

DeleteTagRequest
The request to delete a tag.

DeleteVersionRequest
The request to delete a version.

File
Files store content that is potentially associated with Packages or Versions.

GetFileRequest
The request to retrieve a file.

GetPackageRequest
The request to retrieve a package.

GetRepositoryRequest
The request to retrieve a repository.

GetTagRequest
The request to retrieve a tag.

GetVersionRequest
The request to retrieve a version.

Hash
A hash of file content.

ListFilesRequest
The request to list files.

ListFilesResponse
The response from listing files.

ListPackagesRequest
The request to list packages.

ListPackagesResponse
The response from listing packages.

ListRepositoriesRequest
The request to list repositories.

ListRepositoriesResponse
The response from listing repositories.

ListTagsRequest
The request to list tags.

ListTagsResponse
The response from listing tags.

ListVersionsRequest
The request to list versions.

ListVersionsResponse
The response from listing versions.

OperationMetadata
Metadata type for longrunning-operations, currently empty.

Package
Packages are named collections of versions.

Repository
A Repository for storing artifacts with a specific format.

Tag
Tags point to a version and represent an alternative name that can be used to access the version.

UpdateRepositoryRequest
The request to update a repository.

UpdateTagRequest
The request to create or update a tag.

Version
The body of a version resource. A version resource represents a collection of components, such as files and other data. This may correspond to a version in many package management schemes.

Enums

VersionView
The view, which determines what version information is returned in a response.