Struct googapis::google::firestore::admin::v1::firestore_admin_client::FirestoreAdminClient

pub struct FirestoreAdminClient<T> { /* fields omitted */ }

Operations are created by service FirestoreAdmin, but are accessed via service google.longrunning.Operations.

Implementations

impl<T> FirestoreAdminClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + Sync + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> FirestoreAdminClient<InterceptedService<T, F>> where
F: FnMut(Request<()>) -> Result<Request<()>, Status>,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_index(
&mut self,
request: impl IntoRequest<CreateIndexRequest>
) -> Result<Response<Operation>, Status>

Creates a composite index. This returns a [google.longrunning.Operation][google.longrunning.Operation] which may be used to track the status of the creation. The metadata for the operation will be the type [IndexOperationMetadata][google.firestore.admin.v1.IndexOperationMetadata].

pub async fn list_indexes(
&mut self,
request: impl IntoRequest<ListIndexesRequest>
) -> Result<Response<ListIndexesResponse>, Status>

Lists composite indexes.

pub async fn get_index(
&mut self,
request: impl IntoRequest<GetIndexRequest>
) -> Result<Response<Index>, Status>

Gets a composite index.

pub async fn delete_index(
&mut self,
request: impl IntoRequest<DeleteIndexRequest>
) -> Result<Response<()>, Status>

Deletes a composite index.

pub async fn get_field(
&mut self,
request: impl IntoRequest<GetFieldRequest>
) -> Result<Response<Field>, Status>

Gets the metadata and configuration for a Field.

pub async fn update_field(
&mut self,
request: impl IntoRequest<UpdateFieldRequest>
) -> Result<Response<Operation>, Status>

Updates a field configuration. Currently, field updates apply only to single field index configuration. However, calls to [FirestoreAdmin.UpdateField][google.firestore.admin.v1.FirestoreAdmin.UpdateField] should provide a field mask to avoid changing any configuration that the caller isn’t aware of. The field mask should be specified as: { paths: "index_config" }.

This call returns a [google.longrunning.Operation][google.longrunning.Operation] which may be used to track the status of the field update. The metadata for the operation will be the type [FieldOperationMetadata][google.firestore.admin.v1.FieldOperationMetadata].

To configure the default field settings for the database, use the special Field with resource name: projects/{project_id}/databases/{database_id}/collectionGroups/__default__/fields/*.

pub async fn list_fields(
&mut self,
request: impl IntoRequest<ListFieldsRequest>
) -> Result<Response<ListFieldsResponse>, Status>

Lists the field configuration and metadata for this database.

Currently, [FirestoreAdmin.ListFields][google.firestore.admin.v1.FirestoreAdmin.ListFields] only supports listing fields that have been explicitly overridden. To issue this query, call [FirestoreAdmin.ListFields][google.firestore.admin.v1.FirestoreAdmin.ListFields] with the filter set to indexConfig.usesAncestorConfig:false.

pub async fn export_documents(
&mut self,
request: impl IntoRequest<ExportDocumentsRequest>
) -> Result<Response<Operation>, Status>

Exports a copy of all or a subset of documents from Google Cloud Firestore to another storage system, such as Google Cloud Storage. Recent updates to documents may not be reflected in the export. The export occurs in the background and its progress can be monitored and managed via the Operation resource that is created. The output of an export may only be used once the associated operation is done. If an export operation is cancelled before completion it may leave partial data behind in Google Cloud Storage.

pub async fn import_documents(
&mut self,
request: impl IntoRequest<ImportDocumentsRequest>
) -> Result<Response<Operation>, Status>

Imports documents into Google Cloud Firestore. Existing documents with the same name are overwritten. The import occurs in the background and its progress can be monitored and managed via the Operation resource that is created. If an ImportDocuments operation is cancelled, it is possible that a subset of the data has already been imported to Cloud Firestore.