Module googapis::google::longrunning

Modules

operation
Nested message and enum types in Operation.

operations_client
Generated client implementations.

Structs

CancelOperationRequest
The request message for [Operations.CancelOperation][google.longrunning.Operations.CancelOperation].

DeleteOperationRequest
The request message for [Operations.DeleteOperation][google.longrunning.Operations.DeleteOperation].

GetOperationRequest
The request message for [Operations.GetOperation][google.longrunning.Operations.GetOperation].

ListOperationsRequest
The request message for [Operations.ListOperations][google.longrunning.Operations.ListOperations].

ListOperationsResponse
The response message for [Operations.ListOperations][google.longrunning.Operations.ListOperations].

Operation
This resource represents a long-running operation that is the result of a network API call.

OperationInfo
A message representing the message types used by a long-running operation.

WaitOperationRequest
The request message for [Operations.WaitOperation][google.longrunning.Operations.WaitOperation].