Module googapis::google::storage::v2

Modules

bucket

Nested message and enum types in Bucket.

object

Nested message and enum types in Object.

query_write_status_response

Nested message and enum types in QueryWriteStatusResponse.

service_constants

Nested message and enum types in ServiceConstants.

storage_client

Generated client implementations.

write_object_request

Nested message and enum types in WriteObjectRequest.

write_object_response

Nested message and enum types in WriteObjectResponse.

Structs

Bucket

A bucket.

BucketAccessControl

An access-control entry.

ChecksummedData

Message used to convey content being read or written, along with an optional checksum.

CommonObjectRequestParams

Parameters that can be passed to any object request.

CommonRequestParams

Parameters that can be passed to any request.

ContentRange

Specifies a requested range of bytes to download.

Object

An object.

ObjectAccessControl

An access-control entry.

ObjectChecksums

Message used for storing full (not subrange) object checksums.

Owner

The owner of a specific resource.

ProjectTeam

Represents the Viewers, Editors, or Owners of a given project.

QueryWriteStatusRequest

Request object for QueryWriteStatus.

QueryWriteStatusResponse

Response object for QueryWriteStatus.

ReadObjectRequest

Request message for ReadObject.

ReadObjectResponse

Response message for GetObject.

ServiceConstants

Shared constants.

StartResumableWriteRequest

Request message StartResumableWrite.

StartResumableWriteResponse

Response object for StartResumableWrite.

WriteObjectRequest

Request message for WriteObject.

WriteObjectResponse

Response message for WriteObject.

WriteObjectSpec

Describes an attempt to insert an object, possibly over multiple requests.

Enums

PredefinedObjectAcl

Predefined or “canned” aliases for sets of specific object ACL entries.