Module googapis::maps::fleetengine::v1

Modules

report_billable_trip_request

Nested message and enum types in ReportBillableTripRequest.

search_vehicles_request

Nested message and enum types in SearchVehiclesRequest.

status

Nested message and enum types in Status.

terminal_point_id

Nested message and enum types in TerminalPointId.

trip_service_client

Generated client implementations.

vehicle

Nested message and enum types in Vehicle.

vehicle_match

Nested message and enum types in VehicleMatch.

vehicle_service_client

Generated client implementations.

Structs

Address

Address of a place.

BatteryInfo

Information about the device’s battery.

CreateTripRequest

CreateTrip request message.

CreateVehicleRequest

CreateVehicle request message.

DeviceSettings

Information about various settings on the device.

FormattedAddress

A full, human-readable address for the entity containing this message.

GetTripRequest

GetTrip request message.

GetVehicleRequest

GetVehicle request message.

LicensePlate

The license plate information of the Vehicle. This is used to support congestion pricing restrictions in certain areas. To avoid storing personally-identifiable information, only the minimum information about the license plate is stored as part of the entity.

ListVehiclesRequest

ListVehicles request message.

ListVehiclesResponse

ListVehicles response message.

ReportBillableTripRequest

ReportBillableTrip request message.

RequestHeader

A RequestHeader contains fields common to all Fleet Engine RPC requests.

SearchTripsRequest

SearchTrips request message.

SearchTripsResponse

SearchTrips response message.

SearchVehiclesRequest

SearchVehicles request message.

SearchVehiclesResponse

SearchVehicles response message.

Status

The ‘Status’ defines a FleetEngine custom logical error mode.

StopLocation

The actual location where a stop (pickup/dropoff) happened.

TerminalLocation

Describes the location of a pickup or dropoff.

TerminalPointId

Identifies a terminal point.

Trip

Trip metadata.

TripWaypoint

Describes a stopping point on a vehicle’s route or an ending point on a vehicle’s trip.

UpdateTripRequest

UpdateTrip request message.

UpdateVehicleAttributesRequest

UpdateVehicleAttributes request message.

UpdateVehicleAttributesResponse

UpdateVehicleAttributes response message.

UpdateVehicleLocationRequest

UpdateVehicleLocation request message.

UpdateVehicleRequest

UpdateVehicle request message.

Vehicle

Vehicle metadata.

VehicleAttribute

Describes a vehicle attribute as a key-value pair.

VehicleAttributeList

This messages allows a list-of-list datatype for VehicleAttribute.

VehicleLocation

The location, speed, and heading of a vehicle at a point in time.

VehicleMatch

VehicleMatch contains the vehicle, ETA, and distance calculations for a vehicle that matches the SearchVehiclesRequest.

Waypoint

Waypoint describes intermediate points along a route.

Enums

BatteryStatus

Status of the battery, whether full or charging etc.

BillingPlatformIdentifier

A set of values that indicate upon which platform the request was issued.

LocationPowerSaveMode

How location features are set to behave on the device when battery saver is on. (https://developer.android.com/reference/android/os/PowerManager#getLocationPowerSaveMode())

LocationSensor

Possible location providers.

NavigationStatus

A set of values that specify the vehicle’s navigation status.

PolylineFormatType

The type of polyline format.

PowerSource

Type of the charger being used to charge the battery.

TripStatus

The status of a trip indicating its progression.

TripType

The type of a trip.

TripView

Selector for different sets of Trip fields in a GetTrip response. See [AIP-157](https://google.aip.dev/157) for context. Additional views are likely to be added.

VehicleState

The state of a Vehicle.

WaypointType

The type of waypoint.